Small Functions to Make R Programming More Efficient

A small set of functions to addresses some common inefficiencies in R, such as the creation of logical vectors to compare quantities, unnecessary copies of objects in elementary mathematical or subsetting operations, obtaining information about objects (esp. data frames), or dealing with missing values.

anyv(x, value)              # Faster than any(x == value)
allv(x, value)              # Faster than all(x == value)
allNA(x)                    # Faster than all(is.na(x))
whichv(x, value,            # Faster than which(x == value)
       invert = FALSE)      # or which(x != value)
whichNA(x, invert = FALSE)  # Faster than which((!)is.na(x))
x %==% value                # Infix for whichv(v, value, FALSE), use e.g. in fsubset
x %!=% value                # Infix for whichv(v, value, TRUE)
alloc(value, n)             # Faster than rep_len(value, n)
copyv(X, v, R, ..., invert  # Faster than replace(x, x == v, r) or replace(x, v, r[v])
  = FALSE, vind1 = FALSE)   # or replace(x, x != v, r) or replace(x, !v, r[!v])
setv(X, v, R, ..., invert   # Same for x[x (!/=)= v] <- r or x[(!)v] <- r[(!)v]
  = FALSE, vind1 = FALSE)   # modifies x by reference, fastest
setop(X, op, V, ...,        # Faster than X <- X +\-\*\/ V (modifies by reference)
      rowwise = FALSE)      # optionally can also add v to rows of a matrix
X %+=% V                    # Infix for setop(X, "+", V)
X %-=% V                    # Infix for setop(X, "-", V)
X %*=% V                    # Infix for setop(X, "*", V)
X %/=% V                    # Infix for setop(X, "/", V)
na_rm(x)                    # Fast: if(anyNA(x)) x[!is.na(x)] else x,
                            # also removes NULL / empty elements from list
na_omit(X, cols = NULL,     # Faster na.omit for matrices and data frames,
        na.attr = FALSE)    # can use selected columns and attach indices
na_insert(X, prop = 0.1,    # Insert missing values at random
          value = NA)
missing_cases(X,            # The oposite of complete.cases(), faster for
              cols = NULL)  # data frames
vlengths(X, use.names=TRUE) # Faster version of lengths() (in C, no method dispatch)
vtypes(X, use.names = TRUE) # Get data storage types (faster vapply(X, typeof, ...))
fnlevels(x)                 # Faster version of nlevels(x) (for factors)
fnrow(X)                    # Faster nrow for data frames (not faster for matrices)
fncol(X)                    # Faster ncol for data frames (not faster for matrices)
fdim(X)                     # Faster dim for data frames (not faster for matrices)
seq_row(X)                  # Fast integer sequences along rows of X
seq_col(X)                  # Fast integer sequences along columns of X
cinv(x)                     # Choleski (fast) inverse of symmetric PD matrix, e.g. X'X

Arguments

X, V, R

a vector, matrix or data frame.

x, v

a (atomic) vector or matrix (na_rm also supports lists).

value

a single value of any (atomic) vector type.

invert

logical. TRUE considers elements x != value.

vind1

logical. If length(v) == 1L, setting vind1 = TRUE will interpret v as an index of X and R, rather than a value to search and replace.

op

an integer or character string indicating the operation to perform.

Int. 		 String 		 Description

1		"+"		add v

2		"-"		subtract v

3		"*"		multiply by v

4		"/"		divide by v

rowwise

logical. TRUE performs the operation between v and each row of x. Only applicable if x is a matrix and v a vector such that length(v) == ncol(x).

cols

select columns to check for missing values using column names, indices, a logical vector or a function (e.g. is.numeric). The default is to check all columns, which could be inefficient.

n

integer. The length of the vector to allocate with value.

na.attr

logical. TRUE adds an attribute containing the removed cases. For compatibility reasons this is exactly the same format as na.omit i.e. the attribute is called "na.action" and of class "omit".

prop

double. Specify the proportion of observations randomly replaced with NA.

use.names

logical. Preserve names if X is a list.

...

not used, reserved for possible future arguments.

Details

copyv and setv are designed to optimize operations that require replacing a single value in an object e.g. X[X == value] <- r or X[X == value] <- R[R == value] or simply copying parts of an existing object into another object e.g. X[v] <- R[v]. Thus they only cover cases where base R is inefficient by either creating a logical vector or materializing a subset to do some replacement. No alternative is provided in cases where base R is efficient e.g. x[v] <- r or cases provided by set and copy from the data.table package. Both functions work equivalently, with the difference that copyv creates a deep copy of the data before making the replacements and returns the copy, whereas setv modifies the data directly without creating a copy and returns the modified object invisibly. Thus setv is considerably more efficient.

copyv and setv perform different tasks, depending on the input. If v is a scalar, the elements of X are compared to v, and the matching ones (or non-matching ones if invert = TRUE) are replaced with R, where R can be either a scalar or an object of the same dimensions as X. If X is a data frame, R can also be a column-vector matching fnrow(X). The second option is if v is either a logical or integer vector of indices with length(v) > 1L, indicating the elements of a vector / matrix (or rows if X is a data frame) to replace with corresponding elements from R. Thus R has to be of equal dimensions as X, but could also be a column-vector if X is a data frame. Setting vind1 = TRUE ensures that v is always interpreted as an index, even if length(v) == 1L.

Note

(1) None of these functions currently support complex vectors. (2) It is possible to compare factors by the levels (e.g. iris$Species %==% "setosa")) or using integers (iris$Species %==% 1L). The latter is slightly more efficient. (3) Nothing special is implemented for other objects apart from basic types, e.g. for dates (which are stored as doubles) you need to generate a date object e.g. wlddev$date %==% as.Date("2019-01-01"), wlddev$date %==% "2019-01-01" will give integer(0).

See also

Examples

## Which value
whichNA(wlddev$PCGDP)                # Same as which(is.na(wlddev$PCGDP))
#>    [1]    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>   [15]   15   16   17   18   19   20   21   22   23   24   25   26   27   28
#>   [29]   29   30   31   32   33   34   35   36   37   38   39   40   41   42
#>   [43]   61   62   63   64   65   66   67   68   69   70   71   72   73   74
#>   [57]   75   76   77   78   79   80   81  122  183  184  185  186  187  188
#>   [71]  189  190  191  192  193  194  195  196  197  198  199  200  201  202
#>   [85]  203  204  205  206  207  208  209  210  211  212  213  214  215  216
#>   [99]  217  218  219  220  221  222  223  224  225  243  244  245  246  247
#>  [113]  248  249  250  251  252  253  254  305  306  307  308  309  310  311
#>  [127]  312  313  314  315  316  317  318  319  320  321  322  323  324  325
#>  [141]  366  367  368  369  370  371  372  373  374  375  376  377  378  379
#>  [155]  380  381  382  383  427  488  489  490  491  492  493  494  495  496
#>  [169]  497  498  499  500  501  502  503  504  505  506  507  508  509  510
#>  [183]  511  512  513  514  515  516  517  518  549  550  551  552  553  554
#>  [197]  555  556  557  558  559  560  561  562  563  564  565  566  567  568
#>  [211]  569  570  571  572  573  574  575  608  609  610  671  732  733  734
#>  [225]  735  736  737  738  739  740  741  742  743  744  745  746  747  748
#>  [239]  749  750  751  752  753  754  755  756  757  758  759  760  761  762
#>  [253]  793  854  855  856  857  858  859  860  861  862  863  864  865  866
#>  [267]  867  868  869  870  871  872  873  874  915  976  977  978  979  980
#>  [281]  981  982  983  984  985  986  987  988  989  990 1037 1038 1039 1040
#>  [295] 1041 1042 1043 1044 1045 1046 1047 1048 1049 1050 1051 1052 1053 1054
#>  [309] 1055 1056 1057 1058 1059 1060 1061 1062 1063 1064 1065 1066 1067 1098
#>  [323] 1159 1220 1281 1342 1343 1344 1345 1346 1347 1348 1349 1350 1351 1352
#>  [337] 1353 1354 1355 1356 1357 1358 1359 1360 1361 1362 1403 1464 1465 1466
#>  [351] 1467 1468 1469 1470 1471 1472 1473 1474 1475 1476 1477 1478 1479 1480
#>  [365] 1481 1482 1483 1484 1485 1486 1487 1488 1489 1490 1491 1492 1493 1494
#>  [379] 1495 1496 1497 1498 1525 1586 1647 1648 1649 1650 1651 1652 1653 1654
#>  [393] 1655 1656 1657 1658 1659 1660 1661 1662 1663 1664 1665 1666 1667 1668
#>  [407] 1669 1670 1671 1672 1673 1674 1675 1676 1677 1678 1679 1680 1681 1682
#>  [421] 1683 1684 1685 1686 1687 1688 1689 1690 1691 1692 1693 1694 1695 1696
#>  [435] 1697 1698 1699 1700 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710
#>  [449] 1711 1712 1713 1714 1715 1716 1717 1718 1719 1720 1721 1722 1769 1770
#>  [463] 1771 1772 1773 1774 1775 1776 1777 1778 1779 1780 1781 1782 1783 1784
#>  [477] 1785 1786 1787 1788 1789 1830 1891 1952 1953 1954 1955 1956 1957 1958
#>  [491] 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972
#>  [505] 2013 2014 2015 2016 2017 2018 2019 2020 2021 2022 2023 2024 2025 2026
#>  [519] 2027 2028 2029 2030 2031 2032 2033 2034 2035 2036 2037 2038 2039 2040
#>  [533] 2041 2042 2043 2044 2045 2046 2074 2135 2196 2197 2198 2199 2200 2201
#>  [547] 2202 2203 2204 2205 2206 2207 2208 2209 2210 2211 2212 2213 2214 2215
#>  [561] 2216 2217 2218 2219 2220 2221 2222 2223 2224 2225 2226 2227 2228 2229
#>  [575] 2230 2231 2232 2233 2234 2235 2236 2237 2238 2239 2240 2241 2242 2256
#>  [589] 2257 2318 2379 2380 2381 2382 2383 2384 2385 2386 2387 2388 2389 2390
#>  [603] 2391 2392 2393 2394 2395 2396 2397 2398 2399 2400 2401 2402 2403 2404
#>  [617] 2405 2406 2407 2408 2409 2410 2411 2412 2413 2414 2415 2416 2417 2418
#>  [631] 2419 2420 2421 2422 2423 2424 2425 2426 2427 2428 2429 2430 2431 2432
#>  [645] 2433 2434 2435 2436 2437 2438 2439 2440 2501 2562 2623 2624 2625 2626
#>  [659] 2627 2628 2629 2630 2631 2632 2633 2634 2635 2636 2637 2638 2639 2640
#>  [673] 2641 2642 2643 2684 2745 2806 2867 2928 2929 2930 2931 2932 2933 2934
#>  [687] 2935 2936 2937 2938 2939 2940 2941 2942 2943 2944 2945 2946 2947 2948
#>  [701] 2949 2950 2951 2952 2953 2954 2955 2956 2957 2958 2959 2960 2961 2962
#>  [715] 2963 2989 2990 2991 2992 2993 2994 2995 2996 2997 2998 2999 3049 3050
#>  [729] 3051 3052 3053 3054 3055 3056 3057 3058 3059 3060 3061 3062 3063 3064
#>  [743] 3065 3066 3067 3068 3069 3070 3071 3072 3073 3074 3075 3076 3077 3078
#>  [757] 3079 3080 3081 3082 3083 3084 3085 3086 3087 3088 3089 3090 3091 3092
#>  [771] 3093 3094 3095 3096 3097 3098 3099 3100 3101 3102 3103 3104 3105 3106
#>  [785] 3107 3108 3109 3110 3111 3112 3113 3114 3115 3116 3117 3118 3119 3120
#>  [799] 3121 3122 3123 3124 3125 3126 3172 3173 3174 3175 3176 3177 3178 3179
#>  [813] 3180 3181 3182 3183 3184 3185 3186 3187 3188 3189 3190 3191 3192 3193
#>  [827] 3194 3195 3196 3197 3198 3199 3200 3201 3202 3233 3294 3295 3296 3297
#>  [841] 3298 3299 3300 3301 3302 3303 3304 3305 3306 3307 3308 3309 3310 3311
#>  [855] 3312 3313 3314 3315 3316 3317 3318 3319 3320 3321 3322 3323 3324 3325
#>  [869] 3326 3327 3328 3329 3330 3331 3332 3333 3334 3335 3336 3337 3338 3339
#>  [883] 3340 3341 3342 3343 3344 3346 3347 3348 3349 3350 3351 3352 3353 3354
#>  [897] 3355 3356 3357 3358 3359 3360 3361 3362 3363 3364 3365 3366 3367 3368
#>  [911] 3369 3370 3371 3372 3416 3477 3538 3599 3600 3601 3602 3603 3604 3660
#>  [925] 3661 3662 3663 3664 3665 3666 3667 3668 3669 3670 3671 3672 3673 3674
#>  [939] 3675 3676 3677 3678 3679 3680 3721 3722 3723 3724 3725 3726 3727 3728
#>  [953] 3729 3730 3731 3732 3733 3734 3735 3736 3737 3738 3739 3740 3741 3742
#>  [967] 3743 3744 3745 3746 3747 3748 3749 3750 3751 3752 3753 3774 3775 3776
#>  [981] 3777 3778 3779 3780 3781 3782 3783 3784 3785 3786 3787 3788 3789 3790
#>  [995] 3791 3792 3793 3794 3795 3796
#>  [ reached getOption("max.print") -- omitted 2706 entries ]
whichNA(wlddev$PCGDP, invert = TRUE) # Same as which(!is.na(wlddev$PCGDP))
#>    [1]   43   44   45   46   47   48   49   50   51   52   53   54   55   56
#>   [15]   57   58   59   60   82   83   84   85   86   87   88   89   90   91
#>   [29]   92   93   94   95   96   97   98   99  100  101  102  103  104  105
#>   [43]  106  107  108  109  110  111  112  113  114  115  116  117  118  119
#>   [57]  120  121  123  124  125  126  127  128  129  130  131  132  133  134
#>   [71]  135  136  137  138  139  140  141  142  143  144  145  146  147  148
#>   [85]  149  150  151  152  153  154  155  156  157  158  159  160  161  162
#>   [99]  163  164  165  166  167  168  169  170  171  172  173  174  175  176
#>  [113]  177  178  179  180  181  182  226  227  228  229  230  231  232  233
#>  [127]  234  235  236  237  238  239  240  241  242  255  256  257  258  259
#>  [141]  260  261  262  263  264  265  266  267  268  269  270  271  272  273
#>  [155]  274  275  276  277  278  279  280  281  282  283  284  285  286  287
#>  [169]  288  289  290  291  292  293  294  295  296  297  298  299  300  301
#>  [183]  302  303  304  326  327  328  329  330  331  332  333  334  335  336
#>  [197]  337  338  339  340  341  342  343  344  345  346  347  348  349  350
#>  [211]  351  352  353  354  355  356  357  358  359  360  361  362  363  364
#>  [225]  365  384  385  386  387  388  389  390  391  392  393  394  395  396
#>  [239]  397  398  399  400  401  402  403  404  405  406  407  408  409  410
#>  [253]  411  412  413  414  415  416  417  418  419  420  421  422  423  424
#>  [267]  425  426  428  429  430  431  432  433  434  435  436  437  438  439
#>  [281]  440  441  442  443  444  445  446  447  448  449  450  451  452  453
#>  [295]  454  455  456  457  458  459  460  461  462  463  464  465  466  467
#>  [309]  468  469  470  471  472  473  474  475  476  477  478  479  480  481
#>  [323]  482  483  484  485  486  487  519  520  521  522  523  524  525  526
#>  [337]  527  528  529  530  531  532  533  534  535  536  537  538  539  540
#>  [351]  541  542  543  544  545  546  547  548  576  577  578  579  580  581
#>  [365]  582  583  584  585  586  587  588  589  590  591  592  593  594  595
#>  [379]  596  597  598  599  600  601  602  603  604  605  606  607  611  612
#>  [393]  613  614  615  616  617  618  619  620  621  622  623  624  625  626
#>  [407]  627  628  629  630  631  632  633  634  635  636  637  638  639  640
#>  [421]  641  642  643  644  645  646  647  648  649  650  651  652  653  654
#>  [435]  655  656  657  658  659  660  661  662  663  664  665  666  667  668
#>  [449]  669  670  672  673  674  675  676  677  678  679  680  681  682  683
#>  [463]  684  685  686  687  688  689  690  691  692  693  694  695  696  697
#>  [477]  698  699  700  701  702  703  704  705  706  707  708  709  710  711
#>  [491]  712  713  714  715  716  717  718  719  720  721  722  723  724  725
#>  [505]  726  727  728  729  730  731  763  764  765  766  767  768  769  770
#>  [519]  771  772  773  774  775  776  777  778  779  780  781  782  783  784
#>  [533]  785  786  787  788  789  790  791  792  794  795  796  797  798  799
#>  [547]  800  801  802  803  804  805  806  807  808  809  810  811  812  813
#>  [561]  814  815  816  817  818  819  820  821  822  823  824  825  826  827
#>  [575]  828  829  830  831  832  833  834  835  836  837  838  839  840  841
#>  [589]  842  843  844  845  846  847  848  849  850  851  852  853  875  876
#>  [603]  877  878  879  880  881  882  883  884  885  886  887  888  889  890
#>  [617]  891  892  893  894  895  896  897  898  899  900  901  902  903  904
#>  [631]  905  906  907  908  909  910  911  912  913  914  916  917  918  919
#>  [645]  920  921  922  923  924  925  926  927  928  929  930  931  932  933
#>  [659]  934  935  936  937  938  939  940  941  942  943  944  945  946  947
#>  [673]  948  949  950  951  952  953  954  955  956  957  958  959  960  961
#>  [687]  962  963  964  965  966  967  968  969  970  971  972  973  974  975
#>  [701]  991  992  993  994  995  996  997  998  999 1000 1001 1002 1003 1004
#>  [715] 1005 1006 1007 1008 1009 1010 1011 1012 1013 1014 1015 1016 1017 1018
#>  [729] 1019 1020 1021 1022 1023 1024 1025 1026 1027 1028 1029 1030 1031 1032
#>  [743] 1033 1034 1035 1036 1068 1069 1070 1071 1072 1073 1074 1075 1076 1077
#>  [757] 1078 1079 1080 1081 1082 1083 1084 1085 1086 1087 1088 1089 1090 1091
#>  [771] 1092 1093 1094 1095 1096 1097 1099 1100 1101 1102 1103 1104 1105 1106
#>  [785] 1107 1108 1109 1110 1111 1112 1113 1114 1115 1116 1117 1118 1119 1120
#>  [799] 1121 1122 1123 1124 1125 1126 1127 1128 1129 1130 1131 1132 1133 1134
#>  [813] 1135 1136 1137 1138 1139 1140 1141 1142 1143 1144 1145 1146 1147 1148
#>  [827] 1149 1150 1151 1152 1153 1154 1155 1156 1157 1158 1160 1161 1162 1163
#>  [841] 1164 1165 1166 1167 1168 1169 1170 1171 1172 1173 1174 1175 1176 1177
#>  [855] 1178 1179 1180 1181 1182 1183 1184 1185 1186 1187 1188 1189 1190 1191
#>  [869] 1192 1193 1194 1195 1196 1197 1198 1199 1200 1201 1202 1203 1204 1205
#>  [883] 1206 1207 1208 1209 1210 1211 1212 1213 1214 1215 1216 1217 1218 1219
#>  [897] 1221 1222 1223 1224 1225 1226 1227 1228 1229 1230 1231 1232 1233 1234
#>  [911] 1235 1236 1237 1238 1239 1240 1241 1242 1243 1244 1245 1246 1247 1248
#>  [925] 1249 1250 1251 1252 1253 1254 1255 1256 1257 1258 1259 1260 1261 1262
#>  [939] 1263 1264 1265 1266 1267 1268 1269 1270 1271 1272 1273 1274 1275 1276
#>  [953] 1277 1278 1279 1280 1282 1283 1284 1285 1286 1287 1288 1289 1290 1291
#>  [967] 1292 1293 1294 1295 1296 1297 1298 1299 1300 1301 1302 1303 1304 1305
#>  [981] 1306 1307 1308 1309 1310 1311 1312 1313 1314 1315 1316 1317 1318 1319
#>  [995] 1320 1321 1322 1323 1324 1325
#>  [ reached getOption("max.print") -- omitted 8470 entries ]
whichv(wlddev$country, "Chad")       # Same as which(wlddev$county == "Chad")
#>  [1] 2319 2320 2321 2322 2323 2324 2325 2326 2327 2328 2329 2330 2331 2332 2333
#> [16] 2334 2335 2336 2337 2338 2339 2340 2341 2342 2343 2344 2345 2346 2347 2348
#> [31] 2349 2350 2351 2352 2353 2354 2355 2356 2357 2358 2359 2360 2361 2362 2363
#> [46] 2364 2365 2366 2367 2368 2369 2370 2371 2372 2373 2374 2375 2376 2377 2378
#> [61] 2379
wlddev$country %==% "Chad"           # Same thing
#>  [1] 2319 2320 2321 2322 2323 2324 2325 2326 2327 2328 2329 2330 2331 2332 2333
#> [16] 2334 2335 2336 2337 2338 2339 2340 2341 2342 2343 2344 2345 2346 2347 2348
#> [31] 2349 2350 2351 2352 2353 2354 2355 2356 2357 2358 2359 2360 2361 2362 2363
#> [46] 2364 2365 2366 2367 2368 2369 2370 2371 2372 2373 2374 2375 2376 2377 2378
#> [61] 2379
whichv(wlddev$country, "Chad", TRUE) # Same as which(wlddev$county != "Chad")
#>    [1]    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>   [15]   15   16   17   18   19   20   21   22   23   24   25   26   27   28
#>   [29]   29   30   31   32   33   34   35   36   37   38   39   40   41   42
#>   [43]   43   44   45   46   47   48   49   50   51   52   53   54   55   56
#>   [57]   57   58   59   60   61   62   63   64   65   66   67   68   69   70
#>   [71]   71   72   73   74   75   76   77   78   79   80   81   82   83   84
#>   [85]   85   86   87   88   89   90   91   92   93   94   95   96   97   98
#>   [99]   99  100  101  102  103  104  105  106  107  108  109  110  111  112
#>  [113]  113  114  115  116  117  118  119  120  121  122  123  124  125  126
#>  [127]  127  128  129  130  131  132  133  134  135  136  137  138  139  140
#>  [141]  141  142  143  144  145  146  147  148  149  150  151  152  153  154
#>  [155]  155  156  157  158  159  160  161  162  163  164  165  166  167  168
#>  [169]  169  170  171  172  173  174  175  176  177  178  179  180  181  182
#>  [183]  183  184  185  186  187  188  189  190  191  192  193  194  195  196
#>  [197]  197  198  199  200  201  202  203  204  205  206  207  208  209  210
#>  [211]  211  212  213  214  215  216  217  218  219  220  221  222  223  224
#>  [225]  225  226  227  228  229  230  231  232  233  234  235  236  237  238
#>  [239]  239  240  241  242  243  244  245  246  247  248  249  250  251  252
#>  [253]  253  254  255  256  257  258  259  260  261  262  263  264  265  266
#>  [267]  267  268  269  270  271  272  273  274  275  276  277  278  279  280
#>  [281]  281  282  283  284  285  286  287  288  289  290  291  292  293  294
#>  [295]  295  296  297  298  299  300  301  302  303  304  305  306  307  308
#>  [309]  309  310  311  312  313  314  315  316  317  318  319  320  321  322
#>  [323]  323  324  325  326  327  328  329  330  331  332  333  334  335  336
#>  [337]  337  338  339  340  341  342  343  344  345  346  347  348  349  350
#>  [351]  351  352  353  354  355  356  357  358  359  360  361  362  363  364
#>  [365]  365  366  367  368  369  370  371  372  373  374  375  376  377  378
#>  [379]  379  380  381  382  383  384  385  386  387  388  389  390  391  392
#>  [393]  393  394  395  396  397  398  399  400  401  402  403  404  405  406
#>  [407]  407  408  409  410  411  412  413  414  415  416  417  418  419  420
#>  [421]  421  422  423  424  425  426  427  428  429  430  431  432  433  434
#>  [435]  435  436  437  438  439  440  441  442  443  444  445  446  447  448
#>  [449]  449  450  451  452  453  454  455  456  457  458  459  460  461  462
#>  [463]  463  464  465  466  467  468  469  470  471  472  473  474  475  476
#>  [477]  477  478  479  480  481  482  483  484  485  486  487  488  489  490
#>  [491]  491  492  493  494  495  496  497  498  499  500  501  502  503  504
#>  [505]  505  506  507  508  509  510  511  512  513  514  515  516  517  518
#>  [519]  519  520  521  522  523  524  525  526  527  528  529  530  531  532
#>  [533]  533  534  535  536  537  538  539  540  541  542  543  544  545  546
#>  [547]  547  548  549  550  551  552  553  554  555  556  557  558  559  560
#>  [561]  561  562  563  564  565  566  567  568  569  570  571  572  573  574
#>  [575]  575  576  577  578  579  580  581  582  583  584  585  586  587  588
#>  [589]  589  590  591  592  593  594  595  596  597  598  599  600  601  602
#>  [603]  603  604  605  606  607  608  609  610  611  612  613  614  615  616
#>  [617]  617  618  619  620  621  622  623  624  625  626  627  628  629  630
#>  [631]  631  632  633  634  635  636  637  638  639  640  641  642  643  644
#>  [645]  645  646  647  648  649  650  651  652  653  654  655  656  657  658
#>  [659]  659  660  661  662  663  664  665  666  667  668  669  670  671  672
#>  [673]  673  674  675  676  677  678  679  680  681  682  683  684  685  686
#>  [687]  687  688  689  690  691  692  693  694  695  696  697  698  699  700
#>  [701]  701  702  703  704  705  706  707  708  709  710  711  712  713  714
#>  [715]  715  716  717  718  719  720  721  722  723  724  725  726  727  728
#>  [729]  729  730  731  732  733  734  735  736  737  738  739  740  741  742
#>  [743]  743  744  745  746  747  748  749  750  751  752  753  754  755  756
#>  [757]  757  758  759  760  761  762  763  764  765  766  767  768  769  770
#>  [771]  771  772  773  774  775  776  777  778  779  780  781  782  783  784
#>  [785]  785  786  787  788  789  790  791  792  793  794  795  796  797  798
#>  [799]  799  800  801  802  803  804  805  806  807  808  809  810  811  812
#>  [813]  813  814  815  816  817  818  819  820  821  822  823  824  825  826
#>  [827]  827  828  829  830  831  832  833  834  835  836  837  838  839  840
#>  [841]  841  842  843  844  845  846  847  848  849  850  851  852  853  854
#>  [855]  855  856  857  858  859  860  861  862  863  864  865  866  867  868
#>  [869]  869  870  871  872  873  874  875  876  877  878  879  880  881  882
#>  [883]  883  884  885  886  887  888  889  890  891  892  893  894  895  896
#>  [897]  897  898  899  900  901  902  903  904  905  906  907  908  909  910
#>  [911]  911  912  913  914  915  916  917  918  919  920  921  922  923  924
#>  [925]  925  926  927  928  929  930  931  932  933  934  935  936  937  938
#>  [939]  939  940  941  942  943  944  945  946  947  948  949  950  951  952
#>  [953]  953  954  955  956  957  958  959  960  961  962  963  964  965  966
#>  [967]  967  968  969  970  971  972  973  974  975  976  977  978  979  980
#>  [981]  981  982  983  984  985  986  987  988  989  990  991  992  993  994
#>  [995]  995  996  997  998  999 1000
#>  [ reached getOption("max.print") -- omitted 12115 entries ]
wlddev$country %!=% "Chad"           # Same thing
#>    [1]    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>   [15]   15   16   17   18   19   20   21   22   23   24   25   26   27   28
#>   [29]   29   30   31   32   33   34   35   36   37   38   39   40   41   42
#>   [43]   43   44   45   46   47   48   49   50   51   52   53   54   55   56
#>   [57]   57   58   59   60   61   62   63   64   65   66   67   68   69   70
#>   [71]   71   72   73   74   75   76   77   78   79   80   81   82   83   84
#>   [85]   85   86   87   88   89   90   91   92   93   94   95   96   97   98
#>   [99]   99  100  101  102  103  104  105  106  107  108  109  110  111  112
#>  [113]  113  114  115  116  117  118  119  120  121  122  123  124  125  126
#>  [127]  127  128  129  130  131  132  133  134  135  136  137  138  139  140
#>  [141]  141  142  143  144  145  146  147  148  149  150  151  152  153  154
#>  [155]  155  156  157  158  159  160  161  162  163  164  165  166  167  168
#>  [169]  169  170  171  172  173  174  175  176  177  178  179  180  181  182
#>  [183]  183  184  185  186  187  188  189  190  191  192  193  194  195  196
#>  [197]  197  198  199  200  201  202  203  204  205  206  207  208  209  210
#>  [211]  211  212  213  214  215  216  217  218  219  220  221  222  223  224
#>  [225]  225  226  227  228  229  230  231  232  233  234  235  236  237  238
#>  [239]  239  240  241  242  243  244  245  246  247  248  249  250  251  252
#>  [253]  253  254  255  256  257  258  259  260  261  262  263  264  265  266
#>  [267]  267  268  269  270  271  272  273  274  275  276  277  278  279  280
#>  [281]  281  282  283  284  285  286  287  288  289  290  291  292  293  294
#>  [295]  295  296  297  298  299  300  301  302  303  304  305  306  307  308
#>  [309]  309  310  311  312  313  314  315  316  317  318  319  320  321  322
#>  [323]  323  324  325  326  327  328  329  330  331  332  333  334  335  336
#>  [337]  337  338  339  340  341  342  343  344  345  346  347  348  349  350
#>  [351]  351  352  353  354  355  356  357  358  359  360  361  362  363  364
#>  [365]  365  366  367  368  369  370  371  372  373  374  375  376  377  378
#>  [379]  379  380  381  382  383  384  385  386  387  388  389  390  391  392
#>  [393]  393  394  395  396  397  398  399  400  401  402  403  404  405  406
#>  [407]  407  408  409  410  411  412  413  414  415  416  417  418  419  420
#>  [421]  421  422  423  424  425  426  427  428  429  430  431  432  433  434
#>  [435]  435  436  437  438  439  440  441  442  443  444  445  446  447  448
#>  [449]  449  450  451  452  453  454  455  456  457  458  459  460  461  462
#>  [463]  463  464  465  466  467  468  469  470  471  472  473  474  475  476
#>  [477]  477  478  479  480  481  482  483  484  485  486  487  488  489  490
#>  [491]  491  492  493  494  495  496  497  498  499  500  501  502  503  504
#>  [505]  505  506  507  508  509  510  511  512  513  514  515  516  517  518
#>  [519]  519  520  521  522  523  524  525  526  527  528  529  530  531  532
#>  [533]  533  534  535  536  537  538  539  540  541  542  543  544  545  546
#>  [547]  547  548  549  550  551  552  553  554  555  556  557  558  559  560
#>  [561]  561  562  563  564  565  566  567  568  569  570  571  572  573  574
#>  [575]  575  576  577  578  579  580  581  582  583  584  585  586  587  588
#>  [589]  589  590  591  592  593  594  595  596  597  598  599  600  601  602
#>  [603]  603  604  605  606  607  608  609  610  611  612  613  614  615  616
#>  [617]  617  618  619  620  621  622  623  624  625  626  627  628  629  630
#>  [631]  631  632  633  634  635  636  637  638  639  640  641  642  643  644
#>  [645]  645  646  647  648  649  650  651  652  653  654  655  656  657  658
#>  [659]  659  660  661  662  663  664  665  666  667  668  669  670  671  672
#>  [673]  673  674  675  676  677  678  679  680  681  682  683  684  685  686
#>  [687]  687  688  689  690  691  692  693  694  695  696  697  698  699  700
#>  [701]  701  702  703  704  705  706  707  708  709  710  711  712  713  714
#>  [715]  715  716  717  718  719  720  721  722  723  724  725  726  727  728
#>  [729]  729  730  731  732  733  734  735  736  737  738  739  740  741  742
#>  [743]  743  744  745  746  747  748  749  750  751  752  753  754  755  756
#>  [757]  757  758  759  760  761  762  763  764  765  766  767  768  769  770
#>  [771]  771  772  773  774  775  776  777  778  779  780  781  782  783  784
#>  [785]  785  786  787  788  789  790  791  792  793  794  795  796  797  798
#>  [799]  799  800  801  802  803  804  805  806  807  808  809  810  811  812
#>  [813]  813  814  815  816  817  818  819  820  821  822  823  824  825  826
#>  [827]  827  828  829  830  831  832  833  834  835  836  837  838  839  840
#>  [841]  841  842  843  844  845  846  847  848  849  850  851  852  853  854
#>  [855]  855  856  857  858  859  860  861  862  863  864  865  866  867  868
#>  [869]  869  870  871  872  873  874  875  876  877  878  879  880  881  882
#>  [883]  883  884  885  886  887  888  889  890  891  892  893  894  895  896
#>  [897]  897  898  899  900  901  902  903  904  905  906  907  908  909  910
#>  [911]  911  912  913  914  915  916  917  918  919  920  921  922  923  924
#>  [925]  925  926  927  928  929  930  931  932  933  934  935  936  937  938
#>  [939]  939  940  941  942  943  944  945  946  947  948  949  950  951  952
#>  [953]  953  954  955  956  957  958  959  960  961  962  963  964  965  966
#>  [967]  967  968  969  970  971  972  973  974  975  976  977  978  979  980
#>  [981]  981  982  983  984  985  986  987  988  989  990  991  992  993  994
#>  [995]  995  996  997  998  999 1000
#>  [ reached getOption("max.print") -- omitted 12115 entries ]
lvec <- wlddev$country == "Chad"     # If we already have a logical vector...
whichv(lvec, FALSE)                  # is fastver than which(!lvec)
#>    [1]    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>   [15]   15   16   17   18   19   20   21   22   23   24   25   26   27   28
#>   [29]   29   30   31   32   33   34   35   36   37   38   39   40   41   42
#>   [43]   43   44   45   46   47   48   49   50   51   52   53   54   55   56
#>   [57]   57   58   59   60   61   62   63   64   65   66   67   68   69   70
#>   [71]   71   72   73   74   75   76   77   78   79   80   81   82   83   84
#>   [85]   85   86   87   88   89   90   91   92   93   94   95   96   97   98
#>   [99]   99  100  101  102  103  104  105  106  107  108  109  110  111  112
#>  [113]  113  114  115  116  117  118  119  120  121  122  123  124  125  126
#>  [127]  127  128  129  130  131  132  133  134  135  136  137  138  139  140
#>  [141]  141  142  143  144  145  146  147  148  149  150  151  152  153  154
#>  [155]  155  156  157  158  159  160  161  162  163  164  165  166  167  168
#>  [169]  169  170  171  172  173  174  175  176  177  178  179  180  181  182
#>  [183]  183  184  185  186  187  188  189  190  191  192  193  194  195  196
#>  [197]  197  198  199  200  201  202  203  204  205  206  207  208  209  210
#>  [211]  211  212  213  214  215  216  217  218  219  220  221  222  223  224
#>  [225]  225  226  227  228  229  230  231  232  233  234  235  236  237  238
#>  [239]  239  240  241  242  243  244  245  246  247  248  249  250  251  252
#>  [253]  253  254  255  256  257  258  259  260  261  262  263  264  265  266
#>  [267]  267  268  269  270  271  272  273  274  275  276  277  278  279  280
#>  [281]  281  282  283  284  285  286  287  288  289  290  291  292  293  294
#>  [295]  295  296  297  298  299  300  301  302  303  304  305  306  307  308
#>  [309]  309  310  311  312  313  314  315  316  317  318  319  320  321  322
#>  [323]  323  324  325  326  327  328  329  330  331  332  333  334  335  336
#>  [337]  337  338  339  340  341  342  343  344  345  346  347  348  349  350
#>  [351]  351  352  353  354  355  356  357  358  359  360  361  362  363  364
#>  [365]  365  366  367  368  369  370  371  372  373  374  375  376  377  378
#>  [379]  379  380  381  382  383  384  385  386  387  388  389  390  391  392
#>  [393]  393  394  395  396  397  398  399  400  401  402  403  404  405  406
#>  [407]  407  408  409  410  411  412  413  414  415  416  417  418  419  420
#>  [421]  421  422  423  424  425  426  427  428  429  430  431  432  433  434
#>  [435]  435  436  437  438  439  440  441  442  443  444  445  446  447  448
#>  [449]  449  450  451  452  453  454  455  456  457  458  459  460  461  462
#>  [463]  463  464  465  466  467  468  469  470  471  472  473  474  475  476
#>  [477]  477  478  479  480  481  482  483  484  485  486  487  488  489  490
#>  [491]  491  492  493  494  495  496  497  498  499  500  501  502  503  504
#>  [505]  505  506  507  508  509  510  511  512  513  514  515  516  517  518
#>  [519]  519  520  521  522  523  524  525  526  527  528  529  530  531  532
#>  [533]  533  534  535  536  537  538  539  540  541  542  543  544  545  546
#>  [547]  547  548  549  550  551  552  553  554  555  556  557  558  559  560
#>  [561]  561  562  563  564  565  566  567  568  569  570  571  572  573  574
#>  [575]  575  576  577  578  579  580  581  582  583  584  585  586  587  588
#>  [589]  589  590  591  592  593  594  595  596  597  598  599  600  601  602
#>  [603]  603  604  605  606  607  608  609  610  611  612  613  614  615  616
#>  [617]  617  618  619  620  621  622  623  624  625  626  627  628  629  630
#>  [631]  631  632  633  634  635  636  637  638  639  640  641  642  643  644
#>  [645]  645  646  647  648  649  650  651  652  653  654  655  656  657  658
#>  [659]  659  660  661  662  663  664  665  666  667  668  669  670  671  672
#>  [673]  673  674  675  676  677  678  679  680  681  682  683  684  685  686
#>  [687]  687  688  689  690  691  692  693  694  695  696  697  698  699  700
#>  [701]  701  702  703  704  705  706  707  708  709  710  711  712  713  714
#>  [715]  715  716  717  718  719  720  721  722  723  724  725  726  727  728
#>  [729]  729  730  731  732  733  734  735  736  737  738  739  740  741  742
#>  [743]  743  744  745  746  747  748  749  750  751  752  753  754  755  756
#>  [757]  757  758  759  760  761  762  763  764  765  766  767  768  769  770
#>  [771]  771  772  773  774  775  776  777  778  779  780  781  782  783  784
#>  [785]  785  786  787  788  789  790  791  792  793  794  795  796  797  798
#>  [799]  799  800  801  802  803  804  805  806  807  808  809  810  811  812
#>  [813]  813  814  815  816  817  818  819  820  821  822  823  824  825  826
#>  [827]  827  828  829  830  831  832  833  834  835  836  837  838  839  840
#>  [841]  841  842  843  844  845  846  847  848  849  850  851  852  853  854
#>  [855]  855  856  857  858  859  860  861  862  863  864  865  866  867  868
#>  [869]  869  870  871  872  873  874  875  876  877  878  879  880  881  882
#>  [883]  883  884  885  886  887  888  889  890  891  892  893  894  895  896
#>  [897]  897  898  899  900  901  902  903  904  905  906  907  908  909  910
#>  [911]  911  912  913  914  915  916  917  918  919  920  921  922  923  924
#>  [925]  925  926  927  928  929  930  931  932  933  934  935  936  937  938
#>  [939]  939  940  941  942  943  944  945  946  947  948  949  950  951  952
#>  [953]  953  954  955  956  957  958  959  960  961  962  963  964  965  966
#>  [967]  967  968  969  970  971  972  973  974  975  976  977  978  979  980
#>  [981]  981  982  983  984  985  986  987  988  989  990  991  992  993  994
#>  [995]  995  996  997  998  999 1000
#>  [ reached getOption("max.print") -- omitted 12115 entries ]
rm(lvec)

# Using the %==% operator can yield tangible performance gains
fsubset(wlddev, iso3c %==% "DEU") # 3x faster than:
#>    country iso3c       date year decade                region      income OECD
#> 1  Germany   DEU 1961-01-01 1960   1960 Europe & Central Asia High income TRUE
#> 2  Germany   DEU 1962-01-01 1961   1960 Europe & Central Asia High income TRUE
#> 3  Germany   DEU 1963-01-01 1962   1960 Europe & Central Asia High income TRUE
#> 4  Germany   DEU 1964-01-01 1963   1960 Europe & Central Asia High income TRUE
#> 5  Germany   DEU 1965-01-01 1964   1960 Europe & Central Asia High income TRUE
#> 6  Germany   DEU 1966-01-01 1965   1960 Europe & Central Asia High income TRUE
#> 7  Germany   DEU 1967-01-01 1966   1960 Europe & Central Asia High income TRUE
#> 8  Germany   DEU 1968-01-01 1967   1960 Europe & Central Asia High income TRUE
#> 9  Germany   DEU 1969-01-01 1968   1960 Europe & Central Asia High income TRUE
#> 10 Germany   DEU 1970-01-01 1969   1960 Europe & Central Asia High income TRUE
#> 11 Germany   DEU 1971-01-01 1970   1970 Europe & Central Asia High income TRUE
#> 12 Germany   DEU 1972-01-01 1971   1970 Europe & Central Asia High income TRUE
#> 13 Germany   DEU 1973-01-01 1972   1970 Europe & Central Asia High income TRUE
#> 14 Germany   DEU 1974-01-01 1973   1970 Europe & Central Asia High income TRUE
#> 15 Germany   DEU 1975-01-01 1974   1970 Europe & Central Asia High income TRUE
#> 16 Germany   DEU 1976-01-01 1975   1970 Europe & Central Asia High income TRUE
#> 17 Germany   DEU 1977-01-01 1976   1970 Europe & Central Asia High income TRUE
#> 18 Germany   DEU 1978-01-01 1977   1970 Europe & Central Asia High income TRUE
#> 19 Germany   DEU 1979-01-01 1978   1970 Europe & Central Asia High income TRUE
#> 20 Germany   DEU 1980-01-01 1979   1970 Europe & Central Asia High income TRUE
#> 21 Germany   DEU 1981-01-01 1980   1980 Europe & Central Asia High income TRUE
#> 22 Germany   DEU 1982-01-01 1981   1980 Europe & Central Asia High income TRUE
#> 23 Germany   DEU 1983-01-01 1982   1980 Europe & Central Asia High income TRUE
#> 24 Germany   DEU 1984-01-01 1983   1980 Europe & Central Asia High income TRUE
#> 25 Germany   DEU 1985-01-01 1984   1980 Europe & Central Asia High income TRUE
#> 26 Germany   DEU 1986-01-01 1985   1980 Europe & Central Asia High income TRUE
#> 27 Germany   DEU 1987-01-01 1986   1980 Europe & Central Asia High income TRUE
#> 28 Germany   DEU 1988-01-01 1987   1980 Europe & Central Asia High income TRUE
#> 29 Germany   DEU 1989-01-01 1988   1980 Europe & Central Asia High income TRUE
#> 30 Germany   DEU 1990-01-01 1989   1980 Europe & Central Asia High income TRUE
#> 31 Germany   DEU 1991-01-01 1990   1990 Europe & Central Asia High income TRUE
#> 32 Germany   DEU 1992-01-01 1991   1990 Europe & Central Asia High income TRUE
#> 33 Germany   DEU 1993-01-01 1992   1990 Europe & Central Asia High income TRUE
#> 34 Germany   DEU 1994-01-01 1993   1990 Europe & Central Asia High income TRUE
#> 35 Germany   DEU 1995-01-01 1994   1990 Europe & Central Asia High income TRUE
#> 36 Germany   DEU 1996-01-01 1995   1990 Europe & Central Asia High income TRUE
#> 37 Germany   DEU 1997-01-01 1996   1990 Europe & Central Asia High income TRUE
#> 38 Germany   DEU 1998-01-01 1997   1990 Europe & Central Asia High income TRUE
#> 39 Germany   DEU 1999-01-01 1998   1990 Europe & Central Asia High income TRUE
#> 40 Germany   DEU 2000-01-01 1999   1990 Europe & Central Asia High income TRUE
#> 41 Germany   DEU 2001-01-01 2000   2000 Europe & Central Asia High income TRUE
#> 42 Germany   DEU 2002-01-01 2001   2000 Europe & Central Asia High income TRUE
#> 43 Germany   DEU 2003-01-01 2002   2000 Europe & Central Asia High income TRUE
#> 44 Germany   DEU 2004-01-01 2003   2000 Europe & Central Asia High income TRUE
#> 45 Germany   DEU 2005-01-01 2004   2000 Europe & Central Asia High income TRUE
#> 46 Germany   DEU 2006-01-01 2005   2000 Europe & Central Asia High income TRUE
#> 47 Germany   DEU 2007-01-01 2006   2000 Europe & Central Asia High income TRUE
#> 48 Germany   DEU 2008-01-01 2007   2000 Europe & Central Asia High income TRUE
#> 49 Germany   DEU 2009-01-01 2008   2000 Europe & Central Asia High income TRUE
#> 50 Germany   DEU 2010-01-01 2009   2000 Europe & Central Asia High income TRUE
#> 51 Germany   DEU 2011-01-01 2010   2010 Europe & Central Asia High income TRUE
#> 52 Germany   DEU 2012-01-01 2011   2010 Europe & Central Asia High income TRUE
#> 53 Germany   DEU 2013-01-01 2012   2010 Europe & Central Asia High income TRUE
#> 54 Germany   DEU 2014-01-01 2013   2010 Europe & Central Asia High income TRUE
#> 55 Germany   DEU 2015-01-01 2014   2010 Europe & Central Asia High income TRUE
#> 56 Germany   DEU 2016-01-01 2015   2010 Europe & Central Asia High income TRUE
#> 57 Germany   DEU 2017-01-01 2016   2010 Europe & Central Asia High income TRUE
#> 58 Germany   DEU 2018-01-01 2017   2010 Europe & Central Asia High income TRUE
#> 59 Germany   DEU 2019-01-01 2018   2010 Europe & Central Asia High income TRUE
#> 60 Germany   DEU 2020-01-01 2019   2010 Europe & Central Asia High income TRUE
#> 61 Germany   DEU 2021-01-01 2020   2020 Europe & Central Asia High income TRUE
#>       PCGDP   LIFEEX GINI ODA      POP
#> 1        NA 69.31002   NA  NA 72814900
#> 2        NA 69.50800   NA  NA 73377632
#> 3        NA 69.69154   NA  NA 74025784
#> 4        NA 69.85961   NA  NA 74714353
#> 5        NA 70.01371   NA  NA 75318337
#> 6        NA 70.15183   NA  NA 75963695
#> 7        NA 70.27141   NA  NA 76600311
#> 8        NA 70.37400   NA  NA 76951336
#> 9        NA 70.46510   NA  NA 77294314
#> 10       NA 70.55068   NA  NA 77909682
#> 11 19681.32 70.63978   NA  NA 78169289
#> 12 20260.67 70.74288   NA  NA 78312842
#> 13 21031.08 70.86700   NA  NA 78688452
#> 14 21966.55 71.01668   NA  NA 78936666
#> 15 22153.43 71.19541   NA  NA 78967433
#> 16 22043.45 71.40173   NA  NA 78673554
#> 17 23233.84 71.63415   NA  NA 78336950
#> 18 24065.95 71.88368   NA  NA 78159814
#> 19 24811.56 72.14329   NA  NA 78091820
#> 20 25829.91 72.40854   NA  NA 78126350
#> 21 26139.53 72.67790   NA  NA 78288576
#> 22 26237.87 72.95039   NA  NA 78407907
#> 23 26159.15 73.22700   NA  NA 78333366
#> 24 26640.22 73.50520   NA  NA 78128282
#> 25 27487.11 73.78293   NA  NA 77858685
#> 26 28189.92 74.05463   NA  NA 77684873
#> 27 28821.53 74.31373   NA  NA 77720436
#> 28 29180.79 74.55968   NA  NA 77839920
#> 29 30144.59 74.79249   NA  NA 78144619
#> 30 31077.92 75.01315   NA  NA 78751283
#> 31 32430.32 75.22776   NA  NA 79433029
#> 32 33839.49 75.31951 29.2  NA 80013896
#> 33 34229.00 75.81951   NA  NA 80624598
#> 34 33672.54 75.87073   NA  NA 81156363
#> 35 34358.57 76.27073 29.2  NA 81438348
#> 36 34786.73 76.42195 28.9  NA 81678051
#> 37 34965.68 76.67317   NA  NA 81914831
#> 38 35540.29 77.07317   NA  NA 82034771
#> 39 36250.55 77.47561 28.3  NA 82047195
#> 40 36910.83 77.72683   NA  NA 82100243
#> 41 37934.45 77.92683 28.8  NA 82211508
#> 42 38507.47 78.32927 30.3  NA 82349925
#> 43 38366.68 78.22927 29.9  NA 82488495
#> 44 38076.98 78.38049 30.0  NA 82534176
#> 45 38532.78 78.68049 30.4  NA 82516260
#> 46 38836.78 78.93171 32.1  NA 82469422
#> 47 40364.46 79.13171 31.3  NA 82376451
#> 48 41621.51 79.53415 31.3  NA 82266372
#> 49 42101.00 79.73659 31.1  NA 82110097
#> 50 39804.57 79.83659 30.5  NA 81902307
#> 51 41531.93 79.98780 30.2  NA 81776930
#> 52 43969.71 80.43659 30.5  NA 80274983
#> 53 44070.91 80.53902 30.9  NA 80425823
#> 54 44143.13 80.49024 31.1  NA 80645605
#> 55 44930.79 81.09024 30.9  NA 80982500
#> 56 45208.06 80.64146 31.7  NA 81686611
#> 57 45844.64 80.99024 31.9  NA 82348669
#> 58 46862.04 80.99268   NA  NA 82657002
#> 59 47313.85 80.89268   NA  NA 82905782
#> 60 47469.48 80.94146   NA  NA 83092962
#> 61       NA       NA   NA  NA       NA
fsubset(wlddev, iso3c == "DEU")
#>    country iso3c       date year decade                region      income OECD
#> 1  Germany   DEU 1961-01-01 1960   1960 Europe & Central Asia High income TRUE
#> 2  Germany   DEU 1962-01-01 1961   1960 Europe & Central Asia High income TRUE
#> 3  Germany   DEU 1963-01-01 1962   1960 Europe & Central Asia High income TRUE
#> 4  Germany   DEU 1964-01-01 1963   1960 Europe & Central Asia High income TRUE
#> 5  Germany   DEU 1965-01-01 1964   1960 Europe & Central Asia High income TRUE
#> 6  Germany   DEU 1966-01-01 1965   1960 Europe & Central Asia High income TRUE
#> 7  Germany   DEU 1967-01-01 1966   1960 Europe & Central Asia High income TRUE
#> 8  Germany   DEU 1968-01-01 1967   1960 Europe & Central Asia High income TRUE
#> 9  Germany   DEU 1969-01-01 1968   1960 Europe & Central Asia High income TRUE
#> 10 Germany   DEU 1970-01-01 1969   1960 Europe & Central Asia High income TRUE
#> 11 Germany   DEU 1971-01-01 1970   1970 Europe & Central Asia High income TRUE
#> 12 Germany   DEU 1972-01-01 1971   1970 Europe & Central Asia High income TRUE
#> 13 Germany   DEU 1973-01-01 1972   1970 Europe & Central Asia High income TRUE
#> 14 Germany   DEU 1974-01-01 1973   1970 Europe & Central Asia High income TRUE
#> 15 Germany   DEU 1975-01-01 1974   1970 Europe & Central Asia High income TRUE
#> 16 Germany   DEU 1976-01-01 1975   1970 Europe & Central Asia High income TRUE
#> 17 Germany   DEU 1977-01-01 1976   1970 Europe & Central Asia High income TRUE
#> 18 Germany   DEU 1978-01-01 1977   1970 Europe & Central Asia High income TRUE
#> 19 Germany   DEU 1979-01-01 1978   1970 Europe & Central Asia High income TRUE
#> 20 Germany   DEU 1980-01-01 1979   1970 Europe & Central Asia High income TRUE
#> 21 Germany   DEU 1981-01-01 1980   1980 Europe & Central Asia High income TRUE
#> 22 Germany   DEU 1982-01-01 1981   1980 Europe & Central Asia High income TRUE
#> 23 Germany   DEU 1983-01-01 1982   1980 Europe & Central Asia High income TRUE
#> 24 Germany   DEU 1984-01-01 1983   1980 Europe & Central Asia High income TRUE
#> 25 Germany   DEU 1985-01-01 1984   1980 Europe & Central Asia High income TRUE
#> 26 Germany   DEU 1986-01-01 1985   1980 Europe & Central Asia High income TRUE
#> 27 Germany   DEU 1987-01-01 1986   1980 Europe & Central Asia High income TRUE
#> 28 Germany   DEU 1988-01-01 1987   1980 Europe & Central Asia High income TRUE
#> 29 Germany   DEU 1989-01-01 1988   1980 Europe & Central Asia High income TRUE
#> 30 Germany   DEU 1990-01-01 1989   1980 Europe & Central Asia High income TRUE
#> 31 Germany   DEU 1991-01-01 1990   1990 Europe & Central Asia High income TRUE
#> 32 Germany   DEU 1992-01-01 1991   1990 Europe & Central Asia High income TRUE
#> 33 Germany   DEU 1993-01-01 1992   1990 Europe & Central Asia High income TRUE
#> 34 Germany   DEU 1994-01-01 1993   1990 Europe & Central Asia High income TRUE
#> 35 Germany   DEU 1995-01-01 1994   1990 Europe & Central Asia High income TRUE
#> 36 Germany   DEU 1996-01-01 1995   1990 Europe & Central Asia High income TRUE
#> 37 Germany   DEU 1997-01-01 1996   1990 Europe & Central Asia High income TRUE
#> 38 Germany   DEU 1998-01-01 1997   1990 Europe & Central Asia High income TRUE
#> 39 Germany   DEU 1999-01-01 1998   1990 Europe & Central Asia High income TRUE
#> 40 Germany   DEU 2000-01-01 1999   1990 Europe & Central Asia High income TRUE
#> 41 Germany   DEU 2001-01-01 2000   2000 Europe & Central Asia High income TRUE
#> 42 Germany   DEU 2002-01-01 2001   2000 Europe & Central Asia High income TRUE
#> 43 Germany   DEU 2003-01-01 2002   2000 Europe & Central Asia High income TRUE
#> 44 Germany   DEU 2004-01-01 2003   2000 Europe & Central Asia High income TRUE
#> 45 Germany   DEU 2005-01-01 2004   2000 Europe & Central Asia High income TRUE
#> 46 Germany   DEU 2006-01-01 2005   2000 Europe & Central Asia High income TRUE
#> 47 Germany   DEU 2007-01-01 2006   2000 Europe & Central Asia High income TRUE
#> 48 Germany   DEU 2008-01-01 2007   2000 Europe & Central Asia High income TRUE
#> 49 Germany   DEU 2009-01-01 2008   2000 Europe & Central Asia High income TRUE
#> 50 Germany   DEU 2010-01-01 2009   2000 Europe & Central Asia High income TRUE
#> 51 Germany   DEU 2011-01-01 2010   2010 Europe & Central Asia High income TRUE
#> 52 Germany   DEU 2012-01-01 2011   2010 Europe & Central Asia High income TRUE
#> 53 Germany   DEU 2013-01-01 2012   2010 Europe & Central Asia High income TRUE
#> 54 Germany   DEU 2014-01-01 2013   2010 Europe & Central Asia High income TRUE
#> 55 Germany   DEU 2015-01-01 2014   2010 Europe & Central Asia High income TRUE
#> 56 Germany   DEU 2016-01-01 2015   2010 Europe & Central Asia High income TRUE
#> 57 Germany   DEU 2017-01-01 2016   2010 Europe & Central Asia High income TRUE
#> 58 Germany   DEU 2018-01-01 2017   2010 Europe & Central Asia High income TRUE
#> 59 Germany   DEU 2019-01-01 2018   2010 Europe & Central Asia High income TRUE
#> 60 Germany   DEU 2020-01-01 2019   2010 Europe & Central Asia High income TRUE
#> 61 Germany   DEU 2021-01-01 2020   2020 Europe & Central Asia High income TRUE
#>       PCGDP   LIFEEX GINI ODA      POP
#> 1        NA 69.31002   NA  NA 72814900
#> 2        NA 69.50800   NA  NA 73377632
#> 3        NA 69.69154   NA  NA 74025784
#> 4        NA 69.85961   NA  NA 74714353
#> 5        NA 70.01371   NA  NA 75318337
#> 6        NA 70.15183   NA  NA 75963695
#> 7        NA 70.27141   NA  NA 76600311
#> 8        NA 70.37400   NA  NA 76951336
#> 9        NA 70.46510   NA  NA 77294314
#> 10       NA 70.55068   NA  NA 77909682
#> 11 19681.32 70.63978   NA  NA 78169289
#> 12 20260.67 70.74288   NA  NA 78312842
#> 13 21031.08 70.86700   NA  NA 78688452
#> 14 21966.55 71.01668   NA  NA 78936666
#> 15 22153.43 71.19541   NA  NA 78967433
#> 16 22043.45 71.40173   NA  NA 78673554
#> 17 23233.84 71.63415   NA  NA 78336950
#> 18 24065.95 71.88368   NA  NA 78159814
#> 19 24811.56 72.14329   NA  NA 78091820
#> 20 25829.91 72.40854   NA  NA 78126350
#> 21 26139.53 72.67790   NA  NA 78288576
#> 22 26237.87 72.95039   NA  NA 78407907
#> 23 26159.15 73.22700   NA  NA 78333366
#> 24 26640.22 73.50520   NA  NA 78128282
#> 25 27487.11 73.78293   NA  NA 77858685
#> 26 28189.92 74.05463   NA  NA 77684873
#> 27 28821.53 74.31373   NA  NA 77720436
#> 28 29180.79 74.55968   NA  NA 77839920
#> 29 30144.59 74.79249   NA  NA 78144619
#> 30 31077.92 75.01315   NA  NA 78751283
#> 31 32430.32 75.22776   NA  NA 79433029
#> 32 33839.49 75.31951 29.2  NA 80013896
#> 33 34229.00 75.81951   NA  NA 80624598
#> 34 33672.54 75.87073   NA  NA 81156363
#> 35 34358.57 76.27073 29.2  NA 81438348
#> 36 34786.73 76.42195 28.9  NA 81678051
#> 37 34965.68 76.67317   NA  NA 81914831
#> 38 35540.29 77.07317   NA  NA 82034771
#> 39 36250.55 77.47561 28.3  NA 82047195
#> 40 36910.83 77.72683   NA  NA 82100243
#> 41 37934.45 77.92683 28.8  NA 82211508
#> 42 38507.47 78.32927 30.3  NA 82349925
#> 43 38366.68 78.22927 29.9  NA 82488495
#> 44 38076.98 78.38049 30.0  NA 82534176
#> 45 38532.78 78.68049 30.4  NA 82516260
#> 46 38836.78 78.93171 32.1  NA 82469422
#> 47 40364.46 79.13171 31.3  NA 82376451
#> 48 41621.51 79.53415 31.3  NA 82266372
#> 49 42101.00 79.73659 31.1  NA 82110097
#> 50 39804.57 79.83659 30.5  NA 81902307
#> 51 41531.93 79.98780 30.2  NA 81776930
#> 52 43969.71 80.43659 30.5  NA 80274983
#> 53 44070.91 80.53902 30.9  NA 80425823
#> 54 44143.13 80.49024 31.1  NA 80645605
#> 55 44930.79 81.09024 30.9  NA 80982500
#> 56 45208.06 80.64146 31.7  NA 81686611
#> 57 45844.64 80.99024 31.9  NA 82348669
#> 58 46862.04 80.99268   NA  NA 82657002
#> 59 47313.85 80.89268   NA  NA 82905782
#> 60 47469.48 80.94146   NA  NA 83092962
#> 61       NA       NA   NA  NA       NA

## Missing values
mtc_na <- na_insert(mtcars, 0.15)    # Set 15% of values missing at random
fnobs(mtc_na)                        # See observation count
#>  mpg  cyl disp   hp drat   wt qsec   vs   am gear carb 
#>   28   28   28   28   28   28   28   28   28   28   28 
na_omit(mtc_na)                      # 12x faster than na.omit(mtc_na)
#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Duster 360         14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 450SLC        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Toyota Corolla     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
na_omit(mtc_na, na.attr = TRUE)      # Adds attribute with removed cases, like na.omit
#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Duster 360         14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 450SLC        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Toyota Corolla     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
na_omit(mtc_na, cols = c("vs","am")) # Removes only cases missing vs or am
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag       21.0   6 160.0 110   NA 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85    NA    NA  1  1    4    1
#> Hornet 4 Drive      21.4  NA 258.0  NA 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440    NA  0  0    3    2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4    NA  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123   NA 3.440 18.90  1  0   NA   NA
#> Merc 450SE            NA   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental   NA   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Honda Civic         30.4   4  75.7  NA 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4    NA  97 3.70    NA 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0  NA   NA 3.520 16.87  0  0   NA    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Pontiac Firebird    19.2  NA 400.0 175 3.08    NA    NA  0  0   NA    2
#> Fiat X1-9           27.3  NA  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  NA 4.43 2.140    NA  0  1    5   NA
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Volvo 142E          21.4   4 121.0 109   NA    NA 18.60  1  1    4    2
na_omit(qM(mtc_na))                  # Also works for matrices
#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Duster 360         14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 450SLC        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Toyota Corolla     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Lotus Europa       30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
na_omit(mtc_na$vs, na.attr = TRUE)   # Also works with vectors
#>  [1] 0 0 1 1 0 1 0 1 1 1 1 0 0 0 0 1 1 1 1 0 0 0 1 0 1 0 0 1
#> attr(,"na.action")
#> [1] 13 17 24 31
#> attr(,"class")
#> [1] "omit"
na_rm(mtc_na$vs)                     # For vectors na_rm is faster ...
#>  [1] 0 0 1 1 0 1 0 1 1 1 1 0 0 0 0 1 1 1 1 0 0 0 1 0 1 0 0 1
rm(mtc_na)