Pairwise Correlations, Covariances and Observation Count

Computes (pairwise) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method. For an equivalent and faster implementation of pwcor see Hmisc::rcorr (written in Fortran), for weighted correlations see weights::wtd.cors (written in C).

pwcor(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwNobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.85  -.85  -.75   .70  -.89   .29   .63   .50   .49  -.51
#> cyl   -.85    1    .91   .81  -.70   .77  -.66  -.90  -.44  -.44   .58
#> disp  -.85   .91    1    .84  -.74   .89  -.42  -.75  -.59  -.63   .56
#> hp    -.75   .81   .84    1   -.48   .71  -.74  -.72  -.22  -.15   .77
#> drat   .70  -.70  -.74  -.48    1   -.72   .14   .59   .65   .72  -.20
#> wt    -.89   .77   .89   .71  -.72    1   -.09  -.57  -.71  -.66   .56
#> qsec   .29  -.66  -.42  -.74   .14  -.09    1    .76  -.28  -.22  -.66
#> vs     .63  -.90  -.75  -.72   .59  -.57   .76    1    .23   .28  -.53
#> am     .50  -.44  -.59  -.22   .65  -.71  -.28   .23    1    .77   .02
#> gear   .49  -.44  -.63  -.15   .72  -.66  -.22   .28   .77    1    .16
#> carb  -.51   .58   .56   .77  -.20   .56  -.66  -.53   .02   .16    1 

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      25.89     -7.29   -552.63   -261.54      1.95     -4.62      2.84
#> cyl      -7.29      3.24     20.36     98.62      -.64      1.29     -2.34
#> disp   -552.63     20.36  16207.06   6687.79    -51.67    115.39    -95.70
#> hp     -261.54     98.62   6687.79   4741.36    -18.34     44.06     -8.68
#> drat      1.95      -.64    -51.67    -18.34       .27      -.39       .15
#> wt       -4.62      1.29    115.39     44.06      -.39       .98      -.16
#> qsec      2.84     -2.34    -95.70     -8.68       .15      -.16      3.50
#> vs        1.67      -.83    -49.82    -24.99       .16      -.30       .62
#> am        1.26      -.40    -35.68     -8.03       .17      -.33      -.28
#> gear      1.86      -.58    -59.05     -8.00       .25      -.48      -.33
#> carb     -4.18      1.70     85.19     85.06      -.17       .63     -1.96
#>             vs        am      gear      carb
#> mpg       1.67      1.26      1.86     -4.18
#> cyl       -.83      -.40      -.58      1.70
#> disp    -49.82    -35.68    -59.05     85.19
#> hp      -24.99     -8.03     -8.00     85.06
#> drat       .16       .17       .25      -.17
#> wt        -.30      -.33      -.48       .63
#> qsec       .62      -.28      -.33     -1.96
#> vs         .25       .06       .11      -.43
#> am         .06       .26       .29       .02
#> gear       .11       .29       .56       .19
#> carb      -.43       .02       .19      2.44

pwNobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 27   26 26 26   26   26
#> cyl   26  29   26 26   26 26   26 27 26   26   26
#> disp  26  26   29 26   27 27   27 26 26   27   26
#> hp    27  26   26 29   26 27   26 27 26   26   26
#> drat  26  26   27 26   29 26   26 26 26   28   26
#> wt    27  26   27 27   26 29   26 26 26   26   26
#> qsec  26  26   27 26   26 26   29 26 26   26   27
#> vs    26  27   26 27   26 26   26 29 26   26   26
#> am    26  26   26 26   26 26   26 26 29   26   27
#> gear  26  26   27 26   28 26   26 26 26   29   26
#> carb  26  26   26 26   26 26   27 26 27   26   29

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (26) -.85 (26) -.75 (27)  .70 (26) -.89 (27)  .29 (26)
#> cyl  -.85 (26)   1  (29)  .91 (26)  .81 (26) -.70 (26)  .77 (26) -.66 (26)
#> disp -.85 (26)  .91 (26)   1  (29)  .84 (26) -.74 (27)  .89 (27) -.42 (27)
#> hp   -.75 (27)  .81 (26)  .84 (26)   1  (29) -.48 (26)  .71 (27) -.74 (26)
#> drat  .70 (26) -.70 (26) -.74 (27) -.48 (26)   1  (29) -.72 (26)  .14 (26)
#> wt   -.89 (27)  .77 (26)  .89 (27)  .71 (27) -.72 (26)   1  (29) -.09 (26)
#> qsec  .29 (26) -.66 (26) -.42 (27) -.74 (26)  .14 (26) -.09 (26)   1  (29)
#> vs    .63 (26) -.90 (27) -.75 (26) -.72 (27)  .59 (26) -.57 (26)  .76 (26)
#> am    .50 (26) -.44 (26) -.59 (26) -.22 (26)  .65 (26) -.71 (26) -.28 (26)
#> gear  .49 (26) -.44 (26) -.63 (27) -.15 (26)  .72 (28) -.66 (26) -.22 (26)
#> carb -.51 (26)  .58 (26)  .56 (26)  .77 (26) -.20 (26)  .56 (26) -.66 (27)
#>             vs        am      gear      carb
#> mpg   .63 (26)  .50 (26)  .49 (26) -.51 (26)
#> cyl  -.90 (27) -.44 (26) -.44 (26)  .58 (26)
#> disp -.75 (26) -.59 (26) -.63 (27)  .56 (26)
#> hp   -.72 (27) -.22 (26) -.15 (26)  .77 (26)
#> drat  .59 (26)  .65 (26)  .72 (28) -.20 (26)
#> wt   -.57 (26) -.71 (26) -.66 (26)  .56 (26)
#> qsec  .76 (26) -.28 (26) -.22 (26) -.66 (27)
#> vs     1  (29)  .23 (26)  .28 (26) -.53 (26)
#> am    .23 (26)   1  (29)  .77 (26)  .02 (27)
#> gear  .28 (26)  .77 (26)   1  (29)  .16 (26)
#> carb -.53 (26)  .02 (27)  .16 (26)   1  (29)

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.85*  -.75*   .70*  -.89*   .29    .63*   .50*   .49*
#> cyl   -.85*    1     .91*   .81*  -.70*   .77*  -.66*  -.90*  -.44*  -.44*
#> disp  -.85*   .91*    1     .84*  -.74*   .89*  -.42*  -.75*  -.59*  -.63*
#> hp    -.75*   .81*   .84*    1    -.48*   .71*  -.74*  -.72*  -.22   -.15 
#> drat   .70*  -.70*  -.74*  -.48*    1    -.72*   .14    .59*   .65*   .72*
#> wt    -.89*   .77*   .89*   .71*  -.72*    1    -.09   -.57*  -.71*  -.66*
#> qsec   .29   -.66*  -.42*  -.74*   .14   -.09     1     .76*  -.28   -.22 
#> vs     .63*  -.90*  -.75*  -.72*   .59*  -.57*   .76*    1     .23    .28 
#> am     .50*  -.44*  -.59*  -.22    .65*  -.71*  -.28    .23     1     .77*
#> gear   .49*  -.44*  -.63*  -.15    .72*  -.66*  -.22    .28    .77*    1  
#> carb  -.51*   .58*   .56*   .77*  -.20    .56*  -.66*  -.53*   .02    .16 
#>        carb
#> mpg   -.51*
#> cyl    .58*
#> disp   .56*
#> hp     .77*
#> drat  -.20 
#> wt     .56*
#> qsec  -.66*
#> vs    -.53*
#> am     .02 
#> gear   .16 
#> carb    1  

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (26) -.85* (26) -.75* (27)  .70* (26) -.89* (27)
#> cyl  -.85* (26)   1   (29)  .91* (26)  .81* (26) -.70* (26)  .77* (26)
#> disp -.85* (26)  .91* (26)   1   (29)  .84* (26) -.74* (27)  .89* (27)
#> hp   -.75* (27)  .81* (26)  .84* (26)   1   (29) -.48* (26)  .71* (27)
#> drat  .70* (26) -.70* (26) -.74* (27) -.48* (26)   1   (29) -.72* (26)
#> wt   -.89* (27)  .77* (26)  .89* (27)  .71* (27) -.72* (26)   1   (29)
#> qsec  .29  (26) -.66* (26) -.42* (27) -.74* (26)  .14  (26) -.09  (26)
#> vs    .63* (26) -.90* (27) -.75* (26) -.72* (27)  .59* (26) -.57* (26)
#> am    .50* (26) -.44* (26) -.59* (26) -.22  (26)  .65* (26) -.71* (26)
#> gear  .49* (26) -.44* (26) -.63* (27) -.15  (26)  .72* (28) -.66* (26)
#> carb -.51* (26)  .58* (26)  .56* (26)  .77* (26) -.20  (26)  .56* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .29  (26)  .63* (26)  .50* (26)  .49* (26) -.51* (26)
#> cyl  -.66* (26) -.90* (27) -.44* (26) -.44* (26)  .58* (26)
#> disp -.42* (27) -.75* (26) -.59* (26) -.63* (27)  .56* (26)
#> hp   -.74* (26) -.72* (27) -.22  (26) -.15  (26)  .77* (26)
#> drat  .14  (26)  .59* (26)  .65* (26)  .72* (28) -.20  (26)
#> wt   -.09  (26) -.57* (26) -.71* (26) -.66* (26)  .56* (26)
#> qsec   1   (29)  .76* (26) -.28  (26) -.22  (26) -.66* (27)
#> vs    .76* (26)   1   (29)  .23  (26)  .28  (26) -.53* (26)
#> am   -.28  (26)  .23  (26)   1   (29)  .77* (26)  .02  (27)
#> gear -.22  (26)  .28  (26)  .77* (26)   1   (29)  .16  (26)
#> carb -.66* (27) -.53* (26)  .02  (27)  .16  (26)   1   (29)

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.85  -.85  -.75   .70  -.89   .29   .63   .50   .49  -.51
#> N  29    26    26    27    26    27    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .15   .00   .01   .01   .01
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   1   .91   .81  -.70   .77  -.66  -.90  -.44  -.44   .58
#> N    26  29    26    26    26    26    26    27    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .03   .00
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   .91    1   .84  -.74   .89  -.42  -.75  -.59  -.63   .56
#> N    26    26   29    26    27    27    27    26    26    27    26
#> P   .00   .00        .00   .00   .00   .03   .00   .00   .00   .00
#> 
#> , , hp
#> 
#>     mpg   cyl  disp  hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.75   .81   .84   1  -.48   .71  -.74  -.72  -.22  -.15   .77
#> N    27    26    26  29    26    27    26    27    26    26    26
#> P   .00   .00   .00       .01   .00   .00   .00   .28   .47   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .70  -.70  -.74  -.48    1  -.72   .14   .59   .65   .72  -.20
#> N    26    26    27    26   29    26    26    26    26    28    26
#> P   .00   .00   .00   .01        .00   .49   .00   .00   .00   .34
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.89   .77   .89   .71  -.72   1  -.09  -.57  -.71  -.66   .56
#> N    27    26    27    27    26  29    26    26    26    26    26
#> P   .00   .00   .00   .00   .00       .65   .00   .00   .00   .00
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .29  -.66  -.42  -.74   .14  -.09    1   .76  -.28  -.22  -.66
#> N    26    26    27    26    26    26   29    26    26    26    27
#> P   .15   .00   .03   .00   .49   .65        .00   .16   .28   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .63  -.90  -.75  -.72   .59  -.57   .76   1   .23   .28  -.53
#> N    26    27    26    27    26    26    26  29    26    26    26
#> P   .00   .00   .00   .00   .00   .00   .00       .27   .17   .01
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .50  -.44  -.59  -.22   .65  -.71  -.28   .23   1   .77   .02
#> N    26    26    26    26    26    26    26    26  29    26    27
#> P   .01   .02   .00   .28   .00   .00   .16   .27       .00   .92
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .49  -.44  -.63  -.15   .72  -.66  -.22   .28   .77    1   .16
#> N    26    26    27    26    28    26    26    26    26   29    26
#> P   .01   .03   .00   .47   .00   .00   .28   .17   .00        .43
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.51   .58   .56   .77  -.20   .56  -.66  -.53   .02   .16    1
#> N    26    26    26    26    26    26    27    26    27    26   29
#> P   .01   .00   .00   .00   .34   .00   .00   .01   .92   .43     
#> 

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.847* (26)   1    (29)                                                
#> disp -.847* (26)  .910* (26)   1    (29)                                    
#> hp   -.748* (27)  .811* (26)  .845* (26)   1    (29)                        
#> drat  .699* (26) -.705* (26) -.737* (27) -.476  (26)   1    (29)            
#> wt   -.886* (27)  .767* (26)  .895* (27)  .707* (27) -.716* (26)   1    (29)
#> qsec  .291  (26) -.662* (26) -.420  (27) -.741* (26)  .142  (26) -.095  (26)
#> vs    .633* (26) -.902* (27) -.746* (26) -.722* (27)  .594* (26) -.572* (26)
#> am    .502* (26) -.440  (26) -.589* (26) -.221  (26)  .650* (26) -.711* (26)
#> gear  .494  (26) -.437  (26) -.626* (27) -.149  (26)  .724* (28) -.662* (26)
#> carb -.507* (26)  .585* (26)  .565* (26)  .767* (26) -.195  (26)  .556* (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .759* (26)   1    (29)                                    
#> am   -.281  (26)  .226  (26)   1    (29)                        
#> gear -.220  (26)  .279  (26)  .767* (26)   1    (29)            
#> carb -.655* (27) -.533* (26)  .021  (27)  .161  (26)   1    (29)

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.85  -.85  -.75   .70  -.89   .29   .63   .50   .49  -.51
#> cyl   -.85     1   .91   .81  -.70   .77  -.66  -.90  -.44  -.44   .58
#> disp  -.85   .91     1   .84  -.74   .89  -.42  -.75  -.59  -.63   .56
#> hp    -.75   .81   .84     1  -.48   .71  -.74  -.72  -.22  -.15   .77
#> drat   .70  -.70  -.74  -.48     1  -.72   .14   .59   .65   .72  -.20
#> wt    -.89   .77   .89   .71  -.72     1  -.09  -.57  -.71  -.66   .56
#> qsec   .29  -.66  -.42  -.74   .14  -.09     1   .76  -.28  -.22  -.66
#> vs     .63  -.90  -.75  -.72   .59  -.57   .76     1   .23   .28  -.53
#> am     .50  -.44  -.59  -.22   .65  -.71  -.28   .23     1   .77   .02
#> gear   .49  -.44  -.63  -.15   .72  -.66  -.22   .28   .77     1   .16
#> carb  -.51   .58   .56   .77  -.20   .56  -.66  -.53   .02   .16     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  27   26  26  26   26   26
#> cyl   26  29   26  26   26  26   26  27  26   26   26
#> disp  26  26   29  26   27  27   27  26  26   27   26
#> hp    27  26   26  29   26  27   26  27  26   26   26
#> drat  26  26   27  26   29  26   26  26  26   28   26
#> wt    27  26   27  27   26  29   26  26  26   26   26
#> qsec  26  26   27  26   26  26   29  26  26   26   27
#> vs    26  27   26  27   26  26   26  29  26   26   26
#> am    26  26   26  26   26  26   26  26  29   26   27
#> gear  26  26   27  26   28  26   26  26  26   29   26
#> carb  26  26   26  26   26  26   27  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .15  .00  .01  .01  .01
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .03  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .00
#> hp    .00  .00  .00       .01  .00  .00  .00  .28  .47  .00
#> drat  .00  .00  .00  .01       .00  .49  .00  .00  .00  .34
#> wt    .00  .00  .00  .00  .00       .65  .00  .00  .00  .00
#> qsec  .15  .00  .03  .00  .49  .65       .00  .16  .28  .00
#> vs    .00  .00  .00  .00  .00  .00  .00       .27  .17  .01
#> am    .01  .02  .00  .28  .00  .00  .16  .27       .00  .92
#> gear  .01  .03  .00  .47  .00  .00  .28  .17  .00       .43
#> carb  .01  .00  .00  .00  .34  .00  .00  .01  .92  .43     
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.85     1                                                     
#> disp  -.85   .91     1                                               
#> hp    -.75   .81   .84     1                                         
#> drat   .70  -.70  -.74  -.48     1                                   
#> wt    -.89   .77   .89   .71  -.72     1                             
#> qsec   .29  -.66  -.42  -.74   .14  -.09     1                       
#> vs     .63  -.90  -.75  -.72   .59  -.57   .76     1                 
#> am     .50  -.44  -.59  -.22   .65  -.71  -.28   .23     1           
#> gear   .49  -.44  -.63  -.15   .72  -.66  -.22   .28   .77     1     
#> carb  -.51   .58   .56   .77  -.20   .56  -.66  -.53   .02   .16    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    27  26   26  29                                
#> drat  26  26   27  26   29                           
#> wt    27  26   27  27   26  29                       
#> qsec  26  26   27  26   26  26   29                  
#> vs    26  27   26  27   26  26   26  29              
#> am    26  26   26  26   26  26   26  26  29          
#> gear  26  26   27  26   28  26   26  26  26   29     
#> carb  26  26   26  26   26  26   27  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .15  .00  .03  .00  .49  .65                         
#> vs    .00  .00  .00  .00  .00  .00  .00                    
#> am    .01  .02  .00  .28  .00  .00  .16  .27               
#> gear  .01  .03  .00  .47  .00  .00  .28  .17  .00          
#> carb  .01  .00  .00  .00  .34  .00  .00  .01  .92  .43     
#>