(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.87  -.82  -.73   .59  -.85   .34   .63   .57   .51  -.54
#> cyl   -.87    1    .91   .81  -.62   .80  -.56  -.77  -.46  -.46   .54
#> disp  -.82   .91    1    .80  -.67   .85  -.44  -.67  -.53  -.48   .37
#> hp    -.73   .81   .80    1   -.35   .67  -.72  -.74  -.28  -.09   .75
#> drat   .59  -.62  -.67  -.35    1   -.65   .05   .29   .69   .73  -.05
#> wt    -.85   .80   .85   .67  -.65    1   -.16  -.55  -.71  -.56   .43
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      28.90     -7.73   -486.01   -265.15      1.36     -4.00      3.40
#> cyl      -7.73      3.12    166.76     99.00     -0.51      1.30     -1.83
#> disp   -486.01    166.76  13995.81   6809.15    -37.89     82.34    -96.84
#> hp     -265.15     99.00   6809.15   4750.96    -10.98     44.66    -88.45
#> drat      1.36     -0.51    -37.89    -10.98      0.24     -0.30      0.05
#> wt       -4.00      1.30     82.34     44.66     -0.30      0.88     -0.28
#>             vs        am      gear      carb
#> mpg       1.80      1.54      2.25     -4.99
#> cyl      -0.68     -0.41     -0.63      1.61
#> disp    -40.27    -32.73    -42.85     76.29
#> hp      -26.89     -9.69     -4.85     87.34
#> drat      0.07      0.17      0.26     -0.04
#> wt       -0.27     -0.35     -0.42      0.70
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   28 26   26 26 26   26   26
#> cyl   26  29   26 26   26 27   26 27 26   27   26
#> disp  26  26   29 26   26 27   27 26 26   26   26
#> hp    26  26   26 29   26 27   26 26 27   26   27
#> drat  28  26   26 26   29 26   26 26 26   26   26
#> wt    26  27   27 27   26 29   26 26 26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.87 (26) -.82 (26) -.73 (26)  .59 (28) -.85 (26)  .34 (26)
#> cyl  -.87 (26)   1  (29)  .91 (26)  .81 (26) -.62 (26)  .80 (27) -.56 (26)
#> disp -.82 (26)  .91 (26)   1  (29)  .80 (26) -.67 (26)  .85 (27) -.44 (27)
#> hp   -.73 (26)  .81 (26)  .80 (26)   1  (29) -.35 (26)  .67 (27) -.72 (26)
#> drat  .59 (28) -.62 (26) -.67 (26) -.35 (26)   1  (29) -.65 (26)  .05 (26)
#> wt   -.85 (26)  .80 (27)  .85 (27)  .67 (27) -.65 (26)   1  (29) -.16 (26)
#>             vs        am      gear      carb
#> mpg   .63 (26)  .57 (26)  .51 (26) -.54 (26)
#> cyl  -.77 (27) -.46 (26) -.46 (27)  .54 (26)
#> disp -.67 (26) -.53 (26) -.48 (26)  .37 (26)
#> hp   -.74 (26) -.28 (27) -.09 (26)  .75 (27)
#> drat  .29 (26)  .69 (26)  .73 (26) -.05 (26)
#> wt   -.55 (26) -.71 (26) -.56 (26)  .43 (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.87*  -.82*  -.73*   .59*  -.85*   .34    .63*   .57*   .51*
#> cyl   -.87*    1     .91*   .81*  -.62*   .80*  -.56*  -.77*  -.46*  -.46*
#> disp  -.82*   .91*    1     .80*  -.67*   .85*  -.44*  -.67*  -.53*  -.48*
#> hp    -.73*   .81*   .80*    1    -.35    .67*  -.72*  -.74*  -.28   -.09 
#> drat   .59*  -.62*  -.67*  -.35     1    -.65*   .05    .29    .69*   .73*
#> wt    -.85*   .80*   .85*   .67*  -.65*    1    -.16   -.55*  -.71*  -.56*
#>        carb
#> mpg   -.54*
#> cyl    .54*
#> disp   .37 
#> hp     .75*
#> drat  -.05 
#> wt     .43*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.87* (26) -.82* (26) -.73* (26)  .59* (28) -.85* (26)
#> cyl  -.87* (26)   1   (29)  .91* (26)  .81* (26) -.62* (26)  .80* (27)
#> disp -.82* (26)  .91* (26)   1   (29)  .80* (26) -.67* (26)  .85* (27)
#> hp   -.73* (26)  .81* (26)  .80* (26)   1   (29) -.35  (26)  .67* (27)
#> drat  .59* (28) -.62* (26) -.67* (26) -.35  (26)   1   (29) -.65* (26)
#> wt   -.85* (26)  .80* (27)  .85* (27)  .67* (27) -.65* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .34  (26)  .63* (26)  .57* (26)  .51* (26) -.54* (26)
#> cyl  -.56* (26) -.77* (27) -.46* (26) -.46* (27)  .54* (26)
#> disp -.44* (27) -.67* (26) -.53* (26) -.48* (26)  .37  (26)
#> hp   -.72* (26) -.74* (26) -.28  (27) -.09  (26)  .75* (27)
#> drat  .05  (26)  .29  (26)  .69* (26)  .73* (26) -.05  (26)
#> wt   -.16  (26) -.55* (26) -.71* (26) -.56* (26)  .43* (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.87  -.82  -.73   .59  -.85   .34   .63   .57   .51  -.54
#> N  29    26    26    26    28    26    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .09   .00   .00   .01   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.87   1   .91   .81  -.62   .80  -.56  -.77  -.46  -.46   .54
#> N    26  29    26    26    26    27    26    27    26    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .01   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.867* (26)   1    (29)                                                
#> disp -.822* (26)  .911* (26)   1    (29)                                    
#> hp   -.730* (26)  .808* (26)  .798* (26)   1    (29)                        
#> drat  .589* (28) -.618* (26) -.672* (26) -.345  (26)   1    (29)            
#> wt   -.849* (26)  .801* (27)  .853* (27)  .666* (27) -.654* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.87  -.82  -.73   .59  -.85   .34   .63   .57   .51  -.54
#> cyl   -.87     1   .91   .81  -.62   .80  -.56  -.77  -.46  -.46   .54
#> disp  -.82   .91     1   .80  -.67   .85  -.44  -.67  -.53  -.48   .37
#> hp    -.73   .81   .80     1  -.35   .67  -.72  -.74  -.28  -.09   .75
#> drat   .59  -.62  -.67  -.35     1  -.65   .05   .29   .69   .73  -.05
#> wt    -.85   .80   .85   .67  -.65     1  -.16  -.55  -.71  -.56   .43
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   28  26   26  26  26   26   26
#> cyl   26  29   26  26   26  27   26  27  26   27   26
#> disp  26  26   29  26   26  27   27  26  26   26   26
#> hp    26  26   26  29   26  27   26  26  27   26   27
#> drat  28  26   26  26   29  26   26  26  26   26   26
#> wt    26  27   27  27   26  29   26  26  26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .09  .00  .00  .01  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .01  .00
#> disp  .00  .00       .00  .00  .00  .02  .00  .01  .01  .06
#> hp    .00  .00  .00       .08  .00  .00  .00  .16  .65  .00
#> drat  .00  .00  .00  .08       .00  .80  .14  .00  .00  .82
#> wt    .00  .00  .00  .00  .00       .44  .00  .00  .00  .03
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.87     1                                                     
#> disp  -.82   .91     1                                               
#> hp    -.73   .81   .80     1                                         
#> drat   .59  -.62  -.67  -.35     1                                   
#> wt    -.85   .80   .85   .67  -.65     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  26   26  29                                
#> drat  28  26   26  26   29                           
#> wt    26  27   27  27   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .08                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>