Recursive Splitting

rsplit recursively splits a vector or data frame into subsets according to combinations of (multiple) vectors / factors - by default returning a (nested) list. If flatten = TRUE, the list is flattened yielding the same result as split. rsplit is implemented as a wrapper around gsplit, and faster than split.

rsplit(x, ...)

# S3 method for default
rsplit(x, fl, drop = TRUE, flatten = FALSE, use.names = TRUE, ...)

# S3 method for data.frame
rsplit(x, by, drop = TRUE, flatten = FALSE, cols = NULL,
       keep.by = FALSE, simplify = TRUE, use.names = TRUE, ...)

Arguments

x

a vector, data.frame or list.

fl

a vector / factor, GRP object, or list of vectors / factors used to split x.

by

data.frame method: Same as fl, but also allows one- or two-sided formulas i.e. ~ group1 or var1 + var2 ~ group1 + group2. See Examples.

drop

logical. TRUE removes unused levels or combinations of levels from factors before splitting; FALSE retains those combinations yielding empty list elements in the output.

flatten

logical. If fl is a list of vectors / factors, TRUE calls GRP on the list, creating a single grouping used for splitting; FALSE yields recursive splitting.

use.names

logical. TRUE returns a named list (like split); FALSE returns a plain list.

cols

data.frame method: Select columns to split using a function, column names, indices or a logical vector. Note: cols is ignored if a two-sided formula is passed to by.

keep.by

logical. If a formula is passed to by, then TRUE preserves the splitting (right-hand-side) variables in the data frame.

simplify

data.frame method: Logical. TRUE calls rsplit.default if a single column is split e.g. rsplit(data, col1 ~ group1) becomes the same as rsplit(data$col1, data$group1).

...

further arguments passed to GRP. Sensible choices would be sort = FALSE, decreasing = TRUE or na.last = FALSE. Note that these options only apply if fl is not already a factor.

Value

a (nested) list containing the subsets of x.

See also

Examples

rsplit(mtcars$mpg, mtcars$cyl)
#> $`4`
#>  [1] 22.8 24.4 22.8 32.4 30.4 33.9 21.5 27.3 26.0 30.4 21.4
#> 
#> $`6`
#> [1] 21.0 21.0 21.4 18.1 19.2 17.8 19.7
#> 
#> $`8`
#>  [1] 18.7 14.3 16.4 17.3 15.2 10.4 10.4 14.7 15.5 15.2 13.3 19.2 15.8 15.0
#> 
rsplit(mtcars, mtcars$cyl)
#> $`4`
#>                 mpg cyl  disp hp drat    wt  qsec vs am gear carb           X
#> Datsun 710     22.8   4 108.0 93 3.85 2.320 18.61  1  1    4    1  1.38728007
#> Merc 240D      24.4   4 146.7 62 3.69 3.190 20.00  1  0    4    2  0.22512721
#> Merc 230       22.8   4 140.8 95 3.92 3.150 22.90  1  0    4    2 -0.97389777
#> Fiat 128       32.4   4  78.7 66 4.08 2.200 19.47  1  1    4    1  0.54689909
#> Honda Civic    30.4   4  75.7 52 4.93 1.615 18.52  1  1    4    2 -0.03580896
#> Toyota Corolla 33.9   4  71.1 65 4.22 1.835 19.90  1  1    4    1  0.78018236
#> Toyota Corona  21.5   4 120.1 97 3.70 2.465 20.01  1  0    3    1 -0.90357957
#> Fiat X1-9      27.3   4  79.0 66 4.08 1.935 18.90  1  1    4    1  1.36428721
#>  [ reached 'max' / getOption("max.print") -- omitted 3 rows ]
#> 
#> $`6`
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb           X
#> Mazda RX4      21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 -0.76272049
#> Mazda RX4 Wag  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 -0.08641398
#> Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 -0.65361593
#> Valiant        18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1  0.29165057
#> Merc 280       19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 -1.49434781
#> Merc 280C      17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 -0.72575648
#> Ferrari Dino   19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6  0.32232900
#> 
#> $`8`
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#>                               X
#> Hornet Sportabout   -0.26870764
#> Duster 360          -0.86896897
#> Merc 450SE          -1.80350382
#> Merc 450SL          -1.96684935
#> Merc 450SLC         -0.31206217
#> Cadillac Fleetwood  -0.07331864
#> Lincoln Continental  0.14952405
#> Chrysler Imperial    1.75274156
#>  [ reached 'max' / getOption("max.print") -- omitted 6 rows ]
#> 

rsplit(mtcars, mtcars[.c(cyl, vs, am)])
#> $`4`
#> $`4`$`0`
#> $`4`$`0`$`1`
#>               mpg cyl  disp hp drat   wt qsec vs am gear carb          X
#> Porsche 914-2  26   4 120.3 91 4.43 2.14 16.7  0  1    5    2 -0.5538141
#> 
#> 
#> $`4`$`1`
#> $`4`$`1`$`0`
#>                mpg cyl  disp hp drat    wt  qsec vs am gear carb          X
#> Merc 240D     24.4   4 146.7 62 3.69 3.190 20.00  1  0    4    2  0.2251272
#> Merc 230      22.8   4 140.8 95 3.92 3.150 22.90  1  0    4    2 -0.9738978
#> Toyota Corona 21.5   4 120.1 97 3.70 2.465 20.01  1  0    3    1 -0.9035796
#> 
#> $`4`$`1`$`1`
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb           X
#> Datsun 710     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1  1.38728007
#> Fiat 128       32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1  0.54689909
#> Honda Civic    30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 -0.03580896
#> Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1  0.78018236
#> Fiat X1-9      27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1  1.36428721
#> Lotus Europa   30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 -0.40406286
#> Volvo 142E     21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2  1.34516688
#> 
#> 
#> 
#> $`6`
#> $`6`$`0`
#> $`6`$`0`$`1`
#>                mpg cyl disp  hp drat    wt  qsec vs am gear carb           X
#> Mazda RX4     21.0   6  160 110 3.90 2.620 16.46  0  1    4    4 -0.76272049
#> Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4 -0.08641398
#> Ferrari Dino  19.7   6  145 175 3.62 2.770 15.50  0  1    5    6  0.32232900
#> 
#> 
#> $`6`$`1`
#> $`6`$`1`$`0`
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb          X
#> Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 -0.6536159
#> Valiant        18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1  0.2916506
#> Merc 280       19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 -1.4943478
#> Merc 280C      17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 -0.7257565
#> 
#> 
#> 
#> $`8`
#> $`8`$`0`
#> $`8`$`0`$`0`
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#>                               X
#> Hornet Sportabout   -0.26870764
#> Duster 360          -0.86896897
#> Merc 450SE          -1.80350382
#> Merc 450SL          -1.96684935
#> Merc 450SLC         -0.31206217
#> Cadillac Fleetwood  -0.07331864
#> Lincoln Continental  0.14952405
#> Chrysler Imperial    1.75274156
#>  [ reached 'max' / getOption("max.print") -- omitted 4 rows ]
#> 
#> $`8`$`0`$`1`
#>                 mpg cyl disp  hp drat   wt qsec vs am gear carb          X
#> Ford Pantera L 15.8   8  351 264 4.22 3.17 14.5  0  1    5    4 -0.2300432
#> Maserati Bora  15.0   8  301 335 3.54 3.57 14.6  0  1    5    8 -2.1486097
#> 
#> 
#> 
rsplit(mtcars, ~ cyl + vs + am, keep.by = TRUE)  # Same thing
#> $`4`
#> $`4`$`0`
#> $`4`$`0`$`1`
#>               mpg cyl  disp hp drat   wt qsec vs am gear carb          X
#> Porsche 914-2  26   4 120.3 91 4.43 2.14 16.7  0  1    5    2 -0.5538141
#> 
#> 
#> $`4`$`1`
#> $`4`$`1`$`0`
#>                mpg cyl  disp hp drat    wt  qsec vs am gear carb          X
#> Merc 240D     24.4   4 146.7 62 3.69 3.190 20.00  1  0    4    2  0.2251272
#> Merc 230      22.8   4 140.8 95 3.92 3.150 22.90  1  0    4    2 -0.9738978
#> Toyota Corona 21.5   4 120.1 97 3.70 2.465 20.01  1  0    3    1 -0.9035796
#> 
#> $`4`$`1`$`1`
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb           X
#> Datsun 710     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1  1.38728007
#> Fiat 128       32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1  0.54689909
#> Honda Civic    30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 -0.03580896
#> Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1  0.78018236
#> Fiat X1-9      27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1  1.36428721
#> Lotus Europa   30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 -0.40406286
#> Volvo 142E     21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2  1.34516688
#> 
#> 
#> 
#> $`6`
#> $`6`$`0`
#> $`6`$`0`$`1`
#>                mpg cyl disp  hp drat    wt  qsec vs am gear carb           X
#> Mazda RX4     21.0   6  160 110 3.90 2.620 16.46  0  1    4    4 -0.76272049
#> Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4 -0.08641398
#> Ferrari Dino  19.7   6  145 175 3.62 2.770 15.50  0  1    5    6  0.32232900
#> 
#> 
#> $`6`$`1`
#> $`6`$`1`$`0`
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb          X
#> Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 -0.6536159
#> Valiant        18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1  0.2916506
#> Merc 280       19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 -1.4943478
#> Merc 280C      17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 -0.7257565
#> 
#> 
#> 
#> $`8`
#> $`8`$`0`
#> $`8`$`0`$`0`
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#>                               X
#> Hornet Sportabout   -0.26870764
#> Duster 360          -0.86896897
#> Merc 450SE          -1.80350382
#> Merc 450SL          -1.96684935
#> Merc 450SLC         -0.31206217
#> Cadillac Fleetwood  -0.07331864
#> Lincoln Continental  0.14952405
#> Chrysler Imperial    1.75274156
#>  [ reached 'max' / getOption("max.print") -- omitted 4 rows ]
#> 
#> $`8`$`0`$`1`
#>                 mpg cyl disp  hp drat   wt qsec vs am gear carb          X
#> Ford Pantera L 15.8   8  351 264 4.22 3.17 14.5  0  1    5    4 -0.2300432
#> Maserati Bora  15.0   8  301 335 3.54 3.57 14.6  0  1    5    8 -2.1486097
#> 
#> 
#> 
rsplit(mtcars, ~ cyl + vs + am)
#> $`4`
#> $`4`$`0`
#> $`4`$`0`$`1`
#>               mpg  disp hp drat   wt qsec gear carb          X
#> Porsche 914-2  26 120.3 91 4.43 2.14 16.7    5    2 -0.5538141
#> 
#> 
#> $`4`$`1`
#> $`4`$`1`$`0`
#>                mpg  disp hp drat    wt  qsec gear carb          X
#> Merc 240D     24.4 146.7 62 3.69 3.190 20.00    4    2  0.2251272
#> Merc 230      22.8 140.8 95 3.92 3.150 22.90    4    2 -0.9738978
#> Toyota Corona 21.5 120.1 97 3.70 2.465 20.01    3    1 -0.9035796
#> 
#> $`4`$`1`$`1`
#>                 mpg  disp  hp drat    wt  qsec gear carb           X
#> Datsun 710     22.8 108.0  93 3.85 2.320 18.61    4    1  1.38728007
#> Fiat 128       32.4  78.7  66 4.08 2.200 19.47    4    1  0.54689909
#> Honda Civic    30.4  75.7  52 4.93 1.615 18.52    4    2 -0.03580896
#> Toyota Corolla 33.9  71.1  65 4.22 1.835 19.90    4    1  0.78018236
#> Fiat X1-9      27.3  79.0  66 4.08 1.935 18.90    4    1  1.36428721
#> Lotus Europa   30.4  95.1 113 3.77 1.513 16.90    5    2 -0.40406286
#> Volvo 142E     21.4 121.0 109 4.11 2.780 18.60    4    2  1.34516688
#> 
#> 
#> 
#> $`6`
#> $`6`$`0`
#> $`6`$`0`$`1`
#>                mpg disp  hp drat    wt  qsec gear carb           X
#> Mazda RX4     21.0  160 110 3.90 2.620 16.46    4    4 -0.76272049
#> Mazda RX4 Wag 21.0  160 110 3.90 2.875 17.02    4    4 -0.08641398
#> Ferrari Dino  19.7  145 175 3.62 2.770 15.50    5    6  0.32232900
#> 
#> 
#> $`6`$`1`
#> $`6`$`1`$`0`
#>                 mpg  disp  hp drat    wt  qsec gear carb          X
#> Hornet 4 Drive 21.4 258.0 110 3.08 3.215 19.44    3    1 -0.6536159
#> Valiant        18.1 225.0 105 2.76 3.460 20.22    3    1  0.2916506
#> Merc 280       19.2 167.6 123 3.92 3.440 18.30    4    4 -1.4943478
#> Merc 280C      17.8 167.6 123 3.92 3.440 18.90    4    4 -0.7257565
#> 
#> 
#> 
#> $`8`
#> $`8`$`0`
#> $`8`$`0`$`0`
#>                      mpg  disp  hp drat    wt  qsec gear carb           X
#> Hornet Sportabout   18.7 360.0 175 3.15 3.440 17.02    3    2 -0.26870764
#> Duster 360          14.3 360.0 245 3.21 3.570 15.84    3    4 -0.86896897
#> Merc 450SE          16.4 275.8 180 3.07 4.070 17.40    3    3 -1.80350382
#> Merc 450SL          17.3 275.8 180 3.07 3.730 17.60    3    3 -1.96684935
#> Merc 450SLC         15.2 275.8 180 3.07 3.780 18.00    3    3 -0.31206217
#> Cadillac Fleetwood  10.4 472.0 205 2.93 5.250 17.98    3    4 -0.07331864
#> Lincoln Continental 10.4 460.0 215 3.00 5.424 17.82    3    4  0.14952405
#> Chrysler Imperial   14.7 440.0 230 3.23 5.345 17.42    3    4  1.75274156
#> Dodge Challenger    15.5 318.0 150 2.76 3.520 16.87    3    2 -0.86590275
#> AMC Javelin         15.2 304.0 150 3.15 3.435 17.30    3    2  0.46707835
#> Camaro Z28          13.3 350.0 245 3.73 3.840 15.41    3    4  0.38320839
#>  [ reached 'max' / getOption("max.print") -- omitted 1 rows ]
#> 
#> $`8`$`0`$`1`
#>                 mpg disp  hp drat   wt qsec gear carb          X
#> Ford Pantera L 15.8  351 264 4.22 3.17 14.5    5    4 -0.2300432
#> Maserati Bora  15.0  301 335 3.54 3.57 14.6    5    8 -2.1486097
#> 
#> 
#> 

rsplit(mtcars, ~ cyl + vs + am, flatten = TRUE)
#> $`4.0.1`
#>               mpg  disp hp drat   wt qsec gear carb          X
#> Porsche 914-2  26 120.3 91 4.43 2.14 16.7    5    2 -0.5538141
#> 
#> $`4.1.0`
#>                mpg  disp hp drat    wt  qsec gear carb          X
#> Merc 240D     24.4 146.7 62 3.69 3.190 20.00    4    2  0.2251272
#> Merc 230      22.8 140.8 95 3.92 3.150 22.90    4    2 -0.9738978
#> Toyota Corona 21.5 120.1 97 3.70 2.465 20.01    3    1 -0.9035796
#> 
#> $`4.1.1`
#>                 mpg  disp  hp drat    wt  qsec gear carb           X
#> Datsun 710     22.8 108.0  93 3.85 2.320 18.61    4    1  1.38728007
#> Fiat 128       32.4  78.7  66 4.08 2.200 19.47    4    1  0.54689909
#> Honda Civic    30.4  75.7  52 4.93 1.615 18.52    4    2 -0.03580896
#> Toyota Corolla 33.9  71.1  65 4.22 1.835 19.90    4    1  0.78018236
#> Fiat X1-9      27.3  79.0  66 4.08 1.935 18.90    4    1  1.36428721
#> Lotus Europa   30.4  95.1 113 3.77 1.513 16.90    5    2 -0.40406286
#> Volvo 142E     21.4 121.0 109 4.11 2.780 18.60    4    2  1.34516688
#> 
#> $`6.0.1`
#>                mpg disp  hp drat    wt  qsec gear carb           X
#> Mazda RX4     21.0  160 110 3.90 2.620 16.46    4    4 -0.76272049
#> Mazda RX4 Wag 21.0  160 110 3.90 2.875 17.02    4    4 -0.08641398
#> Ferrari Dino  19.7  145 175 3.62 2.770 15.50    5    6  0.32232900
#> 
#> $`6.1.0`
#>                 mpg  disp  hp drat    wt  qsec gear carb          X
#> Hornet 4 Drive 21.4 258.0 110 3.08 3.215 19.44    3    1 -0.6536159
#> Valiant        18.1 225.0 105 2.76 3.460 20.22    3    1  0.2916506
#> Merc 280       19.2 167.6 123 3.92 3.440 18.30    4    4 -1.4943478
#> Merc 280C      17.8 167.6 123 3.92 3.440 18.90    4    4 -0.7257565
#> 
#> $`8.0.0`
#>                      mpg  disp  hp drat    wt  qsec gear carb           X
#> Hornet Sportabout   18.7 360.0 175 3.15 3.440 17.02    3    2 -0.26870764
#> Duster 360          14.3 360.0 245 3.21 3.570 15.84    3    4 -0.86896897
#> Merc 450SE          16.4 275.8 180 3.07 4.070 17.40    3    3 -1.80350382
#> Merc 450SL          17.3 275.8 180 3.07 3.730 17.60    3    3 -1.96684935
#> Merc 450SLC         15.2 275.8 180 3.07 3.780 18.00    3    3 -0.31206217
#> Cadillac Fleetwood  10.4 472.0 205 2.93 5.250 17.98    3    4 -0.07331864
#> Lincoln Continental 10.4 460.0 215 3.00 5.424 17.82    3    4  0.14952405
#> Chrysler Imperial   14.7 440.0 230 3.23 5.345 17.42    3    4  1.75274156
#> Dodge Challenger    15.5 318.0 150 2.76 3.520 16.87    3    2 -0.86590275
#> AMC Javelin         15.2 304.0 150 3.15 3.435 17.30    3    2  0.46707835
#> Camaro Z28          13.3 350.0 245 3.73 3.840 15.41    3    4  0.38320839
#>  [ reached 'max' / getOption("max.print") -- omitted 1 rows ]
#> 
#> $`8.0.1`
#>                 mpg disp  hp drat   wt qsec gear carb          X
#> Ford Pantera L 15.8  351 264 4.22 3.17 14.5    5    4 -0.2300432
#> Maserati Bora  15.0  301 335 3.54 3.57 14.6    5    8 -2.1486097
#> 

rsplit(mtcars, mpg ~ cyl)
#> $`4`
#>  [1] 22.8 24.4 22.8 32.4 30.4 33.9 21.5 27.3 26.0 30.4 21.4
#> 
#> $`6`
#> [1] 21.0 21.0 21.4 18.1 19.2 17.8 19.7
#> 
#> $`8`
#>  [1] 18.7 14.3 16.4 17.3 15.2 10.4 10.4 14.7 15.5 15.2 13.3 19.2 15.8 15.0
#> 
rsplit(mtcars, mpg ~ cyl, simplify = FALSE)
#> $`4`
#>                 mpg
#> Datsun 710     22.8
#> Merc 240D      24.4
#> Merc 230       22.8
#> Fiat 128       32.4
#> Honda Civic    30.4
#> Toyota Corolla 33.9
#> Toyota Corona  21.5
#> Fiat X1-9      27.3
#> Porsche 914-2  26.0
#> Lotus Europa   30.4
#> Volvo 142E     21.4
#> 
#> $`6`
#>                 mpg
#> Mazda RX4      21.0
#> Mazda RX4 Wag  21.0
#> Hornet 4 Drive 21.4
#> Valiant        18.1
#> Merc 280       19.2
#> Merc 280C      17.8
#> Ferrari Dino   19.7
#> 
#> $`8`
#>                      mpg
#> Hornet Sportabout   18.7
#> Duster 360          14.3
#> Merc 450SE          16.4
#> Merc 450SL          17.3
#> Merc 450SLC         15.2
#> Cadillac Fleetwood  10.4
#> Lincoln Continental 10.4
#> Chrysler Imperial   14.7
#> Dodge Challenger    15.5
#> AMC Javelin         15.2
#> Camaro Z28          13.3
#> Pontiac Firebird    19.2
#> Ford Pantera L      15.8
#> Maserati Bora       15.0
#> 
rsplit(mtcars, mpg + hp ~ cyl + vs + am)
#> $`4`
#> $`4`$`0`
#> $`4`$`0`$`1`
#>               mpg hp
#> Porsche 914-2  26 91
#> 
#> 
#> $`4`$`1`
#> $`4`$`1`$`0`
#>                mpg hp
#> Merc 240D     24.4 62
#> Merc 230      22.8 95
#> Toyota Corona 21.5 97
#> 
#> $`4`$`1`$`1`
#>                 mpg  hp
#> Datsun 710     22.8  93
#> Fiat 128       32.4  66
#> Honda Civic    30.4  52
#> Toyota Corolla 33.9  65
#> Fiat X1-9      27.3  66
#> Lotus Europa   30.4 113
#> Volvo 142E     21.4 109
#> 
#> 
#> 
#> $`6`
#> $`6`$`0`
#> $`6`$`0`$`1`
#>                mpg  hp
#> Mazda RX4     21.0 110
#> Mazda RX4 Wag 21.0 110
#> Ferrari Dino  19.7 175
#> 
#> 
#> $`6`$`1`
#> $`6`$`1`$`0`
#>                 mpg  hp
#> Hornet 4 Drive 21.4 110
#> Valiant        18.1 105
#> Merc 280       19.2 123
#> Merc 280C      17.8 123
#> 
#> 
#> 
#> $`8`
#> $`8`$`0`
#> $`8`$`0`$`0`
#>                      mpg  hp
#> Hornet Sportabout   18.7 175
#> Duster 360          14.3 245
#> Merc 450SE          16.4 180
#> Merc 450SL          17.3 180
#> Merc 450SLC         15.2 180
#> Cadillac Fleetwood  10.4 205
#> Lincoln Continental 10.4 215
#> Chrysler Imperial   14.7 230
#> Dodge Challenger    15.5 150
#> AMC Javelin         15.2 150
#> Camaro Z28          13.3 245
#> Pontiac Firebird    19.2 175
#> 
#> $`8`$`0`$`1`
#>                 mpg  hp
#> Ford Pantera L 15.8 264
#> Maserati Bora  15.0 335
#> 
#> 
#> 
rsplit(mtcars, mpg + hp ~ cyl + vs + am, keep.by = TRUE)
#> $`4`
#> $`4`$`0`
#> $`4`$`0`$`1`
#>               cyl vs am mpg hp
#> Porsche 914-2   4  0  1  26 91
#> 
#> 
#> $`4`$`1`
#> $`4`$`1`$`0`
#>               cyl vs am  mpg hp
#> Merc 240D       4  1  0 24.4 62
#> Merc 230        4  1  0 22.8 95
#> Toyota Corona   4  1  0 21.5 97
#> 
#> $`4`$`1`$`1`
#>                cyl vs am  mpg  hp
#> Datsun 710       4  1  1 22.8  93
#> Fiat 128         4  1  1 32.4  66
#> Honda Civic      4  1  1 30.4  52
#> Toyota Corolla   4  1  1 33.9  65
#> Fiat X1-9        4  1  1 27.3  66
#> Lotus Europa     4  1  1 30.4 113
#> Volvo 142E       4  1  1 21.4 109
#> 
#> 
#> 
#> $`6`
#> $`6`$`0`
#> $`6`$`0`$`1`
#>               cyl vs am  mpg  hp
#> Mazda RX4       6  0  1 21.0 110
#> Mazda RX4 Wag   6  0  1 21.0 110
#> Ferrari Dino    6  0  1 19.7 175
#> 
#> 
#> $`6`$`1`
#> $`6`$`1`$`0`
#>                cyl vs am  mpg  hp
#> Hornet 4 Drive   6  1  0 21.4 110
#> Valiant          6  1  0 18.1 105
#> Merc 280         6  1  0 19.2 123
#> Merc 280C        6  1  0 17.8 123
#> 
#> 
#> 
#> $`8`
#> $`8`$`0`
#> $`8`$`0`$`0`
#>                     cyl vs am  mpg  hp
#> Hornet Sportabout     8  0  0 18.7 175
#> Duster 360            8  0  0 14.3 245
#> Merc 450SE            8  0  0 16.4 180
#> Merc 450SL            8  0  0 17.3 180
#> Merc 450SLC           8  0  0 15.2 180
#> Cadillac Fleetwood    8  0  0 10.4 205
#> Lincoln Continental   8  0  0 10.4 215
#> Chrysler Imperial     8  0  0 14.7 230
#> Dodge Challenger      8  0  0 15.5 150
#> AMC Javelin           8  0  0 15.2 150
#> Camaro Z28            8  0  0 13.3 245
#> Pontiac Firebird      8  0  0 19.2 175
#> 
#> $`8`$`0`$`1`
#>                cyl vs am  mpg  hp
#> Ford Pantera L   8  0  1 15.8 264
#> Maserati Bora    8  0  1 15.0 335
#> 
#> 
#>