Fast Row/Column Arithmetic for Matrix-Like Objects

Fast operators to perform row- or column-wise replacing and sweeping operations of vectors on matrices, data frames, lists.

## Perform the operation with v and each row of X

X %rr% v    # Replace rows of X with v
X %r+% v    # Add v to each row of X
X %r-% v    # Subtract v from each row of X
X %r*% v    # Multiply each row of X with v
X %r/% v    # Divide each row of X by v

## Perform a column-wise operation between V and X

X %cr% V    # Replace columns of X with V
X %c+% V    # Add V to columns of X
X %c-% V    # Subtract V from columns of X
X %c*% V    # Multiply columns of X with V
X %c/% V    # Divide columns of X by V

Arguments

X

a vector, matrix, data frame or list like object (with rows (r) columns (c) matching v / V).

v

for row operations: an atomic vector of matching NCOL(X). If X is a data frame, v can also be a list of scalar atomic elements. It is also possible to sweep lists of vectors v out of lists of matrices or data frames X.

V

for column operations: a suitable scalar, vector, or matrix / data frame matching NROW(X). X can also be a list of vectors / matrices in which case V can be a scalar / vector / matrix or matching list of scalars / vectors / matrices.

Details

With a matrix or data frame X, the default behavior of R when calling X op v (such as multiplication X * v) is to perform the operation of v with each column of X. The equivalent operation is performed by X %cop% V, with the difference that it computes significantly faster if X/V is a data frame / list. A more complex but frequently required task is to perform an operation with v on each row of X. This is provided based on efficient C++ code by the %rop% set of functions, e.g. X %r*% v efficiently multiplies v to each row of X.

Value

X where the operation with v / V was performed on each row or column. All attributes of X are preserved.

Note

Computations and Output: These functions are all quite simple, they only work with X on the LHS i.e. v %op% X will likely fail. The row operations are simple wrappers around TRA which provides more operations including grouped replacing and sweeping (where v would be a matrix or data frame with less rows than X being mapped to the rows of X by grouping vectors). One consequence is that just like TRA, row-wise mathematical operations (+, -, *, /) always yield numeric output, even if both X and v may be integer. This is different for column- operations which depend on base R and may also preserve integer data.

Rules of Arithmetic: Since these operators are defined as simple infix functions, the normal rules of arithmetic are not respected. So a %c+% b %c*% c evaluates as (a %c+% b) %c*% c. As with all chained infix operations, they are just evaluated sequentially from left to right.

Performance Notes: The function setop and a related set of %op=% operators can be used to perform these operations by reference, and are faster if copies of the output are not required!! Furthermore, for Fast Statistical Functions, using fmedian(X, TRA = "-") will be a tiny bit faster than X %r-% fmedian(X). Also use fwithin(X) for fast centering using the mean, and fscale(X) for fast scaling and centering or mean-preserving scaling.

See also

Examples

## Using data frame's / lists
v <- mtcars$cyl
mtcars %cr% v
#>                     mpg cyl disp hp drat wt qsec vs am gear carb
#> Mazda RX4             6   6    6  6    6  6    6  6  6    6    6
#> Mazda RX4 Wag         6   6    6  6    6  6    6  6  6    6    6
#> Datsun 710            4   4    4  4    4  4    4  4  4    4    4
#> Hornet 4 Drive        6   6    6  6    6  6    6  6  6    6    6
#> Hornet Sportabout     8   8    8  8    8  8    8  8  8    8    8
#> Valiant               6   6    6  6    6  6    6  6  6    6    6
#> Duster 360            8   8    8  8    8  8    8  8  8    8    8
#> Merc 240D             4   4    4  4    4  4    4  4  4    4    4
#> Merc 230              4   4    4  4    4  4    4  4  4    4    4
#> Merc 280              6   6    6  6    6  6    6  6  6    6    6
#> Merc 280C             6   6    6  6    6  6    6  6  6    6    6
#> Merc 450SE            8   8    8  8    8  8    8  8  8    8    8
#> Merc 450SL            8   8    8  8    8  8    8  8  8    8    8
#> Merc 450SLC           8   8    8  8    8  8    8  8  8    8    8
#> Cadillac Fleetwood    8   8    8  8    8  8    8  8  8    8    8
#> Lincoln Continental   8   8    8  8    8  8    8  8  8    8    8
#> Chrysler Imperial     8   8    8  8    8  8    8  8  8    8    8
#> Fiat 128              4   4    4  4    4  4    4  4  4    4    4
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]
mtcars %c-% v
#>                      mpg cyl  disp  hp  drat     wt  qsec vs am gear carb
#> Mazda RX4           15.0   0 154.0 104 -2.10 -3.380 10.46 -6 -5   -2   -2
#> Mazda RX4 Wag       15.0   0 154.0 104 -2.10 -3.125 11.02 -6 -5   -2   -2
#> Datsun 710          18.8   0 104.0  89 -0.15 -1.680 14.61 -3 -3    0   -3
#> Hornet 4 Drive      15.4   0 252.0 104 -2.92 -2.785 13.44 -5 -6   -3   -5
#> Hornet Sportabout   10.7   0 352.0 167 -4.85 -4.560  9.02 -8 -8   -5   -6
#> Valiant             12.1   0 219.0  99 -3.24 -2.540 14.22 -5 -6   -3   -5
#> Duster 360           6.3   0 352.0 237 -4.79 -4.430  7.84 -8 -8   -5   -4
#> Merc 240D           20.4   0 142.7  58 -0.31 -0.810 16.00 -3 -4    0   -2
#> Merc 230            18.8   0 136.8  91 -0.08 -0.850 18.90 -3 -4    0   -2
#> Merc 280            13.2   0 161.6 117 -2.08 -2.560 12.30 -5 -6   -2   -2
#> Merc 280C           11.8   0 161.6 117 -2.08 -2.560 12.90 -5 -6   -2   -2
#> Merc 450SE           8.4   0 267.8 172 -4.93 -3.930  9.40 -8 -8   -5   -5
#> Merc 450SL           9.3   0 267.8 172 -4.93 -4.270  9.60 -8 -8   -5   -5
#> Merc 450SLC          7.2   0 267.8 172 -4.93 -4.220 10.00 -8 -8   -5   -5
#> Cadillac Fleetwood   2.4   0 464.0 197 -5.07 -2.750  9.98 -8 -8   -5   -4
#> Lincoln Continental  2.4   0 452.0 207 -5.00 -2.576  9.82 -8 -8   -5   -4
#> Chrysler Imperial    6.7   0 432.0 222 -4.77 -2.655  9.42 -8 -8   -5   -4
#> Fiat 128            28.4   0  74.7  62  0.08 -1.800 15.47 -3 -3    0   -3
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]
mtcars %r-% seq_col(mtcars)
#>                      mpg cyl  disp  hp  drat     wt  qsec vs am gear carb
#> Mazda RX4           20.0   4 157.0 106 -1.10 -3.380  9.46 -8 -8   -6   -7
#> Mazda RX4 Wag       20.0   4 157.0 106 -1.10 -3.125 10.02 -8 -8   -6   -7
#> Datsun 710          21.8   2 105.0  89 -1.15 -3.680 11.61 -7 -8   -6  -10
#> Hornet 4 Drive      20.4   4 255.0 106 -1.92 -2.785 12.44 -7 -9   -7  -10
#> Hornet Sportabout   17.7   6 357.0 171 -1.85 -2.560 10.02 -8 -9   -7   -9
#> Valiant             17.1   4 222.0 101 -2.24 -2.540 13.22 -7 -9   -7  -10
#> Duster 360          13.3   6 357.0 241 -1.79 -2.430  8.84 -8 -9   -7   -7
#> Merc 240D           23.4   2 143.7  58 -1.31 -2.810 13.00 -7 -9   -6   -9
#> Merc 230            21.8   2 137.8  91 -1.08 -2.850 15.90 -7 -9   -6   -9
#> Merc 280            18.2   4 164.6 119 -1.08 -2.560 11.30 -7 -9   -6   -7
#> Merc 280C           16.8   4 164.6 119 -1.08 -2.560 11.90 -7 -9   -6   -7
#> Merc 450SE          15.4   6 272.8 176 -1.93 -1.930 10.40 -8 -9   -7   -8
#> Merc 450SL          16.3   6 272.8 176 -1.93 -2.270 10.60 -8 -9   -7   -8
#> Merc 450SLC         14.2   6 272.8 176 -1.93 -2.220 11.00 -8 -9   -7   -8
#> Cadillac Fleetwood   9.4   6 469.0 201 -2.07 -0.750 10.98 -8 -9   -7   -7
#> Lincoln Continental  9.4   6 457.0 211 -2.00 -0.576 10.82 -8 -9   -7   -7
#> Chrysler Imperial   13.7   6 437.0 226 -1.77 -0.655 10.42 -8 -9   -7   -7
#> Fiat 128            31.4   2  75.7  62 -0.92 -3.800 12.47 -7 -8   -6  -10
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]
mtcars %r-% lapply(mtcars, quantile, 0.28)
#>                        mpg cyl    disp     hp    drat     wt    qsec vs am gear
#> Mazda RX4            5.296   2  25.536   7.56  0.7724 -0.102 -0.5216  0  1    1
#> Mazda RX4 Wag        5.296   2  25.536   7.56  0.7724  0.153  0.0384  0  1    1
#> Datsun 710           7.096   0 -26.464  -9.44  0.7224 -0.402  1.6284  1  1    1
#> Hornet 4 Drive       5.696   2 123.536   7.56 -0.0476  0.493  2.4584  1  0    0
#> Hornet Sportabout    2.996   4 225.536  72.56  0.0224  0.718  0.0384  0  0    0
#> Valiant              2.396   2  90.536   2.56 -0.3676  0.738  3.2384  1  0    0
#> Duster 360          -1.404   4 225.536 142.56  0.0824  0.848 -1.1416  0  0    0
#> Merc 240D            8.696   0  12.236 -40.44  0.5624  0.468  3.0184  1  0    1
#> Merc 230             7.096   0   6.336  -7.44  0.7924  0.428  5.9184  1  0    1
#> Merc 280             3.496   2  33.136  20.56  0.7924  0.718  1.3184  1  0    1
#> Merc 280C            2.096   2  33.136  20.56  0.7924  0.718  1.9184  1  0    1
#> Merc 450SE           0.696   4 141.336  77.56 -0.0576  1.348  0.4184  0  0    0
#> Merc 450SL           1.596   4 141.336  77.56 -0.0576  1.008  0.6184  0  0    0
#> Merc 450SLC         -0.504   4 141.336  77.56 -0.0576  1.058  1.0184  0  0    0
#> Cadillac Fleetwood  -5.304   4 337.536 102.56 -0.1976  2.528  0.9984  0  0    0
#> Lincoln Continental -5.304   4 325.536 112.56 -0.1276  2.702  0.8384  0  0    0
#> Chrysler Imperial   -1.004   4 305.536 127.56  0.1024  2.623  0.4384  0  0    0
#> Fiat 128            16.696   0 -55.764 -36.44  0.9524 -0.522  2.4884  1  1    1
#>                     carb
#> Mazda RX4              2
#> Mazda RX4 Wag          2
#> Datsun 710            -1
#> Hornet 4 Drive        -1
#> Hornet Sportabout      0
#> Valiant               -1
#> Duster 360             2
#> Merc 240D              0
#> Merc 230               0
#> Merc 280               2
#> Merc 280C              2
#> Merc 450SE             1
#> Merc 450SL             1
#> Merc 450SLC            1
#> Cadillac Fleetwood     2
#> Lincoln Continental    2
#> Chrysler Imperial      2
#> Fiat 128              -1
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]

mtcars %c*% 5       # Significantly faster than mtcars * 5
#>                       mpg cyl   disp   hp  drat     wt   qsec vs am gear carb
#> Mazda RX4           105.0  30  800.0  550 19.50 13.100  82.30  0  5   20   20
#> Mazda RX4 Wag       105.0  30  800.0  550 19.50 14.375  85.10  0  5   20   20
#> Datsun 710          114.0  20  540.0  465 19.25 11.600  93.05  5  5   20    5
#> Hornet 4 Drive      107.0  30 1290.0  550 15.40 16.075  97.20  5  0   15    5
#> Hornet Sportabout    93.5  40 1800.0  875 15.75 17.200  85.10  0  0   15   10
#> Valiant              90.5  30 1125.0  525 13.80 17.300 101.10  5  0   15    5
#> Duster 360           71.5  40 1800.0 1225 16.05 17.850  79.20  0  0   15   20
#> Merc 240D           122.0  20  733.5  310 18.45 15.950 100.00  5  0   20   10
#> Merc 230            114.0  20  704.0  475 19.60 15.750 114.50  5  0   20   10
#> Merc 280             96.0  30  838.0  615 19.60 17.200  91.50  5  0   20   20
#> Merc 280C            89.0  30  838.0  615 19.60 17.200  94.50  5  0   20   20
#> Merc 450SE           82.0  40 1379.0  900 15.35 20.350  87.00  0  0   15   15
#> Merc 450SL           86.5  40 1379.0  900 15.35 18.650  88.00  0  0   15   15
#> Merc 450SLC          76.0  40 1379.0  900 15.35 18.900  90.00  0  0   15   15
#> Cadillac Fleetwood   52.0  40 2360.0 1025 14.65 26.250  89.90  0  0   15   20
#> Lincoln Continental  52.0  40 2300.0 1075 15.00 27.120  89.10  0  0   15   20
#> Chrysler Imperial    73.5  40 2200.0 1150 16.15 26.725  87.10  0  0   15   20
#> Fiat 128            162.0  20  393.5  330 20.40 11.000  97.35  5  5   20    5
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]
mtcars %c*% mtcars  # Significantly faster than mtcars * mtcars
#>                         mpg cyl      disp    hp    drat        wt     qsec vs
#> Mazda RX4            441.00  36  25600.00 12100 15.2100  6.864400 270.9316  0
#> Mazda RX4 Wag        441.00  36  25600.00 12100 15.2100  8.265625 289.6804  0
#> Datsun 710           519.84  16  11664.00  8649 14.8225  5.382400 346.3321  1
#> Hornet 4 Drive       457.96  36  66564.00 12100  9.4864 10.336225 377.9136  1
#> Hornet Sportabout    349.69  64 129600.00 30625  9.9225 11.833600 289.6804  0
#> Valiant              327.61  36  50625.00 11025  7.6176 11.971600 408.8484  1
#> Duster 360           204.49  64 129600.00 60025 10.3041 12.744900 250.9056  0
#> Merc 240D            595.36  16  21520.89  3844 13.6161 10.176100 400.0000  1
#> Merc 230             519.84  16  19824.64  9025 15.3664  9.922500 524.4100  1
#> Merc 280             368.64  36  28089.76 15129 15.3664 11.833600 334.8900  1
#> Merc 280C            316.84  36  28089.76 15129 15.3664 11.833600 357.2100  1
#> Merc 450SE           268.96  64  76065.64 32400  9.4249 16.564900 302.7600  0
#> Merc 450SL           299.29  64  76065.64 32400  9.4249 13.912900 309.7600  0
#> Merc 450SLC          231.04  64  76065.64 32400  9.4249 14.288400 324.0000  0
#> Cadillac Fleetwood   108.16  64 222784.00 42025  8.5849 27.562500 323.2804  0
#> Lincoln Continental  108.16  64 211600.00 46225  9.0000 29.419776 317.5524  0
#> Chrysler Imperial    216.09  64 193600.00 52900 10.4329 28.569025 303.4564  0
#> Fiat 128            1049.76  16   6193.69  4356 16.6464  4.840000 379.0809  1
#>                     am gear carb
#> Mazda RX4            1   16   16
#> Mazda RX4 Wag        1   16   16
#> Datsun 710           1   16    1
#> Hornet 4 Drive       0    9    1
#> Hornet Sportabout    0    9    4
#> Valiant              0    9    1
#> Duster 360           0    9   16
#> Merc 240D            0   16    4
#> Merc 230             0   16    4
#> Merc 280             0   16   16
#> Merc 280C            0   16   16
#> Merc 450SE           0    9    9
#> Merc 450SL           0    9    9
#> Merc 450SLC          0    9    9
#> Cadillac Fleetwood   0    9   16
#> Lincoln Continental  0    9   16
#> Chrysler Imperial    0    9   16
#> Fiat 128             1   16    1
#>  [ reached 'max' / getOption("max.print") -- omitted 14 rows ]

## Using matrices
X <- qM(mtcars)
X %cr% v
#>                     mpg cyl disp hp drat wt qsec vs am gear carb
#> Mazda RX4             6   6    6  6    6  6    6  6  6    6    6
#> Mazda RX4 Wag         6   6    6  6    6  6    6  6  6    6    6
#> Datsun 710            4   4    4  4    4  4    4  4  4    4    4
#> Hornet 4 Drive        6   6    6  6    6  6    6  6  6    6    6
#> Hornet Sportabout     8   8    8  8    8  8    8  8  8    8    8
#> Valiant               6   6    6  6    6  6    6  6  6    6    6
#> Duster 360            8   8    8  8    8  8    8  8  8    8    8
#> Merc 240D             4   4    4  4    4  4    4  4  4    4    4
#> Merc 230              4   4    4  4    4  4    4  4  4    4    4
#> Merc 280              6   6    6  6    6  6    6  6  6    6    6
#> Merc 280C             6   6    6  6    6  6    6  6  6    6    6
#> Merc 450SE            8   8    8  8    8  8    8  8  8    8    8
#> Merc 450SL            8   8    8  8    8  8    8  8  8    8    8
#> Merc 450SLC           8   8    8  8    8  8    8  8  8    8    8
#> Cadillac Fleetwood    8   8    8  8    8  8    8  8  8    8    8
#> Lincoln Continental   8   8    8  8    8  8    8  8  8    8    8
#> Chrysler Imperial     8   8    8  8    8  8    8  8  8    8    8
#> Fiat 128              4   4    4  4    4  4    4  4  4    4    4
#>  [ reached getOption("max.print") -- omitted 14 rows ]
X %c-% v
#>                      mpg cyl  disp  hp  drat     wt  qsec vs am gear carb
#> Mazda RX4           15.0   0 154.0 104 -2.10 -3.380 10.46 -6 -5   -2   -2
#> Mazda RX4 Wag       15.0   0 154.0 104 -2.10 -3.125 11.02 -6 -5   -2   -2
#> Datsun 710          18.8   0 104.0  89 -0.15 -1.680 14.61 -3 -3    0   -3
#> Hornet 4 Drive      15.4   0 252.0 104 -2.92 -2.785 13.44 -5 -6   -3   -5
#> Hornet Sportabout   10.7   0 352.0 167 -4.85 -4.560  9.02 -8 -8   -5   -6
#> Valiant             12.1   0 219.0  99 -3.24 -2.540 14.22 -5 -6   -3   -5
#> Duster 360           6.3   0 352.0 237 -4.79 -4.430  7.84 -8 -8   -5   -4
#> Merc 240D           20.4   0 142.7  58 -0.31 -0.810 16.00 -3 -4    0   -2
#> Merc 230            18.8   0 136.8  91 -0.08 -0.850 18.90 -3 -4    0   -2
#> Merc 280            13.2   0 161.6 117 -2.08 -2.560 12.30 -5 -6   -2   -2
#> Merc 280C           11.8   0 161.6 117 -2.08 -2.560 12.90 -5 -6   -2   -2
#> Merc 450SE           8.4   0 267.8 172 -4.93 -3.930  9.40 -8 -8   -5   -5
#> Merc 450SL           9.3   0 267.8 172 -4.93 -4.270  9.60 -8 -8   -5   -5
#> Merc 450SLC          7.2   0 267.8 172 -4.93 -4.220 10.00 -8 -8   -5   -5
#> Cadillac Fleetwood   2.4   0 464.0 197 -5.07 -2.750  9.98 -8 -8   -5   -4
#> Lincoln Continental  2.4   0 452.0 207 -5.00 -2.576  9.82 -8 -8   -5   -4
#> Chrysler Imperial    6.7   0 432.0 222 -4.77 -2.655  9.42 -8 -8   -5   -4
#> Fiat 128            28.4   0  74.7  62  0.08 -1.800 15.47 -3 -3    0   -3
#>  [ reached getOption("max.print") -- omitted 14 rows ]
X %r-% dapply(X, quantile, 0.28)
#>                        mpg cyl    disp     hp    drat     wt    qsec vs am gear
#> Mazda RX4            5.296   2  25.536   7.56  0.7724 -0.102 -0.5216  0  1    1
#> Mazda RX4 Wag        5.296   2  25.536   7.56  0.7724  0.153  0.0384  0  1    1
#> Datsun 710           7.096   0 -26.464  -9.44  0.7224 -0.402  1.6284  1  1    1
#> Hornet 4 Drive       5.696   2 123.536   7.56 -0.0476  0.493  2.4584  1  0    0
#> Hornet Sportabout    2.996   4 225.536  72.56  0.0224  0.718  0.0384  0  0    0
#> Valiant              2.396   2  90.536   2.56 -0.3676  0.738  3.2384  1  0    0
#> Duster 360          -1.404   4 225.536 142.56  0.0824  0.848 -1.1416  0  0    0
#> Merc 240D            8.696   0  12.236 -40.44  0.5624  0.468  3.0184  1  0    1
#> Merc 230             7.096   0   6.336  -7.44  0.7924  0.428  5.9184  1  0    1
#> Merc 280             3.496   2  33.136  20.56  0.7924  0.718  1.3184  1  0    1
#> Merc 280C            2.096   2  33.136  20.56  0.7924  0.718  1.9184  1  0    1
#> Merc 450SE           0.696   4 141.336  77.56 -0.0576  1.348  0.4184  0  0    0
#> Merc 450SL           1.596   4 141.336  77.56 -0.0576  1.008  0.6184  0  0    0
#> Merc 450SLC         -0.504   4 141.336  77.56 -0.0576  1.058  1.0184  0  0    0
#> Cadillac Fleetwood  -5.304   4 337.536 102.56 -0.1976  2.528  0.9984  0  0    0
#> Lincoln Continental -5.304   4 325.536 112.56 -0.1276  2.702  0.8384  0  0    0
#> Chrysler Imperial   -1.004   4 305.536 127.56  0.1024  2.623  0.4384  0  0    0
#> Fiat 128            16.696   0 -55.764 -36.44  0.9524 -0.522  2.4884  1  1    1
#>                     carb
#> Mazda RX4              2
#> Mazda RX4 Wag          2
#> Datsun 710            -1
#> Hornet 4 Drive        -1
#> Hornet Sportabout      0
#> Valiant               -1
#> Duster 360             2
#> Merc 240D              0
#> Merc 230               0
#> Merc 280               2
#> Merc 280C              2
#> Merc 450SE             1
#> Merc 450SL             1
#> Merc 450SLC            1
#> Cadillac Fleetwood     2
#> Lincoln Continental    2
#> Chrysler Imperial      2
#> Fiat 128              -1
#>  [ reached getOption("max.print") -- omitted 14 rows ]

## Chained Operations
library(magrittr) # Note: Used because |> is not available on older R versions
mtcars %>% fwithin() %r-% rnorm(11) %c*% 5 %>%
    tfm(mpg = fsum(mpg)) %>% qsu()
#>        N     Mean        SD        Min        Max
#> mpg   32  61.1666         0    61.1666    61.1666
#> cyl   32  -3.6557    8.9296   -14.5932     5.4068
#> disp  32   4.5329  619.6935  -793.5765  1210.9235
#> hp    32  -2.7058  342.8143  -476.1433   938.8567
#> drat  32  11.4505    2.6734     7.2677    18.1177
#> wt    32   4.4104    4.8923    -4.1108    15.4442
#> qsec  32  -7.8889    8.9347   -24.6326    17.3674
#> vs    32   4.8299    2.5201     2.6424     7.6424
#> am    32   1.9248     2.495    -0.1065     4.8935
#> gear  32   1.6345     3.689     -1.803      8.197
#> carb  32  -3.1903     8.076   -12.2528    22.7472