Advanced Data Aggregation

collap is a fast and easy to use multi-purpose data aggregation command.

It performs simple aggregations, multi-type data aggregations applying different functions to numeric and categorical data, weighted aggregations (including weighted multi-type aggregations), multi-function aggregations applying multiple functions to each column, and fully customized aggregations where the user passes a list mapping functions to columns.

collap works with collapse's Fast Statistical Functions, providing extremely fast conventional and weighted aggregation. It also works with other functions but this does not deliver high speeds on large data and does not support weighted aggregations.

# Main function: allows formula and data input to `by` and `w` arguments
collap(X, by, FUN = fmean, catFUN = fmode, cols = NULL, w = NULL, wFUN = fsum,
       custom = NULL, keep.by = TRUE, keep.w = TRUE, keep.col.order = TRUE,
       sort = TRUE, decreasing = FALSE, na.last = TRUE, parallel = FALSE, mc.cores = 1L,
       return = c("wide","list","long","long_dupl"), give.names = "auto", sort.row, ...)

# Programmer function: allows column names and indices input to `by` and `w` arguments
collapv(X, by, FUN = fmean, catFUN = fmode, cols = NULL, w = NULL, wFUN = fsum,
        custom = NULL, keep.by = TRUE, keep.w = TRUE, keep.col.order = TRUE,
        sort = TRUE, decreasing = FALSE, na.last = TRUE, parallel = FALSE, mc.cores = 1L,
        return = c("wide","list","long","long_dupl"), give.names = "auto", sort.row, ...)

# Auxiliary function: for grouped tibble ('grouped_df') input + non-standard evaluation
collapg(X, FUN = fmean, catFUN = fmode, cols = NULL, w = NULL, wFUN = fsum, custom = NULL,
        keep.group_vars = TRUE, keep.w = TRUE, keep.col.order = TRUE,
        parallel = FALSE, mc.cores = 1L,
        return = c("wide","list","long","long_dupl"), give.names = "auto", sort.row, ...)

Arguments

X
a data frame, or an object coercible to data frame using qDF.

by
for collap: a one-or two sided formula, i.e. ~ group1 or var1 + var2 ~ group1 + group2, or a atomic vector, list of vectors or GRP object used to group X. For collapv: names or indices of grouping columns, or a logical vector or selector function such as is.categorical selecting grouping columns.

FUN
a function, list of functions (i.e. list(fsum, fmean, fsd) or list(myfun1 = function(x).., sd = sd)), or a character vector of function names, which are automatically applied only to numeric variables.

catFUN
same as FUN, but applied only to categorical (non-numeric) typed columns (is.categorical).

cols
select columns to aggregate using a function, column names, indices or logical vector. Note: cols is ignored if a two-sided formula is passed to by.

w
weights. Can be passed as numeric vector or alternatively as formula i.e. ~ weightvar in collap or column name / index etc. i.e. "weightvar" in collapv. collapg supports non-standard evaluations so weightvar can be indicated without quotes if found in X.

wFUN
same as FUN: Function(s) to aggregate weight variable if keep.w = TRUE. By default the sum of the weights is computed in each group.

custom
a named list specifying a fully customized aggregation task. The names of the list are function names and the content columns to aggregate using this function (same input as cols). For example custom = list(fmean = 1:6, fsd = 7:9, fmode = 10:11) tells collap to aggregate columns 1-6 of X using the mean, columns 7-9 using the standard deviation etc. Note: custom lets collap ignore any inputs passed to FUN, catFUN or cols.

keep.by, keep.group_vars
logical. FALSE will omit grouping variables from the output. TRUE keeps the variables, even if passed externally in a list or vector (unlike other collapse functions).

keep.w
logical. FALSE will omit weight variable from the output i.e. no aggregation of the weights. TRUE aggregates and adds weights, even if passed externally as a vector (unlike other collapse functions).

keep.col.order
logical. Retain original column order post-aggregation.

sort, decreasing, na.last
logical. Arguments passed to GRP.default and affecting the row-order in the aggregated data frame.

parallel
logical. Use parallel::mclapply instead of lapply for multi-function or custom aggregation.

mc.cores
integer. Argument to parallel::mclapply setting the number of cores to use.

return
character. Control the output format when aggregating with multiple functions or performing custom aggregation. "wide" (default) returns a wider data frame with added columns for each additional function. "list" returns a list of data frames - one for each function. "long" adds a column "Function" and row-binds the results from different functions using data.table::rbindlist. "long.dupl" is a special option for aggregating multi-type data using multiple FUN but only one catFUN or vice-versa. In that case the format is long and data aggregated using only one function is duplicated. See Examples.

give.names
logical. Create unique names of aggregated columns by adding a prefix 'FUN.'. 'auto' will automatically create such prefixes whenever multiple functions are applied to a column or custom is used.

sort.row
depreciated, renamed to sort.

...
additional arguments passed to all functions supplied to FUN, catFUN, wFUN or custom. The behavior of Fast Statistical Functions is regulated by option("collapse_unused_arg_action") and defaults to "warning".

Details

collap automatically checks each function passed to it whether it is a Fast Statistical Function (i.e. whether the function name is contained in .FAST_STAT_FUN). If the function is a fast statistical function, collap only does the grouping and then calls the function to carry out the grouped computations. If the function is not one of .FAST_STAT_FUN, BY is called internally to perform the computation. The resulting computations from each function are put into a list and recombined to produce the desired output format as controlled by the return argument. When multiple functions are used with collap, setting parallel = TRUE and the number of cores with mc.cores will instruct collap to execute these function calls in parallel using parallel::mclapply. If only a single function is used which is not a .FAST_STAT_FUN, the parallel and mc.cores arguments are handed down to BY. See Examples.

Value

X aggregated. If X is not a data frame it is coerced to one using qDF and then aggregated.

Note

Since BY does not check and split additional arguments passed to it, it is presently not possible to create a weighted function in R and apply it to data by groups with collap. Weighted aggregations only work with Fast Statistical Functions supporting weights. User written weighted functions can be applied using the data.table package.

collap by default (keep.by = TRUE, keep.w = TRUE) preserves all arguments passed to the by or w arguments, whether passed in a formula or externally. The names of externally passed vectors and lists are intelligently extracted. So it is possible to write collap(iris, iris$Species), and obtain an aggregated data frame with two Species columns, whereas collap(iris, ~ Species) only has one Species column. Similarly for weight vectors passed to w. In this regard collap is more sophisticated than other collapse functions where preservation of grouping and weight variables is restricted to formula use. For example STD(iris, iris$Species) does not preserve Species in the output, whereas STD(iris, ~ Species) does. This collap feature is there simply for convenience, for example sometimes a survey is disaggregated into several datasets, and this now allows easy pulling of identifiers or weights from other datasets for aggregations. If all information is available in one dataset, just using formulas is highly recommended.

See also

Examples

## A Simple Introduction --------------------------------------
head(iris)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species        wt AWMSL
#> 1          5.1         3.5          1.4         0.2  setosa 0.6846713 FALSE
#> 2          4.9         3.0          1.4         0.2  setosa 0.7469905 FALSE
#> 3          4.7         3.2          1.3         0.2  setosa 0.1387655 FALSE
#> 4          4.6         3.1          1.5         0.2  setosa 0.2333557 FALSE
#> 5          5.0         3.6          1.4         0.2  setosa 0.6550782 FALSE
#> 6          5.4         3.9          1.7         0.4  setosa 0.9068255 FALSE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3  TRUE
#> 4 FALSE
#> 5  TRUE
#> 6  TRUE

collap(iris, ~ Species)                                        # Default: FUN = fmean for numeric

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species        wt AWMSL
#> 1        5.006       3.428        1.462       0.246     setosa 0.8817075 FALSE
#> 2        5.936       2.770        4.260       1.326 versicolor 0.6920420  TRUE
#> 3        6.588       2.974        5.552       2.026  virginica 0.7794701  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

collapv(iris, 5)                                               # Same using collapv

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species        wt AWMSL
#> 1        5.006       3.428        1.462       0.246     setosa 0.8817075 FALSE
#> 2        5.936       2.770        4.260       1.326 versicolor 0.6920420  TRUE
#> 3        6.588       2.974        5.552       2.026  virginica 0.7794701  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

collap(iris, ~ Species, fmedian)                               # Using the median

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species        wt AWMSL
#> 1          5.0         3.4         1.50         0.2     setosa 0.7786559 FALSE
#> 2          5.9         2.8         4.35         1.3 versicolor 0.5669077  TRUE
#> 3          6.5         3.0         5.55         2.0  virginica 0.5852847  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

collap(iris, ~ Species, fmedian, keep.col.order = FALSE)       # Groups in-front

#>      Species Sepal.Length Sepal.Width Petal.Length Petal.Width        wt AWMSL
#> 1     setosa          5.0         3.4         1.50         0.2 0.7786559 FALSE
#> 2 versicolor          5.9         2.8         4.35         1.3 0.5669077  TRUE
#> 3  virginica          6.5         3.0         5.55         2.0 0.5852847  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

collap(iris, Sepal.Width + Petal.Width ~ Species, fmedian)     # Only '.Width' columns

#>   Sepal.Width Petal.Width    Species
#> 1         3.4         0.2     setosa
#> 2         2.8         1.3 versicolor
#> 3         3.0         2.0  virginica

collapv(iris, 5, cols = c(2, 4))                               # Same using collapv

#>   Sepal.Width Petal.Width    Species
#> 1       3.428       0.246     setosa
#> 2       2.770       1.326 versicolor
#> 3       2.974       2.026  virginica

collap(iris, ~ Species, list(fmean, fmedian))                  # Two functions

#>   fmean.Sepal.Length fmedian.Sepal.Length fmean.Sepal.Width fmedian.Sepal.Width
#> 1              5.006                  5.0             3.428                 3.4
#> 2              5.936                  5.9             2.770                 2.8
#> 3              6.588                  6.5             2.974                 3.0
#>   fmean.Petal.Length fmedian.Petal.Length fmean.Petal.Width fmedian.Petal.Width
#> 1              1.462                 1.50             0.246                 0.2
#> 2              4.260                 4.35             1.326                 1.3
#> 3              5.552                 5.55             2.026                 2.0
#>      Species  fmean.wt fmedian.wt fmode.AWMSL fmode.AWMSW
#> 1     setosa 0.8817075  0.7786559       FALSE        TRUE
#> 2 versicolor 0.6920420  0.5669077        TRUE       FALSE
#> 3  virginica 0.7794701  0.5852847        TRUE       FALSE

collap(iris, ~ Species, list(fmean, fmedian), return = "long") # Long format

#>   Function Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1    fmean        5.006       3.428        1.462       0.246     setosa
#> 2    fmean        5.936       2.770        4.260       1.326 versicolor
#> 3    fmean        6.588       2.974        5.552       2.026  virginica
#> 4  fmedian        5.000       3.400        1.500       0.200     setosa
#> 5  fmedian        5.900       2.800        4.350       1.300 versicolor
#> 6  fmedian        6.500       3.000        5.550       2.000  virginica
#> 7    fmode           NA          NA           NA          NA     setosa
#> 8    fmode           NA          NA           NA          NA versicolor
#> 9    fmode           NA          NA           NA          NA  virginica
#>          wt AWMSL AWMSW
#> 1 0.8817075    NA    NA
#> 2 0.6920420    NA    NA
#> 3 0.7794701    NA    NA
#> 4 0.7786559    NA    NA
#> 5 0.5669077    NA    NA
#> 6 0.5852847    NA    NA
#> 7        NA FALSE  TRUE
#> 8        NA  TRUE FALSE
#> 9        NA  TRUE FALSE

collapv(iris, 5, custom = list(fmean = 1:2, fmedian = 3:4))    # Custom aggregation

#>   fmean.Sepal.Length fmean.Sepal.Width fmedian.Petal.Length fmedian.Petal.Width
#> 1              5.006             3.428                 1.50                 0.2
#> 2              5.936             2.770                 4.35                 1.3
#> 3              6.588             2.974                 5.55                 2.0
#>      Species
#> 1     setosa
#> 2 versicolor
#> 3  virginica

collapv(iris, 5, custom = list(fmean = 1:2, fmedian = 3:4),    # Raw output, no column reordering
        return = "list")

#> $fmean
#>      Species Sepal.Length Sepal.Width
#> 1     setosa        5.006       3.428
#> 2 versicolor        5.936       2.770
#> 3  virginica        6.588       2.974
#> 
#> $fmedian
#>      Species Petal.Length Petal.Width
#> 1     setosa         1.50         0.2
#> 2 versicolor         4.35         1.3
#> 3  virginica         5.55         2.0
#> 

collapv(iris, 5, custom = list(fmean = 1:2, fmedian = 3:4),    # A strange choice..
        return = "long")

#>   Function Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1    fmean        5.006       3.428           NA          NA     setosa
#> 2    fmean        5.936       2.770           NA          NA versicolor
#> 3    fmean        6.588       2.974           NA          NA  virginica
#> 4  fmedian           NA          NA         1.50         0.2     setosa
#> 5  fmedian           NA          NA         4.35         1.3 versicolor
#> 6  fmedian           NA          NA         5.55         2.0  virginica

collap(iris, ~ Species, w = ~ Sepal.Length)                    # Using Sepal.Length as weights, ..

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species        wt AWMSL
#> 1        250.3    3.447423     1.465202   0.2480224     setosa 0.8917024 FALSE
#> 2        296.8    2.784063     4.290195   1.3352089 versicolor 0.6893412  TRUE
#> 3        329.4    2.987948     5.597116   2.0333030  virginica 0.7919285  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

weights <- abs(rnorm(fnrow(iris)))
collap(iris, ~ Species, w = weights)                           # Some random weights..

#>    weights Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1 39.56057     4.987909    3.427702     1.444973   0.2415078     setosa
#> 2 39.64675     5.891207    2.785107     4.202130   1.3139783 versicolor
#> 3 39.72540     6.577970    2.999060     5.555977   2.0436818  virginica
#>          wt AWMSL AWMSW
#> 1 0.9812484 FALSE  TRUE
#> 2 0.7306248  TRUE FALSE
#> 3 0.7586074  TRUE FALSE

collap(iris, iris$Species, w = weights)                        # Note this behavior..

#>      Species  weights Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1     setosa 39.56057     4.987909    3.427702     1.444973   0.2415078
#> 2 versicolor 39.64675     5.891207    2.785107     4.202130   1.3139783
#> 3  virginica 39.72540     6.577970    2.999060     5.555977   2.0436818
#>      Species        wt AWMSL AWMSW
#> 1     setosa 0.9812484 FALSE  TRUE
#> 2 versicolor 0.7306248  TRUE FALSE
#> 3  virginica 0.7586074  TRUE FALSE

collap(iris, iris$Species, w = weights,
       keep.by = FALSE, keep.w = FALSE)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width    Species        wt AWMSL
#> 1     4.987909    3.427702     1.444973   0.2415078     setosa 0.9812484 FALSE
#> 2     5.891207    2.785107     4.202130   1.3139783 versicolor 0.7306248  TRUE
#> 3     6.577970    2.999060     5.555977   2.0436818  virginica 0.7586074  TRUE
#>   AWMSW
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE

library(dplyr) # Needed for "%>%"
iris %>% fgroup_by(Species) %>% collapg                        # dplyr style, but faster

#> # A tibble: 3 x 8
#>   Species    Sepal.Length Sepal.Width Petal.Length Petal.Width    wt AWMSL AWMSW
#>   <fct>             <dbl>       <dbl>        <dbl>       <dbl> <dbl> <lgl> <lgl>
#> 1 setosa             5.01        3.43         1.46       0.246 0.882 FALSE TRUE 
#> 2 versicolor         5.94        2.77         4.26       1.33  0.692 TRUE  FALSE
#> 3 virginica          6.59        2.97         5.55       2.03  0.779 TRUE  FALSE

## Multi-Type Aggregation --------------------------------------
head(wlddev)                                                    # World Development Panel Data

#>       country iso3c       date year decade     region     income  OECD PCGDP
#> 1 Afghanistan   AFG 1961-01-01 1960   1960 South Asia Low income FALSE    NA
#> 2 Afghanistan   AFG 1962-01-01 1961   1960 South Asia Low income FALSE    NA
#> 3 Afghanistan   AFG 1963-01-01 1962   1960 South Asia Low income FALSE    NA
#> 4 Afghanistan   AFG 1964-01-01 1963   1960 South Asia Low income FALSE    NA
#> 5 Afghanistan   AFG 1965-01-01 1964   1960 South Asia Low income FALSE    NA
#> 6 Afghanistan   AFG 1966-01-01 1965   1960 South Asia Low income FALSE    NA
#>   LIFEEX GINI       ODA
#> 1 32.292   NA 114440000
#> 2 32.742   NA 233350000
#> 3 33.185   NA 114880000
#> 4 33.624   NA 236450000
#> 5 34.060   NA 302480000
#> 6 34.495   NA 370250000

head(collap(wlddev, ~ country + decade))                        # Aggregate by country and decade

#>       country iso3c       date   year decade     region     income  OECD
#> 1 Afghanistan   AFG 1961-01-01 1962.5   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1967-01-01 1970.0   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1976-01-01 1980.0   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1987-01-01 1990.0   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1996-01-01 2000.0   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2007-01-01 2010.0   2010 South Asia Low income FALSE
#>      PCGDP   LIFEEX GINI        ODA
#> 1       NA 33.39967   NA  228641667
#> 2       NA 36.70089   NA  224881111
#> 3       NA 42.03909   NA  132813636
#> 4       NA 49.69089   NA  249072222
#> 5 349.7596 55.61818   NA 1045170000
#> 6 506.9931 61.12978   NA 5298066667

head(collap(wlddev, ~ country + decade, fmedian, ffirst))       # Different functions

#>       country iso3c       date   year decade     region     income  OECD
#> 1 Afghanistan   AFG 1961-01-01 1962.5   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1967-01-01 1970.0   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1976-01-01 1980.0   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1987-01-01 1990.0   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1996-01-01 2000.0   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2007-01-01 2010.0   2010 South Asia Low income FALSE
#>      PCGDP  LIFEEX GINI        ODA
#> 1       NA 33.4045   NA  234900000
#> 2       NA 36.6780   NA  229150000
#> 3       NA 41.8530   NA   72610000
#> 4       NA 49.8560   NA  238500000
#> 5 346.9282 55.4820   NA  306200000
#> 6 536.0125 61.2260   NA 5033950000

head(collap(wlddev, ~ country + decade, cols = is.numeric))     # Aggregate only numeric columns

#>       country   year decade decade    PCGDP   LIFEEX GINI        ODA
#> 1 Afghanistan 1962.5   1960   1960       NA 33.39967   NA  228641667
#> 2 Afghanistan 1970.0   1970   1970       NA 36.70089   NA  224881111
#> 3 Afghanistan 1980.0   1980   1980       NA 42.03909   NA  132813636
#> 4 Afghanistan 1990.0   1990   1990       NA 49.69089   NA  249072222
#> 5 Afghanistan 2000.0   2000   2000 349.7596 55.61818   NA 1045170000
#> 6 Afghanistan 2010.0   2010   2010 506.9931 61.12978   NA 5298066667

head(collap(wlddev, ~ country + decade, cols = 9:12))           # Only the 4 series

#>       country decade    PCGDP   LIFEEX GINI        ODA
#> 1 Afghanistan   1960       NA 33.39967   NA  228641667
#> 2 Afghanistan   1970       NA 36.70089   NA  224881111
#> 3 Afghanistan   1980       NA 42.03909   NA  132813636
#> 4 Afghanistan   1990       NA 49.69089   NA  249072222
#> 5 Afghanistan   2000 349.7596 55.61818   NA 1045170000
#> 6 Afghanistan   2010 506.9931 61.12978   NA 5298066667

head(collap(wlddev, PCGDP + LIFEEX ~ country + decade))         # Only GDP and life-expactancy

#>       country decade    PCGDP   LIFEEX
#> 1 Afghanistan   1960       NA 33.39967
#> 2 Afghanistan   1970       NA 36.70089
#> 3 Afghanistan   1980       NA 42.03909
#> 4 Afghanistan   1990       NA 49.69089
#> 5 Afghanistan   2000 349.7596 55.61818
#> 6 Afghanistan   2010 506.9931 61.12978

head(collap(wlddev, PCGDP + LIFEEX ~ country + decade, fsum))   # Using the sum instead

#>       country decade    PCGDP  LIFEEX
#> 1 Afghanistan   1960       NA 200.398
#> 2 Afghanistan   1970       NA 330.308
#> 3 Afghanistan   1980       NA 462.430
#> 4 Afghanistan   1990       NA 447.218
#> 5 Afghanistan   2000 1399.038 611.800
#> 6 Afghanistan   2010 4562.938 550.168

head(collap(wlddev, PCGDP + LIFEEX ~ country + decade, sum,     # Same using base::sum -> slower!
            na.rm = TRUE))

#>       country decade    PCGDP  LIFEEX
#> 1 Afghanistan   1960    0.000 200.398
#> 2 Afghanistan   1970    0.000 330.308
#> 3 Afghanistan   1980    0.000 462.430
#> 4 Afghanistan   1990    0.000 447.218
#> 5 Afghanistan   2000 1399.038 611.800
#> 6 Afghanistan   2010 4562.938 550.168

head(collap(wlddev, wlddev[c("country","decade")], fsum,        # Same, exploring different inputs
            cols = 9:10))

#>       country decade    PCGDP  LIFEEX
#> 1 Afghanistan   1960       NA 200.398
#> 2 Afghanistan   1970       NA 330.308
#> 3 Afghanistan   1980       NA 462.430
#> 4 Afghanistan   1990       NA 447.218
#> 5 Afghanistan   2000 1399.038 611.800
#> 6 Afghanistan   2010 4562.938 550.168

head(collap(wlddev[9:10], wlddev[c("country","decade")], fsum))

#>       country decade    PCGDP  LIFEEX
#> 1 Afghanistan   1960       NA 200.398
#> 2 Afghanistan   1970       NA 330.308
#> 3 Afghanistan   1980       NA 462.430
#> 4 Afghanistan   1990       NA 447.218
#> 5 Afghanistan   2000 1399.038 611.800
#> 6 Afghanistan   2010 4562.938 550.168

head(collapv(wlddev, c("country","decade"), fsum))              # ..names/indices with collapv

#>       country iso3c       date  year decade     region     income  OECD
#> 1 Afghanistan   AFG 1961-01-01 11775   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1967-01-01 17730   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1976-01-01 21780   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1987-01-01 17910   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1996-01-01 22000   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2007-01-01 18090   2010 South Asia Low income FALSE
#>      PCGDP  LIFEEX GINI         ODA
#> 1       NA 200.398   NA  1371850000
#> 2       NA 330.308   NA  2023930000
#> 3       NA 462.430   NA  1460950000
#> 4       NA 447.218   NA  2241650000
#> 5 1399.038 611.800   NA 11496870000
#> 6 4562.938 550.168   NA 47682600000

head(collapv(wlddev, c(1,5), fsum))

#>       country iso3c       date  year decade     region     income  OECD
#> 1 Afghanistan   AFG 1961-01-01 11775   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1967-01-01 17730   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1976-01-01 21780   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1987-01-01 17910   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1996-01-01 22000   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2007-01-01 18090   2010 South Asia Low income FALSE
#>      PCGDP  LIFEEX GINI         ODA
#> 1       NA 200.398   NA  1371850000
#> 2       NA 330.308   NA  2023930000
#> 3       NA 462.430   NA  1460950000
#> 4       NA 447.218   NA  2241650000
#> 5 1399.038 611.800   NA 11496870000
#> 6 4562.938 550.168   NA 47682600000

g <- GRP(wlddev, ~ country + decade)                            # Precomputing the grouping
head(collap(wlddev, g, keep.by = FALSE))                        # This is slightly faster now

#>       country iso3c       date   year decade     region     income  OECD
#> 1 Afghanistan   AFG 1961-01-01 1962.5   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1967-01-01 1970.0   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1976-01-01 1980.0   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1987-01-01 1990.0   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1996-01-01 2000.0   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2007-01-01 2010.0   2010 South Asia Low income FALSE
#>      PCGDP   LIFEEX GINI        ODA
#> 1       NA 33.39967   NA  228641667
#> 2       NA 36.70089   NA  224881111
#> 3       NA 42.03909   NA  132813636
#> 4       NA 49.69089   NA  249072222
#> 5 349.7596 55.61818   NA 1045170000
#> 6 506.9931 61.12978   NA 5298066667

# Aggregate categorical data using not the mode but the last element
head(collap(wlddev, ~ country + decade, fmean, flast))

#>       country iso3c       date   year decade     region     income  OECD
#> 1 Afghanistan   AFG 1966-01-01 1962.5   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1975-01-01 1970.0   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1986-01-01 1980.0   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1995-01-01 1990.0   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 2006-01-01 2000.0   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2015-01-01 2010.0   2010 South Asia Low income FALSE
#>      PCGDP   LIFEEX GINI        ODA
#> 1       NA 33.39967   NA  228641667
#> 2       NA 36.70089   NA  224881111
#> 3       NA 42.03909   NA  132813636
#> 4       NA 49.69089   NA  249072222
#> 5 349.7596 55.61818   NA 1045170000
#> 6 506.9931 61.12978   NA 5298066667

head(collap(wlddev, ~ country + decade, catFUN = flast,         # Aggregate only categorical data
            cols = is.categorical))

#>       country     country iso3c       date decade     region     income  OECD
#> 1 Afghanistan Afghanistan   AFG 1966-01-01   1960 South Asia Low income FALSE
#> 2 Afghanistan Afghanistan   AFG 1975-01-01   1970 South Asia Low income FALSE
#> 3 Afghanistan Afghanistan   AFG 1986-01-01   1980 South Asia Low income FALSE
#> 4 Afghanistan Afghanistan   AFG 1995-01-01   1990 South Asia Low income FALSE
#> 5 Afghanistan Afghanistan   AFG 2006-01-01   2000 South Asia Low income FALSE
#> 6 Afghanistan Afghanistan   AFG 2015-01-01   2010 South Asia Low income FALSE

## Weighted Aggregation ----------------------------------------
weights <- abs(rnorm(fnrow(wlddev)))                            # Random weight vector
head(collap(wlddev, ~ country + decade, w = weights))           # Takes weighted mean for numeric..

#>     weights     country iso3c       date     year decade     region     income
#> 1  3.977500 Afghanistan   AFG 1963-01-01 1962.934   1960 South Asia Low income
#> 2  7.180692 Afghanistan   AFG 1971-01-01 1970.823   1970 South Asia Low income
#> 3  8.201658 Afghanistan   AFG 1983-01-01 1980.488   1980 South Asia Low income
#> 4 10.210256 Afghanistan   AFG 1990-01-01 1990.269   1990 South Asia Low income
#> 5 11.736101 Afghanistan   AFG 1998-01-01 1999.144   2000 South Asia Low income
#> 6  7.948094 Afghanistan   AFG 2015-01-01 2010.514   2010 South Asia Low income
#>    OECD    PCGDP   LIFEEX GINI        ODA
#> 1 FALSE       NA 33.59138   NA  228677841
#> 2 FALSE       NA 37.06992   NA  192003634
#> 3 FALSE       NA 42.38328   NA  113094881
#> 4 FALSE       NA 49.93548   NA  273981591
#> 5 FALSE 343.9491 55.15815   NA  838996680
#> 6 FALSE 510.4979 61.35658   NA 5179680196

# ..and weighted mode for categorical data. The weight vector is aggregated using fsum
wlddev$weights <- weights                                       # Adding to data
head(collap(wlddev, ~ country + decade, w = ~ weights))         # Keeps column order

#>       country iso3c       date     year decade     region     income  OECD
#> 1 Afghanistan   AFG 1963-01-01 1962.934   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1971-01-01 1970.823   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1983-01-01 1980.488   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1990-01-01 1990.269   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1998-01-01 1999.144   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2015-01-01 2010.514   2010 South Asia Low income FALSE
#>      PCGDP   LIFEEX GINI        ODA   weights
#> 1       NA 33.59138   NA  228677841  3.977500
#> 2       NA 37.06992   NA  192003634  7.180692
#> 3       NA 42.38328   NA  113094881  8.201658
#> 4       NA 49.93548   NA  273981591 10.210256
#> 5 343.9491 55.15815   NA  838996680 11.736101
#> 6 510.4979 61.35658   NA 5179680196  7.948094

head(collap(wlddev, ~ country + decade, w = ~ weights,          # Aggregating weights using sum
            wFUN = list(fsum, fmax)))                           # and max (corresponding to mode)

#>       country iso3c       date     year decade     region     income  OECD
#> 1 Afghanistan   AFG 1963-01-01 1962.934   1960 South Asia Low income FALSE
#> 2 Afghanistan   AFG 1971-01-01 1970.823   1970 South Asia Low income FALSE
#> 3 Afghanistan   AFG 1983-01-01 1980.488   1980 South Asia Low income FALSE
#> 4 Afghanistan   AFG 1990-01-01 1990.269   1990 South Asia Low income FALSE
#> 5 Afghanistan   AFG 1998-01-01 1999.144   2000 South Asia Low income FALSE
#> 6 Afghanistan   AFG 2015-01-01 2010.514   2010 South Asia Low income FALSE
#>      PCGDP   LIFEEX GINI        ODA fsum.weights fmax.weights
#> 1       NA 33.59138   NA  228677841     3.977500     1.138012
#> 2       NA 37.06992   NA  192003634     7.180692     1.900061
#> 3       NA 42.38328   NA  113094881     8.201658     1.270672
#> 4       NA 49.93548   NA  273981591    10.210256     2.211769
#> 5 343.9491 55.15815   NA  838996680    11.736101     2.564408
#> 6 510.4979 61.35658   NA 5179680196     7.948094     2.446680

wlddev$weights <- NULL

## Multi-Function Aggregation ----------------------------------
head(collap(wlddev, ~ country + decade, list(fmean, fNobs),     # Saving mean and Nobs
            cols = 9:12))

#>       country decade fmean.PCGDP fNobs.PCGDP fmean.LIFEEX fNobs.LIFEEX
#> 1 Afghanistan   1960          NA           0     33.39967            6
#> 2 Afghanistan   1970          NA           0     36.70089            9
#> 3 Afghanistan   1980          NA           0     42.03909           11
#> 4 Afghanistan   1990          NA           0     49.69089            9
#> 5 Afghanistan   2000    349.7596           4     55.61818           11
#> 6 Afghanistan   2010    506.9931           9     61.12978            9
#>   fmean.GINI fNobs.GINI  fmean.ODA fNobs.ODA
#> 1         NA          0  228641667         6
#> 2         NA          0  224881111         9
#> 3         NA          0  132813636        11
#> 4         NA          0  249072222         9
#> 5         NA          0 1045170000        11
#> 6         NA          0 5298066667         9

head(collap(wlddev, ~ country + decade,                         # Same using base R -> slower
            list(mean = mean,
                 Nobs = function(x, ...) sum(!is.na(x))),
            cols = 9:12, na.rm = TRUE))

#>       country decade mean.PCGDP Nobs.PCGDP mean.LIFEEX Nobs.LIFEEX mean.GINI
#> 1 Afghanistan   1960        NaN          0    33.39967           6       NaN
#> 2 Afghanistan   1970        NaN          0    36.70089           9       NaN
#> 3 Afghanistan   1980        NaN          0    42.03909          11       NaN
#> 4 Afghanistan   1990        NaN          0    49.69089           9       NaN
#> 5 Afghanistan   2000   349.7596          4    55.61818          11       NaN
#> 6 Afghanistan   2010   506.9931          9    61.12978           9       NaN
#>   Nobs.GINI   mean.ODA Nobs.ODA
#> 1         0  228641667        6
#> 2         0  224881111        9
#> 3         0  132813636       11
#> 4         0  249072222        9
#> 5         0 1045170000       11
#> 6         0 5298066667        9

lapply(collap(wlddev, ~ country + decade,                       # List output format
            list(fmean, fNobs), cols = 9:12, return = "list"), head)

#> $fmean
#>       country decade    PCGDP   LIFEEX GINI        ODA
#> 1 Afghanistan   1960       NA 33.39967   NA  228641667
#> 2 Afghanistan   1970       NA 36.70089   NA  224881111
#> 3 Afghanistan   1980       NA 42.03909   NA  132813636
#> 4 Afghanistan   1990       NA 49.69089   NA  249072222
#> 5 Afghanistan   2000 349.7596 55.61818   NA 1045170000
#> 6 Afghanistan   2010 506.9931 61.12978   NA 5298066667
#> 
#> $fNobs
#>       country decade PCGDP LIFEEX GINI ODA
#> 1 Afghanistan   1960     0      6    0   6
#> 2 Afghanistan   1970     0      9    0   9
#> 3 Afghanistan   1980     0     11    0  11
#> 4 Afghanistan   1990     0      9    0   9
#> 5 Afghanistan   2000     4     11    0  11
#> 6 Afghanistan   2010     9      9    0   9
#> 

head(collap(wlddev, ~ country + decade,                         # Long output format
            list(fmean, fNobs), cols = 9:12, return = "long"))

#>   Function     country decade    PCGDP   LIFEEX GINI        ODA
#> 1    fmean Afghanistan   1960       NA 33.39967   NA  228641667
#> 2    fmean Afghanistan   1970       NA 36.70089   NA  224881111
#> 3    fmean Afghanistan   1980       NA 42.03909   NA  132813636
#> 4    fmean Afghanistan   1990       NA 49.69089   NA  249072222
#> 5    fmean Afghanistan   2000 349.7596 55.61818   NA 1045170000
#> 6    fmean Afghanistan   2010 506.9931 61.12978   NA 5298066667

head(collap(wlddev, ~ country + decade,                         # Also aggregating categorical data,
            list(fmean, fNobs), return = "long_dupl"))          # and duplicating it 2 times

#>   Function     country iso3c       date   year decade     region     income
#> 1    fmean Afghanistan   AFG 1961-01-01 1962.5   1960 South Asia Low income
#> 2    fmean Afghanistan   AFG 1967-01-01 1970.0   1970 South Asia Low income
#> 3    fmean Afghanistan   AFG 1976-01-01 1980.0   1980 South Asia Low income
#> 4    fmean Afghanistan   AFG 1987-01-01 1990.0   1990 South Asia Low income
#> 5    fmean Afghanistan   AFG 1996-01-01 2000.0   2000 South Asia Low income
#> 6    fmean Afghanistan   AFG 2007-01-01 2010.0   2010 South Asia Low income
#>    OECD    PCGDP   LIFEEX GINI        ODA
#> 1 FALSE       NA 33.39967   NA  228641667
#> 2 FALSE       NA 36.70089   NA  224881111
#> 3 FALSE       NA 42.03909   NA  132813636
#> 4 FALSE       NA 49.69089   NA  249072222
#> 5 FALSE 349.7596 55.61818   NA 1045170000
#> 6 FALSE 506.9931 61.12978   NA 5298066667

head(collap(wlddev, ~ country + decade,                         # Now also using 2 functions on
            list(fmean, fNobs), list(fmode, flast),             # categorical data
            keep.col.order = FALSE))

#>       country decade fmean.year fmean.PCGDP fmean.LIFEEX fmean.GINI  fmean.ODA
#> 1 Afghanistan   1960     1962.5          NA     33.39967         NA  228641667
#> 2 Afghanistan   1970     1970.0          NA     36.70089         NA  224881111
#> 3 Afghanistan   1980     1980.0          NA     42.03909         NA  132813636
#> 4 Afghanistan   1990     1990.0          NA     49.69089         NA  249072222
#> 5 Afghanistan   2000     2000.0    349.7596     55.61818         NA 1045170000
#> 6 Afghanistan   2010     2010.0    506.9931     61.12978         NA 5298066667
#>   fNobs.year fNobs.PCGDP fNobs.LIFEEX fNobs.GINI fNobs.ODA fmode.iso3c
#> 1          6           0            6          0         6         AFG
#> 2          9           0            9          0         9         AFG
#> 3         11           0           11          0        11         AFG
#> 4          9           0            9          0         9         AFG
#> 5         11           4           11          0        11         AFG
#> 6          9           9            9          0         9         AFG
#>   fmode.date fmode.region fmode.income fmode.OECD flast.iso3c flast.date
#> 1 1961-01-01   South Asia   Low income      FALSE         AFG 1966-01-01
#> 2 1967-01-01   South Asia   Low income      FALSE         AFG 1975-01-01
#> 3 1976-01-01   South Asia   Low income      FALSE         AFG 1986-01-01
#> 4 1987-01-01   South Asia   Low income      FALSE         AFG 1995-01-01
#> 5 1996-01-01   South Asia   Low income      FALSE         AFG 2006-01-01
#> 6 2007-01-01   South Asia   Low income      FALSE         AFG 2015-01-01
#>   flast.region flast.income flast.OECD
#> 1   South Asia   Low income      FALSE
#> 2   South Asia   Low income      FALSE
#> 3   South Asia   Low income      FALSE
#> 4   South Asia   Low income      FALSE
#> 5   South Asia   Low income      FALSE
#> 6   South Asia   Low income      FALSE

head(collap(wlddev, ~ country + decade,                         # More functions, string input,
            c("fmean","fsum","fNobs","fsd","fvar"),             # parallelized execution
            c("fmode","ffirst","flast","fNdistinct"),           # (choose more than 1 cores,
            parallel = TRUE, mc.cores = 1L,                     # depending on your machine)
            keep.col.order = FALSE))

#>       country decade fmean.year fmean.PCGDP fmean.LIFEEX fmean.GINI  fmean.ODA
#> 1 Afghanistan   1960     1962.5          NA     33.39967         NA  228641667
#> 2 Afghanistan   1970     1970.0          NA     36.70089         NA  224881111
#> 3 Afghanistan   1980     1980.0          NA     42.03909         NA  132813636
#> 4 Afghanistan   1990     1990.0          NA     49.69089         NA  249072222
#> 5 Afghanistan   2000     2000.0    349.7596     55.61818         NA 1045170000
#> 6 Afghanistan   2010     2010.0    506.9931     61.12978         NA 5298066667
#>   fsum.year fsum.PCGDP fsum.LIFEEX fsum.GINI    fsum.ODA fNobs.year fNobs.PCGDP
#> 1     11775         NA     200.398        NA  1371850000          6           0
#> 2     17730         NA     330.308        NA  2023930000          9           0
#> 3     21780         NA     462.430        NA  1460950000         11           0
#> 4     17910         NA     447.218        NA  2241650000          9           0
#> 5     22000   1399.038     611.800        NA 11496870000         11           4
#> 6     18090   4562.938     550.168        NA 47682600000          9           9
#>   fNobs.LIFEEX fNobs.GINI fNobs.ODA fsd.year fsd.PCGDP fsd.LIFEEX fsd.GINI
#> 1            6          0         6 1.870829        NA  0.8236083       NA
#> 2            9          0         9 2.738613        NA  1.2329864       NA
#> 3           11          0        11 3.316625        NA  2.1599808       NA
#> 4            9          0         9 2.738613        NA  2.1258392       NA
#> 5           11          0        11 3.316625  11.89378  1.8056120       NA
#> 6            9          0         9 2.738613  86.22723  1.2872844       NA
#>      fsd.ODA fvar.year fvar.PCGDP fvar.LIFEEX fvar.GINI     fvar.ODA
#> 1  101559732       3.5         NA   0.6783307        NA 1.031438e+16
#> 2   58515953       7.5         NA   1.5202554        NA 3.424117e+15
#> 3  108271755      11.0         NA   4.6655169        NA 1.172277e+16
#> 4  203018929       7.5         NA   4.5191921        NA 4.121669e+16
#> 5 1113781886      11.0    141.462   3.2602346        NA 1.240510e+18
#> 6 1102034385       7.5   7435.136   1.6571012        NA 1.214480e+18
#>   fmode.iso3c fmode.date fmode.region fmode.income fmode.OECD ffirst.iso3c
#> 1         AFG 1961-01-01   South Asia   Low income      FALSE          AFG
#> 2         AFG 1967-01-01   South Asia   Low income      FALSE          AFG
#> 3         AFG 1976-01-01   South Asia   Low income      FALSE          AFG
#> 4         AFG 1987-01-01   South Asia   Low income      FALSE          AFG
#> 5         AFG 1996-01-01   South Asia   Low income      FALSE          AFG
#> 6         AFG 2007-01-01   South Asia   Low income      FALSE          AFG
#>   ffirst.date ffirst.region ffirst.income ffirst.OECD flast.iso3c flast.date
#> 1  1961-01-01    South Asia    Low income       FALSE         AFG 1966-01-01
#> 2  1967-01-01    South Asia    Low income       FALSE         AFG 1975-01-01
#> 3  1976-01-01    South Asia    Low income       FALSE         AFG 1986-01-01
#> 4  1987-01-01    South Asia    Low income       FALSE         AFG 1995-01-01
#> 5  1996-01-01    South Asia    Low income       FALSE         AFG 2006-01-01
#> 6  2007-01-01    South Asia    Low income       FALSE         AFG 2015-01-01
#>   flast.region flast.income flast.OECD fNdistinct.iso3c fNdistinct.date
#> 1   South Asia   Low income      FALSE                1               6
#> 2   South Asia   Low income      FALSE                1               9
#> 3   South Asia   Low income      FALSE                1              11
#> 4   South Asia   Low income      FALSE                1               9
#> 5   South Asia   Low income      FALSE                1              11
#> 6   South Asia   Low income      FALSE                1               9
#>   fNdistinct.region fNdistinct.income fNdistinct.OECD
#> 1                 1                 1               1
#> 2                 1                 1               1
#> 3                 1                 1               1
#> 4                 1                 1               1
#> 5                 1                 1               1
#> 6                 1                 1               1

## Custom Aggregation ------------------------------------------
head(collap(wlddev, ~ country + decade,                         # Custom aggregation
            custom = list(fmean = 9:12, fsd = 9:10, fmode = 7:8)))

#>       country decade fmode.income fmode.OECD fmean.PCGDP fsd.PCGDP fmean.LIFEEX
#> 1 Afghanistan   1960   Low income      FALSE          NA        NA     33.39967
#> 2 Afghanistan   1970   Low income      FALSE          NA        NA     36.70089
#> 3 Afghanistan   1980   Low income      FALSE          NA        NA     42.03909
#> 4 Afghanistan   1990   Low income      FALSE          NA        NA     49.69089
#> 5 Afghanistan   2000   Low income      FALSE    349.7596  11.89378     55.61818
#> 6 Afghanistan   2010   Low income      FALSE    506.9931  86.22723     61.12978
#>   fsd.LIFEEX fmean.GINI  fmean.ODA
#> 1  0.8236083         NA  228641667
#> 2  1.2329864         NA  224881111
#> 3  2.1599808         NA  132813636
#> 4  2.1258392         NA  249072222
#> 5  1.8056120         NA 1045170000
#> 6  1.2872844         NA 5298066667

head(collap(wlddev, ~ country + decade,                         # Using column names
            custom = list(fmean = "PCGDP", fsd = c("LIFEEX","GINI"),
                          flast = "date")))

#>       country flast.date decade fmean.PCGDP fsd.LIFEEX fsd.GINI
#> 1 Afghanistan 1966-01-01   1960          NA  0.8236083       NA
#> 2 Afghanistan 1975-01-01   1970          NA  1.2329864       NA
#> 3 Afghanistan 1986-01-01   1980          NA  2.1599808       NA
#> 4 Afghanistan 1995-01-01   1990          NA  2.1258392       NA
#> 5 Afghanistan 2006-01-01   2000    349.7596  1.8056120       NA
#> 6 Afghanistan 2015-01-01   2010    506.9931  1.2872844       NA

head(collap(wlddev, ~ country + decade,                         # Weighted parallelized custom
            custom = list(fmean = 9:12, fsd = 9:10,             # aggregation
                          fmode = 7:8), w = weights,
            wFUN = list(fsum, fmax),
            parallel = TRUE, mc.cores = 1L))

#>   fsum.weights fmax.weights     country decade fmode.income fmode.OECD
#> 1     3.977500     1.138012 Afghanistan   1960   Low income      FALSE
#> 2     7.180692     1.900061 Afghanistan   1970   Low income      FALSE
#> 3     8.201658     1.270672 Afghanistan   1980   Low income      FALSE
#> 4    10.210256     2.211769 Afghanistan   1990   Low income      FALSE
#> 5    11.736101     2.564408 Afghanistan   2000   Low income      FALSE
#> 6     7.948094     2.446680 Afghanistan   2010   Low income      FALSE
#>   fmean.PCGDP fsd.PCGDP fmean.LIFEEX fsd.LIFEEX fmean.GINI  fmean.ODA
#> 1          NA        NA     33.59138  0.7744188         NA  228677841
#> 2          NA        NA     37.06992  1.1054627         NA  192003634
#> 3          NA        NA     42.38328  2.3401883         NA  113094881
#> 4          NA        NA     49.93548  1.8482921         NA  273981591
#> 5    343.9491  7.427909     55.15815  1.7398488         NA  838996680
#> 6    510.4979 84.786556     61.35658  1.3745334         NA 5179680196

head(collap(wlddev, ~ country + decade,                         # No column reordering
            custom = list(fmean = 9:12, fsd = 9:10,
                          fmode = 7:8), w = weights,
            wFUN = list(fsum, fmax),
            parallel = TRUE, mc.cores = 1L, keep.col.order = FALSE))

#>       country decade fsum.weights fmax.weights fmean.PCGDP fmean.LIFEEX
#> 1 Afghanistan   1960     3.977500     1.138012          NA     33.59138
#> 2 Afghanistan   1970     7.180692     1.900061          NA     37.06992
#> 3 Afghanistan   1980     8.201658     1.270672          NA     42.38328
#> 4 Afghanistan   1990    10.210256     2.211769          NA     49.93548
#> 5 Afghanistan   2000    11.736101     2.564408    343.9491     55.15815
#> 6 Afghanistan   2010     7.948094     2.446680    510.4979     61.35658
#>   fmean.GINI  fmean.ODA fsd.PCGDP fsd.LIFEEX fmode.income fmode.OECD
#> 1         NA  228677841        NA  0.7744188   Low income      FALSE
#> 2         NA  192003634        NA  1.1054627   Low income      FALSE
#> 3         NA  113094881        NA  2.3401883   Low income      FALSE
#> 4         NA  273981591        NA  1.8482921   Low income      FALSE
#> 5         NA  838996680  7.427909  1.7398488   Low income      FALSE
#> 6         NA 5179680196 84.786556  1.3745334   Low income      FALSE

## Piped Use --------------------------------------------------
wlddev %>% fgroup_by(country, decade) %>% collapg

#> # A tibble: 1,512 x 12
#>    country decade iso3c date        year region income OECD  PCGDP LIFEEX  GINI
#>    <chr>    <dbl> <fct> <date>     <dbl> <fct>  <fct>  <lgl> <dbl>  <dbl> <dbl>
#>  1 Afghan~   1960 AFG   1961-01-01 1962. South~ Low i~ FALSE   NA    33.4    NA
#>  2 Afghan~   1970 AFG   1967-01-01 1970  South~ Low i~ FALSE   NA    36.7    NA
#>  3 Afghan~   1980 AFG   1976-01-01 1980  South~ Low i~ FALSE   NA    42.0    NA
#>  4 Afghan~   1990 AFG   1987-01-01 1990  South~ Low i~ FALSE   NA    49.7    NA
#>  5 Afghan~   2000 AFG   1996-01-01 2000  South~ Low i~ FALSE  350.   55.6    NA
#>  6 Afghan~   2010 AFG   2007-01-01 2010  South~ Low i~ FALSE  507.   61.1    NA
#>  7 Afghan~   2020 AFG   2016-01-01 2016. South~ Low i~ FALSE  584.   63.5    NA
#>  8 Albania   1960 ALB   1961-01-01 1962. Europ~ Upper~ FALSE   NA    64.3    NA
#>  9 Albania   1970 ALB   1967-01-01 1970  Europ~ Upper~ FALSE   NA    67.1    NA
#> 10 Albania   1980 ALB   1976-01-01 1980  Europ~ Upper~ FALSE 2075.   70.2    NA
#> # ... with 1,502 more rows, and 1 more variable: ODA <dbl>

wlddev %>% fgroup_by(country, decade) %>% collapg(w = ODA)

#> # A tibble: 1,512 x 12
#>    country decade      ODA iso3c date        year region income OECD  PCGDP
#>    <chr>    <dbl>    <dbl> <fct> <date>     <dbl> <fct>  <fct>  <lgl> <dbl>
#>  1 Afghan~   1960  1.37e 9 AFG   1966-01-01 1963. South~ Low i~ FALSE   NA 
#>  2 Afghan~   1970  2.02e 9 AFG   1967-01-01 1970. South~ Low i~ FALSE   NA 
#>  3 Afghan~   1980  1.46e 9 AFG   1978-01-01 1978. South~ Low i~ FALSE   NA 
#>  4 Afghan~   1990  2.24e 9 AFG   1992-01-01 1991. South~ Low i~ FALSE   NA 
#>  5 Afghan~   2000  1.15e10 AFG   2006-01-01 2003. South~ Low i~ FALSE  351.
#>  6 Afghan~   2010  4.77e10 AFG   2012-01-01 2010. South~ Low i~ FALSE  515.
#>  7 Afghan~   2020  1.21e10 AFG   2016-01-01 2016. South~ Low i~ FALSE  584.
#>  8 Albania   1960 NA       ALB   1966-01-01   NA  Europ~ Upper~ NA      NA 
#>  9 Albania   1970 NA       ALB   1975-01-01   NA  Europ~ Upper~ NA      NA 
#> 10 Albania   1980 NA       ALB   1986-01-01   NA  Europ~ Upper~ NA      NA 
#> # ... with 1,502 more rows, and 2 more variables: LIFEEX <dbl>, GINI <dbl>

wlddev %>% fgroup_by(country, decade) %>% collapg(fmedian, flast)

#> # A tibble: 1,512 x 12
#>    country decade iso3c date        year region income OECD  PCGDP LIFEEX  GINI
#>    <chr>    <dbl> <fct> <date>     <dbl> <fct>  <fct>  <lgl> <dbl>  <dbl> <dbl>
#>  1 Afghan~   1960 AFG   1966-01-01 1962. South~ Low i~ FALSE   NA    33.4    NA
#>  2 Afghan~   1970 AFG   1975-01-01 1970  South~ Low i~ FALSE   NA    36.7    NA
#>  3 Afghan~   1980 AFG   1986-01-01 1980  South~ Low i~ FALSE   NA    41.9    NA
#>  4 Afghan~   1990 AFG   1995-01-01 1990  South~ Low i~ FALSE   NA    49.9    NA
#>  5 Afghan~   2000 AFG   2006-01-01 2000  South~ Low i~ FALSE  347.   55.5    NA
#>  6 Afghan~   2010 AFG   2015-01-01 2010  South~ Low i~ FALSE  536.   61.2    NA
#>  7 Afghan~   2020 AFG   2019-01-01 2016. South~ Low i~ FALSE  584.   63.5    NA
#>  8 Albania   1960 ALB   1966-01-01 1962. Europ~ Upper~ FALSE   NA    64.5    NA
#>  9 Albania   1970 ALB   1975-01-01 1970  Europ~ Upper~ FALSE   NA    66.9    NA
#> 10 Albania   1980 ALB   1986-01-01 1980  Europ~ Upper~ FALSE 2072.   70.2    NA
#> # ... with 1,502 more rows, and 1 more variable: ODA <dbl>

wlddev %>% fgroup_by(country, decade) %>%
  collapg(custom = list(fmean = 9:12, fmode = 5:7, flast = 3))

#> # A tibble: 1,512 x 10
#>    country decade flast.date fmode.decade fmode.region fmode.income fmean.PCGDP
#>    <chr>    <dbl> <date>            <dbl> <fct>        <fct>              <dbl>
#>  1 Afghan~   1960 1966-01-01         1960 South Asia   Low income           NA 
#>  2 Afghan~   1970 1975-01-01         1970 South Asia   Low income           NA 
#>  3 Afghan~   1980 1986-01-01         1980 South Asia   Low income           NA 
#>  4 Afghan~   1990 1995-01-01         1990 South Asia   Low income           NA 
#>  5 Afghan~   2000 2006-01-01         2000 South Asia   Low income          350.
#>  6 Afghan~   2010 2015-01-01         2010 South Asia   Low income          507.
#>  7 Afghan~   2020 2019-01-01         2020 South Asia   Low income          584.
#>  8 Albania   1960 1966-01-01         1960 Europe & Ce~ Upper middl~         NA 
#>  9 Albania   1970 1975-01-01         1970 Europe & Ce~ Upper middl~         NA 
#> 10 Albania   1980 1986-01-01         1980 Europe & Ce~ Upper middl~       2075.
#> # ... with 1,502 more rows, and 3 more variables: fmean.LIFEEX <dbl>,
#> #   fmean.GINI <dbl>, fmean.ODA <dbl>