Fast Renaming and Relabelling Objects

A fast substitute for dplyr::rename. setrename renames objects by reference. These functions also work with objects other than data frames that have a 'names' attribute. relabel and setrelabel do that same for labels attached to list elements / data frame columns.

frename(.x, ..., cols = NULL)
rnm(.x, ..., cols = NULL)     # Shortcut for frename()

setrename(.x, ..., cols = NULL)

relabel(.x, ..., cols = NULL, attrn = "label")

setrelabel(.x, ..., cols = NULL, attrn = "label")

Arguments

.x

an R object with a 'names' attribute.

...

either tagged vector expressions of the form name = newname / name = newlabel, or a single function (+ optional arguments to the function) applied to all names/labels (of columns/elements selected in cols).

cols

If ... is a function, select a subset of columns/elements to rename/relabel using names, indices, a logical vector or a function applied to the columns if .x is a data frame (e.g. is.numeric).

attrn

character. Name of attribute to store labels or retrieve labels from.

Value

.x renamed / relabelled. setrename and setrelabel return .x invisibly.

Note

Note that both relabel and setrelabel modify .x by reference. This is because labels are attached to columns themselves, making it impossible to avoid permanent modification by taking a shallow copy of the encompassing list / data.frame. On the other hand frename makes a shallow copy whereas setrename also modifies by reference.

See also

Examples

## Using tagged expressions
head(frename(iris, Sepal.Length = SL, Sepal.Width = SW,
                   Petal.Length = PL, Petal.Width = PW))
#>    SL  SW  PL  PW Species
#> 1 5.1 3.5 1.4 0.2  setosa
#> 2 4.9 3.0 1.4 0.2  setosa
#> 3 4.7 3.2 1.3 0.2  setosa
#> 4 4.6 3.1 1.5 0.2  setosa
#> 5 5.0 3.6 1.4 0.2  setosa
#> 6 5.4 3.9 1.7 0.4  setosa
head(frename(iris, Sepal.Length = "S L", Sepal.Width = "S W",
                   Petal.Length = "P L", Petal.Width = "P W"))
#>   S L S W P L P W Species
#> 1 5.1 3.5 1.4 0.2  setosa
#> 2 4.9 3.0 1.4 0.2  setosa
#> 3 4.7 3.2 1.3 0.2  setosa
#> 4 4.6 3.1 1.5 0.2  setosa
#> 5 5.0 3.6 1.4 0.2  setosa
#> 6 5.4 3.9 1.7 0.4  setosa

## Using a function
head(frename(iris, tolower))
#>   sepal.length sepal.width petal.length petal.width species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
head(frename(iris, tolower, cols = 1:2))
#>   sepal.length sepal.width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
head(frename(iris, tolower, cols = is.numeric))
#>   sepal.length sepal.width petal.length petal.width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
head(frename(iris, paste, "new", sep = "_", cols = 1:2))
#>   Sepal.Length_new Sepal.Width_new Petal.Length Petal.Width Species
#> 1              5.1             3.5          1.4         0.2  setosa
#> 2              4.9             3.0          1.4         0.2  setosa
#> 3              4.7             3.2          1.3         0.2  setosa
#> 4              4.6             3.1          1.5         0.2  setosa
#> 5              5.0             3.6          1.4         0.2  setosa
#> 6              5.4             3.9          1.7         0.4  setosa

## Renaming by reference
# setrename(iris, tolower)
# head(iris)
# rm(iris)

## Relabelling (by reference)
# namlab(relabel(wlddev, PCGDP = "GDP per Capita", LIFEEX = "Life Expectancy"))
# namlab(relabel(wlddev, toupper))