Fast Summarise

fsummarize is a much faster version of dplyr::summarise, when used together with the Fast Statistical Functions.

fsummarise(.data, ..., keep.group_vars = TRUE)
smr(.data, ..., keep.group_vars = TRUE)        # Shortcut

Arguments

.data
a (grouped) data frame or named list of columns. Grouped data can be created with fgroup_by or dplyr::group_by.

...
name-value pairs of summary functions. The name will be the name of the variable in the result. Functions when applied to a vector need to return a scalar. For fast performance and weighted aggregation use the Fast Statistical Functions.

keep.group_vars
logical. FALSE removes grouping variables after computation.

Value

If .data is grouped by fgroup_by or dplyr::group_by, the result is a data frame of the same class and attributes with rows reduced to the number of groups. If .data is not grouped, the result is a data frame of the same class and attributes with 1 row.

Note

Functions other than the Fast Statistical Functions cannot be passed weight vectors as fsummarize only splits the .data, not any arguments passed to functions. If you want to summarize data with a weighted function you created yourself, use data.table.

See also

Examples

# Simple use
fsummarise(mtcars, mean_mpg = fmean(mpg),
                   sd_mpg = fsd(mpg))

#>   mean_mpg   sd_mpg
#> 1 20.09062 6.026948

# Using base functions (not a big difference without groups)
fsummarise(mtcars, mean_mpg = mean(mpg),
                   sd_mpg = sd(mpg))

#>   mean_mpg   sd_mpg
#> 1 20.09062 6.026948

# Grouped use
library(magrittr)
mtcars %>% fgroup_by(cyl) %>%
  fsummarise(mean_mpg = fmean(mpg),
             sd_mpg = fsd(mpg))

#>   cyl mean_mpg   sd_mpg
#> 1   4 26.66364 4.509828
#> 2   6 19.74286 1.453567
#> 3   8 15.10000 2.560048

# This is still efficient but quite a bit slower on large data (many groups)
mtcars %>% fgroup_by(cyl) %>%
  fsummarise(mean_mpg = mean(mpg),
             sd_mpg = sd(mpg))

#>   cyl mean_mpg   sd_mpg
#> 1   4 26.66364 4.509828
#> 2   6 19.74286 1.453567
#> 3   8 15.10000 2.560048

# Weighted aggregation
mtcars %>% fgroup_by(cyl) %>%
  fsummarise(w_mean_mpg = fmean(mpg, wt),
             w_sd_mpg = fsd(mpg, wt))

#>   cyl w_mean_mpg w_sd_mpg
#> 1   4   25.93504 4.275234
#> 2   6   19.64578 1.397297
#> 3   8   14.80643 2.638850

 
## Can also group with dplyr::group_by, but at a conversion cost, see ?GRP
library(dplyr)
mtcars %>% group_by(cyl) %>%
  fsummarise(mean_mpg = fmean(mpg),
             sd_mpg = fsd(mpg))

#> # A tibble: 3 x 3
#>     cyl mean_mpg sd_mpg
#>   <dbl>    <dbl>  <dbl>
#> 1     4     26.7   4.51
#> 2     6     19.7   1.45
#> 3     8     15.1   2.56

# Again less efficient...
mtcars %>% group_by(cyl) %>%
  fsummarise(mean_mpg = mean(mpg),
             sd_mpg = sd(mpg))

#> # A tibble: 3 x 3
#>     cyl mean_mpg sd_mpg
#>   <dbl>    <dbl>  <dbl>
#> 1     4     26.7   4.51
#> 2     6     19.7   1.45
#> 3     8     15.1   2.56