Fast Transform and Compute Columns on a Data Frame

ftransform is a much faster update of base::transform for data frames. It returns the data frame with new columns computed and/or existing columns modified or deleted. settransform does all of that by reference i.e. it modifies the data frame in the global environment. fcompute can be used to compute new columns from the columns in a data frame and returns only the computed columns.

# Modify and return 'data.frame'
ftransform(X, ...)
tfm(X, ...)            # Shortcut for ftransform

# Modify 'data.frame' by reference
settransform(X, ...)
settfm(X, ...)         # Shortcut for settransform

# Replace modified columns in a 'data.frame'
ftransform(X) <- value
tfm(X) <- value        # Shortcut for ftransform<-

# Compute and return new 'data.frame' from existing one
fcompute(X, ...)

Arguments

X
a data frame or named list of columns.

...
further arguments of the form column = value. The value can be a combination of other columns, a scalar value, or NULL, which deletes column. Alternatively it is also possible to place a single list here, which will be treated the same as a list of column = value arguments.

value
a named list of replacements, it will be treated like an evaluated list of column = value arguments.

Details

The ... arguments to ftransform are tagged vector expressions, which are evaluated in the data frame X. The tags are matched against names(X), and for those that match, the values replace the corresponding variable in X, whereas the others are appended to X. It is also possible to delete columns by assigning NULL to them, i.e. ftransform(data, colk = NULL) removes colk from the data.

Since collapse v1.3.0, is is also possible to pass a single list to ..., i.e. ftransform(data, newdata) or ftransform(data, fmean(list(col1mean = col1, col2mean = col2), drop = FALSE)) etc. This list will be treated the same as a list of tagged vector expressions. See Examples.

The function settransform does all of that by reference, but uses base-R's copy-on modify semantics, which is equivalent to replacing the data with <- (thus it is still memory efficient but the data will have a different memory address after each call of settransform).

Finally, the function fcompute functions just like ftransform, but returns only the changed / computed columns without modifying or appending the data in X.

Value

The modified data frame X, or, for fcompute, a new data frame with the columns computed on X. All attributes of X are preserved.

See also

Examples

## ftransform modifies and returns a data.frame
head(ftransform(airquality, Ozone = -Ozone))

#>   Ozone Solar.R Wind Temp Month Day
#> 1   -41     190  7.4   67     5   1
#> 2   -36     118  8.0   72     5   2
#> 3   -12     149 12.6   74     5   3
#> 4   -18     313 11.5   62     5   4
#> 5    NA      NA 14.3   56     5   5
#> 6   -28      NA 14.9   66     5   6

head(ftransform(airquality, new = -Ozone, Temp = (Temp-32)/1.8))

#>   Ozone Solar.R Wind     Temp Month Day new
#> 1    41     190  7.4 19.44444     5   1 -41
#> 2    36     118  8.0 22.22222     5   2 -36
#> 3    12     149 12.6 23.33333     5   3 -12
#> 4    18     313 11.5 16.66667     5   4 -18
#> 5    NA      NA 14.3 13.33333     5   5  NA
#> 6    28      NA 14.9 18.88889     5   6 -28

head(ftransform(airquality, new = -Ozone, new2 = 1, Temp = NULL))  # Deleting Temp

#>   Ozone Solar.R Wind Month Day new new2
#> 1    41     190  7.4     5   1 -41    1
#> 2    36     118  8.0     5   2 -36    1
#> 3    12     149 12.6     5   3 -12    1
#> 4    18     313 11.5     5   4 -18    1
#> 5    NA      NA 14.3     5   5  NA    1
#> 6    28      NA 14.9     5   6 -28    1

head(ftransform(airquality, Ozone = NULL, Temp = NULL))            # Deleting columns

#>   Solar.R Wind Month Day
#> 1     190  7.4     5   1
#> 2     118  8.0     5   2
#> 3     149 12.6     5   3
#> 4     313 11.5     5   4
#> 5      NA 14.3     5   5
#> 6      NA 14.9     5   6

# This computes the median and standard-deviation of Ozone in each month
head(ftransform(airquality,
                   Ozone_Month_median = fmedian(Ozone, Month, TRA = "replace_fill"),
                   Ozone_Month_sd = fsd(Ozone, Month, TRA = "replace_fill")))

#>   Ozone Solar.R Wind Temp Month Day Ozone_Month_median Ozone_Month_sd
#> 1    41     190  7.4   67     5   1                 18             22
#> 2    36     118  8.0   72     5   2                 18             22
#> 3    12     149 12.6   74     5   3                 18             22
#> 4    18     313 11.5   62     5   4                 18             22
#> 5    NA      NA 14.3   56     5   5                 18             22
#> 6    28      NA 14.9   66     5   6                 18             22

# Grouping by month and above/below average temperature in each month
head(ftransform(airquality, Ozone_Month_high_median =
     fmedian(Ozone, list(Month, Temp > fbetween(Temp, Month)), TRA = "replace_fill")))

#>   Ozone Solar.R Wind Temp Month Day Ozone_Month_high_median
#> 1    41     190  7.4   67     5   1                      28
#> 2    36     118  8.0   72     5   2                      28
#> 3    12     149 12.6   74     5   3                      28
#> 4    18     313 11.5   62     5   4                      14
#> 5    NA      NA 14.3   56     5   5                      14
#> 6    28      NA 14.9   66     5   6                      28

## Since v1.3.0 one can pass a list of columns, and there is a replacement method
head(ftransform(airquality, STD(airquality, cols = 1:3)))  # Could use magrittr::`%<>%`

#>   Ozone Solar.R Wind Temp Month Day   STD.Ozone STD.Solar.R   STD.Wind
#> 1    41     190  7.4   67     5   1 -0.03423409  0.04517615 -0.7259482
#> 2    36     118  8.0   72     5   2 -0.18580489 -0.75430487 -0.5556388
#> 3    12     149 12.6   74     5   3 -0.91334473 -0.41008388  0.7500660
#> 4    18     313 11.5   62     5   4 -0.73145977  1.41095624  0.4378323
#> 5    NA      NA 14.3   56     5   5          NA          NA  1.2326091
#> 6    28      NA 14.9   66     5   6 -0.42831817          NA  1.4029185

ftransform(airquality) <- fscale(get_vars(airquality, 1:3))
rm(airquality)

# This feature also allows to flexibly do grouped operations creating multiple new columns
head(ftransform(airquality,
   fmedian(list(Wind_Month_median = Wind,
                Ozone_Month_median = Ozone), Month, TRA = "replace_fill")))

#>   Ozone Solar.R Wind Temp Month Day Wind_Month_median Ozone_Month_median
#> 1    41     190  7.4   67     5   1              11.5                 18
#> 2    36     118  8.0   72     5   2              11.5                 18
#> 3    12     149 12.6   74     5   3              11.5                 18
#> 4    18     313 11.5   62     5   4              11.5                 18
#> 5    NA      NA 14.3   56     5   5              11.5                 18
#> 6    28      NA 14.9   66     5   6              11.5                 18

# This performs 2 different multi-column grouped operations (need c() to make it one list)
head(ftransform(airquality, c(fmedian(list(Wind_Day_median = Wind,
                                      Ozone_Day_median = Ozone), Day, TRA = "replace_fill"),
                         fsd(list(Wind_Month_sd = Wind,
                                  Ozone_Month_sd = Ozone), Month, TRA = "replace_fill"))))

#>   Ozone Solar.R Wind Temp Month Day Wind_Day_median Ozone_Day_median
#> 1    41     190  7.4   67     5   1             6.9             68.5
#> 2    36     118  8.0   72     5   2             9.2             42.5
#> 3    12     149 12.6   74     5   3             9.2             24.0
#> 4    18     313 11.5   62     5   4             9.2             78.0
#> 5    NA      NA 14.3   56     5   5             7.4             47.0
#> 6    28      NA 14.9   66     5   6            14.3             36.0
#>   Wind_Month_sd Ozone_Month_sd
#> 1       3.53145       22.22445
#> 2       3.53145       22.22445
#> 3       3.53145       22.22445
#> 4       3.53145       22.22445
#> 5       3.53145       22.22445
#> 6       3.53145       22.22445

## settransform works like ftransform but modifies a data frame in the global environment..
airquality_c <- airquality
settransform(airquality_c, Ratio = Ozone / Temp, Ozone = NULL, Temp = NULL)
settransform(airquality_c, STD(airquality_c, cols = 1:2))    # This also works..
head(airquality_c)

#>   Solar.R Wind Month Day     Ratio STD.Solar.R   STD.Wind
#> 1     190  7.4     5   1 0.6119403  0.04517615 -0.7259482
#> 2     118  8.0     5   2 0.5000000 -0.75430487 -0.5556388
#> 3     149 12.6     5   3 0.1621622 -0.41008388  0.7500660
#> 4     313 11.5     5   4 0.2903226  1.41095624  0.4378323
#> 5      NA 14.3     5   5        NA          NA  1.2326091
#> 6      NA 14.9     5   6 0.4242424          NA  1.4029185

rm(airquality_c)

## fcompute only returns the modified / computed columns
head(fcompute(airquality, Ozone = -Ozone))

#>   Ozone
#> 1   -41
#> 2   -36
#> 3   -12
#> 4   -18
#> 5    NA
#> 6   -28

head(fcompute(airquality, new = -Ozone, Temp = (Temp-32)/1.8))

#>   new     Temp
#> 1 -41 19.44444
#> 2 -36 22.22222
#> 3 -12 23.33333
#> 4 -18 16.66667
#> 5  NA 13.33333
#> 6 -28 18.88889

head(fcompute(airquality, new = -Ozone, new2 = 1))

#>   new new2
#> 1 -41    1
#> 2 -36    1
#> 3 -12    1
#> 4 -18    1
#> 5  NA    1
#> 6 -28    1