Fast Unique Elements / Rows

funique is a substantially faster alternative to unique. It is generic with a default vector method and a data frame method.

funique(x, ...)

# S3 method for default
funique(x, sort = FALSE, method = c("auto", "radix", "hash"), ...)

# S3 method for data.frame
funique(x, cols = NULL, sort = FALSE, ...)

# S3 method for sf
funique(x, cols = NULL, sort = FALSE, ...)

Arguments

x
a atomic vector or data frame / list of equal-length columns.

sort
logical. TRUE orders the unique elements / rows. FALSE returns unique values in order of first occurrence, however numeric values are only provided in order of first occurrence when method = "hash" i.e. the radix method always sorts numeric data.

method
an integer or character string specifying the method of computation:

Int. 		 String 		 Description

1		"auto"		automatic selection: hash for character, logical or if length(x) < 500, else radix.

2		"radix"		use radix ordering to determine unique values.

3		"hash"		use index hashing to determine unique values.

cols
compute unique rows according to a subset of columns. Columns can be selected using column names, indices, a logical vector or a selector function (i.e. is.character). Note: All columns are returned.

...
arguments passed to radixorderv, e.g. decreasing or na.last. Only applicable in funique.default if method = "radix".

Details

If x is a data frame / list and all rows are already unique, then x is returned. Otherwise a copy of x with duplicate rows removed is returned. See qF for some additional computational details.

The sf method simply ignores the geometry column when determining unique values.

Value

x with duplicate elements/rows removed.

See also

Examples

funique(mtcars$cyl)

#> [1] 6 4 8

funique(gv(mtcars, c(2,8,9)))

#>                   cyl vs am
#> Porsche 914-2       4  0  1
#> Merc 240D           4  1  0
#> Datsun 710          4  1  1
#> Mazda RX4           6  0  1
#> Hornet 4 Drive      6  1  0
#> Hornet Sportabout   8  0  0
#> Ford Pantera L      8  0  1

funique(mtcars, cols = c(2,8,9))

#>                    mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Porsche 914-2     26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Merc 240D         24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Datsun 710        22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Mazda RX4         21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Hornet 4 Drive    21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Ford Pantera L    15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4