Fast Hash-Based Grouping

group() scans the rows of a data frame (or atomic vector / list of atomic vectors), assigning to each unique row an integer id - starting with 1 and proceeding in first-appearance order of the rows. The function is written in C and optimized for R's data structures. It is the workhorse behind functions like GRP / fgroup_by, collap, qF, qG, finteraction and funique, when called with argument sort = FALSE.

group(x, starts = FALSE, group.sizes = FALSE)

Arguments

x

an atomic vector or data frame / list of equal-length atomic vectors.

starts

logical. If TRUE, an additional attribute 'starts' is attached giving a vector of group starts (index of first-occurrence of unique rows).

group.sizes

logical. If TRUE, an additional attribute 'group.sizes' is attached giving the size of each group.

Details

A data frame is grouped on a column-by-column basis, starting from the leftmost column. For each new column the grouping vector obtained after the previous column is also fed back into the hash function so that unique values are determined on a running basis. The algorithm terminates as soon as the number of unique rows reaches the size of the data frame. Missing values are also grouped just like any other values. Invoking arguments starts and/or group.sizes requires an additional pass through the final grouping vector.

Value

An object is of class 'qG' see qG.

Author

The Hash Function was taken from the excellent kit package by Morgan Jacob.

See also

Examples

# Let's replicate what funique does
g <- group(wlddev, starts = TRUE)
if(attr(g, "N.groups") == fnrow(wlddev)) wlddev else
   ss(wlddev, attr(g, "starts"))
#>        country iso3c       date year decade                region
#> 1  Afghanistan   AFG 1961-01-01 1960   1960            South Asia
#> 2  Afghanistan   AFG 1962-01-01 1961   1960            South Asia
#> 3  Afghanistan   AFG 1963-01-01 1962   1960            South Asia
#> 4  Afghanistan   AFG 1964-01-01 1963   1960            South Asia
#> 5  Afghanistan   AFG 1965-01-01 1964   1960            South Asia
#> 6  Afghanistan   AFG 1966-01-01 1965   1960            South Asia
#> 7  Afghanistan   AFG 1967-01-01 1966   1960            South Asia
#> 8  Afghanistan   AFG 1968-01-01 1967   1960            South Asia
#> 9  Afghanistan   AFG 1969-01-01 1968   1960            South Asia
#> 10 Afghanistan   AFG 1970-01-01 1969   1960            South Asia
#> 11 Afghanistan   AFG 1971-01-01 1970   1970            South Asia
#> 12 Afghanistan   AFG 1972-01-01 1971   1970            South Asia
#> 13 Afghanistan   AFG 1973-01-01 1972   1970            South Asia
#> 14 Afghanistan   AFG 1974-01-01 1973   1970            South Asia
#> 15 Afghanistan   AFG 1975-01-01 1974   1970            South Asia
#> 16 Afghanistan   AFG 1976-01-01 1975   1970            South Asia
#> 17 Afghanistan   AFG 1977-01-01 1976   1970            South Asia
#> 18 Afghanistan   AFG 1978-01-01 1977   1970            South Asia
#> 19 Afghanistan   AFG 1979-01-01 1978   1970            South Asia
#> 20 Afghanistan   AFG 1980-01-01 1979   1970            South Asia
#> 21 Afghanistan   AFG 1981-01-01 1980   1980            South Asia
#> 22 Afghanistan   AFG 1982-01-01 1981   1980            South Asia
#> 23 Afghanistan   AFG 1983-01-01 1982   1980            South Asia
#> 24 Afghanistan   AFG 1984-01-01 1983   1980            South Asia
#> 25 Afghanistan   AFG 1985-01-01 1984   1980            South Asia
#> 26 Afghanistan   AFG 1986-01-01 1985   1980            South Asia
#> 27 Afghanistan   AFG 1987-01-01 1986   1980            South Asia
#> 28 Afghanistan   AFG 1988-01-01 1987   1980            South Asia
#> 29 Afghanistan   AFG 1989-01-01 1988   1980            South Asia
#> 30 Afghanistan   AFG 1990-01-01 1989   1980            South Asia
#> 31 Afghanistan   AFG 1991-01-01 1990   1990            South Asia
#> 32 Afghanistan   AFG 1992-01-01 1991   1990            South Asia
#> 33 Afghanistan   AFG 1993-01-01 1992   1990            South Asia
#> 34 Afghanistan   AFG 1994-01-01 1993   1990            South Asia
#> 35 Afghanistan   AFG 1995-01-01 1994   1990            South Asia
#> 36 Afghanistan   AFG 1996-01-01 1995   1990            South Asia
#> 37 Afghanistan   AFG 1997-01-01 1996   1990            South Asia
#> 38 Afghanistan   AFG 1998-01-01 1997   1990            South Asia
#> 39 Afghanistan   AFG 1999-01-01 1998   1990            South Asia
#> 40 Afghanistan   AFG 2000-01-01 1999   1990            South Asia
#> 41 Afghanistan   AFG 2001-01-01 2000   2000            South Asia
#> 42 Afghanistan   AFG 2002-01-01 2001   2000            South Asia
#> 43 Afghanistan   AFG 2003-01-01 2002   2000            South Asia
#> 44 Afghanistan   AFG 2004-01-01 2003   2000            South Asia
#> 45 Afghanistan   AFG 2005-01-01 2004   2000            South Asia
#> 46 Afghanistan   AFG 2006-01-01 2005   2000            South Asia
#> 47 Afghanistan   AFG 2007-01-01 2006   2000            South Asia
#> 48 Afghanistan   AFG 2008-01-01 2007   2000            South Asia
#> 49 Afghanistan   AFG 2009-01-01 2008   2000            South Asia
#> 50 Afghanistan   AFG 2010-01-01 2009   2000            South Asia
#> 51 Afghanistan   AFG 2011-01-01 2010   2010            South Asia
#> 52 Afghanistan   AFG 2012-01-01 2011   2010            South Asia
#> 53 Afghanistan   AFG 2013-01-01 2012   2010            South Asia
#> 54 Afghanistan   AFG 2014-01-01 2013   2010            South Asia
#> 55 Afghanistan   AFG 2015-01-01 2014   2010            South Asia
#> 56 Afghanistan   AFG 2016-01-01 2015   2010            South Asia
#> 57 Afghanistan   AFG 2017-01-01 2016   2010            South Asia
#> 58 Afghanistan   AFG 2018-01-01 2017   2010            South Asia
#> 59 Afghanistan   AFG 2019-01-01 2018   2010            South Asia
#> 60 Afghanistan   AFG 2020-01-01 2019   2010            South Asia
#> 61 Afghanistan   AFG 2021-01-01 2020   2020            South Asia
#> 62     Albania   ALB 1961-01-01 1960   1960 Europe & Central Asia
#> 63     Albania   ALB 1962-01-01 1961   1960 Europe & Central Asia
#> 64     Albania   ALB 1963-01-01 1962   1960 Europe & Central Asia
#> 65     Albania   ALB 1964-01-01 1963   1960 Europe & Central Asia
#> 66     Albania   ALB 1965-01-01 1964   1960 Europe & Central Asia
#> 67     Albania   ALB 1966-01-01 1965   1960 Europe & Central Asia
#> 68     Albania   ALB 1967-01-01 1966   1960 Europe & Central Asia
#> 69     Albania   ALB 1968-01-01 1967   1960 Europe & Central Asia
#> 70     Albania   ALB 1969-01-01 1968   1960 Europe & Central Asia
#> 71     Albania   ALB 1970-01-01 1969   1960 Europe & Central Asia
#> 72     Albania   ALB 1971-01-01 1970   1970 Europe & Central Asia
#> 73     Albania   ALB 1972-01-01 1971   1970 Europe & Central Asia
#> 74     Albania   ALB 1973-01-01 1972   1970 Europe & Central Asia
#> 75     Albania   ALB 1974-01-01 1973   1970 Europe & Central Asia
#> 76     Albania   ALB 1975-01-01 1974   1970 Europe & Central Asia
#>                 income  OECD    PCGDP LIFEEX GINI        ODA      POP
#> 1           Low income FALSE       NA 32.446   NA  116769997  8996973
#> 2           Low income FALSE       NA 32.962   NA  232080002  9169410
#> 3           Low income FALSE       NA 33.471   NA  112839996  9351441
#> 4           Low income FALSE       NA 33.971   NA  237720001  9543205
#> 5           Low income FALSE       NA 34.463   NA  295920013  9744781
#> 6           Low income FALSE       NA 34.948   NA  341839996  9956320
#> 7           Low income FALSE       NA 35.430   NA  311609985 10174836
#> 8           Low income FALSE       NA 35.914   NA  247610001 10399926
#> 9           Low income FALSE       NA 36.403   NA  171270004 10637063
#> 10          Low income FALSE       NA 36.900   NA  155229996 10893776
#> 11          Low income FALSE       NA 37.409   NA  164380005 11173642
#> 12          Low income FALSE       NA 37.930   NA  241899994 11475445
#> 13          Low income FALSE       NA 38.461   NA  278799988 11791215
#> 14          Low income FALSE       NA 39.003   NA  250750000 12108963
#> 15          Low income FALSE       NA 39.558   NA  201279999 12412950
#> 16          Low income FALSE       NA 40.128   NA  243979996 12689160
#> 17          Low income FALSE       NA 40.715   NA  249039993 12943093
#> 18          Low income FALSE       NA 41.320   NA  269940002 13171306
#> 19          Low income FALSE       NA 41.944   NA  212190002 13341198
#> 20          Low income FALSE       NA 42.585   NA  249440002 13411056
#> 21          Low income FALSE       NA 43.244   NA   66430000 13356511
#> 22          Low income FALSE       NA 43.923   NA   54529999 13171673
#> 23          Low income FALSE       NA 44.617   NA   25200001 12882528
#> 24          Low income FALSE       NA 45.324   NA   37180000 12537730
#> 25          Low income FALSE       NA 46.040   NA   18190001 12204292
#> 26          Low income FALSE       NA 46.761   NA   42549999 11938208
#> 27          Low income FALSE       NA 47.486   NA    5940000 11736179
#> 28          Low income FALSE       NA 48.211   NA   86529999 11604534
#> 29          Low income FALSE       NA 48.930   NA  126980003 11618005
#> 30          Low income FALSE       NA 49.640   NA  253130005 11868877
#> 31          Low income FALSE       NA 50.331   NA  194039993 12412308
#> 32          Low income FALSE       NA 50.999   NA  748539978 13299017
#> 33          Low income FALSE       NA 51.641   NA  299660004 14485546
#> 34          Low income FALSE       NA 52.256   NA  333450012 15816603
#> 35          Low income FALSE       NA 52.842   NA  329350006 17075727
#> 36          Low income FALSE       NA 53.398   NA  270690002 18110657
#> 37          Low income FALSE       NA 53.924   NA  219210007 18853437
#> 38          Low income FALSE       NA 54.424   NA  330470001 19357126
#> 39          Low income FALSE       NA 54.906   NA  238169998 19737765
#> 40          Low income FALSE       NA 55.376   NA  208970001 20170844
#> 41          Low income FALSE       NA 55.841   NA  215800003 20779953
#> 42          Low income FALSE       NA 56.308   NA  682969971 21606988
#> 43          Low income FALSE 330.3036 56.784   NA 1790479980 22600770
#> 44          Low income FALSE 343.0809 57.271   NA 1972890015 23680871
#> 45          Low income FALSE 333.2167 57.772   NA 2681449951 24726684
#> 46          Low income FALSE 357.2347 58.290   NA 3306389893 25654277
#> 47          Low income FALSE 365.2845 58.826   NA 3287489990 26433049
#> 48          Low income FALSE 405.5490 59.375   NA 5218049805 27100536
#> 49          Low income FALSE 412.0143 59.930   NA 4928439941 27722276
#> 50          Low income FALSE 488.3003 60.484   NA 6456560059 28394813
#> 51          Low income FALSE 543.3030 61.028   NA 6424490234 29185507
#> 52          Low income FALSE 528.7366 61.553   NA 6616870117 30117413
#> 53          Low income FALSE 576.1901 62.054   NA 6505500000 31161376
#> 54          Low income FALSE 587.5651 62.525   NA 5011049805 32269589
#> 55          Low income FALSE 583.6562 62.966   NA 4822640137 33370794
#> 56          Low income FALSE 574.1841 63.377   NA 4479720215 34413603
#> 57          Low income FALSE 571.0738 63.763   NA 4289649902 35383128
#> 58          Low income FALSE 571.4407 64.130   NA 3956239990 36296400
#> 59          Low income FALSE 564.6100 64.486   NA 3792449951 37172386
#> 60          Low income FALSE 573.2876 64.833   NA 4339979980 38041754
#> 61          Low income FALSE       NA     NA   NA         NA       NA
#> 62 Upper middle income FALSE       NA 62.283   NA         NA  1608800
#> 63 Upper middle income FALSE       NA 63.301   NA         NA  1659800
#> 64 Upper middle income FALSE       NA 64.190   NA         NA  1711319
#> 65 Upper middle income FALSE       NA 64.914   NA         NA  1762621
#> 66 Upper middle income FALSE       NA 65.463   NA         NA  1814135
#> 67 Upper middle income FALSE       NA 65.850   NA         NA  1864791
#> 68 Upper middle income FALSE       NA 66.110   NA         NA  1914573
#> 69 Upper middle income FALSE       NA 66.304   NA         NA  1965598
#> 70 Upper middle income FALSE       NA 66.487   NA         NA  2022272
#> 71 Upper middle income FALSE       NA 66.689   NA         NA  2081695
#> 72 Upper middle income FALSE       NA 66.935   NA         NA  2135479
#> 73 Upper middle income FALSE       NA 67.237   NA         NA  2187853
#> 74 Upper middle income FALSE       NA 67.582   NA         NA  2243126
#> 75 Upper middle income FALSE       NA 67.953   NA         NA  2296752
#> 76 Upper middle income FALSE       NA 68.343   NA         NA  2350124
#>  [ reached 'max' / getOption("max.print") -- omitted 13100 rows ]