Fast Hash-Based Grouping

group() scans the rows of a data frame (or atomic vector / list of atomic vectors), assigning to each unique row an integer id - starting with 1 and proceeding in first-appearance order of the rows. The function is written in C and optimized for R's data structures. It is the workhorse behind functions like GRP / fgroup_by, collap, qF, qG, finteraction and funique, when called with argument sort = FALSE.

group(x, starts = FALSE, group.sizes = FALSE)

Arguments

x

an atomic vector or data frame / list of equal-length atomic vectors.

starts

logical. If TRUE, an additional attribute 'starts' is attached giving a vector of group starts (index of first-occurrence of unique rows).

group.sizes

logical. If TRUE, an additional attribute 'group.sizes' is attached giving the size of each group.

Details

A data frame is grouped on a column-by-column basis, starting from the leftmost column. For each new column the grouping vector obtained after the previous column is also fed back into the hash function so that unique values are determined on a running basis. The algorithm terminates as soon as the number of unique rows reaches the size of the data frame. Missing values are also grouped just like any other values. Invoking arguments starts and/or group.sizes requires an additional pass through the final grouping vector.

Value

An object is of class 'qG' see qG.

Author

The Hash Function was taken from the excellent kit package by Morgan Jacob.

See also

Examples

# Let's replicate what funique does
g <- group(wlddev, starts = TRUE)
if(attr(g, "N.groups") == fnrow(wlddev)) wlddev else
   ss(wlddev, attr(g, "starts"))
#>        country iso3c       date year decade     region     income  OECD PCGDP
#> 1  Afghanistan   AFG 1961-01-01 1960   1960 South Asia Low income FALSE    NA
#> 2  Afghanistan   AFG 1962-01-01 1961   1960 South Asia Low income FALSE    NA
#> 3  Afghanistan   AFG 1963-01-01 1962   1960 South Asia Low income FALSE    NA
#> 4  Afghanistan   AFG 1964-01-01 1963   1960 South Asia Low income FALSE    NA
#> 5  Afghanistan   AFG 1965-01-01 1964   1960 South Asia Low income FALSE    NA
#> 6  Afghanistan   AFG 1966-01-01 1965   1960 South Asia Low income FALSE    NA
#> 7  Afghanistan   AFG 1967-01-01 1966   1960 South Asia Low income FALSE    NA
#> 8  Afghanistan   AFG 1968-01-01 1967   1960 South Asia Low income FALSE    NA
#> 9  Afghanistan   AFG 1969-01-01 1968   1960 South Asia Low income FALSE    NA
#> 10 Afghanistan   AFG 1970-01-01 1969   1960 South Asia Low income FALSE    NA
#> 11 Afghanistan   AFG 1971-01-01 1970   1970 South Asia Low income FALSE    NA
#> 12 Afghanistan   AFG 1972-01-01 1971   1970 South Asia Low income FALSE    NA
#> 13 Afghanistan   AFG 1973-01-01 1972   1970 South Asia Low income FALSE    NA
#> 14 Afghanistan   AFG 1974-01-01 1973   1970 South Asia Low income FALSE    NA
#> 15 Afghanistan   AFG 1975-01-01 1974   1970 South Asia Low income FALSE    NA
#>    LIFEEX GINI       ODA      POP
#> 1  32.446   NA 116769997  8996973
#> 2  32.962   NA 232080002  9169410
#> 3  33.471   NA 112839996  9351441
#> 4  33.971   NA 237720001  9543205
#> 5  34.463   NA 295920013  9744781
#> 6  34.948   NA 341839996  9956320
#> 7  35.430   NA 311609985 10174836
#> 8  35.914   NA 247610001 10399926
#> 9  36.403   NA 171270004 10637063
#> 10 36.900   NA 155229996 10893776
#> 11 37.409   NA 164380005 11173642
#> 12 37.930   NA 241899994 11475445
#> 13 38.461   NA 278799988 11791215
#> 14 39.003   NA 250750000 12108963
#> 15 39.558   NA 201279999 12412950
#>  [ reached 'max' / getOption("max.print") -- omitted 13161 rows ]