Pairwise Correlations, Covariances and Observation Count

Computes (pairwise) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method. For an equivalent and faster implementation of pwcor see Hmisc::rcorr (written in Fortran), for weighted correlations see weights::wtd.cors (written in C).

pwcor(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwNobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.83  -.83  -.82   .69  -.85   .50   .67   .68   .54  -.56
#> cyl   -.83    1    .89   .81  -.65   .77  -.64  -.85  -.52  -.44   .48
#> disp  -.83   .89    1    .78  -.69   .92  -.45  -.74  -.56  -.47   .39
#> hp    -.82   .81   .78    1   -.38   .69  -.73  -.75  -.25  -.03   .76
#> drat   .69  -.65  -.69  -.38    1   -.71   .15   .51   .72   .75  -.07
#> wt    -.85   .77   .92   .69  -.71    1   -.20  -.62  -.72  -.54   .42
#> qsec   .50  -.64  -.45  -.73   .15  -.20    1    .73  -.17  -.07  -.68
#> vs     .67  -.85  -.74  -.75   .51  -.62   .73    1    .22   .25  -.62
#> am     .68  -.52  -.56  -.25   .72  -.72  -.17   .22    1    .80   .05
#> gear   .54  -.44  -.47  -.03   .75  -.54  -.07   .25   .80    1    .26
#> carb  -.56   .48   .39   .76  -.07   .42  -.68  -.62   .05   .26    1 

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      32.12     -7.91   -538.02   -295.38      2.10     -4.46      4.75
#> cyl      -7.91      3.06     18.67     98.64      -.56      1.22     -2.12
#> disp   -538.02     18.67  15511.15   7046.58    -48.42     12.64   -106.63
#> hp     -295.38     98.64   7046.58   5014.36    -14.47     48.98    -97.54
#> drat      2.10      -.56    -48.42    -14.47       .29      -.38       .15
#> wt       -4.46      1.22     12.64     48.98      -.38      1.02      -.36
#> qsec      4.75     -2.12   -106.63    -97.54       .15      -.36      3.30
#> vs        1.80      -.73    -47.86    -26.90       .14      -.32       .68
#> am        1.96      -.45    -34.44     -8.84       .20      -.34      -.15
#> gear      1.99      -.54    -38.42     -1.18       .29      -.29      -.09
#> carb     -4.50      1.42     84.65     99.35      -.06       .70     -2.16
#>             vs        am      gear      carb
#> mpg       1.80      1.96      1.99     -4.50
#> cyl       -.73      -.45      -.54      1.42
#> disp    -47.86    -34.44    -38.42     84.65
#> hp      -26.90     -8.84     -1.18     99.35
#> drat       .14       .20       .29      -.06
#> wt        -.32      -.34      -.29       .70
#> qsec       .68      -.15      -.09     -2.16
#> vs         .26       .06       .08      -.53
#> am         .06       .25       .27       .04
#> gear       .08       .27       .46       .28
#> carb      -.53       .04       .28      2.79

pwNobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 26   26 26   26 26 27   28   27
#> cyl   27  29   26 26   26 26   26 26 27   27   26
#> disp  26  26   29 27   27 26   27 27 26   26   26
#> hp    26  26   27 29   27 27   27 28 26   26   26
#> drat  26  26   27 27   29 26   27 28 26   26   26
#> wt    26  26   26 27   26 29   27 27 26   26   27
#> qsec  26  26   27 27   27 27   29 28 26   26   26
#> vs    26  26   27 28   28 27   28 29 26   26   26
#> am    27  27   26 26   26 26   26 26 29   28   26
#> gear  28  27   26 26   26 26   26 26 28   29   27
#> carb  27  26   26 26   26 27   26 26 26   27   29

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.83 (27) -.83 (26) -.82 (26)  .69 (26) -.85 (26)  .50 (26)
#> cyl  -.83 (27)   1  (29)  .89 (26)  .81 (26) -.65 (26)  .77 (26) -.64 (26)
#> disp -.83 (26)  .89 (26)   1  (29)  .78 (27) -.69 (27)  .92 (26) -.45 (27)
#> hp   -.82 (26)  .81 (26)  .78 (27)   1  (29) -.38 (27)  .69 (27) -.73 (27)
#> drat  .69 (26) -.65 (26) -.69 (27) -.38 (27)   1  (29) -.71 (26)  .15 (27)
#> wt   -.85 (26)  .77 (26)  .92 (26)  .69 (27) -.71 (26)   1  (29) -.20 (27)
#> qsec  .50 (26) -.64 (26) -.45 (27) -.73 (27)  .15 (27) -.20 (27)   1  (29)
#> vs    .67 (26) -.85 (26) -.74 (27) -.75 (28)  .51 (28) -.62 (27)  .73 (28)
#> am    .68 (27) -.52 (27) -.56 (26) -.25 (26)  .72 (26) -.72 (26) -.17 (26)
#> gear  .54 (28) -.44 (27) -.47 (26) -.03 (26)  .75 (26) -.54 (26) -.07 (26)
#> carb -.56 (27)  .48 (26)  .39 (26)  .76 (26) -.07 (26)  .42 (27) -.68 (26)
#>             vs        am      gear      carb
#> mpg   .67 (26)  .68 (27)  .54 (28) -.56 (27)
#> cyl  -.85 (26) -.52 (27) -.44 (27)  .48 (26)
#> disp -.74 (27) -.56 (26) -.47 (26)  .39 (26)
#> hp   -.75 (28) -.25 (26) -.03 (26)  .76 (26)
#> drat  .51 (28)  .72 (26)  .75 (26) -.07 (26)
#> wt   -.62 (27) -.72 (26) -.54 (26)  .42 (27)
#> qsec  .73 (28) -.17 (26) -.07 (26) -.68 (26)
#> vs     1  (29)  .22 (26)  .25 (26) -.62 (26)
#> am    .22 (26)   1  (29)  .80 (28)  .05 (26)
#> gear  .25 (26)  .80 (28)   1  (29)  .26 (27)
#> carb -.62 (26)  .05 (26)  .26 (27)   1  (29)

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.83*  -.83*  -.82*   .69*  -.85*   .50*   .67*   .68*   .54*
#> cyl   -.83*    1     .89*   .81*  -.65*   .77*  -.64*  -.85*  -.52*  -.44*
#> disp  -.83*   .89*    1     .78*  -.69*   .92*  -.45*  -.74*  -.56*  -.47*
#> hp    -.82*   .81*   .78*    1    -.38*   .69*  -.73*  -.75*  -.25   -.03 
#> drat   .69*  -.65*  -.69*  -.38*    1    -.71*   .15    .51*   .72*   .75*
#> wt    -.85*   .77*   .92*   .69*  -.71*    1    -.20   -.62*  -.72*  -.54*
#> qsec   .50*  -.64*  -.45*  -.73*   .15   -.20     1     .73*  -.17   -.07 
#> vs     .67*  -.85*  -.74*  -.75*   .51*  -.62*   .73*    1     .22    .25 
#> am     .68*  -.52*  -.56*  -.25    .72*  -.72*  -.17    .22     1     .80*
#> gear   .54*  -.44*  -.47*  -.03    .75*  -.54*  -.07    .25    .80*    1  
#> carb  -.56*   .48*   .39*   .76*  -.07    .42*  -.68*  -.62*   .05    .26 
#>        carb
#> mpg   -.56*
#> cyl    .48*
#> disp   .39*
#> hp     .76*
#> drat  -.07 
#> wt     .42*
#> qsec  -.68*
#> vs    -.62*
#> am     .05 
#> gear   .26 
#> carb    1  

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.83* (27) -.83* (26) -.82* (26)  .69* (26) -.85* (26)
#> cyl  -.83* (27)   1   (29)  .89* (26)  .81* (26) -.65* (26)  .77* (26)
#> disp -.83* (26)  .89* (26)   1   (29)  .78* (27) -.69* (27)  .92* (26)
#> hp   -.82* (26)  .81* (26)  .78* (27)   1   (29) -.38* (27)  .69* (27)
#> drat  .69* (26) -.65* (26) -.69* (27) -.38* (27)   1   (29) -.71* (26)
#> wt   -.85* (26)  .77* (26)  .92* (26)  .69* (27) -.71* (26)   1   (29)
#> qsec  .50* (26) -.64* (26) -.45* (27) -.73* (27)  .15  (27) -.20  (27)
#> vs    .67* (26) -.85* (26) -.74* (27) -.75* (28)  .51* (28) -.62* (27)
#> am    .68* (27) -.52* (27) -.56* (26) -.25  (26)  .72* (26) -.72* (26)
#> gear  .54* (28) -.44* (27) -.47* (26) -.03  (26)  .75* (26) -.54* (26)
#> carb -.56* (27)  .48* (26)  .39* (26)  .76* (26) -.07  (26)  .42* (27)
#>            qsec         vs         am       gear       carb
#> mpg   .50* (26)  .67* (26)  .68* (27)  .54* (28) -.56* (27)
#> cyl  -.64* (26) -.85* (26) -.52* (27) -.44* (27)  .48* (26)
#> disp -.45* (27) -.74* (27) -.56* (26) -.47* (26)  .39* (26)
#> hp   -.73* (27) -.75* (28) -.25  (26) -.03  (26)  .76* (26)
#> drat  .15  (27)  .51* (28)  .72* (26)  .75* (26) -.07  (26)
#> wt   -.20  (27) -.62* (27) -.72* (26) -.54* (26)  .42* (27)
#> qsec   1   (29)  .73* (28) -.17  (26) -.07  (26) -.68* (26)
#> vs    .73* (28)   1   (29)  .22  (26)  .25  (26) -.62* (26)
#> am   -.17  (26)  .22  (26)   1   (29)  .80* (28)  .05  (26)
#> gear -.07  (26)  .25  (26)  .80* (28)   1   (29)  .26  (27)
#> carb -.68* (26) -.62* (26)  .05  (26)  .26  (27)   1   (29)

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.83  -.83  -.82   .69  -.85   .50   .67   .68   .54  -.56
#> N  29    27    26    26    26    26    26    26    27    28    27
#> P       .00   .00   .00   .00   .00   .01   .00   .00   .00   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   1   .89   .81  -.65   .77  -.64  -.85  -.52  -.44   .48
#> N    27  29    26    26    26    26    26    26    27    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .02   .01
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   .89    1   .78  -.69   .92  -.45  -.74  -.56  -.47   .39
#> N    26    26   29    27    27    26    27    27    26    26    26
#> P   .00   .00        .00   .00   .00   .02   .00   .00   .02   .05
#> 
#> , , hp
#> 
#>     mpg   cyl  disp  hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.82   .81   .78   1  -.38   .69  -.73  -.75  -.25  -.03   .76
#> N    26    26    27  29    27    27    27    28    26    26    26
#> P   .00   .00   .00       .05   .00   .00   .00   .23   .90   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .69  -.65  -.69  -.38    1  -.71   .15   .51   .72   .75  -.07
#> N    26    26    27    27   29    26    27    28    26    26    26
#> P   .00   .00   .00   .05        .00   .44   .01   .00   .00   .75
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.85   .77   .92   .69  -.71   1  -.20  -.62  -.72  -.54   .42
#> N    26    26    26    27    26  29    27    27    26    26    27
#> P   .00   .00   .00   .00   .00       .32   .00   .00   .00   .03
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .50  -.64  -.45  -.73   .15  -.20    1   .73  -.17  -.07  -.68
#> N    26    26    27    27    27    27   29    28    26    26    26
#> P   .01   .00   .02   .00   .44   .32        .00   .41   .73   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .67  -.85  -.74  -.75   .51  -.62   .73   1   .22   .25  -.62
#> N    26    26    27    28    28    27    28  29    26    26    26
#> P   .00   .00   .00   .00   .01   .00   .00       .28   .21   .00
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .68  -.52  -.56  -.25   .72  -.72  -.17   .22   1   .80   .05
#> N    27    27    26    26    26    26    26    26  29    28    26
#> P   .00   .01   .00   .23   .00   .00   .41   .28       .00   .79
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .54  -.44  -.47  -.03   .75  -.54  -.07   .25   .80    1   .26
#> N    28    27    26    26    26    26    26    26    28   29    27
#> P   .00   .02   .02   .90   .00   .00   .73   .21   .00        .20
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.56   .48   .39   .76  -.07   .42  -.68  -.62   .05   .26    1
#> N    27    26    26    26    26    27    26    26    26    27   29
#> P   .00   .01   .05   .00   .75   .03   .00   .00   .79   .20     
#> 

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.834* (27)   1    (29)                                                
#> disp -.832* (26)  .893* (26)   1    (29)                                    
#> hp   -.817* (26)  .812* (26)  .775* (27)   1    (29)                        
#> drat  .689* (26) -.647* (26) -.691* (27) -.384  (27)   1    (29)            
#> wt   -.846* (26)  .772* (26)  .920* (26)  .685* (27) -.707* (26)   1    (29)
#> qsec  .498* (26) -.644* (26) -.446  (27) -.727* (27)  .155  (27) -.199  (27)
#> vs    .668* (26) -.852* (26) -.740* (27) -.753* (28)  .512* (28) -.619* (27)
#> am    .683* (27) -.521* (27) -.564* (26) -.245  (26)  .717* (26) -.716* (26)
#> gear  .544* (28) -.439  (27) -.467  (26) -.025  (26)  .745* (26) -.542* (26)
#> carb -.561* (27)  .481  (26)  .391  (26)  .763* (26) -.067  (26)  .415  (27)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .729* (28)   1    (29)                                    
#> am   -.168  (26)  .220  (26)   1    (29)                        
#> gear -.072  (26)  .254  (26)  .796* (28)   1    (29)            
#> carb -.679* (26) -.621* (26)  .054  (26)  .255  (27)   1    (29)

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.83  -.83  -.82   .69  -.85   .50   .67   .68   .54  -.56
#> cyl   -.83     1   .89   .81  -.65   .77  -.64  -.85  -.52  -.44   .48
#> disp  -.83   .89     1   .78  -.69   .92  -.45  -.74  -.56  -.47   .39
#> hp    -.82   .81   .78     1  -.38   .69  -.73  -.75  -.25  -.03   .76
#> drat   .69  -.65  -.69  -.38     1  -.71   .15   .51   .72   .75  -.07
#> wt    -.85   .77   .92   .69  -.71     1  -.20  -.62  -.72  -.54   .42
#> qsec   .50  -.64  -.45  -.73   .15  -.20     1   .73  -.17  -.07  -.68
#> vs     .67  -.85  -.74  -.75   .51  -.62   .73     1   .22   .25  -.62
#> am     .68  -.52  -.56  -.25   .72  -.72  -.17   .22     1   .80   .05
#> gear   .54  -.44  -.47  -.03   .75  -.54  -.07   .25   .80     1   .26
#> carb  -.56   .48   .39   .76  -.07   .42  -.68  -.62   .05   .26     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  26   26  26   26  26  27   28   27
#> cyl   27  29   26  26   26  26   26  26  27   27   26
#> disp  26  26   29  27   27  26   27  27  26   26   26
#> hp    26  26   27  29   27  27   27  28  26   26   26
#> drat  26  26   27  27   29  26   27  28  26   26   26
#> wt    26  26   26  27   26  29   27  27  26   26   27
#> qsec  26  26   27  27   27  27   29  28  26   26   26
#> vs    26  26   27  28   28  27   28  29  26   26   26
#> am    27  27   26  26   26  26   26  26  29   28   26
#> gear  28  27   26  26   26  26   26  26  28   29   27
#> carb  27  26   26  26   26  27   26  26  26   27   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .01  .00  .00  .00  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .02  .01
#> disp  .00  .00       .00  .00  .00  .02  .00  .00  .02  .05
#> hp    .00  .00  .00       .05  .00  .00  .00  .23  .90  .00
#> drat  .00  .00  .00  .05       .00  .44  .01  .00  .00  .75
#> wt    .00  .00  .00  .00  .00       .32  .00  .00  .00  .03
#> qsec  .01  .00  .02  .00  .44  .32       .00  .41  .73  .00
#> vs    .00  .00  .00  .00  .01  .00  .00       .28  .21  .00
#> am    .00  .01  .00  .23  .00  .00  .41  .28       .00  .79
#> gear  .00  .02  .02  .90  .00  .00  .73  .21  .00       .20
#> carb  .00  .01  .05  .00  .75  .03  .00  .00  .79  .20     
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.83     1                                                     
#> disp  -.83   .89     1                                               
#> hp    -.82   .81   .78     1                                         
#> drat   .69  -.65  -.69  -.38     1                                   
#> wt    -.85   .77   .92   .69  -.71     1                             
#> qsec   .50  -.64  -.45  -.73   .15  -.20     1                       
#> vs     .67  -.85  -.74  -.75   .51  -.62   .73     1                 
#> am     .68  -.52  -.56  -.25   .72  -.72  -.17   .22     1           
#> gear   .54  -.44  -.47  -.03   .75  -.54  -.07   .25   .80     1     
#> carb  -.56   .48   .39   .76  -.07   .42  -.68  -.62   .05   .26    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  26   29                                    
#> hp    26  26   27  29                                
#> drat  26  26   27  27   29                           
#> wt    26  26   26  27   26  29                       
#> qsec  26  26   27  27   27  27   29                  
#> vs    26  26   27  28   28  27   28  29              
#> am    27  27   26  26   26  26   26  26  29          
#> gear  28  27   26  26   26  26   26  26  28   29     
#> carb  27  26   26  26   26  27   26  26  26   27   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .05                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .01  .00  .02  .00  .44  .32                         
#> vs    .00  .00  .00  .00  .01  .00  .00                    
#> am    .00  .01  .00  .23  .00  .00  .41  .28               
#> gear  .00  .02  .02  .90  .00  .00  .73  .21  .00          
#> carb  .00  .01  .05  .00  .75  .03  .00  .00  .79  .20     
#>