Pairwise Correlations, Covariances and Observation Count

Computes (pairwise) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method. For an equivalent and faster implementation of pwcor see Hmisc::rcorr (written in Fortran), for weighted correlations see weights::wtd.cors (written in C).

pwcor(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwNobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.87  -.87  -.78   .68  -.84   .47   .72   .58   .45  -.54
#> cyl   -.87    1    .90   .89  -.69   .78  -.59  -.78  -.53  -.55   .49
#> disp  -.87   .90    1    .87  -.70   .85  -.44  -.66  -.58  -.60   .34
#> hp    -.78   .89   .87    1   -.43   .74  -.75  -.75  -.21  -.12   .83
#> drat   .68  -.69  -.70  -.43    1   -.76   .06   .35   .67   .69  -.12
#> wt    -.84   .78   .85   .74  -.76    1   -.21  -.47  -.69  -.57   .45
#> qsec   .47  -.59  -.44  -.75   .06  -.21    1    .76  -.28  -.25  -.69
#> vs     .72  -.78  -.66  -.75   .35  -.47   .76    1    .11   .14  -.65
#> am     .58  -.53  -.58  -.21   .67  -.69  -.28   .11    1    .78   .08
#> gear   .45  -.55  -.60  -.12   .69  -.57  -.25   .14   .78    1    .32
#> carb  -.54   .49   .34   .83  -.12   .45  -.69  -.65   .08   .32    1 

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.83     -9.62   -644.34   -378.93      2.24     -4.41      5.42
#> cyl      -9.62      3.14    194.41     10.81      -.68      1.07     -1.90
#> disp   -644.34    194.41  14423.76   6588.58    -48.01     73.94    -91.69
#> hp     -378.93     10.81   6588.58   5127.41    -17.74     43.70   -102.13
#> drat      2.24      -.68    -48.01    -17.74       .30      -.34       .05
#> wt       -4.41      1.07     73.94     43.70      -.34       .78      -.32
#> qsec      5.42     -1.90    -91.69   -102.13       .05      -.32      3.50
#> vs        2.21      -.69     -4.11    -28.24       .10      -.21       .74
#> am        1.87      -.46    -36.01     -7.66       .18      -.30      -.27
#> gear      2.03      -.72    -54.94     -6.32       .29      -.36      -.38
#> carb     -5.96      1.20     55.52     94.69      -.11       .56     -2.17
#>             vs        am      gear      carb
#> mpg       2.21      1.87      2.03     -5.96
#> cyl       -.69      -.46      -.72      1.20
#> disp     -4.11    -36.01    -54.94     55.52
#> hp      -28.24     -7.66     -6.32     94.69
#> drat       .10       .18       .29      -.11
#> wt        -.21      -.30      -.36       .56
#> qsec       .74      -.27      -.38     -2.17
#> vs         .26       .03       .05      -.57
#> am         .03       .25       .28       .07
#> gear       .05       .28       .58       .38
#> carb      -.57       .07       .38      2.76

pwNobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 26 27   26   27
#> cyl   26  29   28 26   27 27   26 27 27   26   26
#> disp  26  28   29 26   27 27   26 26 26   26   26
#> hp    26  26   26 29   27 26   26 26 26   26   26
#> drat  26  27   27 27   29 27   26 26 26   26   26
#> wt    26  27   27 26   27 29   27 26 26   26   26
#> qsec  26  26   26 26   26 27   29 26 26   26   26
#> vs    26  27   26 26   26 26   26 29 27   27   26
#> am    27  27   26 26   26 26   26 27 29   27   27
#> gear  26  26   26 26   26 26   26 27 27   29   26
#> carb  27  26   26 26   26 26   26 26 27   26   29

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.87 (26) -.87 (26) -.78 (26)  .68 (26) -.84 (26)  .47 (26)
#> cyl  -.87 (26)   1  (29)  .90 (28)  .89 (26) -.69 (27)  .78 (27) -.59 (26)
#> disp -.87 (26)  .90 (28)   1  (29)  .87 (26) -.70 (27)  .85 (27) -.44 (26)
#> hp   -.78 (26)  .89 (26)  .87 (26)   1  (29) -.43 (27)  .74 (26) -.75 (26)
#> drat  .68 (26) -.69 (27) -.70 (27) -.43 (27)   1  (29) -.76 (27)  .06 (26)
#> wt   -.84 (26)  .78 (27)  .85 (27)  .74 (26) -.76 (27)   1  (29) -.21 (27)
#> qsec  .47 (26) -.59 (26) -.44 (26) -.75 (26)  .06 (26) -.21 (27)   1  (29)
#> vs    .72 (26) -.78 (27) -.66 (26) -.75 (26)  .35 (26) -.47 (26)  .76 (26)
#> am    .58 (27) -.53 (27) -.58 (26) -.21 (26)  .67 (26) -.69 (26) -.28 (26)
#> gear  .45 (26) -.55 (26) -.60 (26) -.12 (26)  .69 (26) -.57 (26) -.25 (26)
#> carb -.54 (27)  .49 (26)  .34 (26)  .83 (26) -.12 (26)  .45 (26) -.69 (26)
#>             vs        am      gear      carb
#> mpg   .72 (26)  .58 (27)  .45 (26) -.54 (27)
#> cyl  -.78 (27) -.53 (27) -.55 (26)  .49 (26)
#> disp -.66 (26) -.58 (26) -.60 (26)  .34 (26)
#> hp   -.75 (26) -.21 (26) -.12 (26)  .83 (26)
#> drat  .35 (26)  .67 (26)  .69 (26) -.12 (26)
#> wt   -.47 (26) -.69 (26) -.57 (26)  .45 (26)
#> qsec  .76 (26) -.28 (26) -.25 (26) -.69 (26)
#> vs     1  (29)  .11 (27)  .14 (27) -.65 (26)
#> am    .11 (27)   1  (29)  .78 (27)  .08 (27)
#> gear  .14 (27)  .78 (27)   1  (29)  .32 (26)
#> carb -.65 (26)  .08 (27)  .32 (26)   1  (29)

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.87*  -.87*  -.78*   .68*  -.84*   .47*   .72*   .58*   .45*
#> cyl   -.87*    1     .90*   .89*  -.69*   .78*  -.59*  -.78*  -.53*  -.55*
#> disp  -.87*   .90*    1     .87*  -.70*   .85*  -.44*  -.66*  -.58*  -.60*
#> hp    -.78*   .89*   .87*    1    -.43*   .74*  -.75*  -.75*  -.21   -.12 
#> drat   .68*  -.69*  -.70*  -.43*    1    -.76*   .06    .35    .67*   .69*
#> wt    -.84*   .78*   .85*   .74*  -.76*    1    -.21   -.47*  -.69*  -.57*
#> qsec   .47*  -.59*  -.44*  -.75*   .06   -.21     1     .76*  -.28   -.25 
#> vs     .72*  -.78*  -.66*  -.75*   .35   -.47*   .76*    1     .11    .14 
#> am     .58*  -.53*  -.58*  -.21    .67*  -.69*  -.28    .11     1     .78*
#> gear   .45*  -.55*  -.60*  -.12    .69*  -.57*  -.25    .14    .78*    1  
#> carb  -.54*   .49*   .34    .83*  -.12    .45*  -.69*  -.65*   .08    .32 
#>        carb
#> mpg   -.54*
#> cyl    .49*
#> disp   .34 
#> hp     .83*
#> drat  -.12 
#> wt     .45*
#> qsec  -.69*
#> vs    -.65*
#> am     .08 
#> gear   .32 
#> carb    1  

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.87* (26) -.87* (26) -.78* (26)  .68* (26) -.84* (26)
#> cyl  -.87* (26)   1   (29)  .90* (28)  .89* (26) -.69* (27)  .78* (27)
#> disp -.87* (26)  .90* (28)   1   (29)  .87* (26) -.70* (27)  .85* (27)
#> hp   -.78* (26)  .89* (26)  .87* (26)   1   (29) -.43* (27)  .74* (26)
#> drat  .68* (26) -.69* (27) -.70* (27) -.43* (27)   1   (29) -.76* (27)
#> wt   -.84* (26)  .78* (27)  .85* (27)  .74* (26) -.76* (27)   1   (29)
#> qsec  .47* (26) -.59* (26) -.44* (26) -.75* (26)  .06  (26) -.21  (27)
#> vs    .72* (26) -.78* (27) -.66* (26) -.75* (26)  .35  (26) -.47* (26)
#> am    .58* (27) -.53* (27) -.58* (26) -.21  (26)  .67* (26) -.69* (26)
#> gear  .45* (26) -.55* (26) -.60* (26) -.12  (26)  .69* (26) -.57* (26)
#> carb -.54* (27)  .49* (26)  .34  (26)  .83* (26) -.12  (26)  .45* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .47* (26)  .72* (26)  .58* (27)  .45* (26) -.54* (27)
#> cyl  -.59* (26) -.78* (27) -.53* (27) -.55* (26)  .49* (26)
#> disp -.44* (26) -.66* (26) -.58* (26) -.60* (26)  .34  (26)
#> hp   -.75* (26) -.75* (26) -.21  (26) -.12  (26)  .83* (26)
#> drat  .06  (26)  .35  (26)  .67* (26)  .69* (26) -.12  (26)
#> wt   -.21  (27) -.47* (26) -.69* (26) -.57* (26)  .45* (26)
#> qsec   1   (29)  .76* (26) -.28  (26) -.25  (26) -.69* (26)
#> vs    .76* (26)   1   (29)  .11  (27)  .14  (27) -.65* (26)
#> am   -.28  (26)  .11  (27)   1   (29)  .78* (27)  .08  (27)
#> gear -.25  (26)  .14  (27)  .78* (27)   1   (29)  .32  (26)
#> carb -.69* (26) -.65* (26)  .08  (27)  .32  (26)   1   (29)

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.87  -.87  -.78   .68  -.84   .47   .72   .58   .45  -.54
#> N  29    26    26    26    26    26    26    26    27    26    27
#> P       .00   .00   .00   .00   .00   .02   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.87   1   .90   .89  -.69   .78  -.59  -.78  -.53  -.55   .49
#> N    26  29    28    26    27    27    26    27    27    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .01
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.87   .90    1   .87  -.70   .85  -.44  -.66  -.58  -.60   .34
#> N    26    28   29    26    27    27    26    26    26    26    26
#> P   .00   .00        .00   .00   .00   .02   .00   .00   .00   .09
#> 
#> , , hp
#> 
#>     mpg   cyl  disp  hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.78   .89   .87   1  -.43   .74  -.75  -.75  -.21  -.12   .83
#> N    26    26    26  29    27    26    26    26    26    26    26
#> P   .00   .00   .00       .02   .00   .00   .00   .31   .57   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .68  -.69  -.70  -.43    1  -.76   .06   .35   .67   .69  -.12
#> N    26    27    27    27   29    27    26    26    26    26    26
#> P   .00   .00   .00   .02        .00   .79   .08   .00   .00   .57
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.84   .78   .85   .74  -.76   1  -.21  -.47  -.69  -.57   .45
#> N    26    27    27    26    27  29    27    26    26    26    26
#> P   .00   .00   .00   .00   .00       .30   .02   .00   .00   .02
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .47  -.59  -.44  -.75   .06  -.21    1   .76  -.28  -.25  -.69
#> N    26    26    26    26    26    27   29    26    26    26    26
#> P   .02   .00   .02   .00   .79   .30        .00   .17   .22   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .72  -.78  -.66  -.75   .35  -.47   .76   1   .11   .14  -.65
#> N    26    27    26    26    26    26    26  29    27    27    26
#> P   .00   .00   .00   .00   .08   .02   .00       .60   .50   .00
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .58  -.53  -.58  -.21   .67  -.69  -.28   .11   1   .78   .08
#> N    27    27    26    26    26    26    26    27  29    27    27
#> P   .00   .00   .00   .31   .00   .00   .17   .60       .00   .68
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .45  -.55  -.60  -.12   .69  -.57  -.25   .14   .78    1   .32
#> N    26    26    26    26    26    26    26    27    27   29    26
#> P   .02   .00   .00   .57   .00   .00   .22   .50   .00        .12
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.54   .49   .34   .83  -.12   .45  -.69  -.65   .08   .32    1
#> N    27    26    26    26    26    26    26    26    27    26   29
#> P   .00   .01   .09   .00   .57   .02   .00   .00   .68   .12     
#> 

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.865* (26)   1    (29)                                                
#> disp -.868* (26)  .905* (28)   1    (29)                                    
#> hp   -.782* (26)  .886* (26)  .867* (26)   1    (29)                        
#> drat  .682* (26) -.688* (27) -.704* (27) -.432  (27)   1    (29)            
#> wt   -.840* (26)  .780* (27)  .848* (27)  .745* (26) -.763* (27)   1    (29)
#> qsec  .468  (26) -.586* (26) -.442  (26) -.751* (26)  .055  (26) -.209  (27)
#> vs    .718* (26) -.775* (27) -.656* (26) -.753* (26)  .354  (26) -.471  (26)
#> am    .583* (27) -.525* (27) -.583* (26) -.208  (26)  .668* (26) -.694* (26)
#> gear  .450  (26) -.547* (26) -.603* (26) -.116  (26)  .693* (26) -.568* (26)
#> carb -.543* (27)  .492  (26)  .341  (26)  .832* (26) -.117  (26)  .454  (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .755* (26)   1    (29)                                    
#> am   -.280  (26)  .106  (27)   1    (29)                        
#> gear -.247  (26)  .135  (27)  .781* (27)   1    (29)            
#> carb -.693* (26) -.650* (26)  .083  (27)  .316  (26)   1    (29)

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.87  -.87  -.78   .68  -.84   .47   .72   .58   .45  -.54
#> cyl   -.87     1   .90   .89  -.69   .78  -.59  -.78  -.53  -.55   .49
#> disp  -.87   .90     1   .87  -.70   .85  -.44  -.66  -.58  -.60   .34
#> hp    -.78   .89   .87     1  -.43   .74  -.75  -.75  -.21  -.12   .83
#> drat   .68  -.69  -.70  -.43     1  -.76   .06   .35   .67   .69  -.12
#> wt    -.84   .78   .85   .74  -.76     1  -.21  -.47  -.69  -.57   .45
#> qsec   .47  -.59  -.44  -.75   .06  -.21     1   .76  -.28  -.25  -.69
#> vs     .72  -.78  -.66  -.75   .35  -.47   .76     1   .11   .14  -.65
#> am     .58  -.53  -.58  -.21   .67  -.69  -.28   .11     1   .78   .08
#> gear   .45  -.55  -.60  -.12   .69  -.57  -.25   .14   .78     1   .32
#> carb  -.54   .49   .34   .83  -.12   .45  -.69  -.65   .08   .32     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  26  27   26   27
#> cyl   26  29   28  26   27  27   26  27  27   26   26
#> disp  26  28   29  26   27  27   26  26  26   26   26
#> hp    26  26   26  29   27  26   26  26  26   26   26
#> drat  26  27   27  27   29  27   26  26  26   26   26
#> wt    26  27   27  26   27  29   27  26  26   26   26
#> qsec  26  26   26  26   26  27   29  26  26   26   26
#> vs    26  27   26  26   26  26   26  29  27   27   26
#> am    27  27   26  26   26  26   26  27  29   27   27
#> gear  26  26   26  26   26  26   26  27  27   29   26
#> carb  27  26   26  26   26  26   26  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .02  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .01
#> disp  .00  .00       .00  .00  .00  .02  .00  .00  .00  .09
#> hp    .00  .00  .00       .02  .00  .00  .00  .31  .57  .00
#> drat  .00  .00  .00  .02       .00  .79  .08  .00  .00  .57
#> wt    .00  .00  .00  .00  .00       .30  .02  .00  .00  .02
#> qsec  .02  .00  .02  .00  .79  .30       .00  .17  .22  .00
#> vs    .00  .00  .00  .00  .08  .02  .00       .60  .50  .00
#> am    .00  .00  .00  .31  .00  .00  .17  .60       .00  .68
#> gear  .02  .00  .00  .57  .00  .00  .22  .50  .00       .12
#> carb  .00  .01  .09  .00  .57  .02  .00  .00  .68  .12     
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.87     1                                                     
#> disp  -.87   .90     1                                               
#> hp    -.78   .89   .87     1                                         
#> drat   .68  -.69  -.70  -.43     1                                   
#> wt    -.84   .78   .85   .74  -.76     1                             
#> qsec   .47  -.59  -.44  -.75   .06  -.21     1                       
#> vs     .72  -.78  -.66  -.75   .35  -.47   .76     1                 
#> am     .58  -.53  -.58  -.21   .67  -.69  -.28   .11     1           
#> gear   .45  -.55  -.60  -.12   .69  -.57  -.25   .14   .78     1     
#> carb  -.54   .49   .34   .83  -.12   .45  -.69  -.65   .08   .32    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  28   29                                    
#> hp    26  26   26  29                                
#> drat  26  27   27  27   29                           
#> wt    26  27   27  26   27  29                       
#> qsec  26  26   26  26   26  27   29                  
#> vs    26  27   26  26   26  26   26  29              
#> am    27  27   26  26   26  26   26  27  29          
#> gear  26  26   26  26   26  26   26  27  27   29     
#> carb  27  26   26  26   26  26   26  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .02                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .02  .00  .02  .00  .79  .30                         
#> vs    .00  .00  .00  .00  .08  .02  .00                    
#> am    .00  .00  .00  .31  .00  .00  .17  .60               
#> gear  .02  .00  .00  .57  .00  .00  .22  .50  .00          
#> carb  .00  .01  .09  .00  .57  .02  .00  .00  .68  .12     
#>