(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Notes: weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran)

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwNobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

w
numeric. A vector of (frequency) weights.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.85  -.83  -.79   .69  -.84   .43   .64   .57   .43  -.54
#> cyl   -.85    1    .91   .84  -.66   .70  -.66  -.86  -.38  -.37   .52
#> disp  -.83   .91    1    .80  -.69   .86  -.53  -.71  -.55  -.48   .35
#> hp    -.79   .84   .80    1   -.48   .64  -.76  -.70  -.22  -.14   .77
#> drat   .69  -.66  -.69  -.48    1   -.70   .14   .38   .76   .75  -.11
#> wt    -.84   .70   .86   .64  -.70    1   -.25  -.41  -.68  -.59   .38
#> qsec   .43  -.66  -.53  -.76   .14  -.25    1    .77  -.15  -.07  -.66
#> vs     .64  -.86  -.71  -.70   .38  -.41   .77    1    .01   .14  -.53
#> am     .57  -.38  -.55  -.22   .76  -.68  -.15   .01    1    .78   .08
#> gear   .43  -.37  -.48  -.14   .75  -.59  -.07   .14   .78    1    .34
#> carb  -.54   .52   .35   .77  -.11   .38  -.66  -.53   .08   .34    1 

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      34.52     -9.14   -581.29   -333.81      2.19     -4.17      4.85
#> cyl      -9.14      3.06    189.48    104.41      -.63      1.04     -2.16
#> disp   -581.29    189.48   1444.07   6873.60    -46.11     84.26   -116.50
#> hp     -333.81    104.41   6873.60   5082.36    -17.17     41.74   -102.56
#> drat      2.19      -.63    -46.11    -17.17       .30      -.31       .13
#> wt       -4.17      1.04     84.26     41.74      -.31       .82      -.41
#> qsec      4.85     -2.16   -116.50   -102.56       .13      -.41      3.26
#> vs        1.74      -.73    -41.03    -24.41       .11      -.17       .71
#> am        1.59      -.31    -32.17     -7.55       .21      -.30      -.13
#> gear      1.81      -.45     -4.69     -6.84       .29      -.40      -.08
#> carb     -5.39      1.54     74.62     88.75      -.08       .62     -1.92
#>             vs        am      gear      carb
#> mpg       1.74      1.59      1.81     -5.39
#> cyl       -.73      -.31      -.45      1.54
#> disp    -41.03    -32.17     -4.69     74.62
#> hp      -24.41     -7.55     -6.84     88.75
#> drat       .11       .21       .29      -.08
#> wt        -.17      -.30      -.40       .62
#> qsec       .71      -.13      -.08     -1.92
#> vs         .24       .00       .05      -.44
#> am         .00       .23       .28       .07
#> gear       .05       .28       .51       .42
#> carb      -.44       .07       .42      2.79

pwNobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 26   27 26   26 27 26   27   26
#> cyl   27  29   27 26   26 26   27 27 26   26   26
#> disp  26  27   29 26   27 26   27 26 26   26   26
#> hp    26  26   26 29   26 27   27 27 28   27   27
#> drat  27  26   27 26   29 26   26 26 26   26   26
#> wt    26  26   26 27   26 29   26 27 27   27   26
#> qsec  26  27   27 27   26 26   29 27 26   26   26
#> vs    27  27   26 27   26 27   27 29 27   27   26
#> am    26  26   26 28   26 27   26 27 29   27   27
#> gear  27  26   26 27   26 27   26 27 27   29   26
#> carb  26  26   26 27   26 26   26 26 27   26   29

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (27) -.83 (26) -.79 (26)  .69 (27) -.84 (26)  .43 (26)
#> cyl  -.85 (27)   1  (29)  .91 (27)  .84 (26) -.66 (26)  .70 (26) -.66 (27)
#> disp -.83 (26)  .91 (27)   1  (29)  .80 (26) -.69 (27)  .86 (26) -.53 (27)
#> hp   -.79 (26)  .84 (26)  .80 (26)   1  (29) -.48 (26)  .64 (27) -.76 (27)
#> drat  .69 (27) -.66 (26) -.69 (27) -.48 (26)   1  (29) -.70 (26)  .14 (26)
#> wt   -.84 (26)  .70 (26)  .86 (26)  .64 (27) -.70 (26)   1  (29) -.25 (26)
#> qsec  .43 (26) -.66 (27) -.53 (27) -.76 (27)  .14 (26) -.25 (26)   1  (29)
#> vs    .64 (27) -.86 (27) -.71 (26) -.70 (27)  .38 (26) -.41 (27)  .77 (27)
#> am    .57 (26) -.38 (26) -.55 (26) -.22 (28)  .76 (26) -.68 (27) -.15 (26)
#> gear  .43 (27) -.37 (26) -.48 (26) -.14 (27)  .75 (26) -.59 (27) -.07 (26)
#> carb -.54 (26)  .52 (26)  .35 (26)  .77 (27) -.11 (26)  .38 (26) -.66 (26)
#>             vs        am      gear      carb
#> mpg   .64 (27)  .57 (26)  .43 (27) -.54 (26)
#> cyl  -.86 (27) -.38 (26) -.37 (26)  .52 (26)
#> disp -.71 (26) -.55 (26) -.48 (26)  .35 (26)
#> hp   -.70 (27) -.22 (28) -.14 (27)  .77 (27)
#> drat  .38 (26)  .76 (26)  .75 (26) -.11 (26)
#> wt   -.41 (27) -.68 (27) -.59 (27)  .38 (26)
#> qsec  .77 (27) -.15 (26) -.07 (26) -.66 (26)
#> vs     1  (29)  .01 (27)  .14 (27) -.53 (26)
#> am    .01 (27)   1  (29)  .78 (27)  .08 (27)
#> gear  .14 (27)  .78 (27)   1  (29)  .34 (26)
#> carb -.53 (26)  .08 (27)  .34 (26)   1  (29)

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.83*  -.79*   .69*  -.84*   .43*   .64*   .57*   .43*
#> cyl   -.85*    1     .91*   .84*  -.66*   .70*  -.66*  -.86*  -.38   -.37 
#> disp  -.83*   .91*    1     .80*  -.69*   .86*  -.53*  -.71*  -.55*  -.48*
#> hp    -.79*   .84*   .80*    1    -.48*   .64*  -.76*  -.70*  -.22   -.14 
#> drat   .69*  -.66*  -.69*  -.48*    1    -.70*   .14    .38    .76*   .75*
#> wt    -.84*   .70*   .86*   .64*  -.70*    1    -.25   -.41*  -.68*  -.59*
#> qsec   .43*  -.66*  -.53*  -.76*   .14   -.25     1     .77*  -.15   -.07 
#> vs     .64*  -.86*  -.71*  -.70*   .38   -.41*   .77*    1     .01    .14 
#> am     .57*  -.38   -.55*  -.22    .76*  -.68*  -.15    .01     1     .78*
#> gear   .43*  -.37   -.48*  -.14    .75*  -.59*  -.07    .14    .78*    1  
#> carb  -.54*   .52*   .35    .77*  -.11    .38   -.66*  -.53*   .08    .34 
#>        carb
#> mpg   -.54*
#> cyl    .52*
#> disp   .35 
#> hp     .77*
#> drat  -.11 
#> wt     .38 
#> qsec  -.66*
#> vs    -.53*
#> am     .08 
#> gear   .34 
#> carb    1  

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (27) -.83* (26) -.79* (26)  .69* (27) -.84* (26)
#> cyl  -.85* (27)   1   (29)  .91* (27)  .84* (26) -.66* (26)  .70* (26)
#> disp -.83* (26)  .91* (27)   1   (29)  .80* (26) -.69* (27)  .86* (26)
#> hp   -.79* (26)  .84* (26)  .80* (26)   1   (29) -.48* (26)  .64* (27)
#> drat  .69* (27) -.66* (26) -.69* (27) -.48* (26)   1   (29) -.70* (26)
#> wt   -.84* (26)  .70* (26)  .86* (26)  .64* (27) -.70* (26)   1   (29)
#> qsec  .43* (26) -.66* (27) -.53* (27) -.76* (27)  .14  (26) -.25  (26)
#> vs    .64* (27) -.86* (27) -.71* (26) -.70* (27)  .38  (26) -.41* (27)
#> am    .57* (26) -.38  (26) -.55* (26) -.22  (28)  .76* (26) -.68* (27)
#> gear  .43* (27) -.37  (26) -.48* (26) -.14  (27)  .75* (26) -.59* (27)
#> carb -.54* (26)  .52* (26)  .35  (26)  .77* (27) -.11  (26)  .38  (26)
#>            qsec         vs         am       gear       carb
#> mpg   .43* (26)  .64* (27)  .57* (26)  .43* (27) -.54* (26)
#> cyl  -.66* (27) -.86* (27) -.38  (26) -.37  (26)  .52* (26)
#> disp -.53* (27) -.71* (26) -.55* (26) -.48* (26)  .35  (26)
#> hp   -.76* (27) -.70* (27) -.22  (28) -.14  (27)  .77* (27)
#> drat  .14  (26)  .38  (26)  .76* (26)  .75* (26) -.11  (26)
#> wt   -.25  (26) -.41* (27) -.68* (27) -.59* (27)  .38  (26)
#> qsec   1   (29)  .77* (27) -.15  (26) -.07  (26) -.66* (26)
#> vs    .77* (27)   1   (29)  .01  (27)  .14  (27) -.53* (26)
#> am   -.15  (26)  .01  (27)   1   (29)  .78* (27)  .08  (27)
#> gear -.07  (26)  .14  (27)  .78* (27)   1   (29)  .34  (26)
#> carb -.66* (26) -.53* (26)  .08  (27)  .34  (26)   1   (29)

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.85  -.83  -.79   .69  -.84   .43   .64   .57   .43  -.54
#> N  29    27    26    26    27    26    26    27    26    27    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   1   .91   .84  -.66   .70  -.66  -.86  -.38  -.37   .52
#> N    27  29    27    26    26    26    27    27    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .06   .06   .01
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   .91    1   .80  -.69   .86  -.53  -.71  -.55  -.48   .35
#> N    26    27   29    26    27    26    27    26    26    26    26
#> P   .00   .00        .00   .00   .00   .00   .00   .00   .01   .08
#> 
#> , , hp
#> 
#>     mpg   cyl  disp  hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.79   .84   .80   1  -.48   .64  -.76  -.70  -.22  -.14   .77
#> N    26    26    26  29    26    27    27    27    28    27    27
#> P   .00   .00   .00       .01   .00   .00   .00   .26   .50   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .69  -.66  -.69  -.48    1  -.70   .14   .38   .76   .75  -.11
#> N    27    26    27    26   29    26    26    26    26    26    26
#> P   .00   .00   .00   .01        .00   .51   .06   .00   .00   .58
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.84   .70   .86   .64  -.70   1  -.25  -.41  -.68  -.59   .38
#> N    26    26    26    27    26  29    26    27    27    27    26
#> P   .00   .00   .00   .00   .00       .22   .04   .00   .00   .05
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .43  -.66  -.53  -.76   .14  -.25    1   .77  -.15  -.07  -.66
#> N    26    27    27    27    26    26   29    27    26    26    26
#> P   .03   .00   .00   .00   .51   .22        .00   .46   .74   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .64  -.86  -.71  -.70   .38  -.41   .77   1   .01   .14  -.53
#> N    27    27    26    27    26    27    27  29    27    27    26
#> P   .00   .00   .00   .00   .06   .04   .00       .98   .48   .01
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .57  -.38  -.55  -.22   .76  -.68  -.15   .01   1   .78   .08
#> N    26    26    26    28    26    27    26    27  29    27    27
#> P   .00   .06   .00   .26   .00   .00   .46   .98       .00   .67
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .43  -.37  -.48  -.14   .75  -.59  -.07   .14   .78    1   .34
#> N    27    26    26    27    26    27    26    27    27   29    26
#> P   .02   .06   .01   .50   .00   .00   .74   .48   .00        .09
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.54   .52   .35   .77  -.11   .38  -.66  -.53   .08   .34    1
#> N    26    26    26    27    26    26    26    26    27    26   29
#> P   .00   .01   .08   .00   .58   .05   .00   .01   .67   .09     
#> 

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.847* (27)   1    (29)                                                
#> disp -.829* (26)  .908* (27)   1    (29)                                    
#> hp   -.790* (26)  .836* (26)  .800* (26)   1    (29)                        
#> drat  .695* (27) -.656* (26) -.688* (27) -.476  (26)   1    (29)            
#> wt   -.840* (26)  .704* (26)  .864* (26)  .638* (27) -.702* (26)   1    (29)
#> qsec  .432  (26) -.664* (27) -.526* (27) -.764* (27)  .136  (26) -.248  (26)
#> vs    .641* (27) -.856* (27) -.707* (26) -.697* (27)  .379  (26) -.406  (27)
#> am    .567* (26) -.379  (26) -.548* (26) -.220  (28)  .757* (26) -.680* (27)
#> gear  .434  (27) -.370  (26) -.477  (26) -.136  (27)  .751* (26) -.588* (27)
#> carb -.537* (26)  .518* (26)  .353  (26)  .773* (27) -.113  (26)  .383  (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .773* (27)   1    (29)                                    
#> am   -.151  (26)  .006  (27)   1    (29)                        
#> gear -.067  (26)  .142  (27)  .781* (27)   1    (29)            
#> carb -.659* (26) -.530* (26)  .085  (27)  .336  (26)   1    (29)

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.85  -.83  -.79   .69  -.84   .43   .64   .57   .43  -.54
#> cyl   -.85     1   .91   .84  -.66   .70  -.66  -.86  -.38  -.37   .52
#> disp  -.83   .91     1   .80  -.69   .86  -.53  -.71  -.55  -.48   .35
#> hp    -.79   .84   .80     1  -.48   .64  -.76  -.70  -.22  -.14   .77
#> drat   .69  -.66  -.69  -.48     1  -.70   .14   .38   .76   .75  -.11
#> wt    -.84   .70   .86   .64  -.70     1  -.25  -.41  -.68  -.59   .38
#> qsec   .43  -.66  -.53  -.76   .14  -.25     1   .77  -.15  -.07  -.66
#> vs     .64  -.86  -.71  -.70   .38  -.41   .77     1   .01   .14  -.53
#> am     .57  -.38  -.55  -.22   .76  -.68  -.15   .01     1   .78   .08
#> gear   .43  -.37  -.48  -.14   .75  -.59  -.07   .14   .78     1   .34
#> carb  -.54   .52   .35   .77  -.11   .38  -.66  -.53   .08   .34     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  26   27  26   26  27  26   27   26
#> cyl   27  29   27  26   26  26   27  27  26   26   26
#> disp  26  27   29  26   27  26   27  26  26   26   26
#> hp    26  26   26  29   26  27   27  27  28   27   27
#> drat  27  26   27  26   29  26   26  26  26   26   26
#> wt    26  26   26  27   26  29   26  27  27   27   26
#> qsec  26  27   27  27   26  26   29  27  26   26   26
#> vs    27  27   26  27   26  27   27  29  27   27   26
#> am    26  26   26  28   26  27   26  27  29   27   27
#> gear  27  26   26  27   26  27   26  27  27   29   26
#> carb  26  26   26  27   26  26   26  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .06  .06  .01
#> disp  .00  .00       .00  .00  .00  .00  .00  .00  .01  .08
#> hp    .00  .00  .00       .01  .00  .00  .00  .26  .50  .00
#> drat  .00  .00  .00  .01       .00  .51  .06  .00  .00  .58
#> wt    .00  .00  .00  .00  .00       .22  .04  .00  .00  .05
#> qsec  .03  .00  .00  .00  .51  .22       .00  .46  .74  .00
#> vs    .00  .00  .00  .00  .06  .04  .00       .98  .48  .01
#> am    .00  .06  .00  .26  .00  .00  .46  .98       .00  .67
#> gear  .02  .06  .01  .50  .00  .00  .74  .48  .00       .09
#> carb  .00  .01  .08  .00  .58  .05  .00  .01  .67  .09     
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.85     1                                                     
#> disp  -.83   .91     1                                               
#> hp    -.79   .84   .80     1                                         
#> drat   .69  -.66  -.69  -.48     1                                   
#> wt    -.84   .70   .86   .64  -.70     1                             
#> qsec   .43  -.66  -.53  -.76   .14  -.25     1                       
#> vs     .64  -.86  -.71  -.70   .38  -.41   .77     1                 
#> am     .57  -.38  -.55  -.22   .76  -.68  -.15   .01     1           
#> gear   .43  -.37  -.48  -.14   .75  -.59  -.07   .14   .78     1     
#> carb  -.54   .52   .35   .77  -.11   .38  -.66  -.53   .08   .34    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  27   29                                    
#> hp    26  26   26  29                                
#> drat  27  26   27  26   29                           
#> wt    26  26   26  27   26  29                       
#> qsec  26  27   27  27   26  26   29                  
#> vs    27  27   26  27   26  27   27  29              
#> am    26  26   26  28   26  27   26  27  29          
#> gear  27  26   26  27   26  27   26  27  27   29     
#> carb  26  26   26  27   26  26   26  26  27   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .03  .00  .00  .00  .51  .22                         
#> vs    .00  .00  .00  .00  .06  .04  .00                    
#> am    .00  .06  .00  .26  .00  .00  .46  .98               
#> gear  .02  .06  .01  .50  .00  .00  .74  .48  .00          
#> carb  .00  .01  .08  .00  .58  .05  .00  .01  .67  .09     
#>