(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.85  -.80  -.74   .67  -.86   .44   .67   .62   .46  -.59
#> cyl   -.85    1    .91   .85  -.74   .81  -.58  -.83  -.63  -.55   .60
#> disp  -.80   .91    1    .79  -.72   .89  -.49  -.69  -.60  -.47   .40
#> hp    -.74   .85   .79    1   -.47   .67  -.70  -.67  -.31  -.05   .73
#> drat   .67  -.74  -.72  -.47    1   -.72   .15   .56   .68   .67  -.24
#> wt    -.86   .81   .89   .67  -.72    1   -.21  -.61  -.72  -.57   .41
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      35.35     -9.59   -489.77   -269.88      2.27     -4.44      5.01
#> cyl      -9.59      3.28    192.40     91.70     -0.74      1.46     -1.95
#> disp   -489.77    192.40  14039.38   6530.86    -47.34    100.44   -116.80
#> hp     -269.88     91.70   6530.86   4580.32    -17.02     43.63    -93.19
#> drat      2.27     -0.74    -47.34    -17.02      0.28     -0.38      0.15
#> wt       -4.44      1.46    100.44     43.63     -0.38      0.90     -0.39
#>             vs        am      gear      carb
#> mpg       2.12      1.73      1.90     -5.69
#> cyl      -0.77     -0.56     -0.72      1.46
#> disp    -40.83    -33.16    -40.18     81.76
#> hp      -22.76     -9.86     -2.52     83.39
#> drat      0.15      0.18      0.28     -0.20
#> wt       -0.31     -0.32     -0.43      0.68
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 26 26   26   26
#> cyl   26  29   26 26   26 26   26 27 26   26   26
#> disp  26  26   29 27   26 27   26 26 26   26   26
#> hp    26  26   27 29   27 28   26 26 27   27   26
#> drat  26  26   26 27   29 27   26 27 26   26   26
#> wt    26  26   27 28   27 29   27 26 26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (26) -.80 (26) -.74 (26)  .67 (26) -.86 (26)  .44 (26)
#> cyl  -.85 (26)   1  (29)  .91 (26)  .85 (26) -.74 (26)  .81 (26) -.58 (26)
#> disp -.80 (26)  .91 (26)   1  (29)  .79 (27) -.72 (26)  .89 (27) -.49 (26)
#> hp   -.74 (26)  .85 (26)  .79 (27)   1  (29) -.47 (27)  .67 (28) -.70 (26)
#> drat  .67 (26) -.74 (26) -.72 (26) -.47 (27)   1  (29) -.72 (27)  .15 (26)
#> wt   -.86 (26)  .81 (26)  .89 (27)  .67 (28) -.72 (27)   1  (29) -.21 (27)
#>             vs        am      gear      carb
#> mpg   .67 (26)  .62 (26)  .46 (26) -.59 (26)
#> cyl  -.83 (27) -.63 (26) -.55 (26)  .60 (26)
#> disp -.69 (26) -.60 (26) -.47 (26)  .40 (26)
#> hp   -.67 (26) -.31 (27) -.05 (27)  .73 (26)
#> drat  .56 (27)  .68 (26)  .67 (26) -.24 (26)
#> wt   -.61 (26) -.72 (26) -.57 (26)  .41 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.80*  -.74*   .67*  -.86*   .44*   .67*   .62*   .46*
#> cyl   -.85*    1     .91*   .85*  -.74*   .81*  -.58*  -.83*  -.63*  -.55*
#> disp  -.80*   .91*    1     .79*  -.72*   .89*  -.49*  -.69*  -.60*  -.47*
#> hp    -.74*   .85*   .79*    1    -.47*   .67*  -.70*  -.67*  -.31   -.05 
#> drat   .67*  -.74*  -.72*  -.47*    1    -.72*   .15    .56*   .68*   .67*
#> wt    -.86*   .81*   .89*   .67*  -.72*    1    -.21   -.61*  -.72*  -.57*
#>        carb
#> mpg   -.59*
#> cyl    .60*
#> disp   .40*
#> hp     .73*
#> drat  -.24 
#> wt     .41*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (26) -.80* (26) -.74* (26)  .67* (26) -.86* (26)
#> cyl  -.85* (26)   1   (29)  .91* (26)  .85* (26) -.74* (26)  .81* (26)
#> disp -.80* (26)  .91* (26)   1   (29)  .79* (27) -.72* (26)  .89* (27)
#> hp   -.74* (26)  .85* (26)  .79* (27)   1   (29) -.47* (27)  .67* (28)
#> drat  .67* (26) -.74* (26) -.72* (26) -.47* (27)   1   (29) -.72* (27)
#> wt   -.86* (26)  .81* (26)  .89* (27)  .67* (28) -.72* (27)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .44* (26)  .67* (26)  .62* (26)  .46* (26) -.59* (26)
#> cyl  -.58* (26) -.83* (27) -.63* (26) -.55* (26)  .60* (26)
#> disp -.49* (26) -.69* (26) -.60* (26) -.47* (26)  .40* (26)
#> hp   -.70* (26) -.67* (26) -.31  (27) -.05  (27)  .73* (26)
#> drat  .15  (26)  .56* (27)  .68* (26)  .67* (26) -.24  (26)
#> wt   -.21  (27) -.61* (26) -.72* (26) -.57* (26)  .41* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.85  -.80  -.74   .67  -.86   .44   .67   .62   .46  -.59
#> N  29    26    26    26    26    26    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   1   .91   .85  -.74   .81  -.58  -.83  -.63  -.55   .60
#> N    26  29    26    26    26    26    26    27    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.849* (26)   1    (29)                                                
#> disp -.799* (26)  .907* (26)   1    (29)                                    
#> hp   -.741* (26)  .850* (26)  .790* (27)   1    (29)                        
#> drat  .673* (26) -.742* (26) -.722* (26) -.468  (27)   1    (29)            
#> wt   -.857* (26)  .807* (26)  .889* (27)  .668* (28) -.723* (27)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.85  -.80  -.74   .67  -.86   .44   .67   .62   .46  -.59
#> cyl   -.85     1   .91   .85  -.74   .81  -.58  -.83  -.63  -.55   .60
#> disp  -.80   .91     1   .79  -.72   .89  -.49  -.69  -.60  -.47   .40
#> hp    -.74   .85   .79     1  -.47   .67  -.70  -.67  -.31  -.05   .73
#> drat   .67  -.74  -.72  -.47     1  -.72   .15   .56   .68   .67  -.24
#> wt    -.86   .81   .89   .67  -.72     1  -.21  -.61  -.72  -.57   .41
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  26  26   26   26
#> cyl   26  29   26  26   26  26   26  27  26   26   26
#> disp  26  26   29  27   26  27   26  26  26   26   26
#> hp    26  26   27  29   27  28   26  26  27   27   26
#> drat  26  26   26  27   29  27   26  27  26   26   26
#> wt    26  26   27  28   27  29   27  26  26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .00
#> disp  .00  .00       .00  .00  .00  .01  .00  .00  .02  .04
#> hp    .00  .00  .00       .01  .00  .00  .00  .12  .82  .00
#> drat  .00  .00  .00  .01       .00  .47  .00  .00  .00  .25
#> wt    .00  .00  .00  .00  .00       .29  .00  .00  .00  .04
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.85     1                                                     
#> disp  -.80   .91     1                                               
#> hp    -.74   .85   .79     1                                         
#> drat   .67  -.74  -.72  -.47     1                                   
#> wt    -.86   .81   .89   .67  -.72     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  26   27  29                                
#> drat  26  26   26  27   29                           
#> wt    26  26   27  28   27  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>