(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.84  -.84  -.78   .72  -.87   .36   .70   .63   .52  -.51
#> cyl   -.84    1    .90   .83  -.74   .73  -.56  -.88  -.53  -.49   .53
#> disp  -.84   .90    1    .85  -.76   .88  -.40  -.76  -.63  -.62   .37
#> hp    -.78   .83   .85    1   -.52   .66  -.75  -.78  -.24  -.04   .77
#> drat   .72  -.74  -.76  -.52    1   -.73   .07   .54   .67   .67  -.08
#> wt    -.87   .73   .88   .66  -.73    1   -.16  -.57  -.70  -.53   .36
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      33.81     -7.30   -622.37   -316.43      2.11     -4.96      3.69
#> cyl      -7.30      3.16    204.17    105.41     -0.71      1.20     -1.74
#> disp   -622.37    204.17  15800.94   6724.96    -52.15    110.68    -89.39
#> hp     -316.43    105.41   6724.96   5074.61    -19.14     48.01   -102.63
#> drat      2.11     -0.71    -52.15    -19.14      0.28     -0.38      0.05
#> wt       -4.96      1.20    110.68     48.01     -0.38      0.94     -0.29
#>             vs        am      gear      carb
#> mpg       1.82      1.88      2.00     -4.23
#> cyl      -0.79     -0.47     -0.63      1.41
#> disp    -48.15    -39.25    -57.74     59.61
#> hp      -27.42     -8.77     -2.16     93.61
#> drat      0.14      0.18      0.26     -0.07
#> wt       -0.27     -0.32     -0.37      0.56
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 26   26 26   26 26 27   26   26
#> cyl   27  29   27 26   27 26   26 27 26   28   27
#> disp  26  27   29 27   26 26   26 26 26   27   26
#> hp    26  26   27 29   27 27   26 26 26   27   26
#> drat  26  27   26 27   29 26   26 28 26   28   27
#> wt    26  26   26 27   26 29   26 26 27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.84 (27) -.84 (26) -.78 (26)  .72 (26) -.87 (26)  .36 (26)
#> cyl  -.84 (27)   1  (29)  .90 (27)  .83 (26) -.74 (27)  .73 (26) -.56 (26)
#> disp -.84 (26)  .90 (27)   1  (29)  .85 (27) -.76 (26)  .88 (26) -.40 (26)
#> hp   -.78 (26)  .83 (26)  .85 (27)   1  (29) -.52 (27)  .66 (27) -.75 (26)
#> drat  .72 (26) -.74 (27) -.76 (26) -.52 (27)   1  (29) -.73 (26)  .07 (26)
#> wt   -.87 (26)  .73 (26)  .88 (26)  .66 (27) -.73 (26)   1  (29) -.16 (26)
#>             vs        am      gear      carb
#> mpg   .70 (26)  .63 (27)  .52 (26) -.51 (26)
#> cyl  -.88 (27) -.53 (26) -.49 (28)  .53 (27)
#> disp -.76 (26) -.63 (26) -.62 (27)  .37 (26)
#> hp   -.78 (26) -.24 (26) -.04 (27)  .77 (26)
#> drat  .54 (28)  .67 (26)  .67 (28) -.08 (27)
#> wt   -.57 (26) -.70 (27) -.53 (26)  .36 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.84*  -.84*  -.78*   .72*  -.87*   .36    .70*   .63*   .52*
#> cyl   -.84*    1     .90*   .83*  -.74*   .73*  -.56*  -.88*  -.53*  -.49*
#> disp  -.84*   .90*    1     .85*  -.76*   .88*  -.40*  -.76*  -.63*  -.62*
#> hp    -.78*   .83*   .85*    1    -.52*   .66*  -.75*  -.78*  -.24   -.04 
#> drat   .72*  -.74*  -.76*  -.52*    1    -.73*   .07    .54*   .67*   .67*
#> wt    -.87*   .73*   .88*   .66*  -.73*    1    -.16   -.57*  -.70*  -.53*
#>        carb
#> mpg   -.51*
#> cyl    .53*
#> disp   .37 
#> hp     .77*
#> drat  -.08 
#> wt     .36 
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.84* (27) -.84* (26) -.78* (26)  .72* (26) -.87* (26)
#> cyl  -.84* (27)   1   (29)  .90* (27)  .83* (26) -.74* (27)  .73* (26)
#> disp -.84* (26)  .90* (27)   1   (29)  .85* (27) -.76* (26)  .88* (26)
#> hp   -.78* (26)  .83* (26)  .85* (27)   1   (29) -.52* (27)  .66* (27)
#> drat  .72* (26) -.74* (27) -.76* (26) -.52* (27)   1   (29) -.73* (26)
#> wt   -.87* (26)  .73* (26)  .88* (26)  .66* (27) -.73* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .36  (26)  .70* (26)  .63* (27)  .52* (26) -.51* (26)
#> cyl  -.56* (26) -.88* (27) -.53* (26) -.49* (28)  .53* (27)
#> disp -.40* (26) -.76* (26) -.63* (26) -.62* (27)  .37  (26)
#> hp   -.75* (26) -.78* (26) -.24  (26) -.04  (27)  .77* (26)
#> drat  .07  (26)  .54* (28)  .67* (26)  .67* (28) -.08  (27)
#> wt   -.16  (26) -.57* (26) -.70* (27) -.53* (26)  .36  (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.84  -.84  -.78   .72  -.87   .36   .70   .63   .52  -.51
#> N  29    27    26    26    26    26    26    26    27    26    26
#> P       .00   .00   .00   .00   .00   .07   .00   .00   .01   .01
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.84   1   .90   .83  -.74   .73  -.56  -.88  -.53  -.49   .53
#> N    27  29    27    26    27    26    26    27    26    28    27
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .01   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.842* (27)   1    (29)                                                
#> disp -.841* (26)  .902* (27)   1    (29)                                    
#> hp   -.780* (26)  .826* (26)  .853* (27)   1    (29)                        
#> drat  .720* (26) -.739* (27) -.758* (26) -.516* (27)   1    (29)            
#> wt   -.872* (26)  .734* (26)  .881* (26)  .661* (27) -.735* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.84  -.84  -.78   .72  -.87   .36   .70   .63   .52  -.51
#> cyl   -.84     1   .90   .83  -.74   .73  -.56  -.88  -.53  -.49   .53
#> disp  -.84   .90     1   .85  -.76   .88  -.40  -.76  -.63  -.62   .37
#> hp    -.78   .83   .85     1  -.52   .66  -.75  -.78  -.24  -.04   .77
#> drat   .72  -.74  -.76  -.52     1  -.73   .07   .54   .67   .67  -.08
#> wt    -.87   .73   .88   .66  -.73     1  -.16  -.57  -.70  -.53   .36
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  26   26  26   26  26  27   26   26
#> cyl   27  29   27  26   27  26   26  27  26   28   27
#> disp  26  27   29  27   26  26   26  26  26   27   26
#> hp    26  26   27  29   27  27   26  26  26   27   26
#> drat  26  27   26  27   29  26   26  28  26   28   27
#> wt    26  26   26  27   26  29   26  26  27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .07  .00  .00  .01  .01
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .01  .00
#> disp  .00  .00       .00  .00  .00  .05  .00  .00  .00  .06
#> hp    .00  .00  .00       .01  .00  .00  .00  .23  .84  .00
#> drat  .00  .00  .00  .01       .00  .75  .00  .00  .00  .70
#> wt    .00  .00  .00  .00  .00       .43  .00  .00  .01  .07
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.84     1                                                     
#> disp  -.84   .90     1                                               
#> hp    -.78   .83   .85     1                                         
#> drat   .72  -.74  -.76  -.52     1                                   
#> wt    -.87   .73   .88   .66  -.73     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  27   29                                    
#> hp    26  26   27  29                                
#> drat  26  27   26  27   29                           
#> wt    26  26   26  27   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>