(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.83  -.86  -.78   .64  -.87   .46   .68   .56   .47  -.53
#> cyl   -.83    1    .90   .86  -.63   .78  -.57  -.78  -.53  -.63   .51
#> disp  -.86   .90    1    .86  -.67   .88  -.44  -.75  -.60  -.64   .45
#> hp    -.78   .86   .86    1   -.39   .66  -.73  -.74  -.24  -.12   .75
#> drat   .64  -.63  -.67  -.39    1   -.66   .10   .48   .64   .67  -.04
#> wt    -.87   .78   .88   .66  -.66    1   -.18  -.56  -.68  -.61   .35
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      35.73     -8.58   -641.28   -360.61      1.79     -5.19      5.46
#> cyl      -8.58      3.12    204.40     91.80     -0.51      1.26     -1.81
#> disp   -641.28    204.40  15593.08   7014.39    -44.84    102.66   -101.35
#> hp     -360.61     91.80   7014.39   5047.92    -15.01     45.66    -96.85
#> drat      1.79     -0.51    -44.84    -15.01      0.27     -0.32      0.09
#> wt       -5.19      1.26    102.66     45.66     -0.32      0.88     -0.33
#>             vs        am      gear      carb
#> mpg       2.15      1.73      2.19     -4.76
#> cyl      -0.70     -0.46     -0.81      1.22
#> disp    -48.21    -39.46    -59.56     76.19
#> hp      -27.31     -9.02     -6.19     89.74
#> drat      0.13      0.16      0.27     -0.04
#> wt       -0.27     -0.34     -0.45      0.56
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   27 27   26 26 26   26   27
#> cyl   26  29   27 27   26 26   27 26 27   26   26
#> disp  26  27   29 26   26 26   26 26 26   26   26
#> hp    26  27   26 29   26 26   28 26 27   28   26
#> drat  27  26   26 26   29 26   26 26 26   26   28
#> wt    27  26   26 26   26 29   26 27 26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.83 (26) -.86 (26) -.78 (26)  .64 (27) -.87 (27)  .46 (26)
#> cyl  -.83 (26)   1  (29)  .90 (27)  .86 (27) -.63 (26)  .78 (26) -.57 (27)
#> disp -.86 (26)  .90 (27)   1  (29)  .86 (26) -.67 (26)  .88 (26) -.44 (26)
#> hp   -.78 (26)  .86 (27)  .86 (26)   1  (29) -.39 (26)  .66 (26) -.73 (28)
#> drat  .64 (27) -.63 (26) -.67 (26) -.39 (26)   1  (29) -.66 (26)  .10 (26)
#> wt   -.87 (27)  .78 (26)  .88 (26)  .66 (26) -.66 (26)   1  (29) -.18 (26)
#>             vs        am      gear      carb
#> mpg   .68 (26)  .56 (26)  .47 (26) -.53 (27)
#> cyl  -.78 (26) -.53 (27) -.63 (26)  .51 (26)
#> disp -.75 (26) -.60 (26) -.64 (26)  .45 (26)
#> hp   -.74 (26) -.24 (27) -.12 (28)  .75 (26)
#> drat  .48 (26)  .64 (26)  .67 (26) -.04 (28)
#> wt   -.56 (27) -.68 (26) -.61 (27)  .35 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.83*  -.86*  -.78*   .64*  -.87*   .46*   .68*   .56*   .47*
#> cyl   -.83*    1     .90*   .86*  -.63*   .78*  -.57*  -.78*  -.53*  -.63*
#> disp  -.86*   .90*    1     .86*  -.67*   .88*  -.44*  -.75*  -.60*  -.64*
#> hp    -.78*   .86*   .86*    1    -.39*   .66*  -.73*  -.74*  -.24   -.12 
#> drat   .64*  -.63*  -.67*  -.39*    1    -.66*   .10    .48*   .64*   .67*
#> wt    -.87*   .78*   .88*   .66*  -.66*    1    -.18   -.56*  -.68*  -.61*
#>        carb
#> mpg   -.53*
#> cyl    .51*
#> disp   .45*
#> hp     .75*
#> drat  -.04 
#> wt     .35 
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.83* (26) -.86* (26) -.78* (26)  .64* (27) -.87* (27)
#> cyl  -.83* (26)   1   (29)  .90* (27)  .86* (27) -.63* (26)  .78* (26)
#> disp -.86* (26)  .90* (27)   1   (29)  .86* (26) -.67* (26)  .88* (26)
#> hp   -.78* (26)  .86* (27)  .86* (26)   1   (29) -.39* (26)  .66* (26)
#> drat  .64* (27) -.63* (26) -.67* (26) -.39* (26)   1   (29) -.66* (26)
#> wt   -.87* (27)  .78* (26)  .88* (26)  .66* (26) -.66* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .46* (26)  .68* (26)  .56* (26)  .47* (26) -.53* (27)
#> cyl  -.57* (27) -.78* (26) -.53* (27) -.63* (26)  .51* (26)
#> disp -.44* (26) -.75* (26) -.60* (26) -.64* (26)  .45* (26)
#> hp   -.73* (28) -.74* (26) -.24  (27) -.12  (28)  .75* (26)
#> drat  .10  (26)  .48* (26)  .64* (26)  .67* (26) -.04  (28)
#> wt   -.18  (26) -.56* (27) -.68* (26) -.61* (27)  .35  (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.83  -.86  -.78   .64  -.87   .46   .68   .56   .47  -.53
#> N  29    26    26    26    27    27    26    26    26    26    27
#> P       .00   .00   .00   .00   .00   .02   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   1   .90   .86  -.63   .78  -.57  -.78  -.53  -.63   .51
#> N    26  29    27    27    26    26    27    26    27    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.826* (26)   1    (29)                                                
#> disp -.856* (26)  .905* (27)   1    (29)                                    
#> hp   -.785* (26)  .862* (27)  .862* (26)   1    (29)                        
#> drat  .637* (27) -.631* (26) -.665* (26) -.389  (26)   1    (29)            
#> wt   -.869* (27)  .783* (26)  .879* (26)  .655* (26) -.656* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.83  -.86  -.78   .64  -.87   .46   .68   .56   .47  -.53
#> cyl   -.83     1   .90   .86  -.63   .78  -.57  -.78  -.53  -.63   .51
#> disp  -.86   .90     1   .86  -.67   .88  -.44  -.75  -.60  -.64   .45
#> hp    -.78   .86   .86     1  -.39   .66  -.73  -.74  -.24  -.12   .75
#> drat   .64  -.63  -.67  -.39     1  -.66   .10   .48   .64   .67  -.04
#> wt    -.87   .78   .88   .66  -.66     1  -.18  -.56  -.68  -.61   .35
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   27  27   26  26  26   26   27
#> cyl   26  29   27  27   26  26   27  26  27   26   26
#> disp  26  27   29  26   26  26   26  26  26   26   26
#> hp    26  27   26  29   26  26   28  26  27   28   26
#> drat  27  26   26  26   29  26   26  26  26   26   28
#> wt    27  26   26  26   26  29   26  27  26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .02  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .01
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .02
#> hp    .00  .00  .00       .05  .00  .00  .00  .23  .55  .00
#> drat  .00  .00  .00  .05       .00  .64  .01  .00  .00  .82
#> wt    .00  .00  .00  .00  .00       .37  .00  .00  .00  .08
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.83     1                                                     
#> disp  -.86   .90     1                                               
#> hp    -.78   .86   .86     1                                         
#> drat   .64  -.63  -.67  -.39     1                                   
#> wt    -.87   .78   .88   .66  -.66     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    26  27   26  29                                
#> drat  27  26   26  26   29                           
#> wt    27  26   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .05                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>