(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.89  -.82  -.77   .70  -.86   .36   .69   .61   .49  -.47
#> cyl   -.89    1    .92   .84  -.70   .77  -.60  -.88  -.49  -.45   .49
#> disp  -.82   .92    1    .79  -.68   .87  -.54  -.77  -.56  -.49   .33
#> hp    -.77   .84   .79    1   -.46   .65  -.75  -.75  -.27  -.11   .75
#> drat   .70  -.70  -.68  -.46    1   -.71   .03   .51   .71   .66  -.11
#> wt    -.86   .77   .87   .65  -.71    1   -.13  -.57  -.67  -.54   .35
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      37.87     -9.94   -599.31   -348.62      2.54     -5.45      2.76
#> cyl      -9.94      3.24    199.06    114.70     -0.72      1.39     -2.02
#> disp   -599.31    199.06  14308.94   7201.26    -45.03     94.99   -111.84
#> hp     -348.62    114.70   7201.26   5025.54    -18.08     49.48    -80.98
#> drat      2.54     -0.72    -45.03    -18.08      0.29     -0.40      0.03
#> wt       -5.45      1.39     94.99     49.48     -0.40      0.97     -0.22
#>             vs        am      gear      carb
#> mpg       1.88      1.98      2.52     -4.02
#> cyl      -0.76     -0.45     -0.65      1.49
#> disp    -46.55    -32.22    -42.57     64.73
#> hp      -25.89     -9.61     -6.29     87.49
#> drat      0.13      0.20      0.29     -0.10
#> wt       -0.28     -0.31     -0.40      0.54
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 26   26 27 26   26   26
#> cyl   26  29   27 26   27 26   26 26 27   27   26
#> disp  26  27   29 26   26 27   26 27 26   27   26
#> hp    27  26   26 29   28 26   26 27 27   26   26
#> drat  26  27   26 28   29 26   26 26 27   26   26
#> wt    26  26   27 26   26 29   26 27 26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.89 (26) -.82 (26) -.77 (27)  .70 (26) -.86 (26)  .36 (26)
#> cyl  -.89 (26)   1  (29)  .92 (27)  .84 (26) -.70 (27)  .77 (26) -.60 (26)
#> disp -.82 (26)  .92 (27)   1  (29)  .79 (26) -.68 (26)  .87 (27) -.54 (26)
#> hp   -.77 (27)  .84 (26)  .79 (26)   1  (29) -.46 (28)  .65 (26) -.75 (26)
#> drat  .70 (26) -.70 (27) -.68 (26) -.46 (28)   1  (29) -.71 (26)  .03 (26)
#> wt   -.86 (26)  .77 (26)  .87 (27)  .65 (26) -.71 (26)   1  (29) -.13 (26)
#>             vs        am      gear      carb
#> mpg   .69 (27)  .61 (26)  .49 (26) -.47 (26)
#> cyl  -.88 (26) -.49 (27) -.45 (27)  .49 (26)
#> disp -.77 (27) -.56 (26) -.49 (27)  .33 (26)
#> hp   -.75 (27) -.27 (27) -.11 (26)  .75 (26)
#> drat  .51 (26)  .71 (27)  .66 (26) -.11 (26)
#> wt   -.57 (27) -.67 (26) -.54 (27)  .35 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.89*  -.82*  -.77*   .70*  -.86*   .36    .69*   .61*   .49*
#> cyl   -.89*    1     .92*   .84*  -.70*   .77*  -.60*  -.88*  -.49*  -.45*
#> disp  -.82*   .92*    1     .79*  -.68*   .87*  -.54*  -.77*  -.56*  -.49*
#> hp    -.77*   .84*   .79*    1    -.46*   .65*  -.75*  -.75*  -.27   -.11 
#> drat   .70*  -.70*  -.68*  -.46*    1    -.71*   .03    .51*   .71*   .66*
#> wt    -.86*   .77*   .87*   .65*  -.71*    1    -.13   -.57*  -.67*  -.54*
#>        carb
#> mpg   -.47*
#> cyl    .49*
#> disp   .33 
#> hp     .75*
#> drat  -.11 
#> wt     .35 
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.89* (26) -.82* (26) -.77* (27)  .70* (26) -.86* (26)
#> cyl  -.89* (26)   1   (29)  .92* (27)  .84* (26) -.70* (27)  .77* (26)
#> disp -.82* (26)  .92* (27)   1   (29)  .79* (26) -.68* (26)  .87* (27)
#> hp   -.77* (27)  .84* (26)  .79* (26)   1   (29) -.46* (28)  .65* (26)
#> drat  .70* (26) -.70* (27) -.68* (26) -.46* (28)   1   (29) -.71* (26)
#> wt   -.86* (26)  .77* (26)  .87* (27)  .65* (26) -.71* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .36  (26)  .69* (27)  .61* (26)  .49* (26) -.47* (26)
#> cyl  -.60* (26) -.88* (26) -.49* (27) -.45* (27)  .49* (26)
#> disp -.54* (26) -.77* (27) -.56* (26) -.49* (27)  .33  (26)
#> hp   -.75* (26) -.75* (27) -.27  (27) -.11  (26)  .75* (26)
#> drat  .03  (26)  .51* (26)  .71* (27)  .66* (26) -.11  (26)
#> wt   -.13  (26) -.57* (27) -.67* (26) -.54* (27)  .35  (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.89  -.82  -.77   .70  -.86   .36   .69   .61   .49  -.47
#> N  29    26    26    27    26    26    26    27    26    26    26
#> P       .00   .00   .00   .00   .00   .08   .00   .00   .01   .02
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.89   1   .92   .84  -.70   .77  -.60  -.88  -.49  -.45   .49
#> N    26  29    27    26    27    26    26    26    27    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .02   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.889* (26)   1    (29)                                                
#> disp -.825* (26)  .915* (27)   1    (29)                                    
#> hp   -.774* (27)  .840* (26)  .794* (26)   1    (29)                        
#> drat  .699* (26) -.698* (27) -.679* (26) -.459  (28)   1    (29)            
#> wt   -.865* (26)  .768* (26)  .866* (27)  .650* (26) -.712* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.89  -.82  -.77   .70  -.86   .36   .69   .61   .49  -.47
#> cyl   -.89     1   .92   .84  -.70   .77  -.60  -.88  -.49  -.45   .49
#> disp  -.82   .92     1   .79  -.68   .87  -.54  -.77  -.56  -.49   .33
#> hp    -.77   .84   .79     1  -.46   .65  -.75  -.75  -.27  -.11   .75
#> drat   .70  -.70  -.68  -.46     1  -.71   .03   .51   .71   .66  -.11
#> wt    -.86   .77   .87   .65  -.71     1  -.13  -.57  -.67  -.54   .35
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  26   26  27  26   26   26
#> cyl   26  29   27  26   27  26   26  26  27   27   26
#> disp  26  27   29  26   26  27   26  27  26   27   26
#> hp    27  26   26  29   28  26   26  27  27   26   26
#> drat  26  27   26  28   29  26   26  26  27   26   26
#> wt    26  26   27  26   26  29   26  27  26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .08  .00  .00  .01  .02
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .02  .01
#> disp  .00  .00       .00  .00  .00  .00  .00  .00  .01  .10
#> hp    .00  .00  .00       .01  .00  .00  .00  .18  .60  .00
#> drat  .00  .00  .00  .01       .00  .87  .01  .00  .00  .60
#> wt    .00  .00  .00  .00  .00       .51  .00  .00  .00  .08
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.89     1                                                     
#> disp  -.82   .92     1                                               
#> hp    -.77   .84   .79     1                                         
#> drat   .70  -.70  -.68  -.46     1                                   
#> wt    -.86   .77   .87   .65  -.71     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    27  26   26  29                                
#> drat  26  27   26  28   29                           
#> wt    26  26   27  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>