(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.86  -.84   .67  -.87   .44   .66   .63   .56  -.56
#> cyl   -.86    1    .91   .83  -.67   .80  -.61  -.79  -.53  -.41   .54
#> disp  -.86   .91    1    .78  -.73   .90  -.44  -.69  -.61  -.53   .39
#> hp    -.84   .83   .78    1   -.50   .68  -.76  -.68  -.30  -.20   .60
#> drat   .67  -.67  -.73  -.50    1   -.68   .10   .35   .66   .65   .00
#> wt    -.87   .80   .90   .68  -.68    1   -.20  -.52  -.72  -.60   .48
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      39.13    -10.15   -741.54   -325.11      2.16     -6.06      5.06
#> cyl     -10.15      3.24    222.29    100.39     -0.66      1.55     -2.01
#> disp   -741.54    222.29  16678.27   6949.04    -47.70    126.00   -100.84
#> hp     -325.11    100.39   6949.04   4514.49    -18.25     49.35    -77.93
#> drat      2.16     -0.66    -47.70    -18.25      0.29     -0.35      0.10
#> wt       -6.06      1.55    126.00     49.35     -0.35      1.00     -0.37
#>             vs        am      gear      carb
#> mpg       1.96      2.02      2.66     -4.82
#> cyl      -0.71     -0.48     -0.56      1.34
#> disp    -43.63    -40.41    -52.70     70.56
#> hp      -22.54    -10.38    -10.10     44.72
#> drat      0.10      0.17      0.26      0.00
#> wt       -0.27     -0.38     -0.46      0.67
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 26 26   26   27
#> cyl   26  29   27 26   26 26   26 26 26   26   26
#> disp  26  27   29 26   26 26   27 26 27   26   26
#> hp    26  26   26 29   26 26   26 26 26   26   26
#> drat  26  26   26 26   29 26   26 26 26   27   26
#> wt    26  26   26 26   26 29   27 26 26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.86 (26) -.84 (26)  .67 (26) -.87 (26)  .44 (26)
#> cyl  -.86 (26)   1  (29)  .91 (27)  .83 (26) -.67 (26)  .80 (26) -.61 (26)
#> disp -.86 (26)  .91 (27)   1  (29)  .78 (26) -.73 (26)  .90 (26) -.44 (27)
#> hp   -.84 (26)  .83 (26)  .78 (26)   1  (29) -.50 (26)  .68 (26) -.76 (26)
#> drat  .67 (26) -.67 (26) -.73 (26) -.50 (26)   1  (29) -.68 (26)  .10 (26)
#> wt   -.87 (26)  .80 (26)  .90 (26)  .68 (26) -.68 (26)   1  (29) -.20 (27)
#>             vs        am      gear      carb
#> mpg   .66 (26)  .63 (26)  .56 (26) -.56 (27)
#> cyl  -.79 (26) -.53 (26) -.41 (26)  .54 (26)
#> disp -.69 (26) -.61 (27) -.53 (26)  .39 (26)
#> hp   -.68 (26) -.30 (26) -.20 (26)  .60 (26)
#> drat  .35 (26)  .66 (26)  .65 (27)  .00 (26)
#> wt   -.52 (26) -.72 (26) -.60 (27)  .48 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.86*  -.84*   .67*  -.87*   .44*   .66*   .63*   .56*
#> cyl   -.86*    1     .91*   .83*  -.67*   .80*  -.61*  -.79*  -.53*  -.41*
#> disp  -.86*   .91*    1     .78*  -.73*   .90*  -.44*  -.69*  -.61*  -.53*
#> hp    -.84*   .83*   .78*    1    -.50*   .68*  -.76*  -.68*  -.30   -.20 
#> drat   .67*  -.67*  -.73*  -.50*    1    -.68*   .10    .35    .66*   .65*
#> wt    -.87*   .80*   .90*   .68*  -.68*    1    -.20   -.52*  -.72*  -.60*
#>        carb
#> mpg   -.56*
#> cyl    .54*
#> disp   .39 
#> hp     .60*
#> drat   .00 
#> wt     .48*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.86* (26) -.84* (26)  .67* (26) -.87* (26)
#> cyl  -.86* (26)   1   (29)  .91* (27)  .83* (26) -.67* (26)  .80* (26)
#> disp -.86* (26)  .91* (27)   1   (29)  .78* (26) -.73* (26)  .90* (26)
#> hp   -.84* (26)  .83* (26)  .78* (26)   1   (29) -.50* (26)  .68* (26)
#> drat  .67* (26) -.67* (26) -.73* (26) -.50* (26)   1   (29) -.68* (26)
#> wt   -.87* (26)  .80* (26)  .90* (26)  .68* (26) -.68* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .44* (26)  .66* (26)  .63* (26)  .56* (26) -.56* (27)
#> cyl  -.61* (26) -.79* (26) -.53* (26) -.41* (26)  .54* (26)
#> disp -.44* (27) -.69* (26) -.61* (27) -.53* (26)  .39  (26)
#> hp   -.76* (26) -.68* (26) -.30  (26) -.20  (26)  .60* (26)
#> drat  .10  (26)  .35  (26)  .66* (26)  .65* (27)  .00  (26)
#> wt   -.20  (27) -.52* (26) -.72* (26) -.60* (27)  .48* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.86  -.84   .67  -.87   .44   .66   .63   .56  -.56
#> N  29    26    26    26    26    26    26    26    26    26    27
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .00   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .91   .83  -.67   .80  -.61  -.79  -.53  -.41   .54
#> N    26  29    27    26    26    26    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .04   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.857* (26)   1    (29)                                                
#> disp -.863* (26)  .914* (27)   1    (29)                                    
#> hp   -.841* (26)  .829* (26)  .779* (26)   1    (29)                        
#> drat  .672* (26) -.670* (26) -.729* (26) -.496  (26)   1    (29)            
#> wt   -.875* (26)  .805* (26)  .896* (26)  .680* (26) -.679* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.86  -.86  -.84   .67  -.87   .44   .66   .63   .56  -.56
#> cyl   -.86     1   .91   .83  -.67   .80  -.61  -.79  -.53  -.41   .54
#> disp  -.86   .91     1   .78  -.73   .90  -.44  -.69  -.61  -.53   .39
#> hp    -.84   .83   .78     1  -.50   .68  -.76  -.68  -.30  -.20   .60
#> drat   .67  -.67  -.73  -.50     1  -.68   .10   .35   .66   .65   .00
#> wt    -.87   .80   .90   .68  -.68     1  -.20  -.52  -.72  -.60   .48
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  26  26   26   27
#> cyl   26  29   27  26   26  26   26  26  26   26   26
#> disp  26  27   29  26   26  26   27  26  27   26   26
#> hp    26  26   26  29   26  26   26  26  26   26   26
#> drat  26  26   26  26   29  26   26  26  26   27   26
#> wt    26  26   26  26   26  29   27  26  26   27   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp  drat   wt qsec   vs   am gear  carb
#> mpg        .00  .00  .00   .00  .00  .03  .00  .00  .00   .00
#> cyl   .00       .00  .00   .00  .00  .00  .00  .01  .04   .00
#> disp  .00  .00       .00   .00  .00  .02  .00  .00  .00   .05
#> hp    .00  .00  .00        .01  .00  .00  .00  .13  .34   .00
#> drat  .00  .00  .00  .01        .00  .61  .08  .00  .00  1.00
#> wt    .00  .00  .00  .00   .00       .33  .01  .00  .00   .01
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.86     1                                                     
#> disp  -.86   .91     1                                               
#> hp    -.84   .83   .78     1                                         
#> drat   .67  -.67  -.73  -.50     1                                   
#> wt    -.87   .80   .90   .68  -.68     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    26  26   26  29                                
#> drat  26  26   26  26   29                           
#> wt    26  26   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp  drat   wt qsec   vs   am gear carb
#> mpg                                                         
#> cyl   .00                                                   
#> disp  .00  .00                                              
#> hp    .00  .00  .00                                         
#> drat  .00  .00  .00  .01                                    
#> wt    .00  .00  .00  .00   .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>