(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.85  -.79   .65  -.82   .43   .65   .55   .42  -.58
#> cyl   -.86    1    .90   .85  -.72   .74  -.58  -.77  -.44  -.46   .53
#> disp  -.85   .90    1    .88  -.72   .88  -.47  -.73  -.66  -.66   .58
#> hp    -.79   .85   .88    1   -.43   .62  -.68  -.71  -.14  -.05   .77
#> drat   .65  -.72  -.72  -.43    1   -.73   .15   .39   .74   .67  -.12
#> wt    -.82   .74   .88   .62  -.73    1   -.21  -.50  -.66  -.51   .46
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      35.62     -9.47   -646.00   -331.29      2.20     -4.05      4.82
#> cyl      -9.47      3.16    206.14    106.87     -0.68      1.15     -1.78
#> disp   -646.00    206.14  16111.15   7026.12    -50.60    106.01   -100.29
#> hp     -331.29    106.87   7026.12   4915.90    -16.50     39.59    -75.32
#> drat      2.20     -0.68    -50.60    -16.50      0.30     -0.31      0.14
#> wt       -4.05      1.15    106.01     39.59     -0.31      0.85     -0.34
#>             vs        am      gear      carb
#> mpg       2.04      1.64      1.98     -6.06
#> cyl      -0.65     -0.39     -0.66      1.60
#> disp    -47.64    -41.83    -58.47     91.34
#> hp      -25.57     -5.05     -2.88     93.86
#> drat      0.11      0.19      0.28     -0.12
#> wt       -0.23     -0.32     -0.36      0.76
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 27 27   26   27
#> cyl   26  29   26 27   26 27   27 26 26   26   26
#> disp  26  26   29 26   26 26   27 27 26   26   26
#> hp    26  27   26 29   26 27   26 26 26   26   26
#> drat  26  26   26 26   29 26   26 26 26   27   27
#> wt    26  27   26 27   26 29   27 26 27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.85 (26) -.79 (26)  .65 (26) -.82 (26)  .43 (26)
#> cyl  -.86 (26)   1  (29)  .90 (26)  .85 (27) -.72 (26)  .74 (27) -.58 (27)
#> disp -.85 (26)  .90 (26)   1  (29)  .88 (26) -.72 (26)  .88 (26) -.47 (27)
#> hp   -.79 (26)  .85 (27)  .88 (26)   1  (29) -.43 (26)  .62 (27) -.68 (26)
#> drat  .65 (26) -.72 (26) -.72 (26) -.43 (26)   1  (29) -.73 (26)  .15 (26)
#> wt   -.82 (26)  .74 (27)  .88 (26)  .62 (27) -.73 (26)   1  (29) -.21 (27)
#>             vs        am      gear      carb
#> mpg   .65 (27)  .55 (27)  .42 (26) -.58 (27)
#> cyl  -.77 (26) -.44 (26) -.46 (26)  .53 (26)
#> disp -.73 (27) -.66 (26) -.66 (26)  .58 (26)
#> hp   -.71 (26) -.14 (26) -.05 (26)  .77 (26)
#> drat  .39 (26)  .74 (26)  .67 (27) -.12 (27)
#> wt   -.50 (26) -.66 (27) -.51 (26)  .46 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.85*  -.79*   .65*  -.82*   .43*   .65*   .55*   .42*
#> cyl   -.86*    1     .90*   .85*  -.72*   .74*  -.58*  -.77*  -.44*  -.46*
#> disp  -.85*   .90*    1     .88*  -.72*   .88*  -.47*  -.73*  -.66*  -.66*
#> hp    -.79*   .85*   .88*    1    -.43*   .62*  -.68*  -.71*  -.14   -.05 
#> drat   .65*  -.72*  -.72*  -.43*    1    -.73*   .15    .39*   .74*   .67*
#> wt    -.82*   .74*   .88*   .62*  -.73*    1    -.21   -.50*  -.66*  -.51*
#>        carb
#> mpg   -.58*
#> cyl    .53*
#> disp   .58*
#> hp     .77*
#> drat  -.12 
#> wt     .46*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.85* (26) -.79* (26)  .65* (26) -.82* (26)
#> cyl  -.86* (26)   1   (29)  .90* (26)  .85* (27) -.72* (26)  .74* (27)
#> disp -.85* (26)  .90* (26)   1   (29)  .88* (26) -.72* (26)  .88* (26)
#> hp   -.79* (26)  .85* (27)  .88* (26)   1   (29) -.43* (26)  .62* (27)
#> drat  .65* (26) -.72* (26) -.72* (26) -.43* (26)   1   (29) -.73* (26)
#> wt   -.82* (26)  .74* (27)  .88* (26)  .62* (27) -.73* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .43* (26)  .65* (27)  .55* (27)  .42* (26) -.58* (27)
#> cyl  -.58* (27) -.77* (26) -.44* (26) -.46* (26)  .53* (26)
#> disp -.47* (27) -.73* (27) -.66* (26) -.66* (26)  .58* (26)
#> hp   -.68* (26) -.71* (26) -.14  (26) -.05  (26)  .77* (26)
#> drat  .15  (26)  .39* (26)  .74* (26)  .67* (27) -.12  (27)
#> wt   -.21  (27) -.50* (26) -.66* (27) -.51* (26)  .46* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.85  -.79   .65  -.82   .43   .65   .55   .42  -.58
#> N  29    26    26    26    26    26    26    27    27    26    27
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .03   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .90   .85  -.72   .74  -.58  -.77  -.44  -.46   .53
#> N    26  29    26    27    26    27    27    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .02   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.865* (26)   1    (29)                                                
#> disp -.852* (26)  .898* (26)   1    (29)                                    
#> hp   -.791* (26)  .848* (27)  .883* (26)   1    (29)                        
#> drat  .654* (26) -.716* (26) -.716* (26) -.431  (26)   1    (29)            
#> wt   -.823* (26)  .736* (27)  .878* (26)  .623* (27) -.731* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.86  -.85  -.79   .65  -.82   .43   .65   .55   .42  -.58
#> cyl   -.86     1   .90   .85  -.72   .74  -.58  -.77  -.44  -.46   .53
#> disp  -.85   .90     1   .88  -.72   .88  -.47  -.73  -.66  -.66   .58
#> hp    -.79   .85   .88     1  -.43   .62  -.68  -.71  -.14  -.05   .77
#> drat   .65  -.72  -.72  -.43     1  -.73   .15   .39   .74   .67  -.12
#> wt    -.82   .74   .88   .62  -.73     1  -.21  -.50  -.66  -.51   .46
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  27  27   26   27
#> cyl   26  29   26  27   26  27   27  26  26   26   26
#> disp  26  26   29  26   26  26   27  27  26   26   26
#> hp    26  27   26  29   26  27   26  26  26   26   26
#> drat  26  26   26  26   29  26   26  26  26   27   27
#> wt    26  27   26  27   26  29   27  26  27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .03  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .02  .00
#> disp  .00  .00       .00  .00  .00  .01  .00  .00  .00  .00
#> hp    .00  .00  .00       .03  .00  .00  .00  .49  .80  .00
#> drat  .00  .00  .00  .03       .00  .47  .05  .00  .00  .54
#> wt    .00  .00  .00  .00  .00       .30  .01  .00  .01  .02
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.86     1                                                     
#> disp  -.85   .90     1                                               
#> hp    -.79   .85   .88     1                                         
#> drat   .65  -.72  -.72  -.43     1                                   
#> wt    -.82   .74   .88   .62  -.73     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  27   26  29                                
#> drat  26  26   26  26   29                           
#> wt    26  27   26  27   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .03                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>