(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.86  -.80   .68  -.85   .46   .63   .67   .43  -.55
#> cyl   -.86    1    .90   .83  -.68   .75  -.67  -.88  -.63  -.41   .59
#> disp  -.86   .90    1    .81  -.71   .85  -.55  -.72  -.66  -.49   .47
#> hp    -.80   .83   .81    1   -.54   .67  -.74  -.70  -.59  -.14   .76
#> drat   .68  -.68  -.71  -.54    1   -.66   .09   .47   .79   .68   .06
#> wt    -.85   .75   .85   .67  -.66    1   -.14  -.41  -.71  -.55   .33
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      39.06    -10.17   -674.69   -330.77      2.09     -4.26      5.56
#> cyl     -10.17      3.24    202.44    103.07     -0.65      1.19     -2.24
#> disp   -674.69    202.44  15868.49   6724.19    -48.39     85.02   -131.55
#> hp     -330.77    103.07   6724.19   4477.40    -18.40     33.31    -92.90
#> drat      2.09     -0.65    -48.39    -18.40      0.27     -0.30      0.08
#> wt       -4.26      1.19     85.02     33.31     -0.30      0.78     -0.24
#>             vs        am      gear      carb
#> mpg       1.74      2.20      2.15     -5.55
#> cyl      -0.79     -0.57     -0.50      1.68
#> disp    -45.06    -42.85    -43.62     94.01
#> hp      -24.12    -17.12     -6.44     85.26
#> drat      0.13      0.21      0.26      0.05
#> wt       -0.17     -0.33     -0.39      0.46
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   27 26   26 26   27 27 26   26   26
#> cyl   26  29   27 26   27 26   27 26 26   26   26
#> disp  27  27   29 27   26 26   26 28 26   26   26
#> hp    26  26   27 29   26 26   26 27 26   26   26
#> drat  26  27   26 26   29 27   27 26 26   27   26
#> wt    26  26   26 26   27 29   26 26 26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.86 (27) -.80 (26)  .68 (26) -.85 (26)  .46 (27)
#> cyl  -.86 (26)   1  (29)  .90 (27)  .83 (26) -.68 (27)  .75 (26) -.67 (27)
#> disp -.86 (27)  .90 (27)   1  (29)  .81 (27) -.71 (26)  .85 (26) -.55 (26)
#> hp   -.80 (26)  .83 (26)  .81 (27)   1  (29) -.54 (26)  .67 (26) -.74 (26)
#> drat  .68 (26) -.68 (27) -.71 (26) -.54 (26)   1  (29) -.66 (27)  .09 (27)
#> wt   -.85 (26)  .75 (26)  .85 (26)  .67 (26) -.66 (27)   1  (29) -.14 (26)
#>             vs        am      gear      carb
#> mpg   .63 (27)  .67 (26)  .43 (26) -.55 (26)
#> cyl  -.88 (26) -.63 (26) -.41 (26)  .59 (26)
#> disp -.72 (28) -.66 (26) -.49 (26)  .47 (26)
#> hp   -.70 (27) -.59 (26) -.14 (26)  .76 (26)
#> drat  .47 (26)  .79 (26)  .68 (27)  .06 (26)
#> wt   -.41 (26) -.71 (26) -.55 (26)  .33 (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.86*  -.80*   .68*  -.85*   .46*   .63*   .67*   .43*
#> cyl   -.86*    1     .90*   .83*  -.68*   .75*  -.67*  -.88*  -.63*  -.41*
#> disp  -.86*   .90*    1     .81*  -.71*   .85*  -.55*  -.72*  -.66*  -.49*
#> hp    -.80*   .83*   .81*    1    -.54*   .67*  -.74*  -.70*  -.59*  -.14 
#> drat   .68*  -.68*  -.71*  -.54*    1    -.66*   .09    .47*   .79*   .68*
#> wt    -.85*   .75*   .85*   .67*  -.66*    1    -.14   -.41*  -.71*  -.55*
#>        carb
#> mpg   -.55*
#> cyl    .59*
#> disp   .47*
#> hp     .76*
#> drat   .06 
#> wt     .33 
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.86* (27) -.80* (26)  .68* (26) -.85* (26)
#> cyl  -.86* (26)   1   (29)  .90* (27)  .83* (26) -.68* (27)  .75* (26)
#> disp -.86* (27)  .90* (27)   1   (29)  .81* (27) -.71* (26)  .85* (26)
#> hp   -.80* (26)  .83* (26)  .81* (27)   1   (29) -.54* (26)  .67* (26)
#> drat  .68* (26) -.68* (27) -.71* (26) -.54* (26)   1   (29) -.66* (27)
#> wt   -.85* (26)  .75* (26)  .85* (26)  .67* (26) -.66* (27)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .46* (27)  .63* (27)  .67* (26)  .43* (26) -.55* (26)
#> cyl  -.67* (27) -.88* (26) -.63* (26) -.41* (26)  .59* (26)
#> disp -.55* (26) -.72* (28) -.66* (26) -.49* (26)  .47* (26)
#> hp   -.74* (26) -.70* (27) -.59* (26) -.14  (26)  .76* (26)
#> drat  .09  (27)  .47* (26)  .79* (26)  .68* (27)  .06  (26)
#> wt   -.14  (26) -.41* (26) -.71* (26) -.55* (26)  .33  (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.86  -.80   .68  -.85   .46   .63   .67   .43  -.55
#> N  29    26    27    26    26    26    27    27    26    26    26
#> P       .00   .00   .00   .00   .00   .02   .00   .00   .03   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .90   .83  -.68   .75  -.67  -.88  -.63  -.41   .59
#> N    26  29    27    26    27    26    27    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .04   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.861* (26)   1    (29)                                                
#> disp -.860* (27)  .898* (27)   1    (29)                                    
#> hp   -.799* (26)  .833* (26)  .809* (27)   1    (29)                        
#> drat  .676* (26) -.679* (27) -.709* (26) -.539* (26)   1    (29)            
#> wt   -.845* (26)  .749* (26)  .850* (26)  .668* (26) -.656* (27)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.86  -.86  -.80   .68  -.85   .46   .63   .67   .43  -.55
#> cyl   -.86     1   .90   .83  -.68   .75  -.67  -.88  -.63  -.41   .59
#> disp  -.86   .90     1   .81  -.71   .85  -.55  -.72  -.66  -.49   .47
#> hp    -.80   .83   .81     1  -.54   .67  -.74  -.70  -.59  -.14   .76
#> drat   .68  -.68  -.71  -.54     1  -.66   .09   .47   .79   .68   .06
#> wt    -.85   .75   .85   .67  -.66     1  -.14  -.41  -.71  -.55   .33
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   27  26   26  26   27  27  26   26   26
#> cyl   26  29   27  26   27  26   27  26  26   26   26
#> disp  27  27   29  27   26  26   26  28  26   26   26
#> hp    26  26   27  29   26  26   26  27  26   26   26
#> drat  26  27   26  26   29  27   27  26  26   27   26
#> wt    26  26   26  26   27  29   26  26  26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .02  .00  .00  .03  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .04  .00
#> disp  .00  .00       .00  .00  .00  .00  .00  .00  .01  .01
#> hp    .00  .00  .00       .00  .00  .00  .00  .00  .49  .00
#> drat  .00  .00  .00  .00       .00  .67  .02  .00  .00  .76
#> wt    .00  .00  .00  .00  .00       .49  .04  .00  .00  .10
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.86     1                                                     
#> disp  -.86   .90     1                                               
#> hp    -.80   .83   .81     1                                         
#> drat   .68  -.68  -.71  -.54     1                                   
#> wt    -.85   .75   .85   .67  -.66     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  27  27   29                                    
#> hp    26  26   27  29                                
#> drat  26  27   26  26   29                           
#> wt    26  26   26  26   27  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .00                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>