(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.83  -.83  -.77   .67  -.89   .35   .64   .58   .43  -.51
#> cyl   -.83    1    .93   .82  -.67   .82  -.60  -.79  -.50  -.42   .48
#> disp  -.83   .93    1    .80  -.70   .88  -.44  -.68  -.61  -.47   .36
#> hp    -.77   .82   .80    1   -.41   .65  -.68  -.72  -.24  -.10   .78
#> drat   .67  -.67  -.70  -.41    1   -.72   .05   .39   .78   .76  -.07
#> wt    -.89   .82   .88   .65  -.72    1   -.13  -.55  -.74  -.58   .41
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      33.12     -7.71   -565.59   -332.41      2.31     -5.62      3.76
#> cyl      -7.71      3.14    174.49    103.14     -0.67      1.06     -1.99
#> disp   -565.59    174.49  14168.12   6580.09    -48.99    101.13    -95.76
#> hp     -332.41    103.14   6580.09   5114.46    -17.36     49.08    -80.18
#> drat      2.31     -0.67    -48.99    -17.36      0.30     -0.43      0.05
#> wt       -5.62      1.06    101.13     49.08     -0.43      1.02     -0.25
#>             vs        am      gear      carb
#> mpg       1.91      1.73      1.70     -4.87
#> cyl      -0.72     -0.44     -0.53      1.42
#> disp    -40.97    -37.58    -39.92     74.35
#> hp      -26.30     -8.62     -5.34     98.02
#> drat      0.11      0.23      0.30     -0.07
#> wt       -0.30     -0.40     -0.40      0.67
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 26 27   26   27
#> cyl   26  29   27 26   26 26   26 26 26   26   27
#> disp  26  27   29 27   26 26   26 27 26   26   27
#> hp    26  26   27 29   26 27   26 27 26   26   26
#> drat  26  26   26 26   29 26   26 26 26   27   26
#> wt    26  26   26 27   26 29   26 26 26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.83 (26) -.83 (26) -.77 (26)  .67 (26) -.89 (26)  .35 (26)
#> cyl  -.83 (26)   1  (29)  .93 (27)  .82 (26) -.67 (26)  .82 (26) -.60 (26)
#> disp -.83 (26)  .93 (27)   1  (29)  .80 (27) -.70 (26)  .88 (26) -.44 (26)
#> hp   -.77 (26)  .82 (26)  .80 (27)   1  (29) -.41 (26)  .65 (27) -.68 (26)
#> drat  .67 (26) -.67 (26) -.70 (26) -.41 (26)   1  (29) -.72 (26)  .05 (26)
#> wt   -.89 (26)  .82 (26)  .88 (26)  .65 (27) -.72 (26)   1  (29) -.13 (26)
#>             vs        am      gear      carb
#> mpg   .64 (26)  .58 (27)  .43 (26) -.51 (27)
#> cyl  -.79 (26) -.50 (26) -.42 (26)  .48 (27)
#> disp -.68 (27) -.61 (26) -.47 (26)  .36 (27)
#> hp   -.72 (27) -.24 (26) -.10 (26)  .78 (26)
#> drat  .39 (26)  .78 (26)  .76 (27) -.07 (26)
#> wt   -.55 (26) -.74 (26) -.58 (26)  .41 (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.83*  -.83*  -.77*   .67*  -.89*   .35    .64*   .58*   .43*
#> cyl   -.83*    1     .93*   .82*  -.67*   .82*  -.60*  -.79*  -.50*  -.42*
#> disp  -.83*   .93*    1     .80*  -.70*   .88*  -.44*  -.68*  -.61*  -.47*
#> hp    -.77*   .82*   .80*    1    -.41*   .65*  -.68*  -.72*  -.24   -.10 
#> drat   .67*  -.67*  -.70*  -.41*    1    -.72*   .05    .39*   .78*   .76*
#> wt    -.89*   .82*   .88*   .65*  -.72*    1    -.13   -.55*  -.74*  -.58*
#>        carb
#> mpg   -.51*
#> cyl    .48*
#> disp   .36 
#> hp     .78*
#> drat  -.07 
#> wt     .41*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.83* (26) -.83* (26) -.77* (26)  .67* (26) -.89* (26)
#> cyl  -.83* (26)   1   (29)  .93* (27)  .82* (26) -.67* (26)  .82* (26)
#> disp -.83* (26)  .93* (27)   1   (29)  .80* (27) -.70* (26)  .88* (26)
#> hp   -.77* (26)  .82* (26)  .80* (27)   1   (29) -.41* (26)  .65* (27)
#> drat  .67* (26) -.67* (26) -.70* (26) -.41* (26)   1   (29) -.72* (26)
#> wt   -.89* (26)  .82* (26)  .88* (26)  .65* (27) -.72* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .35  (26)  .64* (26)  .58* (27)  .43* (26) -.51* (27)
#> cyl  -.60* (26) -.79* (26) -.50* (26) -.42* (26)  .48* (27)
#> disp -.44* (26) -.68* (27) -.61* (26) -.47* (26)  .36  (27)
#> hp   -.68* (26) -.72* (27) -.24  (26) -.10  (26)  .78* (26)
#> drat  .05  (26)  .39* (26)  .78* (26)  .76* (27) -.07  (26)
#> wt   -.13  (26) -.55* (26) -.74* (26) -.58* (26)  .41* (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.83  -.83  -.77   .67  -.89   .35   .64   .58   .43  -.51
#> N  29    26    26    26    26    26    26    26    27    26    27
#> P       .00   .00   .00   .00   .00   .08   .00   .00   .03   .01
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   1   .93   .82  -.67   .82  -.60  -.79  -.50  -.42   .48
#> N    26  29    27    26    26    26    26    26    26    26    27
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .03   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.832* (26)   1    (29)                                                
#> disp -.833* (26)  .931* (27)   1    (29)                                    
#> hp   -.769* (26)  .816* (26)  .795* (27)   1    (29)                        
#> drat  .674* (26) -.668* (26) -.701* (26) -.415  (26)   1    (29)            
#> wt   -.889* (26)  .821* (26)  .884* (26)  .649* (27) -.721* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.83  -.83  -.77   .67  -.89   .35   .64   .58   .43  -.51
#> cyl   -.83     1   .93   .82  -.67   .82  -.60  -.79  -.50  -.42   .48
#> disp  -.83   .93     1   .80  -.70   .88  -.44  -.68  -.61  -.47   .36
#> hp    -.77   .82   .80     1  -.41   .65  -.68  -.72  -.24  -.10   .78
#> drat   .67  -.67  -.70  -.41     1  -.72   .05   .39   .78   .76  -.07
#> wt    -.89   .82   .88   .65  -.72     1  -.13  -.55  -.74  -.58   .41
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  26  27   26   27
#> cyl   26  29   27  26   26  26   26  26  26   26   27
#> disp  26  27   29  27   26  26   26  27  26   26   27
#> hp    26  26   27  29   26  27   26  27  26   26   26
#> drat  26  26   26  26   29  26   26  26  26   27   26
#> wt    26  26   26  27   26  29   26  26  26   26   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .08  .00  .00  .03  .01
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .03  .01
#> disp  .00  .00       .00  .00  .00  .02  .00  .00  .02  .06
#> hp    .00  .00  .00       .04  .00  .00  .00  .24  .62  .00
#> drat  .00  .00  .00  .04       .00  .80  .05  .00  .00  .74
#> wt    .00  .00  .00  .00  .00       .52  .00  .00  .00  .03
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.83     1                                                     
#> disp  -.83   .93     1                                               
#> hp    -.77   .82   .80     1                                         
#> drat   .67  -.67  -.70  -.41     1                                   
#> wt    -.89   .82   .88   .65  -.72     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    26  26   27  29                                
#> drat  26  26   26  26   29                           
#> wt    26  26   26  27   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .04                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>