(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.80  -.77   .63  -.83   .48   .71   .56   .37  -.55
#> cyl   -.86    1    .90   .85  -.72   .77  -.57  -.84  -.56  -.48   .61
#> disp  -.80   .90    1    .79  -.70   .89  -.44  -.73  -.55  -.49   .47
#> hp    -.77   .85   .79    1   -.39   .67  -.74  -.71  -.23  -.10   .79
#> drat   .63  -.72  -.70  -.39    1   -.72   .06   .49   .75   .72  -.12
#> wt    -.83   .77   .89   .67  -.72    1   -.17  -.57  -.67  -.49   .52
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      32.97     -9.41   -498.77   -330.67      1.71     -3.92      5.22
#> cyl      -9.41      3.24    202.22    112.28     -0.62      1.36     -1.90
#> disp   -498.77    202.22  15382.72   6841.46    -42.22    106.93    -99.07
#> hp     -330.67    112.28   6841.46   4918.35    -13.54     44.77    -92.88
#> drat      1.71     -0.62    -42.22    -13.54      0.24     -0.34      0.05
#> wt       -3.92      1.36    106.93     44.77     -0.34      0.94     -0.27
#>             vs        am      gear      carb
#> mpg       2.13      1.61      1.54     -5.17
#> cyl      -0.79     -0.52     -0.60      1.79
#> disp    -45.07    -32.91    -43.37     93.61
#> hp      -25.22     -8.13     -4.92     89.47
#> drat      0.12      0.19      0.25     -0.10
#> wt       -0.27     -0.32     -0.33      0.85
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 26   26 27 28   26   27
#> cyl   26  29   26 27   26 27   27 26 26   26   26
#> disp  26  26   29 26   26 27   26 26 26   26   27
#> hp    27  27   26 29   26 26   27 28 26   26   27
#> drat  26  26   26 26   29 27   27 26 27   26   26
#> wt    26  27   27 26   27 29   26 26 26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.80 (26) -.77 (27)  .63 (26) -.83 (26)  .48 (26)
#> cyl  -.86 (26)   1  (29)  .90 (26)  .85 (27) -.72 (26)  .77 (27) -.57 (27)
#> disp -.80 (26)  .90 (26)   1  (29)  .79 (26) -.70 (26)  .89 (27) -.44 (26)
#> hp   -.77 (27)  .85 (27)  .79 (26)   1  (29) -.39 (26)  .67 (26) -.74 (27)
#> drat  .63 (26) -.72 (26) -.70 (26) -.39 (26)   1  (29) -.72 (27)  .06 (27)
#> wt   -.83 (26)  .77 (27)  .89 (27)  .67 (26) -.72 (27)   1  (29) -.17 (26)
#>             vs        am      gear      carb
#> mpg   .71 (27)  .56 (28)  .37 (26) -.55 (27)
#> cyl  -.84 (26) -.56 (26) -.48 (26)  .61 (26)
#> disp -.73 (26) -.55 (26) -.49 (26)  .47 (27)
#> hp   -.71 (28) -.23 (26) -.10 (26)  .79 (27)
#> drat  .49 (26)  .75 (27)  .72 (26) -.12 (26)
#> wt   -.57 (26) -.67 (26) -.49 (26)  .52 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.80*  -.77*   .63*  -.83*   .48*   .71*   .56*   .37 
#> cyl   -.86*    1     .90*   .85*  -.72*   .77*  -.57*  -.84*  -.56*  -.48*
#> disp  -.80*   .90*    1     .79*  -.70*   .89*  -.44*  -.73*  -.55*  -.49*
#> hp    -.77*   .85*   .79*    1    -.39*   .67*  -.74*  -.71*  -.23   -.10 
#> drat   .63*  -.72*  -.70*  -.39*    1    -.72*   .06    .49*   .75*   .72*
#> wt    -.83*   .77*   .89*   .67*  -.72*    1    -.17   -.57*  -.67*  -.49*
#>        carb
#> mpg   -.55*
#> cyl    .61*
#> disp   .47*
#> hp     .79*
#> drat  -.12 
#> wt     .52*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.80* (26) -.77* (27)  .63* (26) -.83* (26)
#> cyl  -.86* (26)   1   (29)  .90* (26)  .85* (27) -.72* (26)  .77* (27)
#> disp -.80* (26)  .90* (26)   1   (29)  .79* (26) -.70* (26)  .89* (27)
#> hp   -.77* (27)  .85* (27)  .79* (26)   1   (29) -.39* (26)  .67* (26)
#> drat  .63* (26) -.72* (26) -.70* (26) -.39* (26)   1   (29) -.72* (27)
#> wt   -.83* (26)  .77* (27)  .89* (27)  .67* (26) -.72* (27)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .48* (26)  .71* (27)  .56* (28)  .37  (26) -.55* (27)
#> cyl  -.57* (27) -.84* (26) -.56* (26) -.48* (26)  .61* (26)
#> disp -.44* (26) -.73* (26) -.55* (26) -.49* (26)  .47* (27)
#> hp   -.74* (27) -.71* (28) -.23  (26) -.10  (26)  .79* (27)
#> drat  .06  (27)  .49* (26)  .75* (27)  .72* (26) -.12  (26)
#> wt   -.17  (26) -.57* (26) -.67* (26) -.49* (26)  .52* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.80  -.77   .63  -.83   .48   .71   .56   .37  -.55
#> N  29    26    26    27    26    26    26    27    28    26    27
#> P       .00   .00   .00   .00   .00   .01   .00   .00   .07   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .90   .85  -.72   .77  -.57  -.84  -.56  -.48   .61
#> N    26  29    26    27    26    27    27    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .01   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.859* (26)   1    (29)                                                
#> disp -.798* (26)  .897* (26)   1    (29)                                    
#> hp   -.775* (27)  .845* (27)  .795* (26)   1    (29)                        
#> drat  .634* (26) -.719* (26) -.701* (26) -.389  (26)   1    (29)            
#> wt   -.830* (26)  .766* (27)  .887* (27)  .670* (26) -.719* (27)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.86  -.80  -.77   .63  -.83   .48   .71   .56   .37  -.55
#> cyl   -.86     1   .90   .85  -.72   .77  -.57  -.84  -.56  -.48   .61
#> disp  -.80   .90     1   .79  -.70   .89  -.44  -.73  -.55  -.49   .47
#> hp    -.77   .85   .79     1  -.39   .67  -.74  -.71  -.23  -.10   .79
#> drat   .63  -.72  -.70  -.39     1  -.72   .06   .49   .75   .72  -.12
#> wt    -.83   .77   .89   .67  -.72     1  -.17  -.57  -.67  -.49   .52
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  26   26  27  28   26   27
#> cyl   26  29   26  27   26  27   27  26  26   26   26
#> disp  26  26   29  26   26  27   26  26  26   26   27
#> hp    27  27   26  29   26  26   27  28  26   26   27
#> drat  26  26   26  26   29  27   27  26  27   26   26
#> wt    26  27   27  26   27  29   26  26  26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .01  .00  .00  .07  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .01  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .01  .01
#> hp    .00  .00  .00       .05  .00  .00  .00  .26  .63  .00
#> drat  .00  .00  .00  .05       .00  .78  .01  .00  .00  .57
#> wt    .00  .00  .00  .00  .00       .41  .00  .00  .01  .01
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.86     1                                                     
#> disp  -.80   .90     1                                               
#> hp    -.77   .85   .79     1                                         
#> drat   .63  -.72  -.70  -.39     1                                   
#> wt    -.83   .77   .89   .67  -.72     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    27  27   26  29                                
#> drat  26  26   26  26   29                           
#> wt    26  27   27  26   27  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .05                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>