(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.84  -.86  -.77   .68  -.86   .43   .67   .55   .57  -.55
#> cyl   -.84    1    .89   .81  -.64   .75  -.54  -.77  -.41  -.63   .50
#> disp  -.86   .89    1    .77  -.71   .87  -.35  -.64  -.59  -.65   .29
#> hp    -.77   .81   .77    1   -.42   .59  -.79  -.71  -.08  -.34   .77
#> drat   .68  -.64  -.71  -.42    1   -.70  -.05   .34   .65   .73  -.05
#> wt    -.86   .75   .87   .59  -.70    1   -.11  -.52  -.63  -.67   .32
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      39.66     -8.70   -741.07   -370.72      2.42     -4.64      4.56
#> cyl      -8.70      3.06    181.54    100.36     -0.55      1.15     -1.46
#> disp   -741.07    181.54  15655.50   6609.97    -51.88     96.33    -64.66
#> hp     -370.72    100.36   6609.97   4873.02    -16.56     33.93    -91.91
#> drat      2.42     -0.55    -51.88    -16.56      0.29     -0.36     -0.04
#> wt       -4.64      1.15     96.33     33.93     -0.36      0.83     -0.16
#>             vs        am      gear      carb
#> mpg       2.03      1.54      2.70     -5.31
#> cyl      -0.66     -0.34     -0.81      1.45
#> disp    -38.61    -34.96    -59.00     56.28
#> hp      -24.94     -2.92    -14.88     93.08
#> drat      0.10      0.16      0.29     -0.05
#> wt       -0.25     -0.27     -0.41      0.44
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 26   27 26   26 26 26   26   26
#> cyl   27  29   26 26   27 27   27 27 27   26   26
#> disp  26  26   29 26   26 26   27 26 26   27   26
#> hp    26  26   26 29   27 26   26 26 26   27   26
#> drat  27  27   26 27   29 26   26 26 26   26   27
#> wt    26  27   26 26   26 29   26 27 26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.84 (27) -.86 (26) -.77 (26)  .68 (27) -.86 (26)  .43 (26)
#> cyl  -.84 (27)   1  (29)  .89 (26)  .81 (26) -.64 (27)  .75 (27) -.54 (27)
#> disp -.86 (26)  .89 (26)   1  (29)  .77 (26) -.71 (26)  .87 (26) -.35 (27)
#> hp   -.77 (26)  .81 (26)  .77 (26)   1  (29) -.42 (27)  .59 (26) -.79 (26)
#> drat  .68 (27) -.64 (27) -.71 (26) -.42 (27)   1  (29) -.70 (26) -.05 (26)
#> wt   -.86 (26)  .75 (27)  .87 (26)  .59 (26) -.70 (26)   1  (29) -.11 (26)
#>             vs        am      gear      carb
#> mpg   .67 (26)  .55 (26)  .57 (26) -.55 (26)
#> cyl  -.77 (27) -.41 (27) -.63 (26)  .50 (26)
#> disp -.64 (26) -.59 (26) -.65 (27)  .29 (26)
#> hp   -.71 (26) -.08 (26) -.34 (27)  .77 (26)
#> drat  .34 (26)  .65 (26)  .73 (26) -.05 (27)
#> wt   -.52 (27) -.63 (26) -.67 (26)  .32 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.84*  -.86*  -.77*   .68*  -.86*   .43*   .67*   .55*   .57*
#> cyl   -.84*    1     .89*   .81*  -.64*   .75*  -.54*  -.77*  -.41*  -.63*
#> disp  -.86*   .89*    1     .77*  -.71*   .87*  -.35   -.64*  -.59*  -.65*
#> hp    -.77*   .81*   .77*    1    -.42*   .59*  -.79*  -.71*  -.08   -.34 
#> drat   .68*  -.64*  -.71*  -.42*    1    -.70*  -.05    .34    .65*   .73*
#> wt    -.86*   .75*   .87*   .59*  -.70*    1    -.11   -.52*  -.63*  -.67*
#>        carb
#> mpg   -.55*
#> cyl    .50*
#> disp   .29 
#> hp     .77*
#> drat  -.05 
#> wt     .32 
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.84* (27) -.86* (26) -.77* (26)  .68* (27) -.86* (26)
#> cyl  -.84* (27)   1   (29)  .89* (26)  .81* (26) -.64* (27)  .75* (27)
#> disp -.86* (26)  .89* (26)   1   (29)  .77* (26) -.71* (26)  .87* (26)
#> hp   -.77* (26)  .81* (26)  .77* (26)   1   (29) -.42* (27)  .59* (26)
#> drat  .68* (27) -.64* (27) -.71* (26) -.42* (27)   1   (29) -.70* (26)
#> wt   -.86* (26)  .75* (27)  .87* (26)  .59* (26) -.70* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .43* (26)  .67* (26)  .55* (26)  .57* (26) -.55* (26)
#> cyl  -.54* (27) -.77* (27) -.41* (27) -.63* (26)  .50* (26)
#> disp -.35  (27) -.64* (26) -.59* (26) -.65* (27)  .29  (26)
#> hp   -.79* (26) -.71* (26) -.08  (26) -.34  (27)  .77* (26)
#> drat -.05  (26)  .34  (26)  .65* (26)  .73* (26) -.05  (27)
#> wt   -.11  (26) -.52* (27) -.63* (26) -.67* (26)  .32  (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.84  -.86  -.77   .68  -.86   .43   .67   .55   .57  -.55
#> N  29    27    26    26    27    26    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .00   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.84   1   .89   .81  -.64   .75  -.54  -.77  -.41  -.63   .50
#> N    27  29    26    26    27    27    27    27    27    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .04   .00   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.843* (27)   1    (29)                                                
#> disp -.864* (26)  .890* (26)   1    (29)                                    
#> hp   -.773* (26)  .813* (26)  .766* (26)   1    (29)                        
#> drat  .679* (27) -.641* (27) -.707* (26) -.417  (27)   1    (29)            
#> wt   -.861* (26)  .750* (27)  .866* (26)  .591* (26) -.699* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.84  -.86  -.77   .68  -.86   .43   .67   .55   .57  -.55
#> cyl   -.84     1   .89   .81  -.64   .75  -.54  -.77  -.41  -.63   .50
#> disp  -.86   .89     1   .77  -.71   .87  -.35  -.64  -.59  -.65   .29
#> hp    -.77   .81   .77     1  -.42   .59  -.79  -.71  -.08  -.34   .77
#> drat   .68  -.64  -.71  -.42     1  -.70  -.05   .34   .65   .73  -.05
#> wt    -.86   .75   .87   .59  -.70     1  -.11  -.52  -.63  -.67   .32
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  26   27  26   26  26  26   26   26
#> cyl   27  29   26  26   27  27   27  27  27   26   26
#> disp  26  26   29  26   26  26   27  26  26   27   26
#> hp    26  26   26  29   27  26   26  26  26   27   26
#> drat  27  27   26  27   29  26   26  26  26   26   27
#> wt    26  27   26  26   26  29   26  27  26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .00  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .04  .00  .01
#> disp  .00  .00       .00  .00  .00  .08  .00  .00  .00  .15
#> hp    .00  .00  .00       .03  .00  .00  .00  .68  .09  .00
#> drat  .00  .00  .00  .03       .00  .82  .09  .00  .00  .79
#> wt    .00  .00  .00  .00  .00       .58  .01  .00  .00  .11
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.84     1                                                     
#> disp  -.86   .89     1                                               
#> hp    -.77   .81   .77     1                                         
#> drat   .68  -.64  -.71  -.42     1                                   
#> wt    -.86   .75   .87   .59  -.70     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  26   29                                    
#> hp    26  26   26  29                                
#> drat  27  27   26  27   29                           
#> wt    26  27   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .03                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>