(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.85  -.85  -.76   .69  -.87   .47   .68   .65   .51  -.61
#> cyl   -.85    1    .92   .84  -.68   .80  -.63  -.79  -.56  -.43   .62
#> disp  -.85   .92    1    .75  -.68   .88  -.48  -.65  -.62  -.51   .36
#> hp    -.76   .84   .75    1   -.37   .62  -.76  -.70  -.17  -.06   .77
#> drat   .69  -.68  -.68  -.37    1   -.72   .01   .28   .77   .67  -.05
#> wt    -.87   .80   .88   .62  -.72    1   -.20  -.52  -.70  -.58   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      40.04    -10.39   -673.49   -301.36      2.25     -6.04      5.46
#> cyl     -10.39      3.28    205.22    108.08     -0.65      1.50     -2.16
#> disp   -673.49    205.22  14490.38   6145.44    -44.52     99.66   -113.53
#> hp     -301.36    108.08   6145.44   4548.36    -11.84     43.99    -99.21
#> drat      2.25     -0.65    -44.52    -11.84      0.28     -0.42      0.00
#> wt       -6.04      1.50     99.66     43.99     -0.42      1.04     -0.39
#>             vs        am      gear      carb
#> mpg       1.90      2.16      2.29     -5.77
#> cyl      -0.71     -0.52     -0.56      1.78
#> disp    -38.10    -39.62    -45.19     67.99
#> hp      -22.97     -5.74     -3.50     90.70
#> drat      0.07      0.21      0.27     -0.05
#> wt       -0.26     -0.38     -0.43      0.67
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 27   26 27   26 26 26   26   26
#> cyl   27  29   26 26   27 27   26 26 26   26   26
#> disp  26  26   29 26   26 26   26 26 26   27   27
#> hp    27  26   26 29   27 26   27 27 26   27   26
#> drat  26  27   26 27   29 26   26 26 26   27   26
#> wt    27  27   26 26   26 29   26 26 26   27   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (27) -.85 (26) -.76 (27)  .69 (26) -.87 (27)  .47 (26)
#> cyl  -.85 (27)   1  (29)  .92 (26)  .84 (26) -.68 (27)  .80 (27) -.63 (26)
#> disp -.85 (26)  .92 (26)   1  (29)  .75 (26) -.68 (26)  .88 (26) -.48 (26)
#> hp   -.76 (27)  .84 (26)  .75 (26)   1  (29) -.37 (27)  .62 (26) -.76 (27)
#> drat  .69 (26) -.68 (27) -.68 (26) -.37 (27)   1  (29) -.72 (26)  .01 (26)
#> wt   -.87 (27)  .80 (27)  .88 (26)  .62 (26) -.72 (26)   1  (29) -.20 (26)
#>             vs        am      gear      carb
#> mpg   .68 (26)  .65 (26)  .51 (26) -.61 (26)
#> cyl  -.79 (26) -.56 (26) -.43 (26)  .62 (26)
#> disp -.65 (26) -.62 (26) -.51 (27)  .36 (27)
#> hp   -.70 (27) -.17 (26) -.06 (27)  .77 (26)
#> drat  .28 (26)  .77 (26)  .67 (27) -.05 (26)
#> wt   -.52 (26) -.70 (26) -.58 (27)  .49 (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.85*  -.76*   .69*  -.87*   .47*   .68*   .65*   .51*
#> cyl   -.85*    1     .92*   .84*  -.68*   .80*  -.63*  -.79*  -.56*  -.43*
#> disp  -.85*   .92*    1     .75*  -.68*   .88*  -.48*  -.65*  -.62*  -.51*
#> hp    -.76*   .84*   .75*    1    -.37    .62*  -.76*  -.70*  -.17   -.06 
#> drat   .69*  -.68*  -.68*  -.37     1    -.72*   .01    .28    .77*   .67*
#> wt    -.87*   .80*   .88*   .62*  -.72*    1    -.20   -.52*  -.70*  -.58*
#>        carb
#> mpg   -.61*
#> cyl    .62*
#> disp   .36 
#> hp     .77*
#> drat  -.05 
#> wt     .49*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (27) -.85* (26) -.76* (27)  .69* (26) -.87* (27)
#> cyl  -.85* (27)   1   (29)  .92* (26)  .84* (26) -.68* (27)  .80* (27)
#> disp -.85* (26)  .92* (26)   1   (29)  .75* (26) -.68* (26)  .88* (26)
#> hp   -.76* (27)  .84* (26)  .75* (26)   1   (29) -.37  (27)  .62* (26)
#> drat  .69* (26) -.68* (27) -.68* (26) -.37  (27)   1   (29) -.72* (26)
#> wt   -.87* (27)  .80* (27)  .88* (26)  .62* (26) -.72* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .47* (26)  .68* (26)  .65* (26)  .51* (26) -.61* (26)
#> cyl  -.63* (26) -.79* (26) -.56* (26) -.43* (26)  .62* (26)
#> disp -.48* (26) -.65* (26) -.62* (26) -.51* (27)  .36  (27)
#> hp   -.76* (27) -.70* (27) -.17  (26) -.06  (27)  .77* (26)
#> drat  .01  (26)  .28  (26)  .77* (26)  .67* (27) -.05  (26)
#> wt   -.20  (26) -.52* (26) -.70* (26) -.58* (27)  .49* (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.85  -.85  -.76   .69  -.87   .47   .68   .65   .51  -.61
#> N  29    27    26    27    26    27    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .01   .00   .00   .01   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   1   .92   .84  -.68   .80  -.63  -.79  -.56  -.43   .62
#> N    27  29    26    26    27    27    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .03   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.854* (27)   1    (29)                                                
#> disp -.853* (26)  .921* (26)   1    (29)                                    
#> hp   -.763* (27)  .837* (26)  .750* (26)   1    (29)                        
#> drat  .693* (26) -.678* (27) -.684* (26) -.366  (27)   1    (29)            
#> wt   -.875* (27)  .795* (27)  .883* (26)  .619* (26) -.723* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.85  -.85  -.76   .69  -.87   .47   .68   .65   .51  -.61
#> cyl   -.85     1   .92   .84  -.68   .80  -.63  -.79  -.56  -.43   .62
#> disp  -.85   .92     1   .75  -.68   .88  -.48  -.65  -.62  -.51   .36
#> hp    -.76   .84   .75     1  -.37   .62  -.76  -.70  -.17  -.06   .77
#> drat   .69  -.68  -.68  -.37     1  -.72   .01   .28   .77   .67  -.05
#> wt    -.87   .80   .88   .62  -.72     1  -.20  -.52  -.70  -.58   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  27   26  27   26  26  26   26   26
#> cyl   27  29   26  26   27  27   26  26  26   26   26
#> disp  26  26   29  26   26  26   26  26  26   27   27
#> hp    27  26   26  29   27  26   27  27  26   27   26
#> drat  26  27   26  27   29  26   26  26  26   27   26
#> wt    27  27   26  26   26  29   26  26  26   27   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .01  .00  .00  .01  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .03  .00
#> disp  .00  .00       .00  .00  .00  .01  .00  .00  .01  .06
#> hp    .00  .00  .00       .06  .00  .00  .00  .42  .75  .00
#> drat  .00  .00  .00  .06       .00  .98  .16  .00  .00  .80
#> wt    .00  .00  .00  .00  .00       .32  .01  .00  .00  .01
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.85     1                                                     
#> disp  -.85   .92     1                                               
#> hp    -.76   .84   .75     1                                         
#> drat   .69  -.68  -.68  -.37     1                                   
#> wt    -.87   .80   .88   .62  -.72     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  26   29                                    
#> hp    27  26   26  29                                
#> drat  26  27   26  27   29                           
#> wt    27  27   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .06                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>