(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearson's correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.84  -.85  -.77   .68  -.88   .40   .62   .58   .50  -.55
#> cyl   -.84    1    .91   .87  -.70   .80  -.58  -.79  -.57  -.58   .49
#> disp  -.85   .91    1    .78  -.73   .88  -.41  -.70  -.55  -.53   .35
#> hp    -.77   .87   .78    1   -.42   .65  -.71  -.68  -.24  -.12   .77
#> drat   .68  -.70  -.73  -.42    1   -.70   .12   .41   .68   .72  -.13
#> wt    -.88   .80   .88   .65  -.70    1   -.16  -.46  -.69  -.58   .40
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      34.57     -9.56   -670.67   -313.56      2.31     -4.87      4.35
#> cyl      -9.56      3.28    215.67    101.24     -0.74      1.39     -1.89
#> disp   -670.67    215.67  16354.11   7185.42    -54.30    103.38    -97.02
#> hp     -313.56    101.24   7185.42   4770.67    -14.72     41.35    -91.48
#> drat      2.31     -0.74    -54.30    -14.72      0.30     -0.38      0.12
#> wt       -4.87      1.39    103.38     41.35     -0.38      0.86     -0.29
#>             vs        am      gear      carb
#> mpg       1.66      1.57      2.25     -5.24
#> cyl      -0.71     -0.52     -0.78      1.16
#> disp    -42.16    -35.45    -52.86     76.98
#> hp      -23.00     -8.53     -6.35     84.97
#> drat      0.11      0.16      0.30     -0.11
#> wt       -0.18     -0.32     -0.43      0.62
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   27 26   28 26 27   26   26
#> cyl   26  29   27 26   26 26   27 26 27   26   26
#> disp  26  27   29 26   26 26   27 26 27   26   26
#> hp    26  26   26 29   26 26   26 26 27   27   26
#> drat  27  26   26 26   29 26   26 26 26   26   26
#> wt    26  26   26 26   26 29   26 26 26   26   28
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.84 (26) -.85 (26) -.77 (26)  .68 (27) -.88 (26)  .40 (28)
#> cyl  -.84 (26)   1  (29)  .91 (27)  .87 (26) -.70 (26)  .80 (26) -.58 (27)
#> disp -.85 (26)  .91 (27)   1  (29)  .78 (26) -.73 (26)  .88 (26) -.41 (27)
#> hp   -.77 (26)  .87 (26)  .78 (26)   1  (29) -.42 (26)  .65 (26) -.71 (26)
#> drat  .68 (27) -.70 (26) -.73 (26) -.42 (26)   1  (29) -.70 (26)  .12 (26)
#> wt   -.88 (26)  .80 (26)  .88 (26)  .65 (26) -.70 (26)   1  (29) -.16 (26)
#>             vs        am      gear      carb
#> mpg   .62 (26)  .58 (27)  .50 (26) -.55 (26)
#> cyl  -.79 (26) -.57 (27) -.58 (26)  .49 (26)
#> disp -.70 (26) -.55 (27) -.53 (26)  .35 (26)
#> hp   -.68 (26) -.24 (27) -.12 (27)  .77 (26)
#> drat  .41 (26)  .68 (26)  .72 (26) -.13 (26)
#> wt   -.46 (26) -.69 (26) -.58 (26)  .40 (28)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.84*  -.85*  -.77*   .68*  -.88*   .40*   .62*   .58*   .50*
#> cyl   -.84*    1     .91*   .87*  -.70*   .80*  -.58*  -.79*  -.57*  -.58*
#> disp  -.85*   .91*    1     .78*  -.73*   .88*  -.41*  -.70*  -.55*  -.53*
#> hp    -.77*   .87*   .78*    1    -.42*   .65*  -.71*  -.68*  -.24   -.12 
#> drat   .68*  -.70*  -.73*  -.42*    1    -.70*   .12    .41*   .68*   .72*
#> wt    -.88*   .80*   .88*   .65*  -.70*    1    -.16   -.46*  -.69*  -.58*
#>        carb
#> mpg   -.55*
#> cyl    .49*
#> disp   .35 
#> hp     .77*
#> drat  -.13 
#> wt     .40*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.84* (26) -.85* (26) -.77* (26)  .68* (27) -.88* (26)
#> cyl  -.84* (26)   1   (29)  .91* (27)  .87* (26) -.70* (26)  .80* (26)
#> disp -.85* (26)  .91* (27)   1   (29)  .78* (26) -.73* (26)  .88* (26)
#> hp   -.77* (26)  .87* (26)  .78* (26)   1   (29) -.42* (26)  .65* (26)
#> drat  .68* (27) -.70* (26) -.73* (26) -.42* (26)   1   (29) -.70* (26)
#> wt   -.88* (26)  .80* (26)  .88* (26)  .65* (26) -.70* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .40* (28)  .62* (26)  .58* (27)  .50* (26) -.55* (26)
#> cyl  -.58* (27) -.79* (26) -.57* (27) -.58* (26)  .49* (26)
#> disp -.41* (27) -.70* (26) -.55* (27) -.53* (26)  .35  (26)
#> hp   -.71* (26) -.68* (26) -.24  (27) -.12  (27)  .77* (26)
#> drat  .12  (26)  .41* (26)  .68* (26)  .72* (26) -.13  (26)
#> wt   -.16  (26) -.46* (26) -.69* (26) -.58* (26)  .40* (28)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.84  -.85  -.77   .68  -.88   .40   .62   .58   .50  -.55
#> N  29    26    26    26    27    26    28    26    27    26    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .01   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.84   1   .91   .87  -.70   .80  -.58  -.79  -.57  -.58   .49
#> N    26  29    27    26    26    26    27    26    27    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.843* (26)   1    (29)                                                
#> disp -.854* (26)  .914* (27)   1    (29)                                    
#> hp   -.767* (26)  .872* (26)  .783* (26)   1    (29)                        
#> drat  .680* (27) -.704* (26) -.731* (26) -.424  (26)   1    (29)            
#> wt   -.878* (26)  .804* (26)  .879* (26)  .652* (26) -.700* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.84  -.85  -.77   .68  -.88   .40   .62   .58   .50  -.55
#> cyl   -.84     1   .91   .87  -.70   .80  -.58  -.79  -.57  -.58   .49
#> disp  -.85   .91     1   .78  -.73   .88  -.41  -.70  -.55  -.53   .35
#> hp    -.77   .87   .78     1  -.42   .65  -.71  -.68  -.24  -.12   .77
#> drat   .68  -.70  -.73  -.42     1  -.70   .12   .41   .68   .72  -.13
#> wt    -.88   .80   .88   .65  -.70     1  -.16  -.46  -.69  -.58   .40
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   27  26   28  26  27   26   26
#> cyl   26  29   27  26   26  26   27  26  27   26   26
#> disp  26  27   29  26   26  26   27  26  27   26   26
#> hp    26  26   26  29   26  26   26  26  27   27   26
#> drat  27  26   26  26   29  26   26  26  26   26   26
#> wt    26  26   26  26   26  29   26  26  26   26   28
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .01  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .01
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .01  .08
#> hp    .00  .00  .00       .03  .00  .00  .00  .23  .56  .00
#> drat  .00  .00  .00  .03       .00  .56  .04  .00  .00  .54
#> wt    .00  .00  .00  .00  .00       .43  .02  .00  .00  .04
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.84     1                                                     
#> disp  -.85   .91     1                                               
#> hp    -.77   .87   .78     1                                         
#> drat   .68  -.70  -.73  -.42     1                                   
#> wt    -.88   .80   .88   .65  -.70     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    26  26   26  29                                
#> drat  27  26   26  26   29                           
#> wt    26  26   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .03                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>