(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Usage

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = .op[["digits"]], sig.level = 0.05,
      show = c("all","lower.tri","upper.tri"), spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.87  -.86  -.78   .66  -.88   .35   .67   .55   .41  -.58
#> cyl   -.87    1    .89   .81  -.65   .76  -.58  -.79  -.44  -.43   .48
#> disp  -.86   .89    1    .77  -.73   .88  -.39  -.66  -.53  -.43   .37
#> hp    -.78   .81   .77    1   -.44   .67  -.78  -.74  -.21  -.02   .77
#> drat   .66  -.65  -.73  -.44    1   -.71   .07   .48   .71   .68  -.12
#> wt    -.88   .76   .88   .67  -.71    1   -.14  -.56  -.67  -.55   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.96     -8.68   -638.66   -362.01      2.25     -5.88      3.79
#> cyl      -8.68      3.16    189.37     99.94     -0.54      1.37     -1.89
#> disp   -638.66    189.37  14449.29   6185.42    -47.12    104.71    -82.19
#> hp     -362.01     99.94   6185.42   4938.49    -14.33     47.12    -82.27
#> drat      2.25     -0.54    -47.12    -14.33      0.29     -0.38      0.06
#> wt       -5.88      1.37    104.71     47.12     -0.38      1.04     -0.26
#>             vs        am      gear      carb
#> mpg       2.09      1.71      1.86     -5.96
#> cyl      -0.68     -0.39     -0.60      1.38
#> disp    -39.95    -32.46    -35.26     77.10
#> hp      -27.62     -7.62     -1.28     92.27
#> drat      0.13      0.19      0.25     -0.08
#> wt       -0.29     -0.34     -0.37      0.81
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 27   26 27 26   26   26
#> cyl   26  29   26 26   26 26   27 27 26   26   27
#> disp  26  26   29 26   27 26   26 26 26   26   26
#> hp    26  26   26 29   26 26   26 26 26   26   26
#> drat  26  26   27 26   29 26   26 26 26   26   26
#> wt    27  26   26 26   26 29   28 26 27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.87 (26) -.86 (26) -.78 (26)  .66 (26) -.88 (27)  .35 (26)
#> cyl  -.87 (26)   1  (29)  .89 (26)  .81 (26) -.65 (26)  .76 (26) -.58 (27)
#> disp -.86 (26)  .89 (26)   1  (29)  .77 (26) -.73 (27)  .88 (26) -.39 (26)
#> hp   -.78 (26)  .81 (26)  .77 (26)   1  (29) -.44 (26)  .67 (26) -.78 (26)
#> drat  .66 (26) -.65 (26) -.73 (27) -.44 (26)   1  (29) -.71 (26)  .07 (26)
#> wt   -.88 (27)  .76 (26)  .88 (26)  .67 (26) -.71 (26)   1  (29) -.14 (28)
#>             vs        am      gear      carb
#> mpg   .67 (27)  .55 (26)  .41 (26) -.58 (26)
#> cyl  -.79 (27) -.44 (26) -.43 (26)  .48 (27)
#> disp -.66 (26) -.53 (26) -.43 (26)  .37 (26)
#> hp   -.74 (26) -.21 (26) -.02 (26)  .77 (26)
#> drat  .48 (26)  .71 (26)  .68 (26) -.12 (26)
#> wt   -.56 (26) -.67 (27) -.55 (26)  .49 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.87*  -.86*  -.78*   .66*  -.88*   .35    .67*   .55*   .41*
#> cyl   -.87*    1     .89*   .81*  -.65*   .76*  -.58*  -.79*  -.44*  -.43*
#> disp  -.86*   .89*    1     .77*  -.73*   .88*  -.39*  -.66*  -.53*  -.43*
#> hp    -.78*   .81*   .77*    1    -.44*   .67*  -.78*  -.74*  -.21   -.02 
#> drat   .66*  -.65*  -.73*  -.44*    1    -.71*   .07    .48*   .71*   .68*
#> wt    -.88*   .76*   .88*   .67*  -.71*    1    -.14   -.56*  -.67*  -.55*
#>        carb
#> mpg   -.58*
#> cyl    .48*
#> disp   .37 
#> hp     .77*
#> drat  -.12 
#> wt     .49*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.87* (26) -.86* (26) -.78* (26)  .66* (26) -.88* (27)
#> cyl  -.87* (26)   1   (29)  .89* (26)  .81* (26) -.65* (26)  .76* (26)
#> disp -.86* (26)  .89* (26)   1   (29)  .77* (26) -.73* (27)  .88* (26)
#> hp   -.78* (26)  .81* (26)  .77* (26)   1   (29) -.44* (26)  .67* (26)
#> drat  .66* (26) -.65* (26) -.73* (27) -.44* (26)   1   (29) -.71* (26)
#> wt   -.88* (27)  .76* (26)  .88* (26)  .67* (26) -.71* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .35  (26)  .67* (27)  .55* (26)  .41* (26) -.58* (26)
#> cyl  -.58* (27) -.79* (27) -.44* (26) -.43* (26)  .48* (27)
#> disp -.39* (26) -.66* (26) -.53* (26) -.43* (26)  .37  (26)
#> hp   -.78* (26) -.74* (26) -.21  (26) -.02  (26)  .77* (26)
#> drat  .07  (26)  .48* (26)  .71* (26)  .68* (26) -.12  (26)
#> wt   -.14  (28) -.56* (26) -.67* (27) -.55* (26)  .49* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.87  -.86  -.78   .66  -.88   .35   .67   .55   .41  -.58
#> N  29    26    26    26    26    27    26    27    26    26    26
#> P       .00   .00   .00   .00   .00   .08   .00   .00   .04   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.87   1   .89   .81  -.65   .76  -.58  -.79  -.44  -.43   .48
#> N    26  29    26    26    26    26    27    27    26    26    27
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .03   .01
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.869* (26)   1    (29)                                                
#> disp -.859* (26)  .889* (26)   1    (29)                                    
#> hp   -.782* (26)  .805* (26)  .775* (26)   1    (29)                        
#> drat  .661* (26) -.653* (26) -.727* (27) -.436  (26)   1    (29)            
#> wt   -.882* (27)  .761* (26)  .880* (26)  .671* (26) -.709* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.87  -.86  -.78   .66  -.88   .35   .67   .55   .41  -.58
#> cyl   -.87     1   .89   .81  -.65   .76  -.58  -.79  -.44  -.43   .48
#> disp  -.86   .89     1   .77  -.73   .88  -.39  -.66  -.53  -.43   .37
#> hp    -.78   .81   .77     1  -.44   .67  -.78  -.74  -.21  -.02   .77
#> drat   .66  -.65  -.73  -.44     1  -.71   .07   .48   .71   .68  -.12
#> wt    -.88   .76   .88   .67  -.71     1  -.14  -.56  -.67  -.55   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  27   26  27  26   26   26
#> cyl   26  29   26  26   26  26   27  27  26   26   27
#> disp  26  26   29  26   27  26   26  26  26   26   26
#> hp    26  26   26  29   26  26   26  26  26   26   26
#> drat  26  26   27  26   29  26   26  26  26   26   26
#> wt    27  26   26  26   26  29   28  26  27   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .08  .00  .00  .04  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .03  .01
#> disp  .00  .00       .00  .00  .00  .05  .00  .01  .03  .06
#> hp    .00  .00  .00       .03  .00  .00  .00  .30  .91  .00
#> drat  .00  .00  .00  .03       .00  .75  .01  .00  .00  .56
#> wt    .00  .00  .00  .00  .00       .46  .00  .00  .00  .01
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.87     1                                                     
#> disp  -.86   .89     1                                               
#> hp    -.78   .81   .77     1                                         
#> drat   .66  -.65  -.73  -.44     1                                   
#> wt    -.88   .76   .88   .67  -.71     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  26   26  29                                
#> drat  26  26   27  26   29                           
#> wt    27  26   26  26   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .03                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>