(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

w
numeric. A vector of (frequency) weights.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return
logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.85  -.86  -.77   .64  -.88   .35   .66   .59   .48  -.50
#> cyl   -.85    1    .90   .83  -.65   .81  -.63  -.93  -.46  -.40   .52
#> disp  -.86   .90    1    .78  -.73   .91  -.41  -.78  -.64  -.58   .36
#> hp    -.77   .83   .78    1   -.44   .73  -.74  -.74  -.31  -.08   .77
#> drat   .64  -.65  -.73  -.44    1   -.73   .02   .40   .71   .67  -.10
#> wt    -.88   .81   .91   .73  -.73    1   -.18  -.62  -.69  -.62   .42
#> qsec   .35  -.63  -.41  -.74   .02  -.18    1    .72  -.31  -.41  -.67
#> vs     .66  -.93  -.78  -.74   .40  -.62   .72    1    .22   .23  -.57
#> am     .59  -.46  -.64  -.31   .71  -.69  -.31   .22    1    .84   .17
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      32.45     -8.68   -628.39   -311.05      2.06     -5.36      3.22
#> cyl      -8.68      3.24    203.95    102.35     -0.62      1.40     -2.18
#> disp   -628.39    203.95  14993.56   6926.67    -49.06    108.14    -94.05
#> hp     -311.05    102.35   6926.67   4830.78    -17.03     52.98    -79.51
#> drat      2.06     -0.62    -49.06    -17.03      0.29     -0.42      0.02
#> wt       -5.36      1.40    108.14     52.98     -0.42      1.05     -0.33
#> qsec      3.22     -2.18    -94.05    -79.51      0.02     -0.33      3.31
#> vs        2.00     -0.89    -49.94    -26.97      0.11     -0.34      0.66
#> am        1.75     -0.41    -41.31    -11.44      0.20     -0.37     -0.30
#>             vs        am      gear      carb
#> mpg       2.00      1.75      2.14     -4.98
#> cyl      -0.89     -0.41     -0.53      1.60
#> disp    -49.94    -41.31    -53.93     72.70
#> hp      -26.97    -11.44     -4.25     90.34
#> drat      0.11      0.20      0.29     -0.09
#> wt       -0.34     -0.37     -0.47      0.74
#> qsec      0.66     -0.30     -0.52     -2.05
#> vs        0.26      0.06      0.09     -0.46
#> am        0.06      0.25      0.34      0.15
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwnobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 26   26 26 27   27   26
#> cyl   26  29   27 26   26 26   26 26 26   26   26
#> disp  26  27   29 26   26 26   26 26 26   26   26
#> hp    27  26   26 29   26 26   27 27 26   27   26
#> drat  26  26   26 26   29 27   27 26 26   26   27
#> wt    26  26   26 26   27 29   27 26 26   27   28
#> qsec  26  26   26 27   27 27   29 28 26   26   27
#> vs    26  26   26 27   26 26   28 29 26   26   26
#> am    27  26   26 26   26 26   26 26 29   26   26
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (26) -.86 (26) -.77 (27)  .64 (26) -.88 (26)  .35 (26)
#> cyl  -.85 (26)   1  (29)  .90 (27)  .83 (26) -.65 (26)  .81 (26) -.63 (26)
#> disp -.86 (26)  .90 (27)   1  (29)  .78 (26) -.73 (26)  .91 (26) -.41 (26)
#> hp   -.77 (27)  .83 (26)  .78 (26)   1  (29) -.44 (26)  .73 (26) -.74 (27)
#> drat  .64 (26) -.65 (26) -.73 (26) -.44 (26)   1  (29) -.73 (27)  .02 (27)
#> wt   -.88 (26)  .81 (26)  .91 (26)  .73 (26) -.73 (27)   1  (29) -.18 (27)
#> qsec  .35 (26) -.63 (26) -.41 (26) -.74 (27)  .02 (27) -.18 (27)   1  (29)
#> vs    .66 (26) -.93 (26) -.78 (26) -.74 (27)  .40 (26) -.62 (26)  .72 (28)
#> am    .59 (27) -.46 (26) -.64 (26) -.31 (26)  .71 (26) -.69 (26) -.31 (26)
#>             vs        am      gear      carb
#> mpg   .66 (26)  .59 (27)  .48 (27) -.50 (26)
#> cyl  -.93 (26) -.46 (26) -.40 (26)  .52 (26)
#> disp -.78 (26) -.64 (26) -.58 (26)  .36 (26)
#> hp   -.74 (27) -.31 (26) -.08 (27)  .77 (26)
#> drat  .40 (26)  .71 (26)  .67 (26) -.10 (27)
#> wt   -.62 (26) -.69 (26) -.62 (27)  .42 (28)
#> qsec  .72 (28) -.31 (26) -.41 (26) -.67 (27)
#> vs     1  (29)  .22 (26)  .23 (26) -.57 (26)
#> am    .22 (26)   1  (29)  .84 (26)  .17 (26)
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.86*  -.77*   .64*  -.88*   .35    .66*   .59*   .48*
#> cyl   -.85*    1     .90*   .83*  -.65*   .81*  -.63*  -.93*  -.46*  -.40*
#> disp  -.86*   .90*    1     .78*  -.73*   .91*  -.41*  -.78*  -.64*  -.58*
#> hp    -.77*   .83*   .78*    1    -.44*   .73*  -.74*  -.74*  -.31   -.08 
#> drat   .64*  -.65*  -.73*  -.44*    1    -.73*   .02    .40*   .71*   .67*
#> wt    -.88*   .81*   .91*   .73*  -.73*    1    -.18   -.62*  -.69*  -.62*
#> qsec   .35   -.63*  -.41*  -.74*   .02   -.18     1     .72*  -.31   -.41*
#> vs     .66*  -.93*  -.78*  -.74*   .40*  -.62*   .72*    1     .22    .23 
#> am     .59*  -.46*  -.64*  -.31    .71*  -.69*  -.31    .22     1     .84*
#>        carb
#> mpg   -.50*
#> cyl    .52*
#> disp   .36 
#> hp     .77*
#> drat  -.10 
#> wt     .42*
#> qsec  -.67*
#> vs    -.57*
#> am     .17 
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (26) -.86* (26) -.77* (27)  .64* (26) -.88* (26)
#> cyl  -.85* (26)   1   (29)  .90* (27)  .83* (26) -.65* (26)  .81* (26)
#> disp -.86* (26)  .90* (27)   1   (29)  .78* (26) -.73* (26)  .91* (26)
#> hp   -.77* (27)  .83* (26)  .78* (26)   1   (29) -.44* (26)  .73* (26)
#> drat  .64* (26) -.65* (26) -.73* (26) -.44* (26)   1   (29) -.73* (27)
#> wt   -.88* (26)  .81* (26)  .91* (26)  .73* (26) -.73* (27)   1   (29)
#> qsec  .35  (26) -.63* (26) -.41* (26) -.74* (27)  .02  (27) -.18  (27)
#> vs    .66* (26) -.93* (26) -.78* (26) -.74* (27)  .40* (26) -.62* (26)
#> am    .59* (27) -.46* (26) -.64* (26) -.31  (26)  .71* (26) -.69* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .35  (26)  .66* (26)  .59* (27)  .48* (27) -.50* (26)
#> cyl  -.63* (26) -.93* (26) -.46* (26) -.40* (26)  .52* (26)
#> disp -.41* (26) -.78* (26) -.64* (26) -.58* (26)  .36  (26)
#> hp   -.74* (27) -.74* (27) -.31  (26) -.08  (27)  .77* (26)
#> drat  .02  (27)  .40* (26)  .71* (26)  .67* (26) -.10  (27)
#> wt   -.18  (27) -.62* (26) -.69* (26) -.62* (27)  .42* (28)
#> qsec   1   (29)  .72* (28) -.31  (26) -.41* (26) -.67* (27)
#> vs    .72* (28)   1   (29)  .22  (26)  .23  (26) -.57* (26)
#> am   -.31  (26)  .22  (26)   1   (29)  .84* (26)  .17  (26)
#>  [ reached getOption("max.print") -- omitted 2 rows ]

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.85  -.86  -.77   .64  -.88   .35   .66   .59   .48  -.50
#> N  29    26    26    27    26    26    26    26    27    27    26
#> P       .00   .00   .00   .00   .00   .08   .00   .00   .01   .01
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   1   .90   .83  -.65   .81  -.63  -.93  -.46  -.40   .52
#> N    26  29    27    26    26    26    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .04   .01
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   .90    1   .78  -.73   .91  -.41  -.78  -.64  -.58   .36
#> N    26    27   29    26    26    26    26    26    26    26    26
#> P   .00   .00        .00   .00   .00   .04   .00   .00   .00   .07
#> 
#>  [ reached getOption("max.print") -- omitted 8 matrix slice(s) ]

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.849* (26)   1    (29)                                                
#> disp -.856* (26)  .903* (27)   1    (29)                                    
#> hp   -.770* (27)  .826* (26)  .780* (26)   1    (29)                        
#> drat  .644* (26) -.645* (26) -.731* (26) -.443  (26)   1    (29)            
#> wt   -.882* (26)  .807* (26)  .912* (26)  .733* (26) -.732* (27)   1    (29)
#> qsec  .345  (26) -.634* (26) -.414  (26) -.739* (27)  .023  (27) -.178  (27)
#> vs    .660* (26) -.932* (26) -.776* (26) -.745* (27)  .396  (26) -.617* (26)
#> am    .593* (27) -.457  (26) -.643* (26) -.312  (26)  .713* (26) -.694* (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .720* (28)   1    (29)                                    
#> am   -.314  (26)  .220  (26)   1    (29)                        
#>  [ reached getOption("max.print") -- omitted 2 rows ]

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.85  -.86  -.77   .64  -.88   .35   .66   .59   .48  -.50
#> cyl   -.85     1   .90   .83  -.65   .81  -.63  -.93  -.46  -.40   .52
#> disp  -.86   .90     1   .78  -.73   .91  -.41  -.78  -.64  -.58   .36
#> hp    -.77   .83   .78     1  -.44   .73  -.74  -.74  -.31  -.08   .77
#> drat   .64  -.65  -.73  -.44     1  -.73   .02   .40   .71   .67  -.10
#> wt    -.88   .81   .91   .73  -.73     1  -.18  -.62  -.69  -.62   .42
#> qsec   .35  -.63  -.41  -.74   .02  -.18     1   .72  -.31  -.41  -.67
#> vs     .66  -.93  -.78  -.74   .40  -.62   .72     1   .22   .23  -.57
#> am     .59  -.46  -.64  -.31   .71  -.69  -.31   .22     1   .84   .17
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  26   26  26  27   27   26
#> cyl   26  29   27  26   26  26   26  26  26   26   26
#> disp  26  27   29  26   26  26   26  26  26   26   26
#> hp    27  26   26  29   26  26   27  27  26   27   26
#> drat  26  26   26  26   29  27   27  26  26   26   27
#> wt    26  26   26  26   27  29   27  26  26   27   28
#> qsec  26  26   26  27   27  27   29  28  26   26   27
#> vs    26  26   26  27   26  26   28  29  26   26   26
#> am    27  26   26  26   26  26   26  26  29   26   26
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .08  .00  .00  .01  .01
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .04  .01
#> disp  .00  .00       .00  .00  .00  .04  .00  .00  .00  .07
#> hp    .00  .00  .00       .02  .00  .00  .00  .12  .70  .00
#> drat  .00  .00  .00  .02       .00  .91  .05  .00  .00  .62
#> wt    .00  .00  .00  .00  .00       .37  .00  .00  .00  .02
#> qsec  .08  .00  .04  .00  .91  .37       .00  .12  .04  .00
#> vs    .00  .00  .00  .00  .05  .00  .00       .28  .26  .00
#> am    .00  .02  .00  .12  .00  .00  .12  .28       .00  .40
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.85     1                                                     
#> disp  -.86   .90     1                                               
#> hp    -.77   .83   .78     1                                         
#> drat   .64  -.65  -.73  -.44     1                                   
#> wt    -.88   .81   .91   .73  -.73     1                             
#> qsec   .35  -.63  -.41  -.74   .02  -.18     1                       
#> vs     .66  -.93  -.78  -.74   .40  -.62   .72     1                 
#> am     .59  -.46  -.64  -.31   .71  -.69  -.31   .22     1           
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    27  26   26  29                                
#> drat  26  26   26  26   29                           
#> wt    26  26   26  26   27  29                       
#> qsec  26  26   26  27   27  27   29                  
#> vs    26  26   26  27   26  26   28  29              
#> am    27  26   26  26   26  26   26  26  29          
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .02                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .08  .00  .04  .00  .91  .37                         
#> vs    .00  .00  .00  .00  .05  .00  .00                    
#> am    .00  .02  .00  .12  .00  .00  .12  .28               
#>  [ reached getOption("max.print") -- omitted 2 rows ]
#>