(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.84  -.76   .69  -.88   .44   .66   .62   .46  -.58
#> cyl   -.86    1    .91   .89  -.67   .80  -.64  -.81  -.61  -.62   .69
#> disp  -.84   .91    1    .79  -.71   .89  -.42  -.72  -.64  -.55   .40
#> hp    -.76   .89   .79    1   -.39   .65  -.72  -.69  -.18  -.04   .81
#> drat   .69  -.67  -.71  -.39    1   -.72   .07   .43   .75   .66  -.11
#> wt    -.88   .80   .89   .65  -.72    1   -.18  -.54  -.69  -.59   .44
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.23     -9.80   -641.84   -334.74      2.39     -5.94      4.33
#> cyl      -9.80      3.28    214.52     96.36     -0.66      1.49     -1.98
#> disp   -641.84    214.52  15855.94   6717.84    -46.72    120.84    -97.25
#> hp     -334.74     96.36   6717.84   4772.52    -15.23     50.32    -95.58
#> drat      2.39     -0.66    -46.72    -15.23      0.29     -0.37      0.07
#> wt       -5.94      1.49    120.84     50.32     -0.37      1.03     -0.35
#>             vs        am      gear      carb
#> mpg       2.05      1.98      2.26     -6.32
#> cyl      -0.76     -0.54     -0.78      1.56
#> disp    -43.03    -41.28    -55.19     88.12
#> hp      -24.92     -6.64     -2.05    101.71
#> drat      0.12      0.21      0.27     -0.10
#> wt       -0.26     -0.36     -0.47      0.78
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   26 26   26 26 27   26   26
#> cyl   26  29   26 26   26 27   27 26 26   27   26
#> disp  26  26   29 26   26 26   26 26 26   26   26
#> hp    26  26   26 29   26 26   27 26 26   26   26
#> drat  26  26   26 26   29 27   26 26 26   26   27
#> wt    26  27   26 26   27 29   27 27 27   27   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.84 (26) -.76 (26)  .69 (26) -.88 (26)  .44 (26)
#> cyl  -.86 (26)   1  (29)  .91 (26)  .89 (26) -.67 (26)  .80 (27) -.64 (27)
#> disp -.84 (26)  .91 (26)   1  (29)  .79 (26) -.71 (26)  .89 (26) -.42 (26)
#> hp   -.76 (26)  .89 (26)  .79 (26)   1  (29) -.39 (26)  .65 (26) -.72 (27)
#> drat  .69 (26) -.67 (26) -.71 (26) -.39 (26)   1  (29) -.72 (27)  .07 (26)
#> wt   -.88 (26)  .80 (27)  .89 (26)  .65 (26) -.72 (27)   1  (29) -.18 (27)
#>             vs        am      gear      carb
#> mpg   .66 (26)  .62 (27)  .46 (26) -.58 (26)
#> cyl  -.81 (26) -.61 (26) -.62 (27)  .69 (26)
#> disp -.72 (26) -.64 (26) -.55 (26)  .40 (26)
#> hp   -.69 (26) -.18 (26) -.04 (26)  .81 (26)
#> drat  .43 (26)  .75 (26)  .66 (26) -.11 (27)
#> wt   -.54 (27) -.69 (27) -.59 (27)  .44 (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.84*  -.76*   .69*  -.88*   .44*   .66*   .62*   .46*
#> cyl   -.86*    1     .91*   .89*  -.67*   .80*  -.64*  -.81*  -.61*  -.62*
#> disp  -.84*   .91*    1     .79*  -.71*   .89*  -.42*  -.72*  -.64*  -.55*
#> hp    -.76*   .89*   .79*    1    -.39*   .65*  -.72*  -.69*  -.18   -.04 
#> drat   .69*  -.67*  -.71*  -.39*    1    -.72*   .07    .43*   .75*   .66*
#> wt    -.88*   .80*   .89*   .65*  -.72*    1    -.18   -.54*  -.69*  -.59*
#>        carb
#> mpg   -.58*
#> cyl    .69*
#> disp   .40*
#> hp     .81*
#> drat  -.11 
#> wt     .44*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.84* (26) -.76* (26)  .69* (26) -.88* (26)
#> cyl  -.86* (26)   1   (29)  .91* (26)  .89* (26) -.67* (26)  .80* (27)
#> disp -.84* (26)  .91* (26)   1   (29)  .79* (26) -.71* (26)  .89* (26)
#> hp   -.76* (26)  .89* (26)  .79* (26)   1   (29) -.39* (26)  .65* (26)
#> drat  .69* (26) -.67* (26) -.71* (26) -.39* (26)   1   (29) -.72* (27)
#> wt   -.88* (26)  .80* (27)  .89* (26)  .65* (26) -.72* (27)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .44* (26)  .66* (26)  .62* (27)  .46* (26) -.58* (26)
#> cyl  -.64* (27) -.81* (26) -.61* (26) -.62* (27)  .69* (26)
#> disp -.42* (26) -.72* (26) -.64* (26) -.55* (26)  .40* (26)
#> hp   -.72* (27) -.69* (26) -.18  (26) -.04  (26)  .81* (26)
#> drat  .07  (26)  .43* (26)  .75* (26)  .66* (26) -.11  (27)
#> wt   -.18  (27) -.54* (27) -.69* (27) -.59* (27)  .44* (27)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.84  -.76   .69  -.88   .44   .66   .62   .46  -.58
#> N  29    26    26    26    26    26    26    26    27    26    26
#> P       .00   .00   .00   .00   .00   .02   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .91   .89  -.67   .80  -.64  -.81  -.61  -.62   .69
#> N    26  29    26    26    26    27    27    26    26    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.863* (26)   1    (29)                                                
#> disp -.842* (26)  .905* (26)   1    (29)                                    
#> hp   -.757* (26)  .888* (26)  .787* (26)   1    (29)                        
#> drat  .694* (26) -.670* (26) -.711* (26) -.392  (26)   1    (29)            
#> wt   -.884* (26)  .795* (27)  .891* (26)  .655* (26) -.720* (27)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl   disp    hp   drat    wt  qsec    vs    am  gear   carb
#> mpg      1  -.86   -.84  -.76    .69  -.88   .44   .66   .62   .46   -.58
#> cyl   -.86     1    .91   .89   -.67   .80  -.64  -.81  -.61  -.62    .69
#> disp  -.84   .91   1.00   .79   -.71   .89  -.42  -.72  -.64  -.55    .40
#> hp    -.76   .89    .79     1   -.39   .65  -.72  -.69  -.18  -.04    .81
#> drat   .69  -.67   -.71  -.39   1.00  -.72   .07   .43   .75   .66   -.11
#> wt    -.88   .80    .89   .65   -.72     1  -.18  -.54  -.69  -.59    .44
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   26  26   26  26  27   26   26
#> cyl   26  29   26  26   26  27   27  26  26   27   26
#> disp  26  26   29  26   26  26   26  26  26   26   26
#> hp    26  26   26  29   26  26   27  26  26   26   26
#> drat  26  26   26  26   29  27   26  26  26   26   27
#> wt    26  27   26  26   27  29   27  27  27   27   27
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .02  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .04
#> hp    .00  .00  .00       .05  .00  .00  .00  .37  .86  .00
#> drat  .00  .00  .00  .05       .00  .72  .03  .00  .00  .60
#> wt    .00  .00  .00  .00  .00       .37  .00  .00  .00  .02
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl   disp    hp   drat    wt  qsec    vs    am  gear   carb
#> mpg      1                                                               
#> cyl   -.86     1                                                         
#> disp  -.84   .91   1.00                                                  
#> hp    -.76   .89    .79     1                                            
#> drat   .69  -.67   -.71  -.39   1.00                                     
#> wt    -.88   .80    .89   .65   -.72     1                               
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  26   26  29                                
#> drat  26  26   26  26   29                           
#> wt    26  27   26  26   27  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .05                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>