(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Notes: weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran)

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.84  -.83  -.77   .71  -.85   .39   .68   .65   .64  -.61
#> cyl   -.84    1    .92   .84  -.75   .76  -.59  -.88  -.62  -.53   .63
#> disp  -.83   .92    1    .79  -.76   .86  -.44  -.77  -.63  -.59   .50
#> hp    -.77   .84   .79    1   -.51   .65  -.75  -.79  -.22  -.28   .77
#> drat   .71  -.75  -.76  -.51    1   -.73   .16   .54   .77   .69  -.23
#> wt    -.85   .76   .86   .65  -.73    1   -.23  -.56  -.73  -.53   .46
#> qsec   .39  -.59  -.44  -.75   .16  -.23    1    .73  -.23  -.09  -.66
#> vs     .68  -.88  -.77  -.79   .54  -.56   .73    1    .30   .33  -.67
#> am     .65  -.62  -.63  -.22   .77  -.73  -.23   .30    1    .79  -.05
#> gear   .64  -.53  -.59  -.28   .69  -.53  -.09   .33   .79    1    .11
#> carb  -.61   .63   .50   .77  -.23   .46  -.66  -.67  -.05   .11    1 
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      35.03     -9.20   -608.96   -329.66      2.11     -4.49      4.11
#> cyl      -9.20      3.28    223.32    109.99     -0.75      1.24     -2.14
#> disp   -608.96    223.32  16004.16   7497.63    -52.22     96.52   -107.44
#> hp     -329.66    109.99   7497.63   5025.47    -18.43     43.14   -104.42
#> drat      2.11     -0.75    -52.22    -18.43      0.29     -0.37      0.16
#> wt       -4.49      1.24     96.52     43.14     -0.37      0.80     -0.41
#> qsec      4.11     -2.14   -107.44   -104.42      0.16     -0.41      3.41
#> vs        2.09     -0.80    -50.46    -25.09      0.15     -0.26      0.68
#> am        2.03     -0.56    -40.60     -8.04      0.22     -0.34     -0.18
#> gear      2.31     -0.66    -52.85    -11.93      0.23     -0.31     -0.11
#> carb     -6.42      1.99    106.03     95.84     -0.20      0.74     -2.09
#>             vs        am      gear      carb
#> mpg       2.09      2.03      2.31     -6.42
#> cyl      -0.80     -0.56     -0.66      1.99
#> disp    -50.46    -40.60    -52.85    106.03
#> hp      -25.09     -8.04    -11.93     95.84
#> drat      0.15      0.22      0.23     -0.20
#> wt       -0.26     -0.34     -0.31      0.74
#> qsec      0.68     -0.18     -0.11     -2.09
#> vs        0.26      0.08      0.12     -0.47
#> am        0.08      0.26      0.27     -0.04
#> gear      0.12      0.27      0.47      0.10
#> carb     -0.47     -0.04      0.10      2.71
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   27 26   27 27   26 26 26   26   26
#> cyl   26  29   26 27   26 26   26 26 26   26   26
#> disp  27  26   29 26   26 26   26 26 26   26   26
#> hp    26  27   26 29   26 26   26 28 26   27   27
#> drat  27  26   26 26   29 26   28 26 26   27   26
#> wt    27  26   26 26   26 29   26 26 26   26   26
#> qsec  26  26   26 26   28 26   29 26 26   27   26
#> vs    26  26   26 28   26 26   26 29 26   28   27
#> am    26  26   26 26   26 26   26 26 29   26   26
#> gear  26  26   26 27   27 26   27 28 26   29   26
#> carb  26  26   26 27   26 26   26 27 26   26   29
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.84 (26) -.83 (27) -.77 (26)  .71 (27) -.85 (27)  .39 (26)
#> cyl  -.84 (26)   1  (29)  .92 (26)  .84 (27) -.75 (26)  .76 (26) -.59 (26)
#> disp -.83 (27)  .92 (26)   1  (29)  .79 (26) -.76 (26)  .86 (26) -.44 (26)
#> hp   -.77 (26)  .84 (27)  .79 (26)   1  (29) -.51 (26)  .65 (26) -.75 (26)
#> drat  .71 (27) -.75 (26) -.76 (26) -.51 (26)   1  (29) -.73 (26)  .16 (28)
#> wt   -.85 (27)  .76 (26)  .86 (26)  .65 (26) -.73 (26)   1  (29) -.23 (26)
#> qsec  .39 (26) -.59 (26) -.44 (26) -.75 (26)  .16 (28) -.23 (26)   1  (29)
#> vs    .68 (26) -.88 (26) -.77 (26) -.79 (28)  .54 (26) -.56 (26)  .73 (26)
#> am    .65 (26) -.62 (26) -.63 (26) -.22 (26)  .77 (26) -.73 (26) -.23 (26)
#> gear  .64 (26) -.53 (26) -.59 (26) -.28 (27)  .69 (27) -.53 (26) -.09 (27)
#> carb -.61 (26)  .63 (26)  .50 (26)  .77 (27) -.23 (26)  .46 (26) -.66 (26)
#>             vs        am      gear      carb
#> mpg   .68 (26)  .65 (26)  .64 (26) -.61 (26)
#> cyl  -.88 (26) -.62 (26) -.53 (26)  .63 (26)
#> disp -.77 (26) -.63 (26) -.59 (26)  .50 (26)
#> hp   -.79 (28) -.22 (26) -.28 (27)  .77 (27)
#> drat  .54 (26)  .77 (26)  .69 (27) -.23 (26)
#> wt   -.56 (26) -.73 (26) -.53 (26)  .46 (26)
#> qsec  .73 (26) -.23 (26) -.09 (27) -.66 (26)
#> vs     1  (29)  .30 (26)  .33 (28) -.67 (27)
#> am    .30 (26)   1  (29)  .79 (26) -.05 (26)
#> gear  .33 (28)  .79 (26)   1  (29)  .11 (26)
#> carb -.67 (27) -.05 (26)  .11 (26)   1  (29)
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.84*  -.83*  -.77*   .71*  -.85*   .39*   .68*   .65*   .64*
#> cyl   -.84*    1     .92*   .84*  -.75*   .76*  -.59*  -.88*  -.62*  -.53*
#> disp  -.83*   .92*    1     .79*  -.76*   .86*  -.44*  -.77*  -.63*  -.59*
#> hp    -.77*   .84*   .79*    1    -.51*   .65*  -.75*  -.79*  -.22   -.28 
#> drat   .71*  -.75*  -.76*  -.51*    1    -.73*   .16    .54*   .77*   .69*
#> wt    -.85*   .76*   .86*   .65*  -.73*    1    -.23   -.56*  -.73*  -.53*
#> qsec   .39*  -.59*  -.44*  -.75*   .16   -.23     1     .73*  -.23   -.09 
#> vs     .68*  -.88*  -.77*  -.79*   .54*  -.56*   .73*    1     .30    .33 
#> am     .65*  -.62*  -.63*  -.22    .77*  -.73*  -.23    .30     1     .79*
#> gear   .64*  -.53*  -.59*  -.28    .69*  -.53*  -.09    .33    .79*    1  
#> carb  -.61*   .63*   .50*   .77*  -.23    .46*  -.66*  -.67*  -.05    .11 
#>        carb
#> mpg   -.61*
#> cyl    .63*
#> disp   .50*
#> hp     .77*
#> drat  -.23 
#> wt     .46*
#> qsec  -.66*
#> vs    -.67*
#> am    -.05 
#> gear   .11 
#> carb    1  
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.84* (26) -.83* (27) -.77* (26)  .71* (27) -.85* (27)
#> cyl  -.84* (26)   1   (29)  .92* (26)  .84* (27) -.75* (26)  .76* (26)
#> disp -.83* (27)  .92* (26)   1   (29)  .79* (26) -.76* (26)  .86* (26)
#> hp   -.77* (26)  .84* (27)  .79* (26)   1   (29) -.51* (26)  .65* (26)
#> drat  .71* (27) -.75* (26) -.76* (26) -.51* (26)   1   (29) -.73* (26)
#> wt   -.85* (27)  .76* (26)  .86* (26)  .65* (26) -.73* (26)   1   (29)
#> qsec  .39* (26) -.59* (26) -.44* (26) -.75* (26)  .16  (28) -.23  (26)
#> vs    .68* (26) -.88* (26) -.77* (26) -.79* (28)  .54* (26) -.56* (26)
#> am    .65* (26) -.62* (26) -.63* (26) -.22  (26)  .77* (26) -.73* (26)
#> gear  .64* (26) -.53* (26) -.59* (26) -.28  (27)  .69* (27) -.53* (26)
#> carb -.61* (26)  .63* (26)  .50* (26)  .77* (27) -.23  (26)  .46* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .39* (26)  .68* (26)  .65* (26)  .64* (26) -.61* (26)
#> cyl  -.59* (26) -.88* (26) -.62* (26) -.53* (26)  .63* (26)
#> disp -.44* (26) -.77* (26) -.63* (26) -.59* (26)  .50* (26)
#> hp   -.75* (26) -.79* (28) -.22  (26) -.28  (27)  .77* (27)
#> drat  .16  (28)  .54* (26)  .77* (26)  .69* (27) -.23  (26)
#> wt   -.23  (26) -.56* (26) -.73* (26) -.53* (26)  .46* (26)
#> qsec   1   (29)  .73* (26) -.23  (26) -.09  (27) -.66* (26)
#> vs    .73* (26)   1   (29)  .30  (26)  .33  (28) -.67* (27)
#> am   -.23  (26)  .30  (26)   1   (29)  .79* (26) -.05  (26)
#> gear -.09  (27)  .33  (28)  .79* (26)   1   (29)  .11  (26)
#> carb -.66* (26) -.67* (27) -.05  (26)  .11  (26)   1   (29)
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.84  -.83  -.77   .71  -.85   .39   .68   .65   .64  -.61
#> N  29    26    27    26    27    27    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .05   .00   .00   .00   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.84   1   .92   .84  -.75   .76  -.59  -.88  -.62  -.53   .63
#> N    26  29    26    27    26    26    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .01   .00
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   .92    1   .79  -.76   .86  -.44  -.77  -.63  -.59   .50
#> N    27    26   29    26    26    26    26    26    26    26    26
#> P   .00   .00        .00   .00   .00   .03   .00   .00   .00   .01
#> 
#> , , hp
#> 
#>     mpg   cyl  disp     hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.77   .84   .79   1.00  -.51   .65  -.75  -.79  -.22  -.28   .77
#> N    26    27    26     29    26    26    26    28    26    27    27
#> P   .00   .00   .00          .01   .00   .00   .00   .27   .15   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .71  -.75  -.76  -.51    1  -.73   .16   .54   .77   .69  -.23
#> N    27    26    26    26   29    26    28    26    26    27    26
#> P   .00   .00   .00   .01        .00   .40   .00   .00   .00   .26
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.85   .76   .86   .65  -.73   1  -.23  -.56  -.73  -.53   .46
#> N    27    26    26    26    26  29    26    26    26    26    26
#> P   .00   .00   .00   .00   .00       .26   .00   .00   .01   .02
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt   qsec    vs    am  gear  carb
#> r   .39  -.59  -.44  -.75   .16  -.23   1.00   .73  -.23  -.09  -.66
#> N    26    26    26    26    28    26     29    26    26    27    26
#> P   .05   .00   .03   .00   .40   .26          .00   .26   .66   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .68  -.88  -.77  -.79   .54  -.56   .73   1   .30   .33  -.67
#> N    26    26    26    28    26    26    26  29    26    28    27
#> P   .00   .00   .00   .00   .00   .00   .00       .14   .09   .00
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs     am  gear  carb
#> r   .65  -.62  -.63  -.22   .77  -.73  -.23   .30   1.00   .79  -.05
#> N    26    26    26    26    26    26    26    26     29    26    26
#> P   .00   .00   .00   .27   .00   .00   .26   .14          .00   .82
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am   gear  carb
#> r   .64  -.53  -.59  -.28   .69  -.53  -.09   .33   .79   1.00   .11
#> N    26    26    26    27    27    26    27    28    26     29    26
#> P   .00   .01   .00   .15   .00   .01   .66   .09   .00          .58
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.61   .63   .50   .77  -.23   .46  -.66  -.67  -.05   .11    1
#> N    26    26    26    27    26    26    26    27    26    26   29
#> P   .00   .00   .01   .00   .26   .02   .00   .00   .82   .58     
#> 
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.843* (26)   1    (29)                                                
#> disp -.829* (27)  .919* (26)   1    (29)                                    
#> hp   -.773* (26)  .838* (27)  .786* (26)   1    (29)                        
#> drat  .715* (27) -.753* (26) -.763* (26) -.507* (26)   1    (29)            
#> wt   -.848* (27)  .757* (26)  .864* (26)  .651* (26) -.732* (26)   1    (29)
#> qsec  .389  (26) -.591* (26) -.437  (26) -.745* (26)  .164  (28) -.228  (26)
#> vs    .679* (26) -.877* (26) -.768* (26) -.794* (28)  .543* (26) -.564* (26)
#> am    .655* (26) -.616* (26) -.629* (26) -.223  (26)  .770* (26) -.726* (26)
#> gear  .643* (26) -.530* (26) -.591* (26) -.282  (27)  .690* (27) -.533* (26)
#> carb -.606* (26)  .629* (26)  .500* (26)  .774* (27) -.228  (26)  .460  (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .732* (26)   1    (29)                                    
#> am   -.227  (26)  .300  (26)   1    (29)                        
#> gear -.089  (27)  .331  (28)  .792* (26)   1    (29)            
#> carb -.664* (26) -.669* (27) -.047  (26)  .113  (26)   1    (29)
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp     hp  drat    wt   qsec    vs     am   gear  carb
#> mpg      1  -.84  -.83   -.77   .71  -.85    .39   .68    .65    .64  -.61
#> cyl   -.84     1   .92    .84  -.75   .76   -.59  -.88   -.62   -.53   .63
#> disp  -.83   .92     1    .79  -.76   .86   -.44  -.77   -.63   -.59   .50
#> hp    -.77   .84   .79   1.00  -.51   .65   -.75  -.79   -.22   -.28   .77
#> drat   .71  -.75  -.76   -.51     1  -.73    .16   .54    .77    .69  -.23
#> wt    -.85   .76   .86    .65  -.73     1   -.23  -.56   -.73   -.53   .46
#> qsec   .39  -.59  -.44   -.75   .16  -.23   1.00   .73   -.23   -.09  -.66
#> vs     .68  -.88  -.77   -.79   .54  -.56    .73     1    .30    .33  -.67
#> am     .65  -.62  -.63   -.22   .77  -.73   -.23   .30   1.00    .79  -.05
#> gear   .64  -.53  -.59   -.28   .69  -.53   -.09   .33    .79   1.00   .11
#> carb  -.61   .63   .50    .77  -.23   .46   -.66  -.67   -.05    .11     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   27  26   27  27   26  26  26   26   26
#> cyl   26  29   26  27   26  26   26  26  26   26   26
#> disp  27  26   29  26   26  26   26  26  26   26   26
#> hp    26  27   26  29   26  26   26  28  26   27   27
#> drat  27  26   26  26   29  26   28  26  26   27   26
#> wt    27  26   26  26   26  29   26  26  26   26   26
#> qsec  26  26   26  26   28  26   29  26  26   27   26
#> vs    26  26   26  28   26  26   26  29  26   28   27
#> am    26  26   26  26   26  26   26  26  29   26   26
#> gear  26  26   26  27   27  26   27  28  26   29   26
#> carb  26  26   26  27   26  26   26  27  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .05  .00  .00  .00  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .01  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .01
#> hp    .00  .00  .00       .01  .00  .00  .00  .27  .15  .00
#> drat  .00  .00  .00  .01       .00  .40  .00  .00  .00  .26
#> wt    .00  .00  .00  .00  .00       .26  .00  .00  .01  .02
#> qsec  .05  .00  .03  .00  .40  .26       .00  .26  .66  .00
#> vs    .00  .00  .00  .00  .00  .00  .00       .14  .09  .00
#> am    .00  .00  .00  .27  .00  .00  .26  .14       .00  .82
#> gear  .00  .01  .00  .15  .00  .01  .66  .09  .00       .58
#> carb  .00  .00  .01  .00  .26  .02  .00  .00  .82  .58     
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp     hp  drat    wt   qsec    vs     am   gear carb
#> mpg      1                                                               
#> cyl   -.84     1                                                         
#> disp  -.83   .92     1                                                   
#> hp    -.77   .84   .79   1.00                                            
#> drat   .71  -.75  -.76   -.51     1                                      
#> wt    -.85   .76   .86    .65  -.73     1                                
#> qsec   .39  -.59  -.44   -.75   .16  -.23   1.00                         
#> vs     .68  -.88  -.77   -.79   .54  -.56    .73     1                   
#> am     .65  -.62  -.63   -.22   .77  -.73   -.23   .30   1.00            
#> gear   .64  -.53  -.59   -.28   .69  -.53   -.09   .33    .79   1.00     
#> carb  -.61   .63   .50    .77  -.23   .46   -.66  -.67   -.05    .11    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  27  26   29                                    
#> hp    26  27   26  29                                
#> drat  27  26   26  26   29                           
#> wt    27  26   26  26   26  29                       
#> qsec  26  26   26  26   28  26   29                  
#> vs    26  26   26  28   26  26   26  29              
#> am    26  26   26  26   26  26   26  26  29          
#> gear  26  26   26  27   27  26   27  28  26   29     
#> carb  26  26   26  27   26  26   26  27  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .01                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .05  .00  .03  .00  .40  .26                         
#> vs    .00  .00  .00  .00  .00  .00  .00                    
#> am    .00  .00  .00  .27  .00  .00  .26  .14               
#> gear  .00  .01  .00  .15  .00  .01  .66  .09  .00          
#> carb  .00  .00  .01  .00  .26  .02  .00  .00  .82  .58     
#>