(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.90  -.78   .67  -.85   .44   .66   .58   .45  -.58
#> cyl   -.86    1    .90   .83  -.65   .75  -.63  -.87  -.45  -.42   .56
#> disp  -.90   .90    1    .79  -.69   .87  -.44  -.68  -.56  -.52   .44
#> hp    -.78   .83   .79    1   -.43   .62  -.72  -.73  -.18  -.10   .81
#> drat   .67  -.65  -.69  -.43    1   -.70   .06   .38   .71   .66  -.04
#> wt    -.85   .75   .87   .62  -.70    1   -.19  -.51  -.69  -.56   .46
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      39.27     -9.22   -719.42   -370.94      2.40     -4.66      5.63
#> cyl      -9.22      3.16    189.42    110.60     -0.64      1.16     -2.23
#> disp   -719.42    189.42  14868.52   7161.92    -47.56     83.08   -110.26
#> hp     -370.94    110.60   7161.92   5077.26    -16.88     39.98   -100.77
#> drat      2.40     -0.64    -47.56    -16.88      0.30     -0.29      0.06
#> wt       -4.66      1.16     83.08     39.98     -0.29      0.79     -0.33
#>             vs        am      gear      carb
#> mpg       2.15      1.87      2.17     -5.52
#> cyl      -0.78     -0.40     -0.57      1.56
#> disp    -41.92    -34.26    -49.37     85.12
#> hp      -26.18     -6.44     -5.61     90.44
#> drat      0.11      0.19      0.27     -0.03
#> wt       -0.22     -0.27     -0.36      0.66
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 26   26 26 26   26   26
#> cyl   26  29   27 26   26 26   26 26 27   26   27
#> disp  26  27   29 26   27 26   26 28 27   27   26
#> hp    27  26   26 29   26 27   27 26 26   26   26
#> drat  26  26   27 26   29 26   27 28 27   27   26
#> wt    26  26   26 27   26 29   27 26 26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (26) -.90 (26) -.78 (27)  .67 (26) -.85 (26)  .44 (26)
#> cyl  -.86 (26)   1  (29)  .90 (27)  .83 (26) -.65 (26)  .75 (26) -.63 (26)
#> disp -.90 (26)  .90 (27)   1  (29)  .79 (26) -.69 (27)  .87 (26) -.44 (26)
#> hp   -.78 (27)  .83 (26)  .79 (26)   1  (29) -.43 (26)  .62 (27) -.72 (27)
#> drat  .67 (26) -.65 (26) -.69 (27) -.43 (26)   1  (29) -.70 (26)  .06 (27)
#> wt   -.85 (26)  .75 (26)  .87 (26)  .62 (27) -.70 (26)   1  (29) -.19 (27)
#>             vs        am      gear      carb
#> mpg   .66 (26)  .58 (26)  .45 (26) -.58 (26)
#> cyl  -.87 (26) -.45 (27) -.42 (26)  .56 (27)
#> disp -.68 (28) -.56 (27) -.52 (27)  .44 (26)
#> hp   -.73 (26) -.18 (26) -.10 (26)  .81 (26)
#> drat  .38 (28)  .71 (27)  .66 (27) -.04 (26)
#> wt   -.51 (26) -.69 (26) -.56 (26)  .46 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.90*  -.78*   .67*  -.85*   .44*   .66*   .58*   .45*
#> cyl   -.86*    1     .90*   .83*  -.65*   .75*  -.63*  -.87*  -.45*  -.42*
#> disp  -.90*   .90*    1     .79*  -.69*   .87*  -.44*  -.68*  -.56*  -.52*
#> hp    -.78*   .83*   .79*    1    -.43*   .62*  -.72*  -.73*  -.18   -.10 
#> drat   .67*  -.65*  -.69*  -.43*    1    -.70*   .06    .38*   .71*   .66*
#> wt    -.85*   .75*   .87*   .62*  -.70*    1    -.19   -.51*  -.69*  -.56*
#>        carb
#> mpg   -.58*
#> cyl    .56*
#> disp   .44*
#> hp     .81*
#> drat  -.04 
#> wt     .46*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (26) -.90* (26) -.78* (27)  .67* (26) -.85* (26)
#> cyl  -.86* (26)   1   (29)  .90* (27)  .83* (26) -.65* (26)  .75* (26)
#> disp -.90* (26)  .90* (27)   1   (29)  .79* (26) -.69* (27)  .87* (26)
#> hp   -.78* (27)  .83* (26)  .79* (26)   1   (29) -.43* (26)  .62* (27)
#> drat  .67* (26) -.65* (26) -.69* (27) -.43* (26)   1   (29) -.70* (26)
#> wt   -.85* (26)  .75* (26)  .87* (26)  .62* (27) -.70* (26)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .44* (26)  .66* (26)  .58* (26)  .45* (26) -.58* (26)
#> cyl  -.63* (26) -.87* (26) -.45* (27) -.42* (26)  .56* (27)
#> disp -.44* (26) -.68* (28) -.56* (27) -.52* (27)  .44* (26)
#> hp   -.72* (27) -.73* (26) -.18  (26) -.10  (26)  .81* (26)
#> drat  .06  (27)  .38* (28)  .71* (27)  .66* (27) -.04  (26)
#> wt   -.19  (27) -.51* (26) -.69* (26) -.56* (26)  .46* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1.00  -.86  -.90  -.78   .67  -.85   .44   .66   .58   .45  -.58
#> N     29    26    26    27    26    26    26    26    26    26    26
#> P          .00   .00   .00   .00   .00   .03   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .90   .83  -.65   .75  -.63  -.87  -.45  -.42   .56
#> N    26  29    27    26    26    26    26    26    27    26    27
#> P   .00       .00   .00   .00   .00   .00   .00   .02   .03   .00
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.863* (26)   1    (29)                                                
#> disp -.897* (26)  .895* (27)   1    (29)                                    
#> hp   -.779* (27)  .828* (26)  .795* (26)   1    (29)                        
#> drat  .672* (26) -.655* (26) -.691* (27) -.428  (26)   1    (29)            
#> wt   -.847* (26)  .748* (26)  .866* (26)  .623* (27) -.700* (26)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>         mpg   cyl  disp    hp  drat     wt   qsec     vs    am  gear  carb
#> mpg    1.00  -.86  -.90  -.78   .67   -.85    .44    .66   .58   .45  -.58
#> cyl    -.86     1   .90   .83  -.65    .75   -.63   -.87  -.45  -.42   .56
#> disp   -.90   .90     1   .79  -.69    .87   -.44   -.68  -.56  -.52   .44
#> hp     -.78   .83   .79     1  -.43    .62   -.72   -.73  -.18  -.10   .81
#> drat    .67  -.65  -.69  -.43     1   -.70    .06    .38   .71   .66  -.04
#> wt     -.85   .75   .87   .62  -.70   1.00   -.19   -.51  -.69  -.56   .46
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  26   26  26  26   26   26
#> cyl   26  29   27  26   26  26   26  26  27   26   27
#> disp  26  27   29  26   27  26   26  28  27   27   26
#> hp    27  26   26  29   26  27   27  26  26   26   26
#> drat  26  26   27  26   29  26   27  28  27   27   26
#> wt    26  26   26  27   26  29   27  26  26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .02  .03  .00
#> disp  .00  .00       .00  .00  .00  .02  .00  .00  .01  .03
#> hp    .00  .00  .00       .03  .00  .00  .00  .39  .62  .00
#> drat  .00  .00  .00  .03       .00  .76  .04  .00  .00  .86
#> wt    .00  .00  .00  .00  .00       .35  .01  .00  .00  .02
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>         mpg   cyl  disp    hp  drat     wt   qsec     vs    am  gear carb
#> mpg    1.00                                                              
#> cyl    -.86     1                                                        
#> disp   -.90   .90     1                                                  
#> hp     -.78   .83   .79     1                                            
#> drat    .67  -.65  -.69  -.43     1                                      
#> wt     -.85   .75   .87   .62  -.70   1.00                               
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    27  26   26  29                                
#> drat  26  26   27  26   29                           
#> wt    26  26   26  27   26  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .03                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>