(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.87  -.85  -.77   .69  -.91   .40   .64   .71   .53  -.63
#> cyl   -.87    1    .90   .83  -.73   .83  -.59  -.81  -.71  -.57   .62
#> disp  -.85   .90    1    .75  -.74   .88  -.42  -.70  -.69  -.65   .47
#> hp    -.77   .83   .75    1   -.47   .67  -.72  -.70  -.34  -.15   .79
#> drat   .69  -.73  -.74  -.47    1   -.77   .14   .52   .73   .70  -.25
#> wt    -.91   .83   .88   .67  -.77    1   -.20  -.56  -.74  -.61   .52
#> qsec   .40  -.59  -.42  -.72   .14  -.20    1    .75  -.13  -.19  -.63
#> vs     .64  -.81  -.70  -.70   .52  -.56   .75    1    .28   .23  -.52
#> am     .71  -.71  -.69  -.34   .73  -.74  -.13   .28    1    .84  -.15
#> gear   .53  -.57  -.65  -.15   .70  -.61  -.19   .23   .84    1    .13
#> carb  -.63   .62   .47   .79  -.25   .52  -.63  -.52  -.15   .13    1 
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      39.17    -10.17   -628.42   -309.80      2.35     -5.72      4.67
#> cyl     -10.17      3.24    204.19    104.03     -0.68      1.51     -1.66
#> disp   -628.42    204.19  15416.85   5742.09    -51.08    100.17    -88.56
#> hp     -309.80    104.03   5742.09   4598.11    -17.25     38.08    -89.44
#> drat      2.35     -0.68    -51.08    -17.25      0.28     -0.41      0.14
#> wt       -5.72      1.51    100.17     38.08     -0.41      0.89     -0.30
#> qsec      4.67     -1.66    -88.56    -89.44      0.14     -0.30      3.38
#> vs        1.96     -0.74    -41.57    -24.31      0.14     -0.24      0.72
#> am        2.29     -0.64    -43.33    -11.07      0.20     -0.36     -0.09
#> gear      2.44     -0.75    -59.59     -7.02      0.24     -0.39     -0.25
#> carb     -6.38      1.82     94.42     88.27     -0.21      0.79     -1.82
#>             vs        am      gear      carb
#> mpg       1.96      2.29      2.44     -6.38
#> cyl      -0.74     -0.64     -0.75      1.82
#> disp    -41.57    -43.33    -59.59     94.42
#> hp      -24.31    -11.07     -7.02     88.27
#> drat      0.14      0.20      0.24     -0.21
#> wt       -0.24     -0.36     -0.39      0.79
#> qsec      0.72     -0.09     -0.25     -1.82
#> vs        0.26      0.07      0.09     -0.42
#> am        0.07      0.25      0.31     -0.12
#> gear      0.09      0.31      0.52      0.14
#> carb     -0.42     -0.12      0.14      2.44
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   27 26   26 26   26 27 26   26   27
#> cyl   27  29   26 26   27 26   26 26 27   26   27
#> disp  27  26   29 26   26 26   26 26 26   27   26
#> hp    26  26   26 29   26 27   27 26 26   26   26
#> drat  26  27   26 26   29 26   26 26 27   26   26
#> wt    26  26   26 27   26 29   27 26 27   27   27
#> qsec  26  26   26 27   26 27   29 27 26   26   26
#> vs    27  26   26 26   26 26   27 29 26   26   26
#> am    26  27   26 26   27 27   26 26 29   27   26
#> gear  26  26   27 26   26 27   26 26 27   29   26
#> carb  27  27   26 26   26 27   26 26 26   26   29
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.87 (27) -.85 (27) -.77 (26)  .69 (26) -.91 (26)  .40 (26)
#> cyl  -.87 (27)   1  (29)  .90 (26)  .83 (26) -.73 (27)  .83 (26) -.59 (26)
#> disp -.85 (27)  .90 (26)   1  (29)  .75 (26) -.74 (26)  .88 (26) -.42 (26)
#> hp   -.77 (26)  .83 (26)  .75 (26)   1  (29) -.47 (26)  .67 (27) -.72 (27)
#> drat  .69 (26) -.73 (27) -.74 (26) -.47 (26)   1  (29) -.77 (26)  .14 (26)
#> wt   -.91 (26)  .83 (26)  .88 (26)  .67 (27) -.77 (26)   1  (29) -.20 (27)
#> qsec  .40 (26) -.59 (26) -.42 (26) -.72 (27)  .14 (26) -.20 (27)   1  (29)
#> vs    .64 (27) -.81 (26) -.70 (26) -.70 (26)  .52 (26) -.56 (26)  .75 (27)
#> am    .71 (26) -.71 (27) -.69 (26) -.34 (26)  .73 (27) -.74 (27) -.13 (26)
#> gear  .53 (26) -.57 (26) -.65 (27) -.15 (26)  .70 (26) -.61 (27) -.19 (26)
#> carb -.63 (27)  .62 (27)  .47 (26)  .79 (26) -.25 (26)  .52 (27) -.63 (26)
#>             vs        am      gear      carb
#> mpg   .64 (27)  .71 (26)  .53 (26) -.63 (27)
#> cyl  -.81 (26) -.71 (27) -.57 (26)  .62 (27)
#> disp -.70 (26) -.69 (26) -.65 (27)  .47 (26)
#> hp   -.70 (26) -.34 (26) -.15 (26)  .79 (26)
#> drat  .52 (26)  .73 (27)  .70 (26) -.25 (26)
#> wt   -.56 (26) -.74 (27) -.61 (27)  .52 (27)
#> qsec  .75 (27) -.13 (26) -.19 (26) -.63 (26)
#> vs     1  (29)  .28 (26)  .23 (26) -.52 (26)
#> am    .28 (26)   1  (29)  .84 (27) -.15 (26)
#> gear  .23 (26)  .84 (27)   1  (29)  .13 (26)
#> carb -.52 (26) -.15 (26)  .13 (26)   1  (29)
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.87*  -.85*  -.77*   .69*  -.91*   .40*   .64*   .71*   .53*
#> cyl   -.87*    1     .90*   .83*  -.73*   .83*  -.59*  -.81*  -.71*  -.57*
#> disp  -.85*   .90*    1     .75*  -.74*   .88*  -.42*  -.70*  -.69*  -.65*
#> hp    -.77*   .83*   .75*    1    -.47*   .67*  -.72*  -.70*  -.34   -.15 
#> drat   .69*  -.73*  -.74*  -.47*    1    -.77*   .14    .52*   .73*   .70*
#> wt    -.91*   .83*   .88*   .67*  -.77*    1    -.20   -.56*  -.74*  -.61*
#> qsec   .40*  -.59*  -.42*  -.72*   .14   -.20     1     .75*  -.13   -.19 
#> vs     .64*  -.81*  -.70*  -.70*   .52*  -.56*   .75*    1     .28    .23 
#> am     .71*  -.71*  -.69*  -.34    .73*  -.74*  -.13    .28     1     .84*
#> gear   .53*  -.57*  -.65*  -.15    .70*  -.61*  -.19    .23    .84*    1  
#> carb  -.63*   .62*   .47*   .79*  -.25    .52*  -.63*  -.52*  -.15    .13 
#>        carb
#> mpg   -.63*
#> cyl    .62*
#> disp   .47*
#> hp     .79*
#> drat  -.25 
#> wt     .52*
#> qsec  -.63*
#> vs    -.52*
#> am    -.15 
#> gear   .13 
#> carb    1  
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.87* (27) -.85* (27) -.77* (26)  .69* (26) -.91* (26)
#> cyl  -.87* (27)   1   (29)  .90* (26)  .83* (26) -.73* (27)  .83* (26)
#> disp -.85* (27)  .90* (26)   1   (29)  .75* (26) -.74* (26)  .88* (26)
#> hp   -.77* (26)  .83* (26)  .75* (26)   1   (29) -.47* (26)  .67* (27)
#> drat  .69* (26) -.73* (27) -.74* (26) -.47* (26)   1   (29) -.77* (26)
#> wt   -.91* (26)  .83* (26)  .88* (26)  .67* (27) -.77* (26)   1   (29)
#> qsec  .40* (26) -.59* (26) -.42* (26) -.72* (27)  .14  (26) -.20  (27)
#> vs    .64* (27) -.81* (26) -.70* (26) -.70* (26)  .52* (26) -.56* (26)
#> am    .71* (26) -.71* (27) -.69* (26) -.34  (26)  .73* (27) -.74* (27)
#> gear  .53* (26) -.57* (26) -.65* (27) -.15  (26)  .70* (26) -.61* (27)
#> carb -.63* (27)  .62* (27)  .47* (26)  .79* (26) -.25  (26)  .52* (27)
#>            qsec         vs         am       gear       carb
#> mpg   .40* (26)  .64* (27)  .71* (26)  .53* (26) -.63* (27)
#> cyl  -.59* (26) -.81* (26) -.71* (27) -.57* (26)  .62* (27)
#> disp -.42* (26) -.70* (26) -.69* (26) -.65* (27)  .47* (26)
#> hp   -.72* (27) -.70* (26) -.34  (26) -.15  (26)  .79* (26)
#> drat  .14  (26)  .52* (26)  .73* (27)  .70* (26) -.25  (26)
#> wt   -.20  (27) -.56* (26) -.74* (27) -.61* (27)  .52* (27)
#> qsec   1   (29)  .75* (27) -.13  (26) -.19  (26) -.63* (26)
#> vs    .75* (27)   1   (29)  .28  (26)  .23  (26) -.52* (26)
#> am   -.13  (26)  .28  (26)   1   (29)  .84* (27) -.15  (26)
#> gear -.19  (26)  .23  (26)  .84* (27)   1   (29)  .13  (26)
#> carb -.63* (26) -.52* (26) -.15  (26)  .13  (26)   1   (29)
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.87  -.85  -.77   .69  -.91   .40   .64   .71   .53  -.63
#> N  29    27    27    26    26    26    26    27    26    26    27
#> P       .00   .00   .00   .00   .00   .04   .00   .00   .01   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.87   1   .90   .83  -.73   .83  -.59  -.81  -.71  -.57   .62
#> N    27  29    26    26    27    26    26    26    27    26    27
#> P   .00       .00   .00   .00   .00   .00   .00   .00   .00   .00
#> 
#> , , disp
#> 
#>     mpg   cyl   disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   .90   1.00   .75  -.74   .88  -.42  -.70  -.69  -.65   .47
#> N    27    26     29    26    26    26    26    26    26    27    26
#> P   .00   .00          .00   .00   .00   .03   .00   .00   .00   .02
#> 
#> , , hp
#> 
#>     mpg   cyl  disp     hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.77   .83   .75   1.00  -.47   .67  -.72  -.70  -.34  -.15   .79
#> N    26    26    26     29    26    27    27    26    26    26    26
#> P   .00   .00   .00          .02   .00   .00   .00   .09   .47   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp   drat    wt  qsec    vs    am  gear  carb
#> r   .69  -.73  -.74  -.47   1.00  -.77   .14   .52   .73   .70  -.25
#> N    26    27    26    26     29    26    26    26    27    26    26
#> P   .00   .00   .00   .02          .00   .49   .01   .00   .00   .22
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.91   .83   .88   .67  -.77   1  -.20  -.56  -.74  -.61   .52
#> N    26    26    26    27    26  29    27    26    27    27    27
#> P   .00   .00   .00   .00   .00       .31   .00   .00   .00   .01
#> 
#>  [ reached getOption("max.print") -- omitted 5 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.867* (27)   1    (29)                                                
#> disp -.851* (27)  .899* (26)   1    (29)                                    
#> hp   -.765* (26)  .828* (26)  .752* (26)   1    (29)                        
#> drat  .685* (26) -.733* (27) -.736* (26) -.470  (26)   1    (29)            
#> wt   -.905* (26)  .829* (26)  .882* (26)  .667* (27) -.767* (26)   1    (29)
#> qsec  .401  (26) -.589* (26) -.415  (26) -.722* (27)  .141  (26) -.202  (27)
#> vs    .642* (27) -.810* (26) -.696* (26) -.701* (26)  .519* (26) -.561* (26)
#> am    .712* (26) -.710* (27) -.690* (26) -.339  (26)  .732* (27) -.735* (27)
#> gear  .531* (26) -.573* (26) -.654* (27) -.148  (26)  .700* (26) -.613* (27)
#> carb -.628* (27)  .618* (27)  .468  (26)  .788* (26) -.251  (26)  .518* (27)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .746* (27)   1    (29)                                    
#> am   -.127  (26)  .283  (26)   1    (29)                        
#> gear -.192  (26)  .227  (26)  .836* (27)   1    (29)            
#> carb -.626* (26) -.524* (26) -.146  (26)  .125  (26)   1    (29)
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl   disp     hp   drat    wt  qsec     vs    am  gear  carb
#> mpg      1  -.87   -.85   -.77    .69  -.91   .40    .64   .71   .53  -.63
#> cyl   -.87     1    .90    .83   -.73   .83  -.59   -.81  -.71  -.57   .62
#> disp  -.85   .90   1.00    .75   -.74   .88  -.42   -.70  -.69  -.65   .47
#> hp    -.77   .83    .75   1.00   -.47   .67  -.72   -.70  -.34  -.15   .79
#> drat   .69  -.73   -.74   -.47   1.00  -.77   .14    .52   .73   .70  -.25
#> wt    -.91   .83    .88    .67   -.77     1  -.20   -.56  -.74  -.61   .52
#> qsec   .40  -.59   -.42   -.72    .14  -.20     1    .75  -.13  -.19  -.63
#> vs     .64  -.81   -.70   -.70    .52  -.56   .75   1.00   .28   .23  -.52
#> am     .71  -.71   -.69   -.34    .73  -.74  -.13    .28     1   .84  -.15
#> gear   .53  -.57   -.65   -.15    .70  -.61  -.19    .23   .84     1   .13
#> carb  -.63   .62    .47    .79   -.25   .52  -.63   -.52  -.15   .13     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   27  26   26  26   26  27  26   26   27
#> cyl   27  29   26  26   27  26   26  26  27   26   27
#> disp  27  26   29  26   26  26   26  26  26   27   26
#> hp    26  26   26  29   26  27   27  26  26   26   26
#> drat  26  27   26  26   29  26   26  26  27   26   26
#> wt    26  26   26  27   26  29   27  26  27   27   27
#> qsec  26  26   26  27   26  27   29  27  26   26   26
#> vs    27  26   26  26   26  26   27  29  26   26   26
#> am    26  27   26  26   27  27   26  26  29   27   26
#> gear  26  26   27  26   26  27   26  26  27   29   26
#> carb  27  27   26  26   26  27   26  26  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .04  .00  .00  .01  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .00  .00  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .02
#> hp    .00  .00  .00       .02  .00  .00  .00  .09  .47  .00
#> drat  .00  .00  .00  .02       .00  .49  .01  .00  .00  .22
#> wt    .00  .00  .00  .00  .00       .31  .00  .00  .00  .01
#> qsec  .04  .00  .03  .00  .49  .31       .00  .54  .35  .00
#> vs    .00  .00  .00  .00  .01  .00  .00       .16  .26  .01
#> am    .00  .00  .00  .09  .00  .00  .54  .16       .00  .48
#> gear  .01  .00  .00  .47  .00  .00  .35  .26  .00       .54
#> carb  .00  .00  .02  .00  .22  .01  .00  .01  .48  .54     
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl   disp     hp   drat    wt  qsec     vs    am  gear carb
#> mpg      1                                                               
#> cyl   -.87     1                                                         
#> disp  -.85   .90   1.00                                                  
#> hp    -.77   .83    .75   1.00                                           
#> drat   .69  -.73   -.74   -.47   1.00                                    
#> wt    -.91   .83    .88    .67   -.77     1                              
#> qsec   .40  -.59   -.42   -.72    .14  -.20     1                        
#> vs     .64  -.81   -.70   -.70    .52  -.56   .75   1.00                 
#> am     .71  -.71   -.69   -.34    .73  -.74  -.13    .28     1           
#> gear   .53  -.57   -.65   -.15    .70  -.61  -.19    .23   .84     1     
#> carb  -.63   .62    .47    .79   -.25   .52  -.63   -.52  -.15   .13    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  27  26   29                                    
#> hp    26  26   26  29                                
#> drat  26  27   26  26   29                           
#> wt    26  26   26  27   26  29                       
#> qsec  26  26   26  27   26  27   29                  
#> vs    27  26   26  26   26  26   27  29              
#> am    26  27   26  26   27  27   26  26  29          
#> gear  26  26   27  26   26  27   26  26  27   29     
#> carb  27  27   26  26   26  27   26  26  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .02                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .04  .00  .03  .00  .49  .31                         
#> vs    .00  .00  .00  .00  .01  .00  .00                    
#> am    .00  .00  .00  .09  .00  .00  .54  .16               
#> gear  .01  .00  .00  .47  .00  .00  .35  .26  .00          
#> carb  .00  .00  .02  .00  .22  .01  .00  .01  .48  .54     
#>