(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

Notes: weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran)

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, ...)

Arguments

X
a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x
an object of class 'pwcor' / 'pwcov'.

w
numeric. A vector of (frequency) weights.

N
logical. TRUE also computes pairwise observation counts.

P
logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array
logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use
argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits
integer. The number of digits to round to in print.

sig.level
numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show
character. The part of the correlation / covariance matrix to display.

spacing
integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

...
other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)

#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.88  -.83  -.77   .68  -.87   .41   .70   .51   .47  -.54
#> cyl   -.88    1    .90   .84  -.65   .83  -.55  -.79  -.48  -.43   .55
#> disp  -.83   .90    1    .77  -.67   .90  -.42  -.69  -.56  -.58   .24
#> hp    -.77   .84   .77    1   -.41   .71  -.70  -.70  -.09  -.03   .75
#> drat   .68  -.65  -.67  -.41    1   -.75   .03   .47   .64   .66   .01
#> wt    -.87   .83   .90   .71  -.75    1   -.15  -.59  -.67  -.61   .39
#> qsec   .41  -.55  -.42  -.70   .03  -.15    1    .70  -.25  -.22  -.64
#> vs     .70  -.79  -.69  -.70   .47  -.59   .70    1    .21   .24  -.51
#> am     .51  -.48  -.56  -.09   .64  -.67  -.25   .21    1    .81   .26
#> gear   .47  -.43  -.58  -.03   .66  -.61  -.22   .24   .81    1    .36
#> carb  -.54   .55   .24   .75   .01   .39  -.64  -.51   .26   .36    1 

pwcov(mna)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.10     -9.96   -556.70   -345.34      2.19     -5.91      4.15
#> cyl      -9.96      3.12    183.72    103.57     -0.60      1.59     -1.45
#> disp   -556.70    183.72  13664.50   6142.45    -41.79    103.87    -87.35
#> hp     -345.34    103.57   6142.45   4721.22    -15.39     49.71    -88.51
#> drat      2.19     -0.60    -41.79    -15.39      0.27     -0.41      0.02
#> wt       -5.91      1.59    103.87     49.71     -0.41      1.06     -0.26
#> qsec      4.15     -1.45    -87.35    -88.51      0.02     -0.26      3.16
#> vs        2.20     -0.71    -40.36    -25.02      0.12     -0.32      0.62
#> am        1.40     -0.42    -32.25     -3.15      0.16     -0.34     -0.22
#> gear      2.32     -0.60    -53.91     -1.51      0.27     -0.46     -0.30
#> carb     -5.15      1.61     42.30     85.65      0.01      0.61     -1.96
#>             vs        am      gear      carb
#> mpg       2.20      1.40      2.32     -5.15
#> cyl      -0.71     -0.42     -0.60      1.61
#> disp    -40.36    -32.25    -53.91     42.30
#> hp      -25.02     -3.15     -1.51     85.65
#> drat      0.12      0.16      0.27      0.01
#> wt       -0.32     -0.34     -0.46      0.61
#> qsec      0.62     -0.22     -0.30     -1.96
#> vs        0.25      0.05      0.09     -0.42
#> am        0.05      0.24      0.30      0.21
#> gear      0.09      0.30      0.58      0.48
#> carb     -0.42      0.21      0.48      2.69

pwnobs(mna)

#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 26   27 26   26 27 26   27   26
#> cyl   26  29   26 27   27 27   26 26 26   26   26
#> disp  26  26   29 26   26 26   27 27 26   26   26
#> hp    26  27   26 29   26 27   26 26 26   26   27
#> drat  27  27   26 26   29 26   26 26 26   27   26
#> wt    26  27   26 27   26 29   26 26 26   27   26
#> qsec  26  26   27 26   26 26   29 27 28   26   26
#> vs    27  26   27 26   26 26   27 29 26   26   26
#> am    26  26   26 26   26 26   28 26 29   26   26
#> gear  27  26   26 26   27 27   26 26 26   29   26
#> carb  26  26   26 27   26 26   26 26 26   26   29

pwcor(mna, N = TRUE)

#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.88 (26) -.83 (26) -.77 (26)  .68 (27) -.87 (26)  .41 (26)
#> cyl  -.88 (26)   1  (29)  .90 (26)  .84 (27) -.65 (27)  .83 (27) -.55 (26)
#> disp -.83 (26)  .90 (26)   1  (29)  .77 (26) -.67 (26)  .90 (26) -.42 (27)
#> hp   -.77 (26)  .84 (27)  .77 (26)   1  (29) -.41 (26)  .71 (27) -.70 (26)
#> drat  .68 (27) -.65 (27) -.67 (26) -.41 (26)   1  (29) -.75 (26)  .03 (26)
#> wt   -.87 (26)  .83 (27)  .90 (26)  .71 (27) -.75 (26)   1  (29) -.15 (26)
#> qsec  .41 (26) -.55 (26) -.42 (27) -.70 (26)  .03 (26) -.15 (26)   1  (29)
#> vs    .70 (27) -.79 (26) -.69 (27) -.70 (26)  .47 (26) -.59 (26)  .70 (27)
#> am    .51 (26) -.48 (26) -.56 (26) -.09 (26)  .64 (26) -.67 (26) -.25 (28)
#> gear  .47 (27) -.43 (26) -.58 (26) -.03 (26)  .66 (27) -.61 (27) -.22 (26)
#> carb -.54 (26)  .55 (26)  .24 (26)  .75 (27)  .01 (26)  .39 (26) -.64 (26)
#>             vs        am      gear      carb
#> mpg   .70 (27)  .51 (26)  .47 (27) -.54 (26)
#> cyl  -.79 (26) -.48 (26) -.43 (26)  .55 (26)
#> disp -.69 (27) -.56 (26) -.58 (26)  .24 (26)
#> hp   -.70 (26) -.09 (26) -.03 (26)  .75 (27)
#> drat  .47 (26)  .64 (26)  .66 (27)  .01 (26)
#> wt   -.59 (26) -.67 (26) -.61 (27)  .39 (26)
#> qsec  .70 (27) -.25 (28) -.22 (26) -.64 (26)
#> vs     1  (29)  .21 (26)  .24 (26) -.51 (26)
#> am    .21 (26)   1  (29)  .81 (26)  .26 (26)
#> gear  .24 (26)  .81 (26)   1  (29)  .36 (26)
#> carb -.51 (26)  .26 (26)  .36 (26)   1  (29)

pwcor(mna, P = TRUE)

#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.88*  -.83*  -.77*   .68*  -.87*   .41*   .70*   .51*   .47*
#> cyl   -.88*    1     .90*   .84*  -.65*   .83*  -.55*  -.79*  -.48*  -.43*
#> disp  -.83*   .90*    1     .77*  -.67*   .90*  -.42*  -.69*  -.56*  -.58*
#> hp    -.77*   .84*   .77*    1    -.41*   .71*  -.70*  -.70*  -.09   -.03 
#> drat   .68*  -.65*  -.67*  -.41*    1    -.75*   .03    .47*   .64*   .66*
#> wt    -.87*   .83*   .90*   .71*  -.75*    1    -.15   -.59*  -.67*  -.61*
#> qsec   .41*  -.55*  -.42*  -.70*   .03   -.15     1     .70*  -.25   -.22 
#> vs     .70*  -.79*  -.69*  -.70*   .47*  -.59*   .70*    1     .21    .24 
#> am     .51*  -.48*  -.56*  -.09    .64*  -.67*  -.25    .21     1     .81*
#> gear   .47*  -.43*  -.58*  -.03    .66*  -.61*  -.22    .24    .81*    1  
#> carb  -.54*   .55*   .24    .75*   .01    .39*  -.64*  -.51*   .26    .36 
#>        carb
#> mpg   -.54*
#> cyl    .55*
#> disp   .24 
#> hp     .75*
#> drat   .01 
#> wt     .39*
#> qsec  -.64*
#> vs    -.51*
#> am     .26 
#> gear   .36 
#> carb    1  

pwcor(mna, N = TRUE, P = TRUE)

#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.88* (26) -.83* (26) -.77* (26)  .68* (27) -.87* (26)
#> cyl  -.88* (26)   1   (29)  .90* (26)  .84* (27) -.65* (27)  .83* (27)
#> disp -.83* (26)  .90* (26)   1   (29)  .77* (26) -.67* (26)  .90* (26)
#> hp   -.77* (26)  .84* (27)  .77* (26)   1   (29) -.41* (26)  .71* (27)
#> drat  .68* (27) -.65* (27) -.67* (26) -.41* (26)   1   (29) -.75* (26)
#> wt   -.87* (26)  .83* (27)  .90* (26)  .71* (27) -.75* (26)   1   (29)
#> qsec  .41* (26) -.55* (26) -.42* (27) -.70* (26)  .03  (26) -.15  (26)
#> vs    .70* (27) -.79* (26) -.69* (27) -.70* (26)  .47* (26) -.59* (26)
#> am    .51* (26) -.48* (26) -.56* (26) -.09  (26)  .64* (26) -.67* (26)
#> gear  .47* (27) -.43* (26) -.58* (26) -.03  (26)  .66* (27) -.61* (27)
#> carb -.54* (26)  .55* (26)  .24  (26)  .75* (27)  .01  (26)  .39* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .41* (26)  .70* (27)  .51* (26)  .47* (27) -.54* (26)
#> cyl  -.55* (26) -.79* (26) -.48* (26) -.43* (26)  .55* (26)
#> disp -.42* (27) -.69* (27) -.56* (26) -.58* (26)  .24  (26)
#> hp   -.70* (26) -.70* (26) -.09  (26) -.03  (26)  .75* (27)
#> drat  .03  (26)  .47* (26)  .64* (26)  .66* (27)  .01  (26)
#> wt   -.15  (26) -.59* (26) -.67* (26) -.61* (27)  .39* (26)
#> qsec   1   (29)  .70* (27) -.25  (28) -.22  (26) -.64* (26)
#> vs    .70* (27)   1   (29)  .21  (26)  .24  (26) -.51* (26)
#> am   -.25  (28)  .21  (26)   1   (29)  .81* (26)  .26  (26)
#> gear -.22  (26)  .24  (26)  .81* (26)   1   (29)  .36  (26)
#> carb -.64* (26) -.51* (26)  .26  (26)  .36  (26)   1   (29)

aperm(pwcor(mna, N = TRUE, P = TRUE))

#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.88  -.83  -.77   .68  -.87   .41   .70   .51   .47  -.54
#> N  29    26    26    26    27    26    26    27    26    27    26
#> P       .00   .00   .00   .00   .00   .04   .00   .01   .01   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.88   1   .90   .84  -.65   .83  -.55  -.79  -.48  -.43   .55
#> N    26  29    26    27    27    27    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .03   .00
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   .90    1   .77  -.67   .90  -.42  -.69  -.56  -.58   .24
#> N    26    26   29    26    26    26    27    27    26    26    26
#> P   .00   .00        .00   .00   .00   .03   .00   .00   .00   .24
#> 
#> , , hp
#> 
#>     mpg   cyl  disp  hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.77   .84   .77   1  -.41   .71  -.70  -.70  -.09  -.03   .75
#> N    26    27    26  29    26    27    26    26    26    26    27
#> P   .00   .00   .00       .04   .00   .00   .00   .66   .89   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .68  -.65  -.67  -.41    1  -.75   .03   .47   .64   .66   .01
#> N    27    27    26    26   29    26    26    26    26    27    26
#> P   .00   .00   .00   .04        .00   .90   .02   .00   .00   .97
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.87   .83   .90   .71  -.75   1  -.15  -.59  -.67  -.61   .39
#> N    26    27    26    27    26  29    26    26    26    27    26
#> P   .00   .00   .00   .00   .00       .47   .00   .00   .00   .05
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .41  -.55  -.42  -.70   .03  -.15    1   .70  -.25  -.22  -.64
#> N    26    26    27    26    26    26   29    27    28    26    26
#> P   .04   .00   .03   .00   .90   .47        .00   .21   .29   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec  vs    am  gear  carb
#> r   .70  -.79  -.69  -.70   .47  -.59   .70   1   .21   .24  -.51
#> N    27    26    27    26    26    26    27  29    26    26    26
#> P   .00   .00   .00   .00   .02   .00   .00       .30   .23   .01
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .51  -.48  -.56  -.09   .64  -.67  -.25   .21   1   .81   .26
#> N    26    26    26    26    26    26    28    26  29    26    26
#> P   .01   .01   .00   .66   .00   .00   .21   .30       .00   .20
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .47  -.43  -.58  -.03   .66  -.61  -.22   .24   .81    1   .36
#> N    27    26    26    26    27    27    26    26    26   29    26
#> P   .01   .03   .00   .89   .00   .00   .29   .23   .00        .07
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.54   .55   .24   .75   .01   .39  -.64  -.51   .26   .36    1
#> N    26    26    26    27    26    26    26    26    26    26   29
#> P   .00   .00   .24   .00   .97   .05   .00   .01   .20   .07     
#> 

print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")

#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.878* (26)   1    (29)                                                
#> disp -.830* (26)  .899* (26)   1    (29)                                    
#> hp   -.771* (26)  .839* (27)  .767* (26)   1    (29)                        
#> drat  .685* (27) -.649* (27) -.671* (26) -.407  (26)   1    (29)            
#> wt   -.870* (26)  .832* (27)  .904* (26)  .706* (27) -.748* (26)   1    (29)
#> qsec  .408  (26) -.550* (26) -.421  (27) -.700* (26)  .026  (26) -.147  (26)
#> vs    .699* (27) -.788* (26) -.692* (27) -.696* (26)  .471  (26) -.593* (26)
#> am    .511* (26) -.484  (26) -.557* (26) -.091  (26)  .639* (26) -.675* (26)
#> gear  .471  (27) -.431  (26) -.584* (26) -.028  (26)  .663* (27) -.611* (27)
#> carb -.537* (26)  .545* (26)  .240  (26)  .755* (27)  .007  (26)  .392  (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .697* (27)   1    (29)                                    
#> am   -.247  (28)  .212  (26)   1    (29)                        
#> gear -.217  (26)  .243  (26)  .807* (26)   1    (29)            
#> carb -.643* (26) -.512* (26)  .262  (26)  .364  (26)   1    (29)

pwcor(mna, N = TRUE, P = TRUE, array = FALSE)

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.88  -.83  -.77   .68  -.87   .41   .70   .51   .47  -.54
#> cyl   -.88     1   .90   .84  -.65   .83  -.55  -.79  -.48  -.43   .55
#> disp  -.83   .90     1   .77  -.67   .90  -.42  -.69  -.56  -.58   .24
#> hp    -.77   .84   .77     1  -.41   .71  -.70  -.70  -.09  -.03   .75
#> drat   .68  -.65  -.67  -.41     1  -.75   .03   .47   .64   .66   .01
#> wt    -.87   .83   .90   .71  -.75     1  -.15  -.59  -.67  -.61   .39
#> qsec   .41  -.55  -.42  -.70   .03  -.15     1   .70  -.25  -.22  -.64
#> vs     .70  -.79  -.69  -.70   .47  -.59   .70     1   .21   .24  -.51
#> am     .51  -.48  -.56  -.09   .64  -.67  -.25   .21     1   .81   .26
#> gear   .47  -.43  -.58  -.03   .66  -.61  -.22   .24   .81     1   .36
#> carb  -.54   .55   .24   .75   .01   .39  -.64  -.51   .26   .36     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  26   27  26   26  27  26   27   26
#> cyl   26  29   26  27   27  27   26  26  26   26   26
#> disp  26  26   29  26   26  26   27  27  26   26   26
#> hp    26  27   26  29   26  27   26  26  26   26   27
#> drat  27  27   26  26   29  26   26  26  26   27   26
#> wt    26  27   26  27   26  29   26  26  26   27   26
#> qsec  26  26   27  26   26  26   29  27  28   26   26
#> vs    27  26   27  26   26  26   27  29  26   26   26
#> am    26  26   26  26   26  26   28  26  29   26   26
#> gear  27  26   26  26   27  27   26  26  26   29   26
#> carb  26  26   26  27   26  26   26  26  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .04  .00  .01  .01  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .03  .00
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .00  .24
#> hp    .00  .00  .00       .04  .00  .00  .00  .66  .89  .00
#> drat  .00  .00  .00  .04       .00  .90  .02  .00  .00  .97
#> wt    .00  .00  .00  .00  .00       .47  .00  .00  .00  .05
#> qsec  .04  .00  .03  .00  .90  .47       .00  .21  .29  .00
#> vs    .00  .00  .00  .00  .02  .00  .00       .30  .23  .01
#> am    .01  .01  .00  .66  .00  .00  .21  .30       .00  .20
#> gear  .01  .03  .00  .89  .00  .00  .29  .23  .00       .07
#> carb  .00  .00  .24  .00  .97  .05  .00  .01  .20  .07     
#> 

print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")

#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.88     1                                                     
#> disp  -.83   .90     1                                               
#> hp    -.77   .84   .77     1                                         
#> drat   .68  -.65  -.67  -.41     1                                   
#> wt    -.87   .83   .90   .71  -.75     1                             
#> qsec   .41  -.55  -.42  -.70   .03  -.15     1                       
#> vs     .70  -.79  -.69  -.70   .47  -.59   .70     1                 
#> am     .51  -.48  -.56  -.09   .64  -.67  -.25   .21     1           
#> gear   .47  -.43  -.58  -.03   .66  -.61  -.22   .24   .81     1     
#> carb  -.54   .55   .24   .75   .01   .39  -.64  -.51   .26   .36    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  26   29                                    
#> hp    26  27   26  29                                
#> drat  27  27   26  26   29                           
#> wt    26  27   26  27   26  29                       
#> qsec  26  26   27  26   26  26   29                  
#> vs    27  26   27  26   26  26   27  29              
#> am    26  26   26  26   26  26   28  26  29          
#> gear  27  26   26  26   27  27   26  26  26   29     
#> carb  26  26   26  27   26  26   26  26  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .04                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .04  .00  .03  .00  .90  .47                         
#> vs    .00  .00  .00  .00  .02  .00  .00                    
#> am    .01  .01  .00  .66  .00  .00  .21  .30               
#> gear  .01  .03  .00  .89  .00  .00  .29  .23  .00          
#> carb  .00  .00  .24  .00  .97  .05  .00  .01  .20  .07     
#>