(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb     X
#> mpg     1   -.85  -.83  -.81   .69  -.88   .41   .66   .56   .42  -.53   .19
#> cyl   -.85    1    .91   .86  -.70   .75  -.60  -.78  -.49  -.43   .55  -.28
#> disp  -.83   .91    1    .86  -.72   .86  -.42  -.71  -.60  -.52   .47  -.04
#> hp    -.81   .86   .86    1   -.54   .66  -.66  -.77  -.41  -.30   .72  -.09
#> drat   .69  -.70  -.72  -.54    1   -.76   .02   .37   .80   .71  -.12   .17
#> wt    -.88   .75   .86   .66  -.76    1   -.09  -.52  -.72  -.61   .35  -.21
#> qsec   .41  -.60  -.42  -.66   .02  -.09    1    .74  -.30  -.45  -.71   .26
#> vs     .66  -.78  -.71  -.77   .37  -.52   .74    1    .12   .12  -.61   .27
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      29.20     -7.59   -552.87   -277.60      2.10     -4.00      4.31
#> cyl      -7.59      3.12    197.81     92.71     -0.63      1.03     -2.05
#> disp   -552.87    197.81  16478.67   7271.39    -52.61     99.80    -99.93
#> hp     -277.60     92.71   7271.39   3775.22    -19.19     36.68    -74.88
#> drat      2.10     -0.63    -52.61    -19.19      0.31     -0.40      0.02
#> wt       -4.00      1.03     99.80     36.68     -0.40      0.81     -0.16
#> qsec      4.31     -2.05    -99.93    -74.88      0.02     -0.16      3.31
#> vs        1.85     -0.68    -46.31    -23.81      0.10     -0.24      0.68
#>             vs        am      gear      carb         X
#> mpg       1.85      1.59      1.82     -4.81      1.19
#> cyl      -0.68     -0.44     -0.58      1.71     -0.57
#> disp    -46.31    -38.94    -52.59    106.07     -6.57
#> hp      -23.81    -13.01    -13.10     60.66     -5.88
#> drat      0.10      0.23      0.32     -0.11      0.10
#> wt       -0.24     -0.34     -0.44      0.52     -0.20
#> qsec      0.68     -0.28     -0.55     -2.30      0.43
#> vs        0.24      0.03      0.04     -0.51      0.13
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb  X
#> mpg   29  28   27 26   26 27   27 27 26   26   26 26
#> cyl   28  29   27 26   26 26   26 26 26   26   27 26
#> disp  27  27   29 26   26 26   27 26 26   26   26 26
#> hp    26  26   26 29   26 26   26 26 26   27   26 26
#> drat  26  26   26 26   29 26   27 27 26   27   27 27
#> wt    27  26   26 26   26 29   27 27 27   26   26 26
#> qsec  27  26   27 26   27 27   29 27 26   26   26 27
#> vs    27  26   26 26   27 27   27 29 26   26   27 26
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.85 (28) -.83 (27) -.81 (26)  .69 (26) -.88 (27)  .41 (27)
#> cyl  -.85 (28)   1  (29)  .91 (27)  .86 (26) -.70 (26)  .75 (26) -.60 (26)
#> disp -.83 (27)  .91 (27)   1  (29)  .86 (26) -.72 (26)  .86 (26) -.42 (27)
#> hp   -.81 (26)  .86 (26)  .86 (26)   1  (29) -.54 (26)  .66 (26) -.66 (26)
#> drat  .69 (26) -.70 (26) -.72 (26) -.54 (26)   1  (29) -.76 (26)  .02 (27)
#> wt   -.88 (27)  .75 (26)  .86 (26)  .66 (26) -.76 (26)   1  (29) -.09 (27)
#> qsec  .41 (27) -.60 (26) -.42 (27) -.66 (26)  .02 (27) -.09 (27)   1  (29)
#> vs    .66 (27) -.78 (26) -.71 (26) -.77 (26)  .37 (27) -.52 (27)  .74 (27)
#>             vs        am      gear      carb         X
#> mpg   .66 (27)  .56 (26)  .42 (26) -.53 (26)  .19 (26)
#> cyl  -.78 (26) -.49 (26) -.43 (26)  .55 (27) -.28 (26)
#> disp -.71 (26) -.60 (26) -.52 (26)  .47 (26) -.04 (26)
#> hp   -.77 (26) -.41 (26) -.30 (27)  .72 (26) -.09 (26)
#> drat  .37 (27)  .80 (26)  .71 (27) -.12 (27)  .17 (27)
#> wt   -.52 (27) -.72 (27) -.61 (26)  .35 (26) -.21 (26)
#> qsec  .74 (27) -.30 (26) -.45 (26) -.71 (26)  .26 (27)
#> vs     1  (29)  .12 (26)  .12 (26) -.61 (27)  .27 (26)
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.85*  -.83*  -.81*   .69*  -.88*   .41*   .66*   .56*   .42*
#> cyl   -.85*    1     .91*   .86*  -.70*   .75*  -.60*  -.78*  -.49*  -.43*
#> disp  -.83*   .91*    1     .86*  -.72*   .86*  -.42*  -.71*  -.60*  -.52*
#> hp    -.81*   .86*   .86*    1    -.54*   .66*  -.66*  -.77*  -.41*  -.30 
#> drat   .69*  -.70*  -.72*  -.54*    1    -.76*   .02    .37    .80*   .71*
#> wt    -.88*   .75*   .86*   .66*  -.76*    1    -.09   -.52*  -.72*  -.61*
#> qsec   .41*  -.60*  -.42*  -.66*   .02   -.09     1     .74*  -.30   -.45*
#> vs     .66*  -.78*  -.71*  -.77*   .37   -.52*   .74*    1     .12    .12 
#>        carb      X
#> mpg   -.53*   .19 
#> cyl    .55*  -.28 
#> disp   .47*  -.04 
#> hp     .72*  -.09 
#> drat  -.12    .17 
#> wt     .35   -.21 
#> qsec  -.71*   .26 
#> vs    -.61*   .27 
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.85* (28) -.83* (27) -.81* (26)  .69* (26) -.88* (27)
#> cyl  -.85* (28)   1   (29)  .91* (27)  .86* (26) -.70* (26)  .75* (26)
#> disp -.83* (27)  .91* (27)   1   (29)  .86* (26) -.72* (26)  .86* (26)
#> hp   -.81* (26)  .86* (26)  .86* (26)   1   (29) -.54* (26)  .66* (26)
#> drat  .69* (26) -.70* (26) -.72* (26) -.54* (26)   1   (29) -.76* (26)
#> wt   -.88* (27)  .75* (26)  .86* (26)  .66* (26) -.76* (26)   1   (29)
#> qsec  .41* (27) -.60* (26) -.42* (27) -.66* (26)  .02  (27) -.09  (27)
#> vs    .66* (27) -.78* (26) -.71* (26) -.77* (26)  .37  (27) -.52* (27)
#>            qsec         vs         am       gear       carb          X
#> mpg   .41* (27)  .66* (27)  .56* (26)  .42* (26) -.53* (26)  .19  (26)
#> cyl  -.60* (26) -.78* (26) -.49* (26) -.43* (26)  .55* (27) -.28  (26)
#> disp -.42* (27) -.71* (26) -.60* (26) -.52* (26)  .47* (26) -.04  (26)
#> hp   -.66* (26) -.77* (26) -.41* (26) -.30  (27)  .72* (26) -.09  (26)
#> drat  .02  (27)  .37  (27)  .80* (26)  .71* (27) -.12  (27)  .17  (27)
#> wt   -.09  (27) -.52* (27) -.72* (27) -.61* (26)  .35  (26) -.21  (26)
#> qsec   1   (29)  .74* (27) -.30  (26) -.45* (26) -.71* (26)  .26  (27)
#> vs    .74* (27)   1   (29)  .12  (26)  .12  (26) -.61* (27)  .27  (26)
#>  [ reached getOption("max.print") -- omitted 4 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb     X
#> r   1  -.85  -.83  -.81   .69  -.88   .41   .66   .56   .42  -.53   .19
#> N  29    28    27    26    26    27    27    27    26    26    26    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .03   .01   .35
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb     X
#> r  -.85   1   .91   .86  -.70   .75  -.60  -.78  -.49  -.43   .55  -.28
#> N    28  29    27    26    26    26    26    26    26    26    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .01   .03   .00   .17
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb     X
#> r  -.83   .91    1   .86  -.72   .86  -.42  -.71  -.60  -.52   .47  -.04
#> N    27    27   29    26    26    26    27    26    26    26    26    26
#> 
#>  [ reached getOption("max.print") -- omitted 1 row(s) and 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.847* (28)   1    (29)                                                
#> disp -.830* (27)  .913* (27)   1    (29)                                    
#> hp   -.807* (26)  .857* (26)  .857* (26)   1    (29)                        
#> drat  .693* (26) -.701* (26) -.722* (26) -.538* (26)   1    (29)            
#> wt   -.876* (27)  .753* (26)  .865* (26)  .658* (26) -.757* (26)   1    (29)
#> qsec  .415  (27) -.599* (26) -.415  (27) -.662* (26)  .017  (27) -.093  (27)
#> vs    .661* (27) -.782* (26) -.712* (26) -.769* (26)  .371  (27) -.523* (27)
#>             qsec          vs          am        gear        carb           X
#> mpg                                                                         
#> cyl                                                                         
#> disp                                                                        
#> hp                                                                          
#> drat                                                                        
#> wt                                                                          
#> qsec   1    (29)                                                            
#> vs    .738* (27)   1    (29)                                                
#>  [ reached getOption("max.print") -- omitted 4 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat     wt  qsec    vs     am  gear   carb     X
#> mpg      1  -.85  -.83  -.81   .69   -.88   .41   .66    .56   .42   -.53   .19
#> cyl   -.85     1   .91   .86  -.70    .75  -.60  -.78   -.49  -.43    .55  -.28
#> disp  -.83   .91     1   .86  -.72    .86  -.42  -.71   -.60  -.52    .47  -.04
#> hp    -.81   .86   .86     1  -.54    .66  -.66  -.77   -.41  -.30    .72  -.09
#> drat   .69  -.70  -.72  -.54     1   -.76   .02   .37    .80   .71   -.12   .17
#> wt    -.88   .75   .86   .66  -.76   1.00  -.09  -.52   -.72  -.61    .35  -.21
#> qsec   .41  -.60  -.42  -.66   .02   -.09     1   .74   -.30  -.45   -.71   .26
#> vs     .66  -.78  -.71  -.77   .37   -.52   .74     1    .12   .12   -.61   .27
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb   X
#> mpg   29  28   27  26   26  27   27  27  26   26   26  26
#> cyl   28  29   27  26   26  26   26  26  26   26   27  26
#> disp  27  27   29  26   26  26   27  26  26   26   26  26
#> hp    26  26   26  29   26  26   26  26  26   27   26  26
#> drat  26  26   26  26   29  26   27  27  26   27   27  27
#> wt    27  26   26  26   26  29   27  27  27   26   26  26
#> qsec  27  26   27  26   27  27   29  27  26   26   26  27
#> vs    27  26   26  26   27  27   27  29  26   26   27  26
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb    X
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .03  .01  .35
#> cyl   .00       .00  .00  .00  .00  .00  .00  .01  .03  .00  .17
#> disp  .00  .00       .00  .00  .00  .03  .00  .00  .01  .02  .83
#> hp    .00  .00  .00       .00  .00  .00  .00  .04  .13  .00  .65
#> drat  .00  .00  .00  .00       .00  .93  .06  .00  .00  .54  .38
#> wt    .00  .00  .00  .00  .00       .65  .01  .00  .00  .08  .31
#> qsec  .03  .00  .03  .00  .93  .65       .00  .14  .02  .00  .19
#> vs    .00  .00  .00  .00  .06  .01  .00       .55  .57  .00  .19
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat     wt  qsec    vs     am  gear   carb   X
#> mpg      1                                                                   
#> cyl   -.85     1                                                             
#> disp  -.83   .91     1                                                       
#> hp    -.81   .86   .86     1                                                 
#> drat   .69  -.70  -.72  -.54     1                                           
#> wt    -.88   .75   .86   .66  -.76   1.00                                    
#> qsec   .41  -.60  -.42  -.66   .02   -.09     1                              
#> vs     .66  -.78  -.71  -.77   .37   -.52   .74     1                        
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb   X
#> mpg   29                                                 
#> cyl   28  29                                             
#> disp  27  27   29                                        
#> hp    26  26   26  29                                    
#> drat  26  26   26  26   29                               
#> wt    27  26   26  26   26  29                           
#> qsec  27  26   27  26   27  27   29                      
#> vs    27  26   26  26   27  27   27  29                  
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb X
#> mpg                                                          
#> cyl   .00                                                    
#> disp  .00  .00                                               
#> hp    .00  .00  .00                                          
#> drat  .00  .00  .00  .00                                     
#> wt    .00  .00  .00  .00  .00                                
#> qsec  .03  .00  .03  .00  .93  .65                           
#> vs    .00  .00  .00  .00  .06  .01  .00                      
#>  [ reached getOption("max.print") -- omitted 4 rows ]
#>