(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.86  -.84  -.78   .72  -.90   .40   .63   .78   .70  -.55
#> cyl   -.86    1    .90   .81  -.69   .78  -.49  -.75  -.69  -.62   .45
#> disp  -.84   .90    1    .77  -.64   .88  -.41  -.69  -.63  -.58   .35
#> hp    -.78   .81   .77    1   -.38   .65  -.73  -.68  -.48  -.26   .75
#> drat   .72  -.69  -.64  -.38    1   -.71  -.01   .34   .77   .75  -.10
#> wt    -.90   .78   .88   .65  -.71    1   -.21  -.57  -.75  -.65   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.39     -9.11   -539.11   -308.87      2.50     -5.62      3.58
#> cyl      -9.11      3.06    178.04     98.63     -0.70      1.34     -1.31
#> disp   -539.11    178.04  14040.56   6693.62    -39.99     99.71    -78.22
#> hp     -308.87     98.63   6693.62   4900.15    -15.00     49.60    -81.44
#> drat      2.50     -0.70    -39.99    -15.00      0.29     -0.38     -0.01
#> wt       -5.62      1.34     99.71     49.60     -0.38      1.04     -0.36
#>             vs        am      gear      carb
#> mpg       1.96      2.48      2.65     -5.52
#> cyl      -0.63     -0.62     -0.78      1.29
#> disp    -40.36    -38.74    -47.50     71.25
#> hp      -23.11    -14.67    -11.18     88.59
#> drat      0.08      0.22      0.31     -0.09
#> wt       -0.29     -0.40     -0.45      0.82
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  27   26 26   26 27   26 26 26   26   26
#> cyl   27  29   26 26   26 26   26 26 26   26   26
#> disp  26  26   29 27   26 26   26 26 26   26   26
#> hp    26  26   27 29   26 26   26 26 26   26   26
#> drat  26  26   26 26   29 27   26 26 26   27   26
#> wt    27  26   26 26   27 29   26 27 26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.86 (27) -.84 (26) -.78 (26)  .72 (26) -.90 (27)  .40 (26)
#> cyl  -.86 (27)   1  (29)  .90 (26)  .81 (26) -.69 (26)  .78 (26) -.49 (26)
#> disp -.84 (26)  .90 (26)   1  (29)  .77 (27) -.64 (26)  .88 (26) -.41 (26)
#> hp   -.78 (26)  .81 (26)  .77 (27)   1  (29) -.38 (26)  .65 (26) -.73 (26)
#> drat  .72 (26) -.69 (26) -.64 (26) -.38 (26)   1  (29) -.71 (27) -.01 (26)
#> wt   -.90 (27)  .78 (26)  .88 (26)  .65 (26) -.71 (27)   1  (29) -.21 (26)
#>             vs        am      gear      carb
#> mpg   .63 (26)  .78 (26)  .70 (26) -.55 (26)
#> cyl  -.75 (26) -.69 (26) -.62 (26)  .45 (26)
#> disp -.69 (26) -.63 (26) -.58 (26)  .35 (26)
#> hp   -.68 (26) -.48 (26) -.26 (26)  .75 (26)
#> drat  .34 (26)  .77 (26)  .75 (27) -.10 (26)
#> wt   -.57 (27) -.75 (26) -.65 (26)  .49 (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.86*  -.84*  -.78*   .72*  -.90*   .40*   .63*   .78*   .70*
#> cyl   -.86*    1     .90*   .81*  -.69*   .78*  -.49*  -.75*  -.69*  -.62*
#> disp  -.84*   .90*    1     .77*  -.64*   .88*  -.41*  -.69*  -.63*  -.58*
#> hp    -.78*   .81*   .77*    1    -.38    .65*  -.73*  -.68*  -.48*  -.26 
#> drat   .72*  -.69*  -.64*  -.38     1    -.71*  -.01    .34    .77*   .75*
#> wt    -.90*   .78*   .88*   .65*  -.71*    1    -.21   -.57*  -.75*  -.65*
#>        carb
#> mpg   -.55*
#> cyl    .45*
#> disp   .35 
#> hp     .75*
#> drat  -.10 
#> wt     .49*
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.86* (27) -.84* (26) -.78* (26)  .72* (26) -.90* (27)
#> cyl  -.86* (27)   1   (29)  .90* (26)  .81* (26) -.69* (26)  .78* (26)
#> disp -.84* (26)  .90* (26)   1   (29)  .77* (27) -.64* (26)  .88* (26)
#> hp   -.78* (26)  .81* (26)  .77* (27)   1   (29) -.38  (26)  .65* (26)
#> drat  .72* (26) -.69* (26) -.64* (26) -.38  (26)   1   (29) -.71* (27)
#> wt   -.90* (27)  .78* (26)  .88* (26)  .65* (26) -.71* (27)   1   (29)
#>            qsec         vs         am       gear       carb
#> mpg   .40* (26)  .63* (26)  .78* (26)  .70* (26) -.55* (26)
#> cyl  -.49* (26) -.75* (26) -.69* (26) -.62* (26)  .45* (26)
#> disp -.41* (26) -.69* (26) -.63* (26) -.58* (26)  .35  (26)
#> hp   -.73* (26) -.68* (26) -.48* (26) -.26  (26)  .75* (26)
#> drat -.01  (26)  .34  (26)  .77* (26)  .75* (27) -.10  (26)
#> wt   -.21  (26) -.57* (27) -.75* (26) -.65* (26)  .49* (26)
#>  [ reached getOption("max.print") -- omitted 5 rows ]
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.86  -.84  -.78   .72  -.90   .40   .63   .78   .70  -.55
#> N  29    27    26    26    26    27    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .05   .00   .00   .00   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.86   1   .90   .81  -.69   .78  -.49  -.75  -.69  -.62   .45
#> N    27  29    26    26    26    26    26    26    26    26    26
#> P   .00       .00   .00   .00   .00   .01   .00   .00   .00   .02
#> 
#>  [ reached getOption("max.print") -- omitted 9 matrix slice(s) ]
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.856* (27)   1    (29)                                                
#> disp -.840* (26)  .902* (26)   1    (29)                                    
#> hp   -.778* (26)  .805* (26)  .772* (27)   1    (29)                        
#> drat  .724* (26) -.685* (26) -.645* (26) -.379  (26)   1    (29)            
#> wt   -.900* (27)  .781* (26)  .875* (26)  .653* (26) -.707* (27)   1    (29)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg      1  -.86  -.84  -.78   .72  -.90   .40   .63   .78   .70  -.55
#> cyl   -.86     1   .90   .81  -.69   .78  -.49  -.75  -.69  -.62   .45
#> disp  -.84   .90     1   .77  -.64   .88  -.41  -.69  -.63  -.58   .35
#> hp    -.78   .81   .77     1  -.38   .65  -.73  -.68  -.48  -.26   .75
#> drat   .72  -.69  -.64  -.38     1  -.71  -.01   .34   .77   .75  -.10
#> wt    -.90   .78   .88   .65  -.71     1  -.21  -.57  -.75  -.65   .49
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  27   26  26   26  27   26  26  26   26   26
#> cyl   27  29   26  26   26  26   26  26  26   26   26
#> disp  26  26   29  27   26  26   26  26  26   26   26
#> hp    26  26   27  29   26  26   26  26  26   26   26
#> drat  26  26   26  26   29  27   26  26  26   27   26
#> wt    27  26   26  26   27  29   26  27  26   26   26
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .05  .00  .00  .00  .00
#> cyl   .00       .00  .00  .00  .00  .01  .00  .00  .00  .02
#> disp  .00  .00       .00  .00  .00  .04  .00  .00  .00  .08
#> hp    .00  .00  .00       .06  .00  .00  .00  .01  .20  .00
#> drat  .00  .00  .00  .06       .00  .95  .09  .00  .00  .64
#> wt    .00  .00  .00  .00  .00       .30  .00  .00  .00  .01
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> mpg      1                                                           
#> cyl   -.86     1                                                     
#> disp  -.84   .90     1                                               
#> hp    -.78   .81   .77     1                                         
#> drat   .72  -.69  -.64  -.38     1                                   
#> wt    -.90   .78   .88   .65  -.71     1                             
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   27  29                                         
#> disp  26  26   29                                    
#> hp    26  26   27  29                                
#> drat  26  26   26  26   29                           
#> wt    27  26   26  26   27  29                       
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .06                                   
#> wt    .00  .00  .00  .00  .00                              
#>  [ reached getOption("max.print") -- omitted 5 rows ]
#>