(Pairwise, Weighted) Correlations, Covariances and Observation Counts

Computes (pairwise, weighted) Pearsons correlations, covariances and observation counts. Pairwise correlations and covariances can be computed together with observation counts and p-values, and output as 3D array (default) or list of matrices. pwcor and pwcov offer an elaborate print method.

pwcor(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwcov(X, ..., w = NULL, N = FALSE, P = FALSE, array = TRUE, use = "pairwise.complete.obs")

pwnobs(X)

# S3 method for pwcor
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

# S3 method for pwcov
print(x, digits = 2L, sig.level = 0.05, show = c("all","lower.tri","upper.tri"),
      spacing = 1L, return = FALSE, ...)

Arguments

X

a matrix or data.frame, for pwcor and pwcov all columns must be numeric. All functions are faster on matrices, so converting is advised for large data (see qM).

x

an object of class 'pwcor' / 'pwcov'.

w

numeric. A vector of (frequency) weights.

N

logical. TRUE also computes pairwise observation counts.

P

logical. TRUE also computes pairwise p-values (same as cor.test and Hmisc::rcorr).

array

logical. If N = TRUE or P = TRUE, TRUE (default) returns output as 3D array whereas FALSE returns a list of matrices.

use

argument passed to cor / cov. If use != "pairwise.complete.obs", sum(complete.cases(X)) is used for N, and p-values are computed accordingly.

digits

integer. The number of digits to round to in print.

sig.level

numeric. P-value threshold below which a '*' is displayed above significant coefficients if P = TRUE.

show

character. The part of the correlation / covariance matrix to display.

spacing

integer. Controls the spacing between different reported quantities in the printout of the matrix: 0 - compressed, 1 - single space, 2 - double space.

return

logical. TRUE returns the formatted object from the print method for exporting. The default is to return x invisibly.

...

other arguments passed to cor or cov. Only sensible if P = FALSE.

Value

a numeric matrix, 3D array or list of matrices with the computed statistics. For pwcor and pwcov the object has a class 'pwcor' and 'pwcov', respectively.

Note

weights::wtd.cors is imported for weighted pairwise correlations (written in C for speed). For weighted correlations with bootstrap SE's see weights::wtd.cor (but bootstrap can be slow). Weighted correlations for complex surveys are implemented in jtools::svycor. An equivalent and faster implementation of pwcor (without weights) is provided in Hmisc::rcorr (written in Fortran).

See also

Examples

mna <- na_insert(mtcars)
pwcor(mna)
#>        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> mpg     1   -.83  -.85  -.77   .66  -.88   .43   .66   .54   .47  -.57
#> cyl   -.83    1    .91   .81  -.70   .80  -.54  -.79  -.41  -.43   .50
#> disp  -.85   .91    1    .77  -.70   .91  -.39  -.74  -.54  -.52   .42
#> hp    -.77   .81   .77    1   -.41   .68  -.71  -.73  -.12  -.06   .77
#> drat   .66  -.70  -.70  -.41    1   -.71  -.02   .35   .73   .67  -.06
#> wt    -.88   .80   .91   .68  -.71    1   -.15  -.55  -.68  -.59   .45
#> qsec   .43  -.54  -.39  -.71  -.02  -.15    1    .74  -.42  -.24  -.65
#> vs     .66  -.79  -.74  -.73   .35  -.55   .74    1    .03   .14  -.56
#> am     .54  -.41  -.54  -.12   .73  -.68  -.42   .03    1    .80   .17
#> gear   .47  -.43  -.52  -.06   .67  -.59  -.24   .14   .80    1    .32
#> carb  -.57   .50   .42   .77  -.06   .45  -.65  -.56   .17   .32    1 
pwcov(mna)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg      38.74     -8.38   -729.84   -351.38      2.08     -5.84      5.11
#> cyl      -8.38      3.12    195.96     98.08     -0.69      1.34     -1.69
#> disp   -729.84    195.96  16354.11   6680.64    -47.04    127.31    -91.03
#> hp     -351.38     98.08   6680.64   4861.07    -15.75     44.26    -94.59
#> drat      2.08     -0.69    -47.04    -15.75      0.29     -0.37     -0.01
#> wt       -5.84      1.34    127.31     44.26     -0.37      1.03     -0.22
#> qsec      5.11     -1.69    -91.03    -94.59     -0.01     -0.22      3.32
#> vs        2.06     -0.71    -46.68    -25.62      0.09     -0.26      0.69
#> am        1.58     -0.34    -34.16     -4.18      0.20     -0.34     -0.37
#> gear      2.22     -0.58    -49.35     -3.31      0.28     -0.45     -0.36
#> carb     -6.17      1.52     88.57     94.50     -0.06      0.76     -2.11
#>             vs        am      gear      carb
#> mpg       2.06      1.58      2.22     -6.17
#> cyl      -0.71     -0.34     -0.58      1.52
#> disp    -46.68    -34.16    -49.35     88.57
#> hp      -25.62     -4.18     -3.31     94.50
#> drat      0.09      0.20      0.28     -0.06
#> wt       -0.26     -0.34     -0.45      0.76
#> qsec      0.69     -0.37     -0.36     -2.11
#> vs        0.26      0.01      0.06     -0.49
#> am        0.01      0.24      0.31      0.14
#> gear      0.06      0.31      0.58      0.42
#> carb     -0.49      0.14      0.42      2.76
pwnobs(mna)
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg   29  26   26 27   26 27   26 26 26   26   26
#> cyl   26  29   27 26   28 26   26 26 27   27   26
#> disp  26  27   29 27   26 26   26 26 27   26   26
#> hp    27  26   27 29   26 27   26 27 26   26   26
#> drat  26  28   26 26   29 26   26 26 26   27   26
#> wt    27  26   26 27   26 29   26 26 26   26   26
#> qsec  26  26   26 26   26 26   29 27 26   26   26
#> vs    26  26   26 27   26 26   27 29 26   26   27
#> am    26  27   27 26   26 26   26 26 29   27   26
#> gear  26  27   26 26   27 26   26 26 27   29   26
#> carb  26  26   26 26   26 26   26 27 26   26   29
pwcor(mna, N = TRUE)
#>            mpg       cyl      disp        hp      drat        wt      qsec
#> mpg    1  (29) -.83 (26) -.85 (26) -.77 (27)  .66 (26) -.88 (27)  .43 (26)
#> cyl  -.83 (26)   1  (29)  .91 (27)  .81 (26) -.70 (28)  .80 (26) -.54 (26)
#> disp -.85 (26)  .91 (27)   1  (29)  .77 (27) -.70 (26)  .91 (26) -.39 (26)
#> hp   -.77 (27)  .81 (26)  .77 (27)   1  (29) -.41 (26)  .68 (27) -.71 (26)
#> drat  .66 (26) -.70 (28) -.70 (26) -.41 (26)   1  (29) -.71 (26) -.02 (26)
#> wt   -.88 (27)  .80 (26)  .91 (26)  .68 (27) -.71 (26)   1  (29) -.15 (26)
#> qsec  .43 (26) -.54 (26) -.39 (26) -.71 (26) -.02 (26) -.15 (26)   1  (29)
#> vs    .66 (26) -.79 (26) -.74 (26) -.73 (27)  .35 (26) -.55 (26)  .74 (27)
#> am    .54 (26) -.41 (27) -.54 (27) -.12 (26)  .73 (26) -.68 (26) -.42 (26)
#> gear  .47 (26) -.43 (27) -.52 (26) -.06 (26)  .67 (27) -.59 (26) -.24 (26)
#> carb -.57 (26)  .50 (26)  .42 (26)  .77 (26) -.06 (26)  .45 (26) -.65 (26)
#>             vs        am      gear      carb
#> mpg   .66 (26)  .54 (26)  .47 (26) -.57 (26)
#> cyl  -.79 (26) -.41 (27) -.43 (27)  .50 (26)
#> disp -.74 (26) -.54 (27) -.52 (26)  .42 (26)
#> hp   -.73 (27) -.12 (26) -.06 (26)  .77 (26)
#> drat  .35 (26)  .73 (26)  .67 (27) -.06 (26)
#> wt   -.55 (26) -.68 (26) -.59 (26)  .45 (26)
#> qsec  .74 (27) -.42 (26) -.24 (26) -.65 (26)
#> vs     1  (29)  .03 (26)  .14 (26) -.56 (27)
#> am    .03 (26)   1  (29)  .80 (27)  .17 (26)
#> gear  .14 (26)  .80 (27)   1  (29)  .32 (26)
#> carb -.56 (27)  .17 (26)  .32 (26)   1  (29)
pwcor(mna, P = TRUE)
#>         mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear
#> mpg     1    -.83*  -.85*  -.77*   .66*  -.88*   .43*   .66*   .54*   .47*
#> cyl   -.83*    1     .91*   .81*  -.70*   .80*  -.54*  -.79*  -.41*  -.43*
#> disp  -.85*   .91*    1     .77*  -.70*   .91*  -.39   -.74*  -.54*  -.52*
#> hp    -.77*   .81*   .77*    1    -.41*   .68*  -.71*  -.73*  -.12   -.06 
#> drat   .66*  -.70*  -.70*  -.41*    1    -.71*  -.02    .35    .73*   .67*
#> wt    -.88*   .80*   .91*   .68*  -.71*    1    -.15   -.55*  -.68*  -.59*
#> qsec   .43*  -.54*  -.39   -.71*  -.02   -.15     1     .74*  -.42*  -.24 
#> vs     .66*  -.79*  -.74*  -.73*   .35   -.55*   .74*    1     .03    .14 
#> am     .54*  -.41*  -.54*  -.12    .73*  -.68*  -.42*   .03     1     .80*
#> gear   .47*  -.43*  -.52*  -.06    .67*  -.59*  -.24    .14    .80*    1  
#> carb  -.57*   .50*   .42*   .77*  -.06    .45*  -.65*  -.56*   .17    .32 
#>        carb
#> mpg   -.57*
#> cyl    .50*
#> disp   .42*
#> hp     .77*
#> drat  -.06 
#> wt     .45*
#> qsec  -.65*
#> vs    -.56*
#> am     .17 
#> gear   .32 
#> carb    1  
pwcor(mna, N = TRUE, P = TRUE)
#>             mpg        cyl       disp         hp       drat         wt
#> mpg    1   (29) -.83* (26) -.85* (26) -.77* (27)  .66* (26) -.88* (27)
#> cyl  -.83* (26)   1   (29)  .91* (27)  .81* (26) -.70* (28)  .80* (26)
#> disp -.85* (26)  .91* (27)   1   (29)  .77* (27) -.70* (26)  .91* (26)
#> hp   -.77* (27)  .81* (26)  .77* (27)   1   (29) -.41* (26)  .68* (27)
#> drat  .66* (26) -.70* (28) -.70* (26) -.41* (26)   1   (29) -.71* (26)
#> wt   -.88* (27)  .80* (26)  .91* (26)  .68* (27) -.71* (26)   1   (29)
#> qsec  .43* (26) -.54* (26) -.39  (26) -.71* (26) -.02  (26) -.15  (26)
#> vs    .66* (26) -.79* (26) -.74* (26) -.73* (27)  .35  (26) -.55* (26)
#> am    .54* (26) -.41* (27) -.54* (27) -.12  (26)  .73* (26) -.68* (26)
#> gear  .47* (26) -.43* (27) -.52* (26) -.06  (26)  .67* (27) -.59* (26)
#> carb -.57* (26)  .50* (26)  .42* (26)  .77* (26) -.06  (26)  .45* (26)
#>            qsec         vs         am       gear       carb
#> mpg   .43* (26)  .66* (26)  .54* (26)  .47* (26) -.57* (26)
#> cyl  -.54* (26) -.79* (26) -.41* (27) -.43* (27)  .50* (26)
#> disp -.39  (26) -.74* (26) -.54* (27) -.52* (26)  .42* (26)
#> hp   -.71* (26) -.73* (27) -.12  (26) -.06  (26)  .77* (26)
#> drat -.02  (26)  .35  (26)  .73* (26)  .67* (27) -.06  (26)
#> wt   -.15  (26) -.55* (26) -.68* (26) -.59* (26)  .45* (26)
#> qsec   1   (29)  .74* (27) -.42* (26) -.24  (26) -.65* (26)
#> vs    .74* (27)   1   (29)  .03  (26)  .14  (26) -.56* (27)
#> am   -.42* (26)  .03  (26)   1   (29)  .80* (27)  .17  (26)
#> gear -.24  (26)  .14  (26)  .80* (27)   1   (29)  .32  (26)
#> carb -.65* (26) -.56* (27)  .17  (26)  .32  (26)   1   (29)
aperm(pwcor(mna, N = TRUE, P = TRUE))
#> , , mpg
#> 
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r   1  -.83  -.85  -.77   .66  -.88   .43   .66   .54   .47  -.57
#> N  29    26    26    27    26    27    26    26    26    26    26
#> P       .00   .00   .00   .00   .00   .03   .00   .00   .02   .00
#> 
#> , , cyl
#> 
#>     mpg cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.83   1   .91   .81  -.70   .80  -.54  -.79  -.41  -.43   .50
#> N    26  29    27    26    28    26    26    26    27    27    26
#> P   .00       .00   .00   .00   .00   .00   .00   .04   .02   .01
#> 
#> , , disp
#> 
#>     mpg   cyl disp    hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.85   .91    1   .77  -.70   .91  -.39  -.74  -.54  -.52   .42
#> N    26    27   29    27    26    26    26    26    27    26    26
#> P   .00   .00        .00   .00   .00   .05   .00   .00   .01   .03
#> 
#> , , hp
#> 
#>     mpg   cyl  disp     hp  drat    wt  qsec    vs    am  gear  carb
#> r  -.77   .81   .77   1.00  -.41   .68  -.71  -.73  -.12  -.06   .77
#> N    27    26    27     29    26    27    26    27    26    26    26
#> P   .00   .00   .00          .04   .00   .00   .00   .56   .77   .00
#> 
#> , , drat
#> 
#>     mpg   cyl  disp    hp drat    wt  qsec    vs    am  gear  carb
#> r   .66  -.70  -.70  -.41    1  -.71  -.02   .35   .73   .67  -.06
#> N    26    28    26    26   29    26    26    26    26    27    26
#> P   .00   .00   .00   .04        .00   .94   .08   .00   .00   .76
#> 
#> , , wt
#> 
#>     mpg   cyl  disp    hp  drat  wt  qsec    vs    am  gear  carb
#> r  -.88   .80   .91   .68  -.71   1  -.15  -.55  -.68  -.59   .45
#> N    27    26    26    27    26  29    26    26    26    26    26
#> P   .00   .00   .00   .00   .00       .48   .00   .00   .00   .02
#> 
#> , , qsec
#> 
#>     mpg   cyl  disp    hp  drat    wt qsec    vs    am  gear  carb
#> r   .43  -.54  -.39  -.71  -.02  -.15    1   .74  -.42  -.24  -.65
#> N    26    26    26    26    26    26   29    27    26    26    26
#> P   .03   .00   .05   .00   .94   .48        .00   .03   .24   .00
#> 
#> , , vs
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec     vs    am  gear  carb
#> r   .66  -.79  -.74  -.73   .35  -.55   .74   1.00   .03   .14  -.56
#> N    26    26    26    27    26    26    27     29    26    26    27
#> P   .00   .00   .00   .00   .08   .00   .00          .88   .50   .00
#> 
#> , , am
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs  am  gear  carb
#> r   .54  -.41  -.54  -.12   .73  -.68  -.42   .03   1   .80   .17
#> N    26    27    27    26    26    26    26    26  29    27    26
#> P   .00   .04   .00   .56   .00   .00   .03   .88       .00   .40
#> 
#> , , gear
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am gear  carb
#> r   .47  -.43  -.52  -.06   .67  -.59  -.24   .14   .80    1   .32
#> N    26    27    26    26    27    26    26    26    27   29    26
#> P   .02   .02   .01   .77   .00   .00   .24   .50   .00        .11
#> 
#> , , carb
#> 
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear carb
#> r  -.57   .50   .42   .77  -.06   .45  -.65  -.56   .17   .32    1
#> N    26    26    26    26    26    26    26    27    26    26   29
#> P   .00   .01   .03   .00   .76   .02   .00   .00   .40   .11     
#> 
print(pwcor(mna, N = TRUE, P = TRUE), digits = 3, sig.level = 0.01, show = "lower.tri")
#>              mpg         cyl        disp          hp        drat          wt
#> mpg    1    (29)                                                            
#> cyl  -.834* (26)   1    (29)                                                
#> disp -.852* (26)  .909* (27)   1    (29)                                    
#> hp   -.769* (27)  .806* (26)  .774* (27)   1    (29)                        
#> drat  .661* (26) -.702* (28) -.704* (26) -.410  (26)   1    (29)            
#> wt   -.875* (27)  .799* (26)  .913* (26)  .679* (27) -.711* (26)   1    (29)
#> qsec  .425  (26) -.543* (26) -.387  (26) -.712* (26) -.016  (26) -.146  (26)
#> vs    .656* (26) -.788* (26) -.740* (26) -.727* (27)  .346  (26) -.552* (26)
#> am    .540* (26) -.405  (27) -.538* (27) -.121  (26)  .730* (26) -.675* (26)
#> gear  .469  (26) -.431  (27) -.515* (26) -.061  (26)  .673* (27) -.589* (26)
#> carb -.575* (26)  .504* (26)  .418  (26)  .770* (26) -.063  (26)  .451  (26)
#>             qsec          vs          am        gear        carb
#> mpg                                                             
#> cyl                                                             
#> disp                                                            
#> hp                                                              
#> drat                                                            
#> wt                                                              
#> qsec   1    (29)                                                
#> vs    .736* (27)   1    (29)                                    
#> am   -.420  (26)  .031  (26)   1    (29)                        
#> gear -.241  (26)  .139  (26)  .803* (27)   1    (29)            
#> carb -.650* (26) -.561* (27)  .171  (26)  .322  (26)   1    (29)
pwcor(mna, N = TRUE, P = TRUE, array = FALSE)
#> $r
#>        mpg   cyl  disp     hp  drat    wt  qsec     vs    am  gear  carb
#> mpg      1  -.83  -.85   -.77   .66  -.88   .43    .66   .54   .47  -.57
#> cyl   -.83     1   .91    .81  -.70   .80  -.54   -.79  -.41  -.43   .50
#> disp  -.85   .91     1    .77  -.70   .91  -.39   -.74  -.54  -.52   .42
#> hp    -.77   .81   .77   1.00  -.41   .68  -.71   -.73  -.12  -.06   .77
#> drat   .66  -.70  -.70   -.41     1  -.71  -.02    .35   .73   .67  -.06
#> wt    -.88   .80   .91    .68  -.71     1  -.15   -.55  -.68  -.59   .45
#> qsec   .43  -.54  -.39   -.71  -.02  -.15     1    .74  -.42  -.24  -.65
#> vs     .66  -.79  -.74   -.73   .35  -.55   .74   1.00   .03   .14  -.56
#> am     .54  -.41  -.54   -.12   .73  -.68  -.42    .03     1   .80   .17
#> gear   .47  -.43  -.52   -.06   .67  -.59  -.24    .14   .80     1   .32
#> carb  -.57   .50   .42    .77  -.06   .45  -.65   -.56   .17   .32     1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29  26   26  27   26  27   26  26  26   26   26
#> cyl   26  29   27  26   28  26   26  26  27   27   26
#> disp  26  27   29  27   26  26   26  26  27   26   26
#> hp    27  26   27  29   26  27   26  27  26   26   26
#> drat  26  28   26  26   29  26   26  26  26   27   26
#> wt    27  26   26  27   26  29   26  26  26   26   26
#> qsec  26  26   26  26   26  26   29  27  26   26   26
#> vs    26  26   26  27   26  26   27  29  26   26   27
#> am    26  27   27  26   26  26   26  26  29   27   26
#> gear  26  27   26  26   27  26   26  26  27   29   26
#> carb  26  26   26  26   26  26   26  27  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg        .00  .00  .00  .00  .00  .03  .00  .00  .02  .00
#> cyl   .00       .00  .00  .00  .00  .00  .00  .04  .02  .01
#> disp  .00  .00       .00  .00  .00  .05  .00  .00  .01  .03
#> hp    .00  .00  .00       .04  .00  .00  .00  .56  .77  .00
#> drat  .00  .00  .00  .04       .00  .94  .08  .00  .00  .76
#> wt    .00  .00  .00  .00  .00       .48  .00  .00  .00  .02
#> qsec  .03  .00  .05  .00  .94  .48       .00  .03  .24  .00
#> vs    .00  .00  .00  .00  .08  .00  .00       .88  .50  .00
#> am    .00  .04  .00  .56  .00  .00  .03  .88       .00  .40
#> gear  .02  .02  .01  .77  .00  .00  .24  .50  .00       .11
#> carb  .00  .01  .03  .00  .76  .02  .00  .00  .40  .11     
#> 
print(pwcor(mna, N = TRUE, P = TRUE, array = FALSE), show = "lower.tri")
#> $r
#>        mpg   cyl  disp     hp  drat    wt  qsec     vs    am  gear carb
#> mpg      1                                                             
#> cyl   -.83     1                                                       
#> disp  -.85   .91     1                                                 
#> hp    -.77   .81   .77   1.00                                          
#> drat   .66  -.70  -.70   -.41     1                                    
#> wt    -.88   .80   .91    .68  -.71     1                              
#> qsec   .43  -.54  -.39   -.71  -.02  -.15     1                        
#> vs     .66  -.79  -.74   -.73   .35  -.55   .74   1.00                 
#> am     .54  -.41  -.54   -.12   .73  -.68  -.42    .03     1           
#> gear   .47  -.43  -.52   -.06   .67  -.59  -.24    .14   .80     1     
#> carb  -.57   .50   .42    .77  -.06   .45  -.65   -.56   .17   .32    1
#> 
#> $N
#>      mpg cyl disp  hp drat  wt qsec  vs  am gear carb
#> mpg   29                                             
#> cyl   26  29                                         
#> disp  26  27   29                                    
#> hp    27  26   27  29                                
#> drat  26  28   26  26   29                           
#> wt    27  26   26  27   26  29                       
#> qsec  26  26   26  26   26  26   29                  
#> vs    26  26   26  27   26  26   27  29              
#> am    26  27   27  26   26  26   26  26  29          
#> gear  26  27   26  26   27  26   26  26  27   29     
#> carb  26  26   26  26   26  26   26  27  26   26   29
#> 
#> $P
#>       mpg  cyl disp   hp drat   wt qsec   vs   am gear carb
#> mpg                                                        
#> cyl   .00                                                  
#> disp  .00  .00                                             
#> hp    .00  .00  .00                                        
#> drat  .00  .00  .00  .04                                   
#> wt    .00  .00  .00  .00  .00                              
#> qsec  .03  .00  .05  .00  .94  .48                         
#> vs    .00  .00  .00  .00  .08  .00  .00                    
#> am    .00  .04  .00  .56  .00  .00  .03  .88               
#> gear  .02  .02  .01  .77  .00  .00  .24  .50  .00          
#> carb  .00  .01  .03  .00  .76  .02  .00  .00  .40  .11     
#>