Quick Data Conversion

Fast, flexible and precise conversion of common data objects, without method dispatch and extensive checks:

qDF, qDT and qTBL convert vectors, matrices, higher-dimensional arrays and suitable lists to data frame, data.table and tibble, respectively.

qM converts vectors, higher-dimensional arrays, data frames and suitable lists to matrix.

mctl and mrtl column- or row-wise convert a matrix to list, data frame or data.table. They are used internally by qDF/qDT/qTBL, dapply, BY, etc...

qF converts atomic vectors to factor (documented on a separate page).

as_numeric_factor and as_character_factor convert factors, or all factor columns in a data frame / list, to character or numeric (by converting the levels).

# Converting between matrices, data frames / tables / tibbles

 qDF(X, row.names.col = FALSE, keep.attr = FALSE, class = "data.frame")
 qDT(X, row.names.col = FALSE, keep.attr = FALSE, class = c("data.table", "data.frame"))
qTBL(X, row.names.col = FALSE, keep.attr = FALSE, class = c("tbl_df","tbl","data.frame"))
  qM(X,                        keep.attr = FALSE, class = NULL)

# Programmer functions: matrix rows or columns to list / DF / DT - fully in C++

mctl(X, names = FALSE, return = "list")
mrtl(X, names = FALSE, return = "list")

# Converting factors or factor columns

  as_numeric_factor(X, keep.attr = TRUE)
as_character_factor(X, keep.attr = TRUE)

Arguments

X
a vector, factor, matrix, higher-dimensional array, data frame or list. mctl and mrtl only accept matrices, as_numeric_factor and as_character_factor only accept factors, data frames or lists.

row.names.col
should a column capturing names or row.names be added? e.g. when converting atomic objects to data frame or data frame to data.table. Can be logical TRUE, which will add a column "row.names" in front, or can supply a name for the column e.g. "variable".

keep.attr
logical. FALSE (default) yields a hard / thorough object conversion: All unnecessary attributes are removed from the object yielding a plain matrix / data.frame / data.table. FALSE yields a soft / minimal object conversion: Only the attributes 'names', 'row.names', 'dim', 'dimnames' and 'levels' are modified in the conversion. Other attributes are preserved. See also class.

class
if a vector of classes is passed here, the converted object will be assigned these classes. If NULL is passed, the default classes are assigned: qM assigns no class, qDF a class "data.frame", and qDT a class c("data.table", "data.frame"). If keep.attr = TRUE and class = NULL and the object already inherits the default classes, further inherited classes are preserved. See Details and the Example.

names
logical. Should the list be named using row/column names from the matrix?

return
an integer or string specifying what to return. The options are:

Int. 		 String 		 Description

1		"list"		returns a plain list

2		"data.frame"		returns a plain data.frame

3		"data.table"		returns a plain data.table

Details

Object conversions using these functions are maximally efficient and involve 3 consecutive steps: (1) Converting the storage mode / dimensions / data of the object, (2) converting / modifying the attributes and (3) modifying the class of the object:

(1) is determined by the choice of function and the optional row.names.col argument to qDF and qDT. Higher-dimensional arrays are converted by expanding the second dimension (adding columns, same as as.matrix, as.data.frame, as.data.table).

(2) is determined by the keep.attr argument: keep.attr = TRUE seeks to preserve the attributes of the object. It's effect is like copying attributes(converted) <- attributes(original), and then modifying the "dim", "dimnames", "names", "row.names" and "levels" attributes as necessitated by the conversion task. keep.attr = FALSE only converts / assigns / removes these attributes and drops all others.

(3) is determined by the class argument: Setting class = "myclass" will yield a converted object of class "myclass", with any other / prior classes being removed by this replacement. Setting class = NULL does NOT mean that a class NULL is assigned (which would remove the class attribute), but rather that the default classes are assigned: qM assigns no class, qDF a class "data.frame", and qDT a class c("data.table", "data.frame"). At this point there is an interaction with keep.attr: If keep.attr = TRUE and class = NULL and the object converted already inherits the respective default classes, then any other inherited classes will also be preserved (with qM(x, keep.attr = TRUE, class = NULL) any class will be preserved if is.matrix(x) evaluates to TRUE.)

The default keep.attr = FALSE ensures hard conversions so that all unnecessary attributes are dropped. Furthermore in qDF/qDT/qTBL the default classes were explicitly assigned. This is to ensure that the default methods apply, even if the user chooses to preserve further attributes. For qM a more lenient default setup was chosen to enable the full preservation of time series matrices with keep.attr = TRUE. If the user wants to keep attributes attached to a matrix but make sure that all default methods work properly, either one of qM(x, keep.attr = TRUE, class = "matrix") or unclass(qM(x, keep.attr = TRUE)) should be employed.

Value

qDF - returns a data.frame
qDT - returns a data.table
qTBL - returns a tibble
qM - returns a matrix
mctl, mrtl - return a list, data frame or data.table
qF - returns a factor
as_numeric_factor - returns X with factors converted to numeric variables
as_character_factor - returns X with factors converted to character variables

See also

Examples

## Basic Examples
mtcarsM <- qM(mtcars)                   # Matrix from data.frame
mtcarsDT <- qDT(mtcarsM)                # data.table from matrix columns
mtcarsTBL <- qTBL(mtcarsM)              # tibble from matrix columns
head(mrtl(mtcarsM, TRUE, "data.frame")) # data.frame from matrix rows, etc..

#>      Mazda RX4 Mazda RX4 Wag Datsun 710 Hornet 4 Drive Hornet Sportabout
#> mpg         21            21       22.8           21.4              18.7
#> cyl          6             6        4.0            6.0               8.0
#> disp       160           160      108.0          258.0             360.0
#>      Valiant Duster 360 Merc 240D Merc 230 Merc 280 Merc 280C Merc 450SE
#> mpg     18.1       14.3      24.4     22.8     19.2      17.8       16.4
#> cyl      6.0        8.0       4.0      4.0      6.0       6.0        8.0
#> disp   225.0      360.0     146.7    140.8    167.6     167.6      275.8
#>      Merc 450SL Merc 450SLC Cadillac Fleetwood Lincoln Continental
#> mpg        17.3        15.2               10.4                10.4
#> cyl         8.0         8.0                8.0                 8.0
#> disp      275.8       275.8              472.0               460.0
#>      Chrysler Imperial Fiat 128 Honda Civic Toyota Corolla Toyota Corona
#> mpg               14.7     32.4        30.4           33.9          21.5
#> cyl                8.0      4.0         4.0            4.0           4.0
#> disp             440.0     78.7        75.7           71.1         120.1
#>      Dodge Challenger AMC Javelin Camaro Z28 Pontiac Firebird Fiat X1-9
#> mpg              15.5        15.2       13.3             19.2      27.3
#> cyl               8.0         8.0        8.0              8.0       4.0
#> disp            318.0       304.0      350.0            400.0      79.0
#>      Porsche 914-2 Lotus Europa Ford Pantera L Ferrari Dino Maserati Bora
#> mpg           26.0         30.4           15.8         19.7            15
#> cyl            4.0          4.0            8.0          6.0             8
#> disp         120.3         95.1          351.0        145.0           301
#>      Volvo 142E
#> mpg        21.4
#> cyl         4.0
#> disp      121.0
#>  [ reached 'max' / getOption("max.print") -- omitted 3 rows ]

head(qDF(mtcarsM, "cars"))              # Adding a row.names column when converting from matrix

#>                cars  mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1         Mazda RX4 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2     Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3        Datsun 710 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> 4    Hornet 4 Drive 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 5 Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> 6           Valiant 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

head(qDT(mtcars, "cars"))               # Saving row.names when converting data frame to data.table

#>                 cars  mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1:         Mazda RX4 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2:     Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3:        Datsun 710 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> 4:    Hornet 4 Drive 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 5: Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> 6:           Valiant 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

cylF <- qF(mtcars$cyl)                  # Factor from atomic vector
cylF

#>  [1] 6 6 4 6 8 6 8 4 4 6 6 8 8 8 8 8 8 4 4 4 4 8 8 8 8 4 4 4 8 6 8 4
#> Levels: 4 6 8

# Factor to numeric conversions
identical(mtcars,  as_numeric_factor(dapply(mtcars, qF)))

#> [1] TRUE