Small (Helper) Functions

Convenience functions in the collapse package that help to deal with attributes such as variable names and labels, missing values, matching and object checking etc.. Some functions are performance improved replacements for base R functions.

.c(...)                       # Non-standard concatenation i.e. .c(a, b) == c("a", "b")
vlabels(X, attrn = "label")   # Get labels of variables in X, in attr(X[[i]], attrn)
vlabels(X, attrn = "label") <- value      # Set labels of variables in X
vclasses(X)                   # Get classes of variables in X
vtypes(X)                     # Get data storage types of variables in X (calling typeof)
namlab(X, class = FALSE,      # Return data frame of names, labels and classes
          attrn = "label")
add_stub(X, stub, pre = TRUE) # Add a stub (i.e. prefix or postfix) to column names
rm_stub(X, stub, pre = TRUE)  # Remove stub from column names
x %!in% table                 # The opposite of %in%
ckmatch(x, table,             # Check-match: throws an informative error if non-matched
     e = "Unknown columns:")
fnlevels(x)                   # Faster version of nlevels(x) (for factors)
fnrow(X)                      # Faster nrow for data frames (not faster for matrices)
fncol(X)                      # Faster ncol for data frames (not faster for matrices)
fdim(X)                       # Faster dim for data frames (not faster for matrices)
allNA(x)                      # Check for all missing cases in a vector
missing_cases(X, cols = NULL) # Find missing cases in a matrix / data framme
na_rm(x)                      # Remove missing values from vector and return vector
na_omit(X, cols = NULL,       # Faster na.omit for matrices and data frames
        na.attr = FALSE)
na_insert(X, prop = 0.1)      # Insert missing values at random in vectors, matrices DF's
cinv(X)                       # Choleski (fast) inverse of symmetric, positive def. matrix
all_identical(...)            # Check exact equality of multiple objects or list-elements
all_obj_equal(...)            # Check near equality of multiple objects or list-elements
seq_row(X)                    # Fast integer sequences along rows of X
seq_col(X)                    # Fast integer sequences along columns of X
setRownames(object,
 nm = if(is.atomic(object))   # Set rownames of object and return object
 seq_row(object) else NULL)
setColnames(object, nm)       # Set colnames of object and return object
setDimnames(object, dn,
            which = NULL)     # Set dimension names of object and return object
unattrib(object)              # Remove all attributes from object
setAttrib(object, a)          # Replace all attributes with list of attributes 'a'
copyAttrib(to, from)          # Copy all attributes from object 'from' to object 'to'
copyMostAttrib(to, from)      # Copy most attributes from object 'from' to object 'to'
is.categorical(x)             # The opposite of is.numeric
is.Date(x)                    # Check if object is of class "Date", "POSIXlt" or "POSIXct"

Arguments

X
a matrix or data frame (some functions also support vectors and arrays although that is less common).

x, table
a atomic vector.

object, to, from
a suitable R object.

a
a suitable list of attributes.

attrn
character. Name of attribute to store labels or retrieve labels from.

value
a matching character vector of variable labels.

class
logical. Also show the classes of variables in X in a column?

stub
a single character stub, i.e. "log.", which by default will be pre-applied to all variables or column names in X.

pre
logical. FALSE will post-apply stub.

cols
only removes rows with missing values on these columns. Columns can be selected using column names, indices, a logical vector or a selector function (i.e. is.numeric).

na.attr
logical. TRUE adds an attribute containing the removed cases. For compatibility reasons this is exactly the same format as na.omit i.e. the attribute is called "na.action" and of class "omit".

nm
a suitable vector of row- or column-names.

dn
a suitable vector or list of names for dimension(s).

which
integer. If NULL, dn has to be a list fully specifying the dimension names of the object. Alternatively, a vector or list of names for dimensions which can be supplied. See Examples.

prop
double. Specify the proportion of observations randomly replaced with NA.

e
the error message thrown by ckmatch for non-matched elements. The message is followed by the comma-separated non-matched elements.

...
for .c: Comma-separated expressions. For all_identical / all_obj_equal: Either multiple comma-separated objects or a single list of objects in which all elements will be checked for exact / numeric equality.

See also

Examples

## Non-standard concatenation
.c(a, b, "c d", e == f)

#> [1] "a"      "b"      "c d"    "e == f"

## Variable labels
namlab(wlddev, class = TRUE)

#>    Variable     Class                                   Label
#> 1   country character                            Country Name
#> 2     iso3c    factor                            Country Code
#> 3      date      Date              Date Recorded (Fictitious)
#> 4      year   integer                                    Year
#> 5    decade   numeric                                  Decade
#> 6    region    factor                                  Region
#> 7    income    factor                            Income Level
#> 8      OECD   logical                 Is OECD Member Country?
#> 9     PCGDP   numeric      GDP per capita (constant 2010 US$)
#> 10   LIFEEX   numeric Life expectancy at birth, total (years)
#> 11     GINI   numeric        GINI index (World Bank estimate)
#> 12      ODA   numeric    Net ODA received (constant 2015 US$)

vlabels(wlddev)

#>                                   country 
#>                            "Country Name" 
#>                                     iso3c 
#>                            "Country Code" 
#>                                      date 
#>              "Date Recorded (Fictitious)" 
#>                                      year 
#>                                    "Year" 
#>                                    decade 
#>                                  "Decade" 
#>                                    region 
#>                                  "Region" 
#>                                    income 
#>                            "Income Level" 
#>                                      OECD 
#>                 "Is OECD Member Country?" 
#>                                     PCGDP 
#>      "GDP per capita (constant 2010 US$)" 
#>                                    LIFEEX 
#> "Life expectancy at birth, total (years)" 
#>                                      GINI 
#>        "GINI index (World Bank estimate)" 
#>                                       ODA 
#>    "Net ODA received (constant 2015 US$)" 

vlabels(wlddev) <- vlabels(wlddev)

## Stub-renaming
log_mtc <- add_stub(log(mtcars), "log.")
head(log_mtc)

#>                    log.mpg  log.cyl log.disp   log.hp log.drat    log.wt
#> Mazda RX4         3.044522 1.791759 5.075174 4.700480 1.360977 0.9631743
#> Mazda RX4 Wag     3.044522 1.791759 5.075174 4.700480 1.360977 1.0560527
#> Datsun 710        3.126761 1.386294 4.682131 4.532599 1.348073 0.8415672
#> Hornet 4 Drive    3.063391 1.791759 5.552960 4.700480 1.124930 1.1678274
#> Hornet Sportabout 2.928524 2.079442 5.886104 5.164786 1.147402 1.2354715
#> Valiant           2.895912 1.791759 5.416100 4.653960 1.015231 1.2412686
#>                   log.qsec log.vs log.am log.gear  log.carb
#> Mazda RX4         2.800933   -Inf      0 1.386294 1.3862944
#> Mazda RX4 Wag     2.834389   -Inf      0 1.386294 1.3862944
#> Datsun 710        2.923699      0      0 1.386294 0.0000000
#> Hornet 4 Drive    2.967333      0   -Inf 1.098612 0.0000000
#> Hornet Sportabout 2.834389   -Inf   -Inf 1.098612 0.6931472
#> Valiant           3.006672      0   -Inf 1.098612 0.0000000

head(rm_stub(log_mtc, "log."))

#>                        mpg      cyl     disp       hp     drat        wt
#> Mazda RX4         3.044522 1.791759 5.075174 4.700480 1.360977 0.9631743
#> Mazda RX4 Wag     3.044522 1.791759 5.075174 4.700480 1.360977 1.0560527
#> Datsun 710        3.126761 1.386294 4.682131 4.532599 1.348073 0.8415672
#> Hornet 4 Drive    3.063391 1.791759 5.552960 4.700480 1.124930 1.1678274
#> Hornet Sportabout 2.928524 2.079442 5.886104 5.164786 1.147402 1.2354715
#> Valiant           2.895912 1.791759 5.416100 4.653960 1.015231 1.2412686
#>                       qsec   vs   am     gear      carb
#> Mazda RX4         2.800933 -Inf    0 1.386294 1.3862944
#> Mazda RX4 Wag     2.834389 -Inf    0 1.386294 1.3862944
#> Datsun 710        2.923699    0    0 1.386294 0.0000000
#> Hornet 4 Drive    2.967333    0 -Inf 1.098612 0.0000000
#> Hornet Sportabout 2.834389 -Inf -Inf 1.098612 0.6931472
#> Valiant           3.006672    0 -Inf 1.098612 0.0000000

rm(log_mtc)

## Setting dimension names of an object
head(setRownames(mtcars))

#>    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> 4 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 5 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> 6 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

ar <- array(1:9, c(3,3,3))
setRownames(ar)

#> , , 1
#> 
#>   [,1] [,2] [,3]
#> 1    1    4    7
#> 2    2    5    8
#> 3    3    6    9
#> 
#> , , 2
#> 
#>   [,1] [,2] [,3]
#> 1    1    4    7
#> 2    2    5    8
#> 3    3    6    9
#> 
#> , , 3
#> 
#>   [,1] [,2] [,3]
#> 1    1    4    7
#> 2    2    5    8
#> 3    3    6    9
#> 

setColnames(ar, c("a","b","c"))

#> , , 1
#> 
#>      a b c
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 
#> , , 2
#> 
#>      a b c
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 
#> , , 3
#> 
#>      a b c
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 

setDimnames(ar, c("a","b","c"), which = 3)

#> , , a
#> 
#>      [,1] [,2] [,3]
#> [1,]    1    4    7
#> [2,]    2    5    8
#> [3,]    3    6    9
#> 
#> , , b
#> 
#>      [,1] [,2] [,3]
#> [1,]    1    4    7
#> [2,]    2    5    8
#> [3,]    3    6    9
#> 
#> , , c
#> 
#>      [,1] [,2] [,3]
#> [1,]    1    4    7
#> [2,]    2    5    8
#> [3,]    3    6    9
#> 

setDimnames(ar, list(c("d","e","f"), c("a","b","c")), which = 2:3)

#> , , a
#> 
#>      d e f
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 
#> , , b
#> 
#>      d e f
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 
#> , , c
#> 
#>      d e f
#> [1,] 1 4 7
#> [2,] 2 5 8
#> [3,] 3 6 9
#> 

setDimnames(ar, list(c("g","h","i"), c("d","e","f"), c("a","b","c")))

#> , , a
#> 
#>   d e f
#> g 1 4 7
#> h 2 5 8
#> i 3 6 9
#> 
#> , , b
#> 
#>   d e f
#> g 1 4 7
#> h 2 5 8
#> i 3 6 9
#> 
#> , , c
#> 
#>   d e f
#> g 1 4 7
#> h 2 5 8
#> i 3 6 9
#> 

## Checking exact equality of multiple objects
all_identical(iris, iris, iris, iris)

#> [1] TRUE

l <- replicate(100, fmean(num_vars(iris), iris$Species), simplify = FALSE)
all_identical(l)

#> [1] TRUE

rm(l)

## Missing values
mtc_na <- na_insert(mtcars, 0.15)    # Set 15% of values missing at random
fNobs(mtc_na)                        # See observation count

#>  mpg  cyl disp   hp drat   wt qsec   vs   am gear carb 
#>   28   28   28   28   28   28   28   28   28   28   28 

na_omit(mtc_na)                      # 12x faster than na.omit(mtc_na)

#>             mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 240D  24.4   4 146.7  62 3.69 3.19 20.0  1  0    4    2
#> Merc 280C  17.8   6 167.6 123 3.92 3.44 18.9  1  0    4    4
#> Merc 450SE 16.4   8 275.8 180 3.07 4.07 17.4  0  0    3    3
#> Merc 450SL 17.3   8 275.8 180 3.07 3.73 17.6  0  0    3    3
#> Volvo 142E 21.4   4 121.0 109 4.11 2.78 18.6  1  1    4    2

na_omit(mtc_na, na.attr = TRUE)      # Adds attribute with removed cases, like na.omit

#>             mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 240D  24.4   4 146.7  62 3.69 3.19 20.0  1  0    4    2
#> Merc 280C  17.8   6 167.6 123 3.92 3.44 18.9  1  0    4    4
#> Merc 450SE 16.4   8 275.8 180 3.07 4.07 17.4  0  0    3    3
#> Merc 450SL 17.3   8 275.8 180 3.07 3.73 17.6  0  0    3    3
#> Volvo 142E 21.4   4 121.0 109 4.11 2.78 18.6  1  1    4    2

na_omit(mtc_na, cols = c("vs","am")) # Removes only cases missing vs or am

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620    NA  0  1   NA    4
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3   NA
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3   NA
#> Valiant             18.1   6 225.0 105   NA 3.460 20.22  1  0    3    1
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230              NA   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6    NA 123 3.92 3.440    NA  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780    NA  0  0    3    3
#> Cadillac Fleetwood  10.4  NA 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4  NA 460.0  NA 3.00 5.424 17.82  0  0    3    4
#> Fiat 128            32.4   4  78.7  NA 4.08 2.200 19.47  1  1    4    1
#> Honda Civic           NA   4  75.7  NA 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22    NA 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70    NA 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76    NA 16.87  0  0   NA    2
#> AMC Javelin         15.2   8    NA 150 3.15 3.435 17.30  0  0    3    2
#> Porsche 914-2       26.0  NA 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4    NA 113 3.77 1.513 16.90  1  1    5   NA
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1   NA    4
#> Maserati Bora       15.0   8 301.0 335 3.54    NA 14.60  0  1   NA    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

na_omit(qM(mtc_na))                  # Also works for matrices

#>             mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 240D  24.4   4 146.7  62 3.69 3.19 20.0  1  0    4    2
#> Merc 280C  17.8   6 167.6 123 3.92 3.44 18.9  1  0    4    4
#> Merc 450SE 16.4   8 275.8 180 3.07 4.07 17.4  0  0    3    3
#> Merc 450SL 17.3   8 275.8 180 3.07 3.73 17.6  0  0    3    3
#> Volvo 142E 21.4   4 121.0 109 4.11 2.78 18.6  1  1    4    2

na_omit(mtc_na$vs, na.attr = TRUE)   # Also works with vectors

#>  [1] 0 1 1 0 1 1 1 1 1 0 0 0 0 0 0 1 1 1 1 0 0 0 0 1 0 0 0 1
#> attr(,"na.action")
#> [1]  2  7 24 26
#> attr(,"class")
#> [1] "omit"

na_rm(mtc_na$vs)                     # For vectors na_rm is faster ...

#>  [1] 0 1 1 0 1 1 1 1 1 0 0 0 0 0 0 1 1 1 1 0 0 0 0 1 0 0 0 1

rm(mtc_na)