World Development Dataset

This dataset contains 5 indicators from the World Bank's World Development Indicators (WDI) database: (1) GDP per capita, (2) Life expectancy at birth, (3) GINI index, (4) Net ODA and official aid received and (5) Population. The panel data is balanced and covers 216 present and historic countries from 1960-2020 (World Bank aggregates and regional entities are excluded).

Apart from the indicators the data contains a number of identifiers (character country name, factor ISO3 country code, World Bank region and income level, numeric year and decade) and 2 generated variables: A logical variable indicating whether the country is an OECD member, and a fictitious variable stating the date the data was recorded. These variables were added so that all common data-types are represented in this dataset, making it an ideal test-dataset for certain collapse functions.

data("wlddev")

Format

A data frame with 13176 observations on the following 13 variables. All variables are labeled e.g. have a 'label' attribute.

country

chr Country Name

iso3c

fct Country Code

date

date Date Recorded (Fictitious)

year

int Year

decade

int Decade

region

fct World Bank Region

income

fct World Bank Income Level

OECD

log Is OECD Member Country?

PCGDP

num GDP per capita (constant 2010 US$)

LIFEEX

num Life expectancy at birth, total (years)

GINI

num GINI index (World Bank estimate)

ODA

num Net official development assistance and official aid received (constant 2018 US$)

POP

num Population, total

Source

https://data.worldbank.org/, accessed via the WDI package. The codes for the series are c("NY.GDP.PCAP.KD", "SP.DYN.LE00.IN", "SI.POV.GINI", "DT.ODA.ALLD.KD", "SP.POP.TOTL").

See also

Examples

data(wlddev)

# Panel-summarizing the 5 series
qsu(wlddev, pid = ~iso3c, cols = 9:13, vlabels = TRUE)

#> , , PCGDP: GDP per capita (constant 2010 US$)
#> 
#>              N/T        Mean          SD          Min         Max
#> Overall     9470   12048.778  19077.6416     132.0776  196061.417
#> Between      206  12962.6054  20189.9007     253.1886   141200.38
#> Within   45.9709   12048.778   6723.6808  -33504.8721  76767.5254
#> 
#> , , LIFEEX: Life expectancy at birth, total (years)
#> 
#>              N/T     Mean       SD      Min      Max
#> Overall    11670  64.2963  11.4764   18.907  85.4171
#> Between      207  64.9537   9.8936  40.9663  85.4171
#> Within   56.3768  64.2963   6.0842  32.9068  84.4198
#> 
#> , , GINI: Gini index (World Bank estimate)
#> 
#>              N/T     Mean      SD      Min      Max
#> Overall     1744  38.5341  9.2006     20.7     65.8
#> Between      167  39.4233  8.1356  24.8667  61.7143
#> Within   10.4431  38.5341  2.9277  25.3917  55.3591
#> 
#> , , ODA: Net official development assistance and official aid received (constant 2018 US$)
#> 
#>              N/T        Mean          SD              Min             Max
#> Overall     8608  454,720131  868,712654      -997,679993  2.56715605e+10
#> Between      178  439,168412  569,049959       468717.916  3.62337432e+09
#> Within   48.3596  454,720131  650,709624  -2.44379420e+09  2.45610972e+10
#> 
#> , , POP: Population, total
#> 
#>              N/T         Mean           SD          Min             Max
#> Overall    12919  24,245971.6   102,120674         2833  1.39771500e+09
#> Between      216    24,178573  98,616506.7    8343.3333  1.08786967e+09
#> Within   59.8102  24,245971.6  26,803077.4  -405,793067      510,077008
#> 

# By Region
qsu(wlddev, by = ~region, cols = 9:13, vlabels = TRUE)

#> , , PCGDP: GDP per capita (constant 2010 US$)
#> 
#>                                N        Mean          SD         Min
#> East Asia & Pacific         1467  10513.2441  14383.5507    132.0776
#> Europe & Central Asia       2243  25992.9618  26435.1316    366.9354
#> Latin America & Caribbean   1976   7628.4477   8818.5055   1005.4085
#> Middle East & North Africa   842  13878.4213  18419.7912    578.5996
#> North America                180    48699.76  24196.2855  16405.9053
#> South Asia                   382   1235.9256   1611.2232    265.9625
#> Sub-Saharan Africa          2380   1840.0259   2596.0104    164.3366
#>                                    Max
#> East Asia & Pacific         71992.1517
#> Europe & Central Asia       196061.417
#> Latin America & Caribbean   88391.3331
#> Middle East & North Africa  116232.753
#> North America               113236.091
#> South Asia                    8476.564
#> Sub-Saharan Africa          20532.9523
#> 
#> , , LIFEEX: Life expectancy at birth, total (years)
#> 
#>                                N     Mean       SD      Min      Max
#> East Asia & Pacific         1807  65.9445  10.1633   18.907   85.078
#> Europe & Central Asia       3046  72.1625   5.7602   45.369  85.4171
#> Latin America & Caribbean   2107  68.3486   7.3768   41.762  82.1902
#> Middle East & North Africa  1226  66.2508   9.8306   29.919  82.8049
#> North America                144  76.2867   3.5734  68.8978  82.0488
#> South Asia                   480  57.5585  11.3004   32.446   78.921
#> Sub-Saharan Africa          2860   51.581   8.6876   26.172  74.5146
#> 
#> , , GINI: Gini index (World Bank estimate)
#> 
#>                               N     Mean      SD   Min   Max
#> East Asia & Pacific         154  37.7571  5.0318  27.8  49.1
#> Europe & Central Asia       798  31.9114  4.5809  20.7  48.4
#> Latin America & Caribbean   413  49.9557  5.4821  34.4  63.3
#> Middle East & North Africa   91  36.0143  5.2073    26  47.4
#> North America                49  37.4816  3.6972    31  41.5
#> South Asia                   46  33.8804  3.9898  25.9  43.8
#> Sub-Saharan Africa          193  44.6606  8.2003  29.8  65.8
#> 
#> , , ODA: Net official development assistance and official aid received (constant 2018 US$)
#> 
#>                                N            Mean              SD           Min
#> East Asia & Pacific         1537      352,017964      622,847624   -997,679993
#> Europe & Central Asia        787      402,455286      568,237036   -322,070007
#> Latin America & Caribbean   1972      172,880081      260,781049   -444,040009
#> Middle East & North Africa  1105      732,380009  1.52108993e+09   -141,789993
#> North America                 39      468717.916     10,653560.8  -15,869999.9
#> South Asia                   466  1.27049955e+09  1.61492889e+09   -247,369995
#> Sub-Saharan Africa          2702      486,371750      656,336230  -18,409999.8
#>                                        Max
#> East Asia & Pacific         4.04487988e+09
#> Europe & Central Asia       4.34612988e+09
#> Latin America & Caribbean   2.99568994e+09
#> Middle East & North Africa  2.56715605e+10
#> North America                  61,509998.3
#> South Asia                  8.75425977e+09
#> Sub-Saharan Africa          1.18790801e+10
#> 
#> , , POP: Population, total
#> 
#>                                N         Mean           SD    Min
#> East Asia & Pacific         2160  47,788194.6   184,121276   4375
#> Europe & Central Asia       3480  14,103754.9  24,686065.8  13411
#> Latin America & Caribbean   2520  10,371370.7    26,728206   2833
#> Middle East & North Africa  1227  12,549479.2  17,750818.2  47384
#> North America                180  93,783656.8   116,925373  44400
#> South Asia                   480   144,180423   300,282162  89863
#> Sub-Saharan Africa          2872  11,597636.9  19,916800.8  41700
#>                                        Max
#> East Asia & Pacific         1.39771500e+09
#> Europe & Central Asia           148,538197
#> Latin America & Caribbean       211,049527
#> Middle East & North Africa      100,388073
#> North America                   326,687501
#> South Asia                  1.36641775e+09
#> Sub-Saharan Africa              200,963599
#> 

# Panel-summary by region
qsu(wlddev, by = ~region, pid = ~iso3c, cols = 9:13, vlabels = TRUE)

#> , , Overall, PCGDP: GDP per capita (constant 2010 US$)
#> 
#>                              N/T        Mean          SD         Min
#> East Asia & Pacific         1467  10513.2441  14383.5507    132.0776
#> Europe & Central Asia       2243  25992.9618  26435.1316    366.9354
#> Latin America & Caribbean   1976   7628.4477   8818.5055   1005.4085
#> Middle East & North Africa   842  13878.4213  18419.7912    578.5996
#> North America                180    48699.76  24196.2855  16405.9053
#> South Asia                   382   1235.9256   1611.2232    265.9625
#> Sub-Saharan Africa          2380   1840.0259   2596.0104    164.3366
#>                                    Max
#> East Asia & Pacific         71992.1517
#> Europe & Central Asia       196061.417
#> Latin America & Caribbean   88391.3331
#> Middle East & North Africa  116232.753
#> North America               113236.091
#> South Asia                    8476.564
#> Sub-Saharan Africa          20532.9523
#> 
#> , , Between, PCGDP: GDP per capita (constant 2010 US$)
#> 
#>                             N/T        Mean          SD         Min         Max
#> East Asia & Pacific          34  10513.2441   12771.742    444.2899  39722.0077
#> Europe & Central Asia        56  25992.9618   24051.035    809.4753   141200.38
#> Latin America & Caribbean    38   7628.4477   8470.9708   1357.3326  77403.7443
#> Middle East & North Africa   20  13878.4213  17251.6962   1069.6596  64878.4021
#> North America                 3    48699.76  18604.4369  35260.4708  74934.5874
#> South Asia                    8   1235.9256   1488.3669      413.68   6621.5002
#> Sub-Saharan Africa           47   1840.0259   2234.3254    253.1886   9922.0052
#> 
#> , , Within, PCGDP: GDP per capita (constant 2010 US$)
#> 
#>                                 N/T       Mean          SD          Min
#> East Asia & Pacific         43.1471  12048.778   6615.8248  -11964.6472
#> Europe & Central Asia       40.0536  12048.778  10971.0483  -33504.8721
#> Latin America & Caribbean        52  12048.778   2451.2636    -354.1639
#> Middle East & North Africa     42.1  12048.778   6455.0512  -18674.4049
#> North America                    60  12048.778  15470.4609  -29523.1017
#> South Asia                    47.75  12048.778    617.0934   10026.9155
#> Sub-Saharan Africa          50.6383  12048.778    1321.764    4846.3834
#>                                    Max
#> East Asia & Pacific          49541.463
#> Europe & Central Asia       76767.5254
#> Latin America & Caribbean   23036.3668
#> Middle East & North Africa  63665.0446
#> North America               50350.2816
#> South Asia                  14390.3768
#> Sub-Saharan Africa          24883.1246
#> 
#> , , Overall, LIFEEX: Life expectancy at birth, total (years)
#> 
#>                              N/T     Mean       SD      Min      Max
#> East Asia & Pacific         1807  65.9445  10.1633   18.907   85.078
#> Europe & Central Asia       3046  72.1625   5.7602   45.369  85.4171
#> Latin America & Caribbean   2107  68.3486   7.3768   41.762  82.1902
#> Middle East & North Africa  1226  66.2508   9.8306   29.919  82.8049
#> North America                144  76.2867   3.5734  68.8978  82.0488
#> South Asia                   480  57.5585  11.3004   32.446   78.921
#> Sub-Saharan Africa          2860   51.581   8.6876   26.172  74.5146
#> 
#> , , Between, LIFEEX: Life expectancy at birth, total (years)
#> 
#>                             N/T     Mean      SD      Min      Max
#> East Asia & Pacific          32  65.9445  7.6833  49.7995  77.9008
#> Europe & Central Asia        55  72.1625  4.4378  60.1129  85.4171
#> Latin America & Caribbean    40  68.3486  4.9199  53.4918  82.1902
#> Middle East & North Africa   21  66.2508   5.922  52.5371  76.7395
#> North America                 3  76.2867  1.3589  74.8065  78.4175
#> South Asia                    8  57.5585  5.6158  49.1972  69.3429
#> Sub-Saharan Africa           48   51.581   5.657  40.9663  71.5749
#> 
#> , , Within, LIFEEX: Life expectancy at birth, total (years)
#> 
#>                                 N/T     Mean      SD      Min      Max
#> East Asia & Pacific         56.4688  64.2963  6.6528  32.9068  83.9918
#> Europe & Central Asia       55.3818  64.2963  3.6723  46.3045  78.6265
#> Latin America & Caribbean    52.675  64.2963  5.4965  46.7831  79.5026
#> Middle East & North Africa   58.381  64.2963  7.8467  41.6187  78.8872
#> North America                    48  64.2963  3.3049  54.7766  69.4306
#> South Asia                       60  64.2963  9.8062  41.4342  83.0122
#> Sub-Saharan Africa          59.5833  64.2963  6.5933  41.5678  84.4198
#> 
#> , , Overall, GINI: Gini index (World Bank estimate)
#> 
#>                             N/T     Mean      SD   Min   Max
#> East Asia & Pacific         154  37.7571  5.0318  27.8  49.1
#> Europe & Central Asia       798  31.9114  4.5809  20.7  48.4
#> Latin America & Caribbean   413  49.9557  5.4821  34.4  63.3
#> Middle East & North Africa   91  36.0143  5.2073    26  47.4
#> North America                49  37.4816  3.6972    31  41.5
#> South Asia                   46  33.8804  3.9898  25.9  43.8
#> Sub-Saharan Africa          193  44.6606  8.2003  29.8  65.8
#> 
#> , , Between, GINI: Gini index (World Bank estimate)
#> 
#>                             N/T     Mean      SD      Min      Max
#> East Asia & Pacific          23  37.7571  4.3005     30.8  45.8857
#> Europe & Central Asia        49  31.9114  4.0611  24.8667   40.935
#> Latin America & Caribbean    25  49.9557  4.0492     41.1     57.9
#> Middle East & North Africa   15  36.0143  4.7002    29.05     42.7
#> North America                 2  37.4816  3.3563  33.1222  40.0129
#> South Asia                    7  33.8804  3.0052  30.3556     38.8
#> Sub-Saharan Africa           46  44.6606  6.8844    34.52  61.7143
#> 
#> , , Within, GINI: Gini index (World Bank estimate)
#> 
#>                                 N/T     Mean      SD      Min      Max
#> East Asia & Pacific          6.6957  38.5341  2.6125  31.0187  45.8901
#> Europe & Central Asia       16.2857  38.5341  2.1195  31.2841  50.1387
#> Latin America & Caribbean     16.52  38.5341  3.6955  25.3917  48.8341
#> Middle East & North Africa   6.0667  38.5341  2.2415  31.7675   45.777
#> North America                  24.5  38.5341  1.5507  33.0212  42.7119
#> South Asia                   6.5714  38.5341  2.6244  32.8341  45.0675
#> Sub-Saharan Africa           4.1957  38.5341  4.4553  27.9452  55.3591
#> 
#> , , Overall, ODA: Net official development assistance and official aid received (constant 2018 US$)
#> 
#>                              N/T            Mean              SD           Min
#> East Asia & Pacific         1537      352,017964      622,847624   -997,679993
#> Europe & Central Asia        787      402,455286      568,237036   -322,070007
#> Latin America & Caribbean   1972      172,880081      260,781049   -444,040009
#> Middle East & North Africa  1105      732,380009  1.52108993e+09   -141,789993
#> North America                 39      468717.916     10,653560.8  -15,869999.9
#> South Asia                   466  1.27049955e+09  1.61492889e+09   -247,369995
#> Sub-Saharan Africa          2702      486,371750      656,336230  -18,409999.8
#>                                        Max
#> East Asia & Pacific         4.04487988e+09
#> Europe & Central Asia       4.34612988e+09
#> Latin America & Caribbean   2.99568994e+09
#> Middle East & North Africa  2.56715605e+10
#> North America                  61,509998.3
#> South Asia                  8.75425977e+09
#> Sub-Saharan Africa          1.18790801e+10
#> 
#> , , Between, ODA: Net official development assistance and official aid received (constant 2018 US$)
#> 
#>                             N/T            Mean              SD          Min
#> East Asia & Pacific          31      352,017964      457,183279  1,654615.38
#> Europe & Central Asia        32      402,455286      438,074771  12,516000.1
#> Latin America & Caribbean    37      172,880081      167,160838  2,225483.88
#> Middle East & North Africa   21      732,380009      775,418887   3,112820.5
#> North America                 1      468717.916               0   468717.916
#> South Asia                    8  1.27049955e+09  1.18347893e+09  27,152499.9
#> Sub-Saharan Africa           48      486,371750      397,995105  28,801206.9
#>                                        Max
#> East Asia & Pacific         1.63585532e+09
#> Europe & Central Asia       2.05456932e+09
#> Latin America & Caribbean       538,386665
#> Middle East & North Africa  2.86174883e+09
#> North America                   468717.916
#> South Asia                  3.62337432e+09
#> Sub-Saharan Africa          1.55049113e+09
#> 
#> , , Within, ODA: Net official development assistance and official aid received (constant 2018 US$)
#> 
#>                                 N/T        Mean              SD
#> East Asia & Pacific         49.5806  454,720131      422,992450
#> Europe & Central Asia       24.5938  454,720131      361,916875
#> Latin America & Caribbean   53.2973  454,720131      200,159960
#> Middle East & North Africa   52.619  454,720131  1.30860235e+09
#> North America                    39  454,720131     10,653560.8
#> South Asia                    58.25  454,720131  1.09880524e+09
#> Sub-Saharan Africa          56.2917  454,720131      521,897637
#>                                         Min             Max
#> East Asia & Pacific         -2.04042108e+09  3.59673152e+09
#> Europe & Central Asia       -1.08796786e+09  3.30549004e+09
#> Latin America & Caribbean       -527,706542  3.28976141e+09
#> Middle East & North Africa  -2.34610870e+09  2.45610972e+10
#> North America                    438,381413      515,761411
#> South Asia                  -2.44379420e+09  5.58560558e+09
#> Sub-Saharan Africa              -952,168698  1.12814455e+10
#> 
#> , , Overall, POP: Population, total
#> 
#>                              N/T         Mean           SD    Min
#> East Asia & Pacific         2160  47,788194.6   184,121276   4375
#> Europe & Central Asia       3480  14,103754.9  24,686065.8  13411
#> Latin America & Caribbean   2520  10,371370.7    26,728206   2833
#> Middle East & North Africa  1227  12,549479.2  17,750818.2  47384
#> North America                180  93,783656.8   116,925373  44400
#> South Asia                   480   144,180423   300,282162  89863
#> Sub-Saharan Africa          2872  11,597636.9  19,916800.8  41700
#>                                        Max
#> East Asia & Pacific         1.39771500e+09
#> Europe & Central Asia           148,538197
#> Latin America & Caribbean       211,049527
#> Middle East & North Africa      100,388073
#> North America                   326,687501
#> South Asia                  1.36641775e+09
#> Sub-Saharan Africa              200,963599
#> 
#> , , Between, POP: Population, total
#> 
#>                             N/T         Mean           SD         Min
#> East Asia & Pacific          36  47,788194.6   179,715591   8343.3333
#> Europe & Central Asia        58  14,103754.9    24,507651     24593.9
#> Latin America & Caribbean    42  10,371370.7  25,405423.8    16704.95
#> Middle East & North Africa   21  12,549479.2  15,905067.3    363388.1
#> North America                 3  93,783656.8   114,001343  58013.9167
#> South Asia                    8   144,180423   281,793860  239010.367
#> Sub-Saharan Africa           48  11,597636.9  17,461323.8  70367.5833
#>                                        Max
#> East Asia & Pacific         1.08786967e+09
#> Europe & Central Asia           139,805322
#> Latin America & Caribbean       145,198074
#> Middle East & North Africa     57,543388.2
#> North America                   253,773772
#> South Asia                      880,586717
#> Sub-Saharan Africa              103,186135
#> 
#> , , Within, POP: Population, total
#> 
#>                                 N/T         Mean           SD           Min
#> East Asia & Pacific              60  24,245971.6    40,036868   -403,293695
#> Europe & Central Asia            60  24,245971.6  2,962581.72  -1,848462.12
#> Latin America & Caribbean        60  24,245971.6   8,304302.6  -48,772876.7
#> Middle East & North Africa  58.4286  24,245971.6  7,881648.49  -6,664522.66
#> North America                    60  24,245971.6  25,985318.3    -48,856800
#> South Asia                       60  24,245971.6   103,738120   -405,793067
#> Sub-Saharan Africa          59.8333  24,245971.6   9,580246.6  -33,801705.6
#>                                     Max
#> East Asia & Pacific          334,091305
#> Europe & Central Asia       54,108807.9
#> Latin America & Caribbean   90,097424.3
#> Middle East & North Africa  67,090656.3
#> North America                 97,159701
#> South Asia                   510,077008
#> Sub-Saharan Africa           122,023435
#> 

# Pairwise correlations: Ovarall
print(pwcor(get_vars(wlddev, 9:13), N = TRUE, P = TRUE), show = "lower.tri")

#>               PCGDP        LIFEEX         GINI          ODA           POP
#> PCGDP    1   (9470)                                                      
#> LIFEEX  .57* (9022)   1   (11670)                                        
#> GINI   -.44* (1735)  -.35* (1742)   1   (1744)                           
#> ODA    -.16* (7128)  -.02  (8142) -.20* (1109)   1   (8608)              
#> POP    -.06* (9470)  .03* (11659)  .04  (1744)  .31* (8597)   1   (12919)

# Pairwise correlations: Between Countries
print(pwcor(fmean(get_vars(wlddev, 9:13), wlddev$iso3c), N = TRUE, P = TRUE), show = "lower.tri")

#>              PCGDP      LIFEEX        GINI         ODA         POP
#> PCGDP    1   (206)                                                
#> LIFEEX  .60* (199)   1   (207)                                    
#> GINI   -.42* (165) -.40* (165)   1   (167)                        
#> ODA    -.25* (172) -.21* (172) -.19* (145)   1   (178)            
#> POP    -.07  (206) -.02  (207) -.04  (167)  .50* (178)   1   (216)

# Pairwise correlations: Within Countries
print(pwcor(fwithin(get_vars(wlddev, 9:13), wlddev$iso3c), N = TRUE, P = TRUE), show = "lower.tri")

#>               PCGDP        LIFEEX         GINI          ODA           POP
#> PCGDP    1   (9470)                                                      
#> LIFEEX  .31* (9022)   1   (11670)                                        
#> GINI   -.01  (1735)  -.16* (1742)   1   (1744)                           
#> ODA    -.01  (7128)   .17* (8142) -.08* (1109)   1   (8608)              
#> POP     .06* (9470)  .29* (11659)  .01  (1744) -.11* (8597)   1   (12919)